% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afs_effrs_afsact1.R
\name{afs_effrs_afsact1}
\alias{afs_effrs_afsact1}
\title{Retrieve effrs afsact1 data from afs database}
\usage{
afs_effrs_afsact1(STATE_COUNTY_COMPLIANCE_SOURCE = NULL,
  ACTION_NUMBER = NULL, KEY_ACTION_NUMBERS = NULL,
  AIR_PROGRAM_CODES = NULL, NATIONAL_ACTION_TYPE = NULL,
  NATIONAL_ACTION_DESC = NULL, ACTION_TYPE = NULL, ACTION_DESC = NULL,
  DATE_ACHIEVED = NULL, PENALTY_AMT = NULL, RESULTS_CODE = NULL,
  RESULTS_CODE_DESC = NULL, POLLUTANT_CODE = NULL,
  REGIONAL_DATA_ELEMENT = NULL, REGIONAL_DATA_ELEMENT_16 = NULL,
  HPV_ADDRESSING_ACTION_DESC = NULL, CREATION_DATE = NULL)
}
\arguments{
\item{STATE_COUNTY_COMPLIANCE_SOURCE}{e.g. '0100100001'. See Details.}

\item{ACTION_NUMBER}{e.g. '00460'. See Details.}

\item{KEY_ACTION_NUMBERS}{e.g. 'NA'. See Details.}

\item{AIR_PROGRAM_CODES}{e.g. 'M'. See Details.}

\item{NATIONAL_ACTION_TYPE}{e.g. 'TS'. See Details.}

\item{NATIONAL_ACTION_DESC}{e.g. 'STATE/LOCAL RECEIPT OF STACK TEST REPORT'. See Details.}

\item{ACTION_TYPE}{e.g. 'TS'. See Details.}

\item{ACTION_DESC}{e.g. 'STATE RECEIPT OF STACK TEST REPORT'. See Details.}

\item{DATE_ACHIEVED}{e.g. '21-JUL-10'. See Details.}

\item{PENALTY_AMT}{e.g. 'NA'. See Details.}

\item{RESULTS_CODE}{e.g. 'PP'. See Details.}

\item{RESULTS_CODE_DESC}{e.g. 'STACK TEST PASSED'. See Details.}

\item{POLLUTANT_CODE}{e.g. 'PT'. See Details.}

\item{REGIONAL_DATA_ELEMENT}{e.g. 'NA'. See Details.}

\item{REGIONAL_DATA_ELEMENT_16}{e.g. 'NA'. See Details.}

\item{HPV_ADDRESSING_ACTION_DESC}{e.g. 'NA'. See Details.}

\item{CREATION_DATE}{e.g. '18-AUG-10'. See Details.}
}
\description{
Retrieve effrs afsact1 data from afs database
}

