% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afs_effrs_afsact2.R
\name{afs_effrs_afsact2}
\alias{afs_effrs_afsact2}
\title{Retrieve effrs afsact2 data from afs database}
\usage{
afs_effrs_afsact2(STATE_COUNTY_COMPLIANCE_SOURCE = NULL,
  ACTION_NUMBER = NULL, AIR_PROGRAM_CODES = NULL,
  NATIONAL_ACTION_TYPE = NULL, ACTION_TYPE = NULL, DATE_ACHIEVED = NULL,
  LEAD_AGENCY = NULL, LEAD_AGENCY_DESC = NULL,
  HPV_DAY_ZERO_ACTION_DATE = NULL, HPV_DAY_ZERO_ACTION_TYPE = NULL,
  HPV_DAY_ZERO_ACTION_DESC = NULL, HPV_DISCOVERY_DATE = NULL,
  HPV_DISCOVERY_ACTION_TYPE = NULL, HPV_DISCOVERY_ACTION_DESC = NULL,
  HPV_ADDRESSING_ACTION_DATE = NULL, HPV_ADDRESSING_ACTION_TYPE = NULL,
  HPV_ADDRESSING_ACTION_DESC = NULL, HPV_ADDRESSING_ACTION_PENALTY = NULL,
  HPV_RESOLUTION_DATE = NULL, HPV_RESOLUTION_ACTION_TYPE = NULL,
  HPV_RESOLUTION_ACTION_DESC = NULL, VIOLATION_POLLUTANTS = NULL,
  VIOLATION_TYPE_CODES = NULL, VIOLATION_IDENTIFIER = NULL,
  DATE_CREATED = NULL)
}
\arguments{
\item{STATE_COUNTY_COMPLIANCE_SOURCE}{e.g. '0100100001'. See Details.}

\item{ACTION_NUMBER}{e.g. '00239'. See Details.}

\item{AIR_PROGRAM_CODES}{e.g. '0'. See Details.}

\item{NATIONAL_ACTION_TYPE}{e.g. '2E'. See Details.}

\item{ACTION_TYPE}{e.g. '04'. See Details.}

\item{DATE_ACHIEVED}{e.g. '11-MAY-00'. See Details.}

\item{LEAD_AGENCY}{e.g. 'S'. See Details.}

\item{LEAD_AGENCY_DESC}{e.g. 'STATE LEAD FOR HPV'. See Details.}

\item{HPV_DAY_ZERO_ACTION_DATE}{e.g. '11-MAY-00'. See Details.}

\item{HPV_DAY_ZERO_ACTION_TYPE}{e.g. '04'. See Details.}

\item{HPV_DAY_ZERO_ACTION_DESC}{e.g. 'STATE DAY ZERO'. See Details.}

\item{HPV_DISCOVERY_DATE}{e.g. 'NA'. See Details.}

\item{HPV_DISCOVERY_ACTION_TYPE}{e.g. 'NA'. See Details.}

\item{HPV_DISCOVERY_ACTION_DESC}{e.g. 'NA'. See Details.}

\item{HPV_ADDRESSING_ACTION_DATE}{e.g. '21-AUG-00'. See Details.}

\item{HPV_ADDRESSING_ACTION_TYPE}{e.g. 'AW'. See Details.}

\item{HPV_ADDRESSING_ACTION_DESC}{e.g. 'COMPL BY STATE/LOCAL, NO ACT REQ'. See Details.}

\item{HPV_ADDRESSING_ACTION_PENALTY}{e.g. '000000000'. See Details.}

\item{HPV_RESOLUTION_DATE}{e.g. '21-AUG-00'. See Details.}

\item{HPV_RESOLUTION_ACTION_TYPE}{e.g. 'VR'. See Details.}

\item{HPV_RESOLUTION_ACTION_DESC}{e.g. 'VIOLATION RESOLVED'. See Details.}

\item{VIOLATION_POLLUTANTS}{e.g. 'NA'. See Details.}

\item{VIOLATION_TYPE_CODES}{e.g. 'NA'. See Details.}

\item{VIOLATION_IDENTIFIER}{e.g. 'NA'. See Details.}

\item{DATE_CREATED}{e.g. 'NA'. See Details.}
}
\description{
Retrieve effrs afsact2 data from afs database
}

