% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afs_effrs_afsap.R
\name{afs_effrs_afsap}
\alias{afs_effrs_afsap}
\title{Retrieve effrs afsap data from afs database}
\usage{
afs_effrs_afsap(STATE_COUNTY_COMPLIANCE_SOURCE = NULL,
  AIR_PROGRAM_CODE = NULL, AIR_PROGRAM_DESC = NULL,
  AIR_PROGRAM_STATUS = NULL, AIR_PROGRAM_STATUS_DESC = NULL,
  AIR_PROGRAM_SUBPART = NULL, AIR_PROGRAM_SUBPART_DESC = NULL,
  ES_CLASS_CODE = NULL, ES_CLASS_CODE_DESC = NULL,
  ES_COMPLIANCE_STATUS = NULL, ES_COMPLIANCE_STATUS_DESC = NULL,
  EPA_CLASS_CODE = NULL, EPA_CLASS_CODE_DESC = NULL,
  EPA_COMPLIANCE_STATUS = NULL, EPA_COMPLIANCE_STATUS_DESC = NULL,
  STATE_CLASS_CODE = NULL, STATE_CLASS_CODE_DESC = NULL,
  STATE_COMPLIANCE_STATUS = NULL, STATE_COMPLIANCE_STATUS_DESC = NULL,
  AIR_PROGRAMS = NULL)
}
\arguments{
\item{STATE_COUNTY_COMPLIANCE_SOURCE}{e.g. '0100100001'. See Details.}

\item{AIR_PROGRAM_CODE}{e.g. '9'. See Details.}

\item{AIR_PROGRAM_DESC}{e.g. 'NSPS'. See Details.}

\item{AIR_PROGRAM_STATUS}{e.g. 'O'. See Details.}

\item{AIR_PROGRAM_STATUS_DESC}{e.g. 'OPERATING'. See Details.}

\item{AIR_PROGRAM_SUBPART}{e.g. 'BB'. See Details.}

\item{AIR_PROGRAM_SUBPART_DESC}{e.g. 'KRAFT PULP MILLS'. See Details.}

\item{ES_CLASS_CODE}{e.g. 'C'. See Details.}

\item{ES_CLASS_CODE_DESC}{e.g. 'CLASS IS UNKNOWN.'. See Details.}

\item{ES_COMPLIANCE_STATUS}{e.g. '3'. See Details.}

\item{ES_COMPLIANCE_STATUS_DESC}{e.g. 'IN COMPLIANCE - INSPECTIO'. See Details.}

\item{EPA_CLASS_CODE}{e.g. 'NA'. See Details.}

\item{EPA_CLASS_CODE_DESC}{e.g. 'NA'. See Details.}

\item{EPA_COMPLIANCE_STATUS}{e.g. 'NA'. See Details.}

\item{EPA_COMPLIANCE_STATUS_DESC}{e.g. 'NA'. See Details.}

\item{STATE_CLASS_CODE}{e.g. 'C'. See Details.}

\item{STATE_CLASS_CODE_DESC}{e.g. 'CLASS IS UNKNOWN.'. See Details.}

\item{STATE_COMPLIANCE_STATUS}{e.g. '3'. See Details.}

\item{STATE_COMPLIANCE_STATUS_DESC}{e.g. 'IN COMPLIANCE - INSPECTIO'. See Details.}

\item{AIR_PROGRAMS}{e.g. 'AMV069'. See Details.}
}
\description{
Retrieve effrs afsap data from afs database
}

