% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afs_effrs_afsapp.R
\name{afs_effrs_afsapp}
\alias{afs_effrs_afsapp}
\title{Retrieve effrs afsapp data from afs database}
\usage{
afs_effrs_afsapp(STATE_COUNTY_COMPLIANCE_SOURCE = NULL,
  AIR_PROGRAM_CODE = NULL, POLLUTANT_CODE = NULL,
  POLLUTANT_CODE_DESC = NULL, CHEMICAL_ABSTRACT_SERVICE_NUM = NULL,
  CHEMICAL_ABSTRACT_SERVICE_DESC = NULL, ES_ATTAIN_IND_DESC = NULL,
  ES_ATTAIN_IND = NULL, ES_PLLT_COMPLIANCE_STATUS = NULL,
  ES_PLLT_COMPLIANCE_DESC = NULL, ES_PLLT_CLASS_CODE = NULL,
  ES_PLLT_CLASS_DESC = NULL, STATE_ATTAIN_IND = NULL,
  STATE_ATTAIN_IND_DESC = NULL, STATE_PLLT_COMPLIANCE_STATUS = NULL,
  STATE_PLLT_COMPLIANCE_DESC = NULL, STATE_PLLT_CLASS_CODE = NULL,
  STATE_PLLT_CLASS_DESC = NULL, EPA_ATTAIN_IND = NULL,
  EPA_ATTAIN_IND_DESC = NULL, EPA_PLLT_COMPLIANCE_STATUS = NULL,
  EPA_PLLT_COMPLIANCE_DESC = NULL, EPA_PLLT_CLASS_CODE = NULL,
  EPA_PLLT_CLASS_DESC = NULL, DATE_RECORD_UPDATED = NULL,
  AIR_PROGRAMS = NULL)
}
\arguments{
\item{STATE_COUNTY_COMPLIANCE_SOURCE}{e.g. '0100100001'. See Details.}

\item{AIR_PROGRAM_CODE}{e.g. '0'. See Details.}

\item{POLLUTANT_CODE}{e.g. 'PM10'. See Details.}

\item{POLLUTANT_CODE_DESC}{e.g. 'PARTICULATE MATTER %3C 10 UM'. See Details.}

\item{CHEMICAL_ABSTRACT_SERVICE_NUM}{e.g. 'NA'. See Details.}

\item{CHEMICAL_ABSTRACT_SERVICE_DESC}{e.g. 'NA'. See Details.}

\item{ES_ATTAIN_IND_DESC}{e.g. 'ATTAINMENT AREA FOR A GIV'. See Details.}

\item{ES_ATTAIN_IND}{e.g. 'A'. See Details.}

\item{ES_PLLT_COMPLIANCE_STATUS}{e.g. '4'. See Details.}

\item{ES_PLLT_COMPLIANCE_DESC}{e.g. 'IN COMPLIANCE - CERTIFICA'. See Details.}

\item{ES_PLLT_CLASS_CODE}{e.g. 'A'. See Details.}

\item{ES_PLLT_CLASS_DESC}{e.g. 'ACTUAL OR POTENTIAL EMISS'. See Details.}

\item{STATE_ATTAIN_IND}{e.g. 'A'. See Details.}

\item{STATE_ATTAIN_IND_DESC}{e.g. 'ATTAINMENT AREA FOR A GIV'. See Details.}

\item{STATE_PLLT_COMPLIANCE_STATUS}{e.g. '4'. See Details.}

\item{STATE_PLLT_COMPLIANCE_DESC}{e.g. 'IN COMPLIANCE - CERTIFICA'. See Details.}

\item{STATE_PLLT_CLASS_CODE}{e.g. 'A'. See Details.}

\item{STATE_PLLT_CLASS_DESC}{e.g. 'ACTUAL OR POTENTIAL EMISS'. See Details.}

\item{EPA_ATTAIN_IND}{e.g. 'NA'. See Details.}

\item{EPA_ATTAIN_IND_DESC}{e.g. 'NA'. See Details.}

\item{EPA_PLLT_COMPLIANCE_STATUS}{e.g. 'NA'. See Details.}

\item{EPA_PLLT_COMPLIANCE_DESC}{e.g. 'NA'. See Details.}

\item{EPA_PLLT_CLASS_CODE}{e.g. 'NA'. See Details.}

\item{EPA_PLLT_CLASS_DESC}{e.g. 'NA'. See Details.}

\item{DATE_RECORD_UPDATED}{e.g. '21-JUN-99'. See Details.}

\item{AIR_PROGRAMS}{e.g. 'AMV069'. See Details.}
}
\description{
Retrieve effrs afsapp data from afs database
}

