% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afs_effrs_afscms.R
\name{afs_effrs_afscms}
\alias{afs_effrs_afscms}
\title{Retrieve effrs afscms data from afs database}
\usage{
afs_effrs_afscms(STATE_COUNTY_COMPLIANCE_SOURCE = NULL, EPA_PLANT_ID = NULL,
  PLANT_ID = NULL, REGION_CODE = NULL, STATE_ABBREV = NULL,
  STATE_NAME = NULL, COUNTY_NAME = NULL, DUN_BRADSTREET_NUMBER = NULL,
  CMS_SOURCE_CATEGORY = NULL, CMS_START_DATE = NULL,
  FY2008_CMS_IND = NULL, FY2008_CMS_IND_DESC = NULL,
  FY2009_CMS_IND = NULL, FY2009_CMS_IND_DESC = NULL,
  FY2010_CMS_IND = NULL, FY2010_CMS_IND_DESC = NULL,
  FY2011_CMS_IND = NULL, FY2011_CMS_IND_DESC = NULL,
  FY2012_CMS_IND = NULL, FY2012_CMS_IND_DESC = NULL,
  FY2013_CMS_IND = NULL, FY2013_CMS_IND_DESC = NULL,
  FY2014_CMS_IND = NULL, FY2014_CMS_IND_DESC = NULL,
  FY2015_CMS_IND = NULL, FY2015_CMS_IND_DESC = NULL)
}
\arguments{
\item{STATE_COUNTY_COMPLIANCE_SOURCE}{e.g. '0100100001'. See Details.}

\item{EPA_PLANT_ID}{e.g. '110011734935'. See Details.}

\item{PLANT_ID}{e.g. '0013025'. See Details.}

\item{REGION_CODE}{e.g. '04'. See Details.}

\item{STATE_ABBREV}{e.g. 'AL'. See Details.}

\item{STATE_NAME}{e.g. 'ALABAMA'. See Details.}

\item{COUNTY_NAME}{e.g. 'AUTAUGA'. See Details.}

\item{DUN_BRADSTREET_NUMBER}{e.g. '039044318'. See Details.}

\item{CMS_SOURCE_CATEGORY}{e.g. 'A'. See Details.}

\item{CMS_START_DATE}{e.g. '01-OCT-01'. See Details.}

\item{FY2008_CMS_IND}{e.g. 'A'. See Details.}

\item{FY2008_CMS_IND_DESC}{e.g. 'TITLE V MAJOR'. See Details.}

\item{FY2009_CMS_IND}{e.g. 'A'. See Details.}

\item{FY2009_CMS_IND_DESC}{e.g. 'TITLE V MAJOR'. See Details.}

\item{FY2010_CMS_IND}{e.g. 'A'. See Details.}

\item{FY2010_CMS_IND_DESC}{e.g. 'TITLE V MAJOR'. See Details.}

\item{FY2011_CMS_IND}{e.g. 'A'. See Details.}

\item{FY2011_CMS_IND_DESC}{e.g. 'TITLE V MAJOR'. See Details.}

\item{FY2012_CMS_IND}{e.g. 'A'. See Details.}

\item{FY2012_CMS_IND_DESC}{e.g. 'TITLE V MAJOR'. See Details.}

\item{FY2013_CMS_IND}{e.g. 'A'. See Details.}

\item{FY2013_CMS_IND_DESC}{e.g. 'TITLE V MAJOR'. See Details.}

\item{FY2014_CMS_IND}{e.g. 'NA'. See Details.}

\item{FY2014_CMS_IND_DESC}{e.g. 'NA'. See Details.}

\item{FY2015_CMS_IND}{e.g. 'NA'. See Details.}

\item{FY2015_CMS_IND_DESC}{e.g. 'NA'. See Details.}
}
\description{
Retrieve effrs afscms data from afs database
}

