% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afs_effrs_afsplt1.R
\name{afs_effrs_afsplt1}
\alias{afs_effrs_afsplt1}
\title{Retrieve effrs afsplt1 data from afs database}
\usage{
afs_effrs_afsplt1(EPA_PLANT_ID = NULL, PLANT_ID = NULL,
  REGION_CODE = NULL, STATE_ABBREV = NULL, STATE_NAME = NULL,
  COUNTY_NAME = NULL, DUN_BRADSTREET_NUMBER = NULL,
  STATE_COUNTY_COMPLIANCE_SOURCE = NULL, STATE_CODE = NULL,
  COUNTY_CODE = NULL, CDS_PLANT_ID = NULL, PLANT_NAME = NULL,
  PLANT_STREET_ADDRESS = NULL, PLANT_CITY = NULL, ZIP_CODE = NULL,
  PRIMARY_SIC_CODE = NULL, PRIMARY_SIC_DESCRIPTION = NULL,
  SECONDARY_SIC_CODE = NULL, TERTIARY_SIC_CODE = NULL,
  SIC_EQUIVALENT_NAICS = NULL, NAICS_EQUIVALENT_SIC = NULL,
  GOVT_FACILITY_CODE_DESC = NULL, NAICS_CODE = NULL,
  NAICS_CODE_DESCRIPTION = NULL, STATE_REGISTRATION_NUMBER = NULL,
  LOCAL_CONTROL_REGION = NULL, LOCAL_CONTROL_REGION_DESC = NULL,
  OPERATING_STATUS = NULL, OPERATING_STATUS_DESCRIPTION = NULL,
  HPV_FLAG = NULL, DATE_PLANT_INFO_LAST_UPDATED = NULL,
  NAT_INIT_AIR_TOXICS = NULL, NAT_INIT_AIR_TOXICS_DESC = NULL,
  AFS_PLANT_ID = NULL)
}
\arguments{
\item{EPA_PLANT_ID}{e.g. '110000307695'. See Details.}

\item{PLANT_ID}{e.g. '1011473'. See Details.}

\item{REGION_CODE}{e.g. '02'. See Details.}

\item{STATE_ABBREV}{e.g. 'PR'. See Details.}

\item{STATE_NAME}{e.g. 'PUERTO RICO'. See Details.}

\item{COUNTY_NAME}{e.g. 'BAYAMON'. See Details.}

\item{DUN_BRADSTREET_NUMBER}{e.g. 'NA'. See Details.}

\item{STATE_COUNTY_COMPLIANCE_SOURCE}{e.g. '7202100240'. See Details.}

\item{STATE_CODE}{e.g. '72'. See Details.}

\item{COUNTY_CODE}{e.g. '021'. See Details.}

\item{CDS_PLANT_ID}{e.g. '00240'. See Details.}

\item{PLANT_NAME}{e.g. 'TEEN CHALLENGE CLEANERS'. See Details.}

\item{PLANT_STREET_ADDRESS}{e.g. 'MARGINAL C-21 URB. STA. CRUZ'. See Details.}

\item{PLANT_CITY}{e.g. 'BAYAMON'. See Details.}

\item{ZIP_CODE}{e.g. '00956'. See Details.}

\item{PRIMARY_SIC_CODE}{e.g. '7216'. See Details.}

\item{PRIMARY_SIC_DESCRIPTION}{e.g. 'DRY CLEANING PLANTS, EXCE'. See Details.}

\item{SECONDARY_SIC_CODE}{e.g. 'NA'. See Details.}

\item{TERTIARY_SIC_CODE}{e.g. 'NA'. See Details.}

\item{SIC_EQUIVALENT_NAICS}{e.g. 'PRIV'. See Details.}

\item{NAICS_EQUIVALENT_SIC}{e.g. 'ATELY'. See Details.}

\item{GOVT_FACILITY_CODE_DESC}{e.g. 'PRIVATELY OWNED/OPERATED'. See Details.}

\item{NAICS_CODE}{e.g. '812320'. See Details.}

\item{NAICS_CODE_DESCRIPTION}{e.g. 'Drycleaning and Laundry Services (except Coin-Operated)'. See Details.}

\item{STATE_REGISTRATION_NUMBER}{e.g. '1001-2240'. See Details.}

\item{LOCAL_CONTROL_REGION}{e.g. 'NA'. See Details.}

\item{LOCAL_CONTROL_REGION_DESC}{e.g. 'NA'. See Details.}

\item{OPERATING_STATUS}{e.g. 'O'. See Details.}

\item{OPERATING_STATUS_DESCRIPTION}{e.g. 'OPERATING'. See Details.}

\item{HPV_FLAG}{e.g. 'NA'. See Details.}

\item{DATE_PLANT_INFO_LAST_UPDATED}{e.g. '14-MAR-14'. See Details.}

\item{NAT_INIT_AIR_TOXICS}{e.g. 'NA'. See Details.}

\item{NAT_INIT_AIR_TOXICS_DESC}{e.g. 'NA'. See Details.}

\item{AFS_PLANT_ID}{e.g. 'NA'. See Details.}
}
\description{
Retrieve effrs afsplt1 data from afs database
}

