% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afs_effrs_afsplt2.R
\name{afs_effrs_afsplt2}
\alias{afs_effrs_afsplt2}
\title{Retrieve effrs afsplt2 data from afs database}
\usage{
afs_effrs_afsplt2(STATE_COUNTY_COMPLIANCE_SOURCE = NULL,
  ES_CLASS_CODE = NULL, ES_CLASS_CODE_DESC = NULL,
  ES_COMPLIANCE_STATUS = NULL, ES_COMPLIANCE_STATUS_DESC = NULL,
  EPA_CLASS_CODE = NULL, EPA_CLASS_CODE_DESC = NULL,
  EPA_COMPLIANCE_STATUS = NULL, EPA_COMPLIANCE_STATUS_DESC = NULL,
  STATE_CLASS_CODE = NULL, STATE_CLASS_CODE_DESC = NULL,
  STATE_COMPLIANCE_STATUS = NULL, STATE_COMPLIANCE_STATUS_DESC = NULL,
  ES_LATEST_FCE_DATE_ACHIEVED = NULL, ES_LATEST_FCE_ACTION_TYPE = NULL,
  STATE_LATEST_FCE_DATE_ACHIEVED = NULL,
  STATE_LATEST_FCE_ACTION_TYPE = NULL, EPA_LATEST_FCE_DATE_ACHIEVED = NULL,
  EPA_LATEST_FCE_ACTION_TYPE = NULL, CONTACT_PERSION_COMPLIANCE = NULL,
  PHONE_AREA_CODE_COMPLIANCE = NULL, PHONE_COMPLIANCE = NULL,
  MAILING_CITY = NULL, MAILING_PLANT_NAME = NULL,
  MAILING_STATE_ABBREV = NULL, MAILING_STATE_NAME = NULL,
  MAILING_STREET_ADDRESS = NULL, MAILING_ZIP = NULL)
}
\arguments{
\item{STATE_COUNTY_COMPLIANCE_SOURCE}{e.g. '0100100001'. See Details.}

\item{ES_CLASS_CODE}{e.g. 'A'. See Details.}

\item{ES_CLASS_CODE_DESC}{e.g. 'ACTUAL OR POTENTIAL EMISS'. See Details.}

\item{ES_COMPLIANCE_STATUS}{e.g. '4'. See Details.}

\item{ES_COMPLIANCE_STATUS_DESC}{e.g. 'IN COMPLIANCE - CERTIFICA'. See Details.}

\item{EPA_CLASS_CODE}{e.g. 'NA'. See Details.}

\item{EPA_CLASS_CODE_DESC}{e.g. 'NA'. See Details.}

\item{EPA_COMPLIANCE_STATUS}{e.g. '4'. See Details.}

\item{EPA_COMPLIANCE_STATUS_DESC}{e.g. 'IN COMPLIANCE - CERTIFICA'. See Details.}

\item{STATE_CLASS_CODE}{e.g. 'A'. See Details.}

\item{STATE_CLASS_CODE_DESC}{e.g. 'ACTUAL OR POTENTIAL EMISS'. See Details.}

\item{STATE_COMPLIANCE_STATUS}{e.g. '4'. See Details.}

\item{STATE_COMPLIANCE_STATUS_DESC}{e.g. 'IN COMPLIANCE - CERTIFICA'. See Details.}

\item{ES_LATEST_FCE_DATE_ACHIEVED}{e.g. '30-SEP-13'. See Details.}

\item{ES_LATEST_FCE_ACTION_TYPE}{e.g. 'FS'. See Details.}

\item{STATE_LATEST_FCE_DATE_ACHIEVED}{e.g. '30-SEP-13'. See Details.}

\item{STATE_LATEST_FCE_ACTION_TYPE}{e.g. 'FS'. See Details.}

\item{EPA_LATEST_FCE_DATE_ACHIEVED}{e.g. 'NA'. See Details.}

\item{EPA_LATEST_FCE_ACTION_TYPE}{e.g. 'NA'. See Details.}

\item{CONTACT_PERSION_COMPLIANCE}{e.g. 'NA'. See Details.}

\item{PHONE_AREA_CODE_COMPLIANCE}{e.g. 'NA'. See Details.}

\item{PHONE_COMPLIANCE}{e.g. 'NA'. See Details.}

\item{MAILING_CITY}{e.g. 'NA'. See Details.}

\item{MAILING_PLANT_NAME}{e.g. 'NA'. See Details.}

\item{MAILING_STATE_ABBREV}{e.g. 'NA'. See Details.}

\item{MAILING_STATE_NAME}{e.g. 'NA'. See Details.}

\item{MAILING_STREET_ADDRESS}{e.g. 'NA'. See Details.}

\item{MAILING_ZIP}{e.g. 'NA'. See Details.}
}
\description{
Retrieve effrs afsplt2 data from afs database
}

