% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afs_effrs_afssubp.R
\name{afs_effrs_afssubp}
\alias{afs_effrs_afssubp}
\title{Retrieve effrs afssubp data from afs database}
\usage{
afs_effrs_afssubp(EPA_PLANT_ID = NULL, PLANT_ID = NULL,
  REGION_CODE = NULL, STATE_ABBREV = NULL, STATE_NAME = NULL,
  COUNTY_NAME = NULL, DUN_BRADSTREET_NUMBER = NULL,
  STATE_COUNTY_COMPLIANCE_SOURCE = NULL, STATE_CODE = NULL,
  COUNTY_CODE = NULL, CDS_PLANT_ID = NULL, AIR_PROGRAM_CODE = NULL,
  SUBPART_HORIZONTAL = NULL)
}
\arguments{
\item{EPA_PLANT_ID}{e.g. '110000307695'. See Details.}

\item{PLANT_ID}{e.g. '1011473'. See Details.}

\item{REGION_CODE}{e.g. '02'. See Details.}

\item{STATE_ABBREV}{e.g. 'PR'. See Details.}

\item{STATE_NAME}{e.g. 'PUERTO RICO'. See Details.}

\item{COUNTY_NAME}{e.g. 'BAYAMON'. See Details.}

\item{DUN_BRADSTREET_NUMBER}{e.g. 'NA'. See Details.}

\item{STATE_COUNTY_COMPLIANCE_SOURCE}{e.g. '7202100240'. See Details.}

\item{STATE_CODE}{e.g. '72'. See Details.}

\item{COUNTY_CODE}{e.g. '021'. See Details.}

\item{CDS_PLANT_ID}{e.g. '00240'. See Details.}

\item{AIR_PROGRAM_CODE}{e.g. '0'. See Details.}

\item{SUBPART_HORIZONTAL}{e.g. 'NA'. See Details.}
}
\description{
Retrieve effrs afssubp data from afs database
}
