% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/br_bgm_offsite_shipment.R
\name{br_bgm_offsite_shipment}
\alias{br_bgm_offsite_shipment}
\title{Retrieve bgm offsite shipment data from br database}
\usage{
br_bgm_offsite_shipment(HANDLER_ID = NULL, ACTIVITY_LOCATION = NULL,
  SOURCE_TYPE = NULL, SEQ_NUMBER = NULL, HZ_PG = NULL,
  IO_PG_NUM_SEQ = NULL, IO_TDR_ID = NULL, MANAGEMENT_METHOD = NULL,
  MANAGEMENT_METHOD_OWNER = NULL, IO_TDR_QTY = NULL,
  SHIPPED_TONS_CALC = NULL, USERID = NULL, LAST_CHANGE = NULL)
}
\arguments{
\item{HANDLER_ID}{e.g. 'AK0000016709'. See Details.}

\item{ACTIVITY_LOCATION}{e.g. 'AK'. See Details.}

\item{SOURCE_TYPE}{e.g. 'R'. See Details.}

\item{SEQ_NUMBER}{e.g. '1'. See Details.}

\item{HZ_PG}{e.g. '1'. See Details.}

\item{IO_PG_NUM_SEQ}{e.g. '1'. See Details.}

\item{IO_TDR_ID}{e.g. 'WAD000812909'. See Details.}

\item{MANAGEMENT_METHOD}{e.g. 'H040'. See Details.}

\item{MANAGEMENT_METHOD_OWNER}{e.g. 'HQ'. See Details.}

\item{IO_TDR_QTY}{e.g. '9267'. See Details.}

\item{SHIPPED_TONS_CALC}{e.g. '4.6335'. See Details.}

\item{USERID}{e.g. 'IYWCVT'. See Details.}

\item{LAST_CHANGE}{e.g. '10-NOV-02'. See Details.}
}
\description{
Retrieve bgm offsite shipment data from br database
}
