% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/br_bgm_waste_code.R
\name{br_bgm_waste_code}
\alias{br_bgm_waste_code}
\title{Retrieve bgm waste code data from br database}
\usage{
br_bgm_waste_code(HANDLER_ID = NULL, ACTIVITY_LOCATION = NULL,
  SOURCE_TYPE = NULL, SEQ_NUMBER = NULL, HZ_PG = NULL,
  WASTE_CODE_OWNER = NULL, WASTE_CODE = NULL, USERID = NULL,
  LAST_CHANGE = NULL)
}
\arguments{
\item{HANDLER_ID}{e.g. 'AK0000016709'. See Details.}

\item{ACTIVITY_LOCATION}{e.g. 'AK'. See Details.}

\item{SOURCE_TYPE}{e.g. 'R'. See Details.}

\item{SEQ_NUMBER}{e.g. '1'. See Details.}

\item{HZ_PG}{e.g. '1'. See Details.}

\item{WASTE_CODE_OWNER}{e.g. 'HQ'. See Details.}

\item{WASTE_CODE}{e.g. 'D008'. See Details.}

\item{USERID}{e.g. 'IYWCVT'. See Details.}

\item{LAST_CHANGE}{e.g. '10-NOV-02'. See Details.}
}
\description{
Retrieve bgm waste code data from br database
}

