% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/br_breport_cycle.R
\name{br_breport_cycle}
\alias{br_breport_cycle}
\title{Retrieve breport cycle data from br database}
\usage{
br_breport_cycle(YEAR = NULL, USERID = NULL, LAST_CHANGE = NULL)
}
\arguments{
\item{YEAR}{e.g. '1989'. See Details.}

\item{USERID}{e.g. 'IYW'. See Details.}

\item{LAST_CHANGE}{e.g. '28-NOV-00'. See Details.}
}
\description{
Retrieve breport cycle data from br database
}

