% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/br_bwr_basic.R
\name{br_bwr_basic}
\alias{br_bwr_basic}
\title{Retrieve bwr basic data from br database}
\usage{
br_bwr_basic(HANDLER_ID = NULL, ACTIVITY_LOCATION = NULL,
  SOURCE_TYPE = NULL, SEQ_NUMBER = NULL, HZ_PG = NULL,
  SUB_PG_NUM = NULL, REPORT_CYCLE = NULL,
  INCLUDE_IN_NATIONAL_REPORT = NULL, IO_TDR_ID = NULL, IO_TDR_QTY = NULL,
  UOM_OWNER = NULL, UNIT_OF_MEASURE = NULL, WST_DENSITY = NULL,
  DENSITY_UOM_OWNER = NULL, DENSITY_UNIT_OF_MEASURE = NULL,
  RECEIVED_TONS_CALC = NULL, FORM_CODE_OWNER = NULL, FORM_CODE = NULL,
  RADIOACTIVE_OWNER = NULL, RADIOACTIVE_MIX = NULL,
  MANAGEMENT_METHOD_OWNER = NULL, MANAGEMENT_METHOD = NULL, USERID = NULL,
  LAST_CHANGE = NULL, DESCRIPTION = NULL, NOTES = NULL)
}
\arguments{
\item{HANDLER_ID}{e.g. 'AK6210022426'. See Details.}

\item{ACTIVITY_LOCATION}{e.g. 'AK'. See Details.}

\item{SOURCE_TYPE}{e.g. 'R'. See Details.}

\item{SEQ_NUMBER}{e.g. '5'. See Details.}

\item{HZ_PG}{e.g. '390'. See Details.}

\item{SUB_PG_NUM}{e.g. '2'. See Details.}

\item{REPORT_CYCLE}{e.g. '1999'. See Details.}

\item{INCLUDE_IN_NATIONAL_REPORT}{e.g. 'U'. See Details.}

\item{IO_TDR_ID}{e.g. 'AK1570028646'. See Details.}

\item{IO_TDR_QTY}{e.g. '87'. See Details.}

\item{UOM_OWNER}{e.g. 'HQ'. See Details.}

\item{UNIT_OF_MEASURE}{e.g. '1'. See Details.}

\item{WST_DENSITY}{e.g. '0'. See Details.}

\item{DENSITY_UOM_OWNER}{e.g. 'NA'. See Details.}

\item{DENSITY_UNIT_OF_MEASURE}{e.g. 'NA'. See Details.}

\item{RECEIVED_TONS_CALC}{e.g. '0.0435'. See Details.}

\item{FORM_CODE_OWNER}{e.g. 'US'. See Details.}

\item{FORM_CODE}{e.g. 'W310'. See Details.}

\item{RADIOACTIVE_OWNER}{e.g. 'HQ'. See Details.}

\item{RADIOACTIVE_MIX}{e.g. 'N'. See Details.}

\item{MANAGEMENT_METHOD_OWNER}{e.g. 'HQ'. See Details.}

\item{MANAGEMENT_METHOD}{e.g. 'H141'. See Details.}

\item{USERID}{e.g. 'IYWCVT'. See Details.}

\item{LAST_CHANGE}{e.g. '10-NOV-02'. See Details.}

\item{DESCRIPTION}{e.g. 'IGNITABLE, TOXIC FUEL FILTERS FROM FILTER REPLACEMENT CONTAINS BENZENE'. See Details.}

\item{NOTES}{e.g. 'NA'. See Details.}
}
\description{
Retrieve bwr basic data from br database
}

