% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/br_rcr_bgm_basic.R
\name{br_rcr_bgm_basic}
\alias{br_rcr_bgm_basic}
\title{Retrieve rcr bgm basic data from br database}
\usage{
br_rcr_bgm_basic(HANDLER_ID = NULL, ACTIVITY_LOCATION = NULL,
  SOURCE_TYPE = NULL, SEQ_NUMBER = NULL, HZ_PG = NULL,
  REPORT_CYCLE = NULL, INCLUDE_IN_NATIONAL_REPORT = NULL,
  WST_DENSITY = NULL, GEN_QTY = NULL, GENERATION_TONS_CALC = NULL,
  MANAGEMENT_METHOD_OWNER = NULL, MANAGEMENT_METHOD = NULL,
  UOM_OWNER = NULL, UNIT_OF_MEASURE = NULL, DENSITY_UOM_OWNER = NULL,
  DENSITY_UNIT_OF_MEASURE = NULL, RADIOACTIVE_OWNER = NULL,
  RADIOACTIVE_MIX = NULL, FORM_CODE_OWNER = NULL, FORM_CODE = NULL,
  SOURCE_CODE_OWNER = NULL, SOURCE_CODE = NULL, DESCRIPTION = NULL,
  NOTES = NULL, USERID = NULL, LAST_CHANGE = NULL)
}
\arguments{
\item{HANDLER_ID}{e.g. 'AK0000016709'. See Details.}

\item{ACTIVITY_LOCATION}{e.g. 'AK'. See Details.}

\item{SOURCE_TYPE}{e.g. 'R'. See Details.}

\item{SEQ_NUMBER}{e.g. '1'. See Details.}

\item{HZ_PG}{e.g. '1'. See Details.}

\item{REPORT_CYCLE}{e.g. '1997'. See Details.}

\item{INCLUDE_IN_NATIONAL_REPORT}{e.g. 'U'. See Details.}

\item{WST_DENSITY}{e.g. 'NA'. See Details.}

\item{GEN_QTY}{e.g. '9267'. See Details.}

\item{GENERATION_TONS_CALC}{e.g. '4.6335'. See Details.}

\item{MANAGEMENT_METHOD_OWNER}{e.g. 'NA'. See Details.}

\item{MANAGEMENT_METHOD}{e.g. 'NA'. See Details.}

\item{UOM_OWNER}{e.g. 'HQ'. See Details.}

\item{UNIT_OF_MEASURE}{e.g. '1'. See Details.}

\item{DENSITY_UOM_OWNER}{e.g. 'NA'. See Details.}

\item{DENSITY_UNIT_OF_MEASURE}{e.g. 'NA'. See Details.}

\item{RADIOACTIVE_OWNER}{e.g. 'HQ'. See Details.}

\item{RADIOACTIVE_MIX}{e.g. 'N'. See Details.}

\item{FORM_CODE_OWNER}{e.g. 'US'. See Details.}

\item{FORM_CODE}{e.g. 'W110'. See Details.}

\item{SOURCE_CODE_OWNER}{e.g. 'HQ'. See Details.}

\item{SOURCE_CODE}{e.g. 'G14'. See Details.}

\item{DESCRIPTION}{e.g. 'WASTEWATER SLUDGE CONTAINING SODIUM METAL SILICATE & LEAD'. See Details.}

\item{NOTES}{e.g. 'NA'. See Details.}

\item{USERID}{e.g. 'IYWCVT'. See Details.}

\item{LAST_CHANGE}{e.g. '10-NOV-02'. See Details.}
}
\description{
Retrieve rcr bgm basic data from br database
}
