% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/br_rcr_bwr_waste_code.R
\name{br_rcr_bwr_waste_code}
\alias{br_rcr_bwr_waste_code}
\title{Retrieve rcr bwr waste code data from br database}
\usage{
br_rcr_bwr_waste_code(HANDLER_ID = NULL, ACTIVITY_LOCATION = NULL,
  SOURCE_TYPE = NULL, SEQ_NUMBER = NULL, HZ_PG = NULL,
  SUB_PAGE_NUM = NULL, WASTE_CODE_OWNER = NULL, WASTE_CODE = NULL)
}
\arguments{
\item{HANDLER_ID}{e.g. 'AK8570028649'. See Details.}

\item{ACTIVITY_LOCATION}{e.g. 'AK'. See Details.}

\item{SOURCE_TYPE}{e.g. 'R'. See Details.}

\item{SEQ_NUMBER}{e.g. '9'. See Details.}

\item{HZ_PG}{e.g. '8'. See Details.}

\item{SUB_PAGE_NUM}{e.g. '1'. See Details.}

\item{WASTE_CODE_OWNER}{e.g. 'HQ'. See Details.}

\item{WASTE_CODE}{e.g. 'D001'. See Details.}
}
\description{
Retrieve rcr bwr waste code data from br database
}
