% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/br_rcr_lu_form_code2.R
\name{br_rcr_lu_form_code2}
\alias{br_rcr_lu_form_code2}
\title{Retrieve rcr lu form code2 data from br database}
\usage{
br_rcr_lu_form_code2(OWNER = NULL, FORM_CODE = NULL,
  FORM_CODE_NAME = NULL, FORM_CODE_DESC = NULL, USAGE = NULL,
  ACTIVE_STATUS = NULL)
}
\arguments{
\item{OWNER}{e.g. 'US'. See Details.}

\item{FORM_CODE}{e.g. 'W101'. See Details.}

\item{FORM_CODE_NAME}{e.g. 'DILUTE AQUEOUS WST'. See Details.}

\item{FORM_CODE_DESC}{e.g. 'VERY DILUTE AQUEOUS WASTE CONTAINING MORE THAN 99% WATER'. See Details.}

\item{USAGE}{e.g. '3'. See Details.}

\item{ACTIVE_STATUS}{e.g. 'Y'. See Details.}
}
\description{
Retrieve rcr lu form code2 data from br database
}

