% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/br_rcr_lu_source_code2.R
\name{br_rcr_lu_source_code2}
\alias{br_rcr_lu_source_code2}
\title{Retrieve rcr lu source code2 data from br database}
\usage{
br_rcr_lu_source_code2(OWNER = NULL, SOURCE_CODE = NULL,
  SOURCE_CODE_NAME = NULL, USAGE = NULL, ACTIVE_STATUS = NULL,
  SOURCE_CODE_DESC = NULL)
}
\arguments{
\item{OWNER}{e.g. 'HQ'. See Details.}

\item{SOURCE_CODE}{e.g. 'G01'. See Details.}

\item{SOURCE_CODE_NAME}{e.g. 'RINSING'. See Details.}

\item{USAGE}{e.g. '1'. See Details.}

\item{ACTIVE_STATUS}{e.g. 'Y'. See Details.}

\item{SOURCE_CODE_DESC}{e.g. 'DIP, FLUSH OR SPRAY RINSING'. See Details.}
}
\description{
Retrieve rcr lu source code2 data from br database
}
