% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerclis_cer_action.R
\name{cerclis_cer_action}
\alias{cerclis_cer_action}
\title{Retrieve cer action data from cerclis database}
\usage{
cerclis_cer_action(RAT_CODE = NULL, SITE_ID = NULL, ACT_CODE_ID = NULL,
  ACT_ACTL_STRT_DATE = NULL, ACT_ACTL_CMPLTN_DATE = NULL,
  RAQ_ACTION_QUAL_CODE = NULL, OU_ID = NULL, RALT_CODE = NULL,
  RPS_CODE = NULL, RACT_CRITICAL_FLAG_CODE = NULL, RAA_CODE = NULL)
}
\arguments{
\item{RAT_CODE}{e.g. '5L'. See Details.}

\item{SITE_ID}{e.g. '0703163'. See Details.}

\item{ACT_CODE_ID}{e.g. '001'. See Details.}

\item{ACT_ACTL_STRT_DATE}{e.g. 'NA'. See Details.}

\item{ACT_ACTL_CMPLTN_DATE}{e.g. '27-OCT-98'. See Details.}

\item{RAQ_ACTION_QUAL_CODE}{e.g. 'NA'. See Details.}

\item{OU_ID}{e.g. '00'. See Details.}

\item{RALT_CODE}{e.g. 'EP'. See Details.}

\item{RPS_CODE}{e.g. 'NA'. See Details.}

\item{RACT_CRITICAL_FLAG_CODE}{e.g. 'NA'. See Details.}

\item{RAA_CODE}{e.g. 'NA'. See Details.}
}
\description{
Retrieve cer action data from cerclis database
}

