% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerclis_cer_cmt_site.R
\name{cerclis_cer_cmt_site}
\alias{cerclis_cer_cmt_site}
\title{Retrieve cer cmt site data from cerclis database}
\usage{
cerclis_cer_cmt_site(SITE_ID = NULL, RCST_CODE = NULL,
  CSITE_SEQ_NMBR = NULL, CSITE_TEXT = NULL)
}
\arguments{
\item{SITE_ID}{e.g. '0100041'. See Details.}

\item{RCST_CODE}{e.g. '01'. See Details.}

\item{CSITE_SEQ_NMBR}{e.g. '1274232'. See Details.}

\item{CSITE_TEXT}{e.g. 'Linemaster Switch is an active manufacturing facility.'}
}
\description{
Retrieve cer cmt site data from cerclis database
}
