% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerclis_cer_contact.R
\name{cerclis_cer_contact}
\alias{cerclis_cer_contact}
\title{Retrieve cer contact data from cerclis database}
\usage{
cerclis_cer_contact(SITE_ID = NULL, PERSON_LAST_NAME = NULL,
  PERSON_FIRST_NAME = NULL, PERSON_PHONE_NMBR1 = NULL, RROLE_DESC = NULL)
}
\arguments{
\item{SITE_ID}{e.g. '0100015'. See Details.}

\item{PERSON_LAST_NAME}{e.g. 'HAWORTH'. See Details.}

\item{PERSON_FIRST_NAME}{e.g. 'RICHARD'. See Details.}

\item{PERSON_PHONE_NMBR1}{e.g. '6179181229'. See Details.}

\item{RROLE_DESC}{e.g. 'On-Scene Coordinator (OSC)'. See Details.}
}
\description{
Retrieve cer contact data from cerclis database
}

