% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerclis_cer_federal_register.R
\name{cerclis_cer_federal_register}
\alias{cerclis_cer_federal_register}
\title{Retrieve cer federal register data from cerclis database}
\usage{
cerclis_cer_federal_register(SITE_ID = NULL, FR_DATE = NULL,
  RFRA_CODE = NULL, FR_NPL_UPDATE_NMBR = NULL)
}
\arguments{
\item{SITE_ID}{e.g. '0100041'. See Details.}

\item{FR_DATE}{e.g. '24-JUN-88'. See Details.}

\item{RFRA_CODE}{e.g. 'P'. See Details.}

\item{FR_NPL_UPDATE_NMBR}{e.g. '7'. See Details.}
}
\description{
Retrieve cer federal register data from cerclis database
}

