% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerclis_cer_financial_transaction.R
\name{cerclis_cer_financial_transaction}
\alias{cerclis_cer_financial_transaction}
\title{Retrieve cer financial transaction data from cerclis database}
\usage{
cerclis_cer_financial_transaction(RAT_CODE = NULL, ACT_CODE_ID = NULL,
  SITE_ID = NULL, RFTT_CODE = NULL, FT_ID = NULL, FT_DATE = NULL,
  FT_AMT = NULL, FT_BUDGET_FY = NULL)
}
\arguments{
\item{RAT_CODE}{e.g. 'AA'. See Details.}

\item{ACT_CODE_ID}{e.g. '001'. See Details.}

\item{SITE_ID}{e.g. '0603590'. See Details.}

\item{RFTT_CODE}{e.g. 'D'. See Details.}

\item{FT_ID}{e.g. '0001'. See Details.}

\item{FT_DATE}{e.g. '21-DEC-98'. See Details.}

\item{FT_AMT}{e.g. '95'. See Details.}

\item{FT_BUDGET_FY}{e.g. '1999'. See Details.}
}
\description{
Retrieve cer financial transaction data from cerclis database
}
