% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerclis_cer_location.R
\name{cerclis_cer_location}
\alias{cerclis_cer_location}
\title{Retrieve cer location data from cerclis database}
\usage{
cerclis_cer_location(SITE_EPA_ID = NULL, LATITUDE = NULL,
  LONGITUDE = NULL, ACCURACY_VALUE = NULL, SUB_ID = NULL,
  SUB_TYPE_CODE = NULL, GEOMETRIC_TYPE_CODE = NULL,
  VERTICAL_MEASURE = NULL, VMETHOD_CODE = NULL, VDATUM_CODE = NULL,
  VACCURACY = NULL, COLLECT_MTH_CODE = NULL, SCALE = NULL,
  SOURCE_CODE = NULL, HDATUM_CODE = NULL, REF_POINT_CODE = NULL,
  COLLECTION_DATE = NULL, CITY_NAME = NULL, COUNTY_NAME = NULL,
  LOCATION_ADDRESS = NULL, POSTAL_CODE = NULL, PRIMARY_NAME = NULL,
  STATE_CODE = NULL, COMMENT_TEXT = NULL)
}
\arguments{
\item{SITE_EPA_ID}{e.g. 'AK0001897602'. See Details.}

\item{LATITUDE}{e.g. '55.626389'. See Details.}

\item{LONGITUDE}{e.g. '-132.558614'. See Details.}

\item{ACCURACY_VALUE}{e.g. '0'. See Details.}

\item{SUB_ID}{e.g. '003'. See Details.}

\item{SUB_TYPE_CODE}{e.g. 'NA'. See Details.}

\item{GEOMETRIC_TYPE_CODE}{e.g. '001'. See Details.}

\item{VERTICAL_MEASURE}{e.g. 'NA'. See Details.}

\item{VMETHOD_CODE}{e.g. '027'. See Details.}

\item{VDATUM_CODE}{e.g. 'NA'. See Details.}

\item{VACCURACY}{e.g. 'NA'. See Details.}

\item{COLLECT_MTH_CODE}{e.g. '027'. See Details.}

\item{SCALE}{e.g. 'NA'. See Details.}

\item{SOURCE_CODE}{e.g. 'NA'. See Details.}

\item{HDATUM_CODE}{e.g. '002'. See Details.}

\item{REF_POINT_CODE}{e.g. '001'. See Details.}

\item{COLLECTION_DATE}{e.g. '31-DEC-93'. See Details.}

\item{CITY_NAME}{e.g. 'THORNE BAY'. See Details.}

\item{COUNTY_NAME}{e.g. 'OUTER KETCHIKAN'. See Details.}

\item{LOCATION_ADDRESS}{e.g. '4 MI SW CY,  T72S R84E S17, COPPER RIVER'. See Details.}

\item{POSTAL_CODE}{e.g. '99919'. See Details.}

\item{PRIMARY_NAME}{e.g. 'SALT CHUCK MINE'. See Details.}

\item{STATE_CODE}{e.g. 'AK'. See Details.}

\item{COMMENT_TEXT}{e.g. 'SOURCE = USGS'. See Details.}
}
\description{
Retrieve cer location data from cerclis database
}
