% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerclis_cer_operable_unit.R
\name{cerclis_cer_operable_unit}
\alias{cerclis_cer_operable_unit}
\title{Retrieve cer operable unit data from cerclis database}
\usage{
cerclis_cer_operable_unit(SITE_ID = NULL, OU_ID = NULL, OU_NAME = NULL,
  OU_DESC_TEXT = NULL)
}
\arguments{
\item{SITE_ID}{e.g. ' or OU4'. See Details.}

\item{OU_ID}{e.g. '.'. See Details.}

\item{OU_NAME}{e.g. 'NA'. See Details.}

\item{OU_DESC_TEXT}{e.g. 'NA'. See Details.}
}
\description{
Retrieve cer operable unit data from cerclis database
}

