% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerclis_cer_ref_action_anomaly.R
\name{cerclis_cer_ref_action_anomaly}
\alias{cerclis_cer_ref_action_anomaly}
\title{Retrieve cer ref action anomaly data from cerclis database}
\usage{
cerclis_cer_ref_action_anomaly(RAA_CODE = NULL, RAA_DESC = NULL)
}
\arguments{
\item{RAA_CODE}{e.g. 'OA'. See Details.}

\item{RAA_DESC}{e.g. 'Other Start and Completion Anomaly'. See Details.}
}
\description{
Retrieve cer ref action anomaly data from cerclis database
}
