% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerclis_cer_ref_action_critical_flag.R
\name{cerclis_cer_ref_action_critical_flag}
\alias{cerclis_cer_ref_action_critical_flag}
\title{Retrieve cer ref action critical flag data from cerclis database}
\usage{
cerclis_cer_ref_action_critical_flag(RACT_CRITICAL_FLAG_CODE = NULL,
  RACT_CRITICAL_FLAG_DESC = NULL)
}
\arguments{
\item{RACT_CRITICAL_FLAG_CODE}{e.g. '1'. See Details.}

\item{RACT_CRITICAL_FLAG_DESC}{e.g. 'Emergency'. See Details.}
}
\description{
Retrieve cer ref action critical flag data from cerclis database
}
