% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerclis_cer_ref_action_lead_type.R
\name{cerclis_cer_ref_action_lead_type}
\alias{cerclis_cer_ref_action_lead_type}
\title{Retrieve cer ref action lead type data from cerclis database}
\usage{
cerclis_cer_ref_action_lead_type(RALT_CODE = NULL, RALT_DESC = NULL)
}
\arguments{
\item{RALT_CODE}{e.g. 'CG'. See Details.}

\item{RALT_DESC}{e.g. 'Coast Guard'. See Details.}
}
\description{
Retrieve cer ref action lead type data from cerclis database
}
