% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerclis_cer_ref_action_qualifier.R
\name{cerclis_cer_ref_action_qualifier}
\alias{cerclis_cer_ref_action_qualifier}
\title{Retrieve cer ref action qualifier data from cerclis database}
\usage{
cerclis_cer_ref_action_qualifier(RAQ_ACTION_QUAL_CODE = NULL,
  RAQ_ACTION_QUAL_DESC = NULL)
}
\arguments{
\item{RAQ_ACTION_QUAL_CODE}{e.g. 'A'. See Details.}

\item{RAQ_ACTION_QUAL_DESC}{e.g. 'Addressed as part of an existing NPL site'. See Details.}
}
\description{
Retrieve cer ref action qualifier data from cerclis database
}
