% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerclis_cer_ref_action_type.R
\name{cerclis_cer_ref_action_type}
\alias{cerclis_cer_ref_action_type}
\title{Retrieve cer ref action type data from cerclis database}
\usage{
cerclis_cer_ref_action_type(RAT_CODE = NULL, RAT_NAME = NULL,
  RAT_SHORT_NAME = NULL, RAT_WITHOUT_START_FLAG = NULL,
  RAT_C2_CODE = NULL, RAT_C2_TYPE = NULL,
  RAT_HISTORICAL_ONLY_FLAG = NULL, RAT_NSI_INDICATOR = NULL,
  RAT_LEVEL = NULL, RAT_DEFAULT_PRED_ACT_TYPE_CODE = NULL,
  RIE_CODE = NULL, RAT_DEF_OU = NULL, RFBS_CODE = NULL,
  RAT_DEF = NULL)
}
\arguments{
\item{RAT_CODE}{e.g. '3D'. See Details.}

\item{RAT_NAME}{e.g. 'IC Report'. See Details.}

\item{RAT_SHORT_NAME}{e.g. 'TEST ACTN'. See Details.}

\item{RAT_WITHOUT_START_FLAG}{e.g. 'N'. See Details.}

\item{RAT_C2_CODE}{e.g. 'NA'. See Details.}

\item{RAT_C2_TYPE}{e.g. 'NA'. See Details.}

\item{RAT_HISTORICAL_ONLY_FLAG}{e.g. 'NA'. See Details.}

\item{RAT_NSI_INDICATOR}{e.g. 'S'. See Details.}

\item{RAT_LEVEL}{e.g. '1'. See Details.}

\item{RAT_DEFAULT_PRED_ACT_TYPE_CODE}{e.g. 'RO'. See Details.}

\item{RIE_CODE}{e.g. 'N'. See Details.}

\item{RAT_DEF_OU}{e.g. '00'. See Details.}

\item{RFBS_CODE}{e.g. 'NA'. See Details.}

\item{RAT_DEF}{e.g. 'tHIS IS A TEST rEGIONALLY dEFINED ACTION THAT i'VE SET UP FOR sa TRAINING'. See Details.}
}
\description{
Retrieve cer ref action type data from cerclis database
}
