% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerclis_cer_ref_cmt_site.R
\name{cerclis_cer_ref_cmt_site}
\alias{cerclis_cer_ref_cmt_site}
\title{Retrieve cer ref cmt site data from cerclis database}
\usage{
cerclis_cer_ref_cmt_site(RCST_CODE = NULL, RCS_DESC = NULL)
}
\arguments{
\item{RCST_CODE}{e.g. '01'. See Details.}

\item{RCS_DESC}{e.g. 'site_dscrptn_text'. See Details.}
}
\description{
Retrieve cer ref cmt site data from cerclis database
}

