% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerclis_cer_ref_fdrl_rgstr_actn.R
\name{cerclis_cer_ref_fdrl_rgstr_actn}
\alias{cerclis_cer_ref_fdrl_rgstr_actn}
\title{Retrieve cer ref fdrl rgstr actn data from cerclis database}
\usage{
cerclis_cer_ref_fdrl_rgstr_actn(RFRA_CODE = NULL, RFRA_DESC = NULL)
}
\arguments{
\item{RFRA_CODE}{e.g. 'D'. See Details.}

\item{RFRA_DESC}{e.g. 'Deleted from the Final Federal Register'. See Details.}
}
\description{
Retrieve cer ref fdrl rgstr actn data from cerclis database
}

