% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerclis_cer_ref_federal_facility.R
\name{cerclis_cer_ref_federal_facility}
\alias{cerclis_cer_ref_federal_facility}
\title{Retrieve cer ref federal facility data from cerclis database}
\usage{
cerclis_cer_ref_federal_facility(RFED_FCLTY_CODE = NULL,
  RFED_FCLTY_DESC = NULL)
}
\arguments{
\item{RFED_FCLTY_CODE}{e.g. 'D'. See Details.}

\item{RFED_FCLTY_DESC}{e.g. 'Status Undetermined'. See Details.}
}
\description{
Retrieve cer ref federal facility data from cerclis database
}

