% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerclis_cer_ref_finan_transaction_type.R
\name{cerclis_cer_ref_finan_transaction_type}
\alias{cerclis_cer_ref_finan_transaction_type}
\title{Retrieve cer ref finan transaction type data from cerclis database}
\usage{
cerclis_cer_ref_finan_transaction_type(RFTT_CODE = NULL,
  RFTT_LONG_NAME = NULL, RFTT_SHORT_NAME = NULL, RFTT_PLAN_ACT = NULL,
  RFTT_ORDER = NULL)
}
\arguments{
\item{RFTT_CODE}{e.g. 'A'. See Details.}

\item{RFTT_LONG_NAME}{e.g. 'Actual Obligation'. See Details.}

\item{RFTT_SHORT_NAME}{e.g. 'Actual Ob'. See Details.}

\item{RFTT_PLAN_ACT}{e.g. 'A'. See Details.}

\item{RFTT_ORDER}{e.g. '05'. See Details.}
}
\description{
Retrieve cer ref finan transaction type data from cerclis database
}

