% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerclis_cer_ref_ownership_type.R
\name{cerclis_cer_ref_ownership_type}
\alias{cerclis_cer_ref_ownership_type}
\title{Retrieve cer ref ownership type data from cerclis database}
\usage{
cerclis_cer_ref_ownership_type(ROT_CODE = NULL, ROT_DESC = NULL)
}
\arguments{
\item{ROT_CODE}{e.g. 'BL'. See Details.}

\item{ROT_DESC}{e.g. 'Bank/Loan Company'. See Details.}
}
\description{
Retrieve cer ref ownership type data from cerclis database
}

