% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerclis_cer_ref_planning_status.R
\name{cerclis_cer_ref_planning_status}
\alias{cerclis_cer_ref_planning_status}
\title{Retrieve cer ref planning status data from cerclis database}
\usage{
cerclis_cer_ref_planning_status(RPS_CODE = NULL, RPS_DESC = NULL)
}
\arguments{
\item{RPS_CODE}{e.g. 'A'. See Details.}

\item{RPS_DESC}{e.g. 'Alternate'. See Details.}
}
\description{
Retrieve cer ref planning status data from cerclis database
}
