% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerclis_cer_ref_region.R
\name{cerclis_cer_ref_region}
\alias{cerclis_cer_ref_region}
\title{Retrieve cer ref region data from cerclis database}
\usage{
cerclis_cer_ref_region(RREGION_CODE = NULL, RREGION_DESC = NULL,
  RREGION_STRT_ADRS1 = NULL, RREGION_STRT_ADRS2 = NULL,
  RREGION_CITY_NAME = NULL, RREGION_ZIP_CODE = NULL,
  RSTATE_CODE = NULL)
}
\arguments{
\item{RREGION_CODE}{e.g. '00'. See Details.}

\item{RREGION_DESC}{e.g. 'National'. See Details.}

\item{RREGION_STRT_ADRS1}{e.g. 'NA'. See Details.}

\item{RREGION_STRT_ADRS2}{e.g. 'NA'. See Details.}

\item{RREGION_CITY_NAME}{e.g. 'NA'. See Details.}

\item{RREGION_ZIP_CODE}{e.g. 'NA'. See Details.}

\item{RSTATE_CODE}{e.g. 'NA'. See Details.}
}
\description{
Retrieve cer ref region data from cerclis database
}
