% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerclis_cer_ref_sic.R
\name{cerclis_cer_ref_sic}
\alias{cerclis_cer_ref_sic}
\title{Retrieve cer ref sic data from cerclis database}
\usage{
cerclis_cer_ref_sic(RSIC_CODE = NULL, RSIC_CODE_DESC = NULL)
}
\arguments{
\item{RSIC_CODE}{e.g. '0000'. See Details.}

\item{RSIC_CODE_DESC}{e.g. 'Other'. See Details.}
}
\description{
Retrieve cer ref sic data from cerclis database
}
