% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerclis_cer_ref_site_incident_category.R
\name{cerclis_cer_ref_site_incident_category}
\alias{cerclis_cer_ref_site_incident_category}
\title{Retrieve cer ref site incident category data from cerclis database}
\usage{
cerclis_cer_ref_site_incident_category(RSITINC_CODE = NULL,
  RSITINC_DESC = NULL)
}
\arguments{
\item{RSITINC_CODE}{e.g. 'A'. See Details.}

\item{RSITINC_DESC}{e.g. ' Abandoned'. See Details.}
}
\description{
Retrieve cer ref site incident category data from cerclis database
}

