% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerclis_cer_ref_site_incident_type.R
\name{cerclis_cer_ref_site_incident_type}
\alias{cerclis_cer_ref_site_incident_type}
\title{Retrieve cer ref site incident type data from cerclis database}
\usage{
cerclis_cer_ref_site_incident_type(RSIT_CODE = NULL, RSIT_DESC = NULL)
}
\arguments{
\item{RSIT_CODE}{e.g. 'N'. See Details.}

\item{RSIT_DESC}{e.g. 'Non-Oil Spill'. See Details.}
}
\description{
Retrieve cer ref site incident type data from cerclis database
}
