% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerclis_cer_ref_state.R
\name{cerclis_cer_ref_state}
\alias{cerclis_cer_ref_state}
\title{Retrieve cer ref state data from cerclis database}
\usage{
cerclis_cer_ref_state(RSTATE_CODE = NULL, RSTATE_NAME = NULL,
  RREGION_CODE = NULL)
}
\arguments{
\item{RSTATE_CODE}{e.g. '99'. See Details.}

\item{RSTATE_NAME}{e.g. 'All'. See Details.}

\item{RREGION_CODE}{e.g. '00'. See Details.}
}
\description{
Retrieve cer ref state data from cerclis database
}
