% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerclis_cer_site.R
\name{cerclis_cer_site}
\alias{cerclis_cer_site}
\title{Retrieve cer site data from cerclis database}
\usage{
cerclis_cer_site(SITE_ID = NULL, SITE_EPA_ID = NULL,
  SITE_SHORT_NAME = NULL, SITE_NAME = NULL, SITE_STRT_ADRS1 = NULL,
  SITE_STRT_ADRS2 = NULL, SITE_CITY_NAME = NULL, SITE_ZIP_CODE = NULL,
  SITE_CNTY_NAME = NULL, SITE_CNGRSNL_DSTRCT_CODE = NULL,
  SITE_RCRA_ID = NULL, SITE_IFMS_SSID_CODE = NULL,
  SITE_USGS_HYDRO_UNIT_NMBR = NULL, SITE_NFRAP_FLAG = NULL,
  SITE_SMSA_NMBR = NULL, RREGION_CODE = NULL, RSTATE_CODE = NULL,
  RNPL_STATUS_CODE = NULL, RFED_FCLTY_CODE = NULL, RSITINC_CODE = NULL,
  RNON_NPL_STATUS_CODE = NULL, SITE_FUDS_FLAG = NULL)
}
\arguments{
\item{SITE_ID}{e.g. '0100002'. See Details.}

\item{SITE_EPA_ID}{e.g. 'CTD000606533'. See Details.}

\item{SITE_SHORT_NAME}{e.g. 'MODOC TOWN DUMP'. See Details.}

\item{SITE_NAME}{e.g. 'MODOC TOWN DUMP'. See Details.}

\item{SITE_STRT_ADRS1}{e.g. 'MODOC ROAD'. See Details.}

\item{SITE_STRT_ADRS2}{e.g. 'NA'. See Details.}

\item{SITE_CITY_NAME}{e.g. 'PUTNAM'. See Details.}

\item{SITE_ZIP_CODE}{e.g. '06260'. See Details.}

\item{SITE_CNTY_NAME}{e.g. 'WINDHAM'. See Details.}

\item{SITE_CNGRSNL_DSTRCT_CODE}{e.g. 'NA'. See Details.}

\item{SITE_RCRA_ID}{e.g. 'NA'. See Details.}

\item{SITE_IFMS_SSID_CODE}{e.g. 'NA'. See Details.}

\item{SITE_USGS_HYDRO_UNIT_NMBR}{e.g. '01100001'. See Details.}

\item{SITE_NFRAP_FLAG}{e.g. 'NA'. See Details.}

\item{SITE_SMSA_NMBR}{e.g. 'NA'. See Details.}

\item{RREGION_CODE}{e.g. '01'. See Details.}

\item{RSTATE_CODE}{e.g. 'CT'. See Details.}

\item{RNPL_STATUS_CODE}{e.g. 'N'. See Details.}

\item{RFED_FCLTY_CODE}{e.g. 'N'. See Details.}

\item{RSITINC_CODE}{e.g. 'NA'. See Details.}

\item{RNON_NPL_STATUS_CODE}{e.g. 'RN'. See Details.}

\item{SITE_FUDS_FLAG}{e.g. 'NA'. See Details.}
}
\description{
Retrieve cer site data from cerclis database
}

