% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerclis_cer_site_alias.R
\name{cerclis_cer_site_alias}
\alias{cerclis_cer_site_alias}
\title{Retrieve cer site alias data from cerclis database}
\usage{
cerclis_cer_site_alias(SITE_ID = NULL, SITE_ALIAS_ID = NULL,
  SA_NAME = NULL, SITE_ALIAS_CITY_NAME = NULL,
  SITE_ALIAS_STRT_ADRS = NULL, SITE_ALIAS_ZIP_CODE = NULL,
  RSTATE_CODE = NULL)
}
\arguments{
\item{SITE_ID}{e.g. '0100002'. See Details.}

\item{SITE_ALIAS_ID}{e.g. '101'. See Details.}

\item{SA_NAME}{e.g. 'PUTNAM TOWN DUMP'. See Details.}

\item{SITE_ALIAS_CITY_NAME}{e.g. 'PUTNAM'. See Details.}

\item{SITE_ALIAS_STRT_ADRS}{e.g. 'MODOC RD'. See Details.}

\item{SITE_ALIAS_ZIP_CODE}{e.g. '06260'. See Details.}

\item{RSTATE_CODE}{e.g. 'CT'. See Details.}
}
\description{
Retrieve cer site alias data from cerclis database
}
