% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerclis_cer_site_rot_ref.R
\name{cerclis_cer_site_rot_ref}
\alias{cerclis_cer_site_rot_ref}
\title{Retrieve cer site rot ref data from cerclis database}
\usage{
cerclis_cer_site_rot_ref(SITE_ID = NULL, ROT_CODE = NULL,
  CREATED_BY = NULL, CREATED_DATE = NULL, LAST_UPDATED_BY = NULL,
  LAST_UPDATED_DATE = NULL, M_ROW_x0024__x0024_ = NULL)
}
\arguments{
\item{SITE_ID}{e.g. '0100001'. See Details.}

\item{ROT_CODE}{e.g. 'ST'. See Details.}

\item{CREATED_BY}{e.g. 'REV CXN'. See Details.}

\item{CREATED_DATE}{e.g. '22-JUN-02'. See Details.}

\item{LAST_UPDATED_BY}{e.g. 'NA'. See Details.}

\item{LAST_UPDATED_DATE}{e.g. 'NA'. See Details.}

\item{M_ROW_x0024__x0024_}{e.g. '01'. See Details.}
}
\description{
Retrieve cer site rot ref data from cerclis database
}

