% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerclis_cer_wsmc_cross_link.R
\name{cerclis_cer_wsmc_cross_link}
\alias{cerclis_cer_wsmc_cross_link}
\title{Retrieve cer wsmc cross link data from cerclis database}
\usage{
cerclis_cer_wsmc_cross_link(SITE_ID = NULL, WSMC_NMBR = NULL,
  WCL_C2501 = NULL)
}
\arguments{
\item{SITE_ID}{e.g. '0300034'. See Details.}

\item{WSMC_NMBR}{e.g. '3000260'. See Details.}

\item{WCL_C2501}{e.g. 'MEDX'. See Details.}
}
\description{
Retrieve cer wsmc cross link data from cerclis database
}
