% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerclis_cer_wst_src_media_contaminated.R
\name{cerclis_cer_wst_src_media_contaminated}
\alias{cerclis_cer_wst_src_media_contaminated}
\title{Retrieve cer wst src media contaminated data from cerclis database}
\usage{
cerclis_cer_wst_src_media_contaminated(SITE_ID = NULL, WSMC_NMBR = NULL)
}
\arguments{
\item{SITE_ID}{e.g. '0100015'. See Details.}

\item{WSMC_NMBR}{e.g. '1272394'. See Details.}
}
\description{
Retrieve cer wst src media contaminated data from cerclis database
}

