% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envirofacts.R
\name{envir_makeFunction}
\alias{envir_makeFunction}
\title{Automatically write a function to get envirofacts data from a given table and service}
\usage{
envir_makeFunction(table, service, description = NULL, dir = "./R",
  tidy = TRUE)
}
\arguments{
\item{table}{Envirofacts table name}

\item{service}{Envirofacts service name}

\item{description}{Roxygen description to add to the function header}

\item{dir}{Directory to put the function script into}

\item{tidy}{If TRUE (default) use formatR to tidy the source code.}
}
\description{
Automatically write a function to get envirofacts data from a given table and service
}
