% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envirofacts.R
\name{envir_serviceInfo}
\alias{envir_serviceInfo}
\title{Information about envirofacts services}
\usage{
envir_serviceInfo(service = NULL)
}
\arguments{
\item{service}{Abbreviation of service. If null (default), returns all service descriptions as a named character.}
}
\description{
Information about envirofacts services
}

