% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fii_frs_affiliation.R
\name{fii_frs_affiliation}
\alias{fii_frs_affiliation}
\title{Retrieve frs affiliation data from fii database}
\usage{
fii_frs_affiliation(AFFILIATION_UIN = NULL, PGM_SYS_ACRNM = NULL,
  PGM_SYS_ID = NULL, AFFILIATION_TYPE = NULL, START_DATE = NULL,
  END_DATE = NULL, ORG_UIN = NULL, CONTACT_UIN = NULL, MAIL_UIN = NULL,
  SOURCE_OF_DATA = NULL, LAST_REPORTED_DATE = NULL, CREATE_DATE = NULL,
  UPDATE_DATE = NULL, USER_ID = NULL, SENSITIVE_IND = NULL,
  PUBLIC_IND = NULL, SUP_INTEREST_ID = NULL, REGISTRY_ID = NULL)
}
\arguments{
\item{AFFILIATION_UIN}{e.g. '2578011'. See Details.}

\item{PGM_SYS_ACRNM}{e.g. 'GEIMS'. See Details.}

\item{PGM_SYS_ID}{e.g. '930052'. See Details.}

\item{AFFILIATION_TYPE}{e.g. 'OWNER'. See Details.}

\item{START_DATE}{e.g. 'NA'. See Details.}

\item{END_DATE}{e.g. 'NA'. See Details.}

\item{ORG_UIN}{e.g. 'NA'. See Details.}

\item{CONTACT_UIN}{e.g. '110015862342'. See Details.}

\item{MAIL_UIN}{e.g. '110015862351'. See Details.}

\item{SOURCE_OF_DATA}{e.g. 'GEIMS'. See Details.}

\item{LAST_REPORTED_DATE}{e.g. 'NA'. See Details.}

\item{CREATE_DATE}{e.g. '24-APR-03'. See Details.}

\item{UPDATE_DATE}{e.g. 'NA'. See Details.}

\item{USER_ID}{e.g. 'REFRESH_CON_INS'. See Details.}

\item{SENSITIVE_IND}{e.g. 'NA'. See Details.}

\item{PUBLIC_IND}{e.g. 'Y'. See Details.}

\item{SUP_INTEREST_ID}{e.g. 'NA'. See Details.}

\item{REGISTRY_ID}{e.g. 'NA'. See Details.}
}
\description{
Retrieve frs affiliation data from fii database
}

