% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fii_frs_agency_ref.R
\name{fii_frs_agency_ref}
\alias{fii_frs_agency_ref}
\title{Retrieve frs agency ref data from fii database}
\usage{
fii_frs_agency_ref(FEDERAL_AGENCY_CODE = NULL,
  FEDERAL_AGENCY_NAME = NULL)
}
\arguments{
\item{FEDERAL_AGENCY_CODE}{e.g. 'C0007'. See Details.}

\item{FEDERAL_AGENCY_NAME}{e.g. 'Red Cross'. See Details.}
}
\description{
Retrieve frs agency ref data from fii database
}
