% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fii_frs_alt_name.R
\name{fii_frs_alt_name}
\alias{fii_frs_alt_name}
\title{Retrieve frs alt name data from fii database}
\usage{
fii_frs_alt_name(PGM_SYS_ACRNM = NULL, PGM_SYS_ID = NULL,
  REGISTRY_ID = NULL, ALTERNATIVE_NAME = NULL,
  ALTERNATIVE_NAME_TYPE = NULL, SOURCE_OF_DATA = NULL,
  LAST_REPORTED_DATE = NULL, STD_ALT_NAME = NULL, CREATE_DATE = NULL,
  USER_ID = NULL, SENSITIVE_IND = NULL, UPDATE_DATE = NULL,
  PUBLIC_IND = NULL)
}
\arguments{
\item{PGM_SYS_ACRNM}{e.g. 'ACES'. See Details.}

\item{PGM_SYS_ID}{e.g. '170001445675'. See Details.}

\item{REGISTRY_ID}{e.g. '110018344255'. See Details.}

\item{ALTERNATIVE_NAME}{e.g. 'GOLF, VILLAGE OF'. See Details.}

\item{ALTERNATIVE_NAME_TYPE}{e.g. 'PROGRAM NAME'. See Details.}

\item{SOURCE_OF_DATA}{e.g. 'ACES'. See Details.}

\item{LAST_REPORTED_DATE}{e.g. 'NA'. See Details.}

\item{STD_ALT_NAME}{e.g. 'GOLF VLG OF'. See Details.}

\item{CREATE_DATE}{e.g. '19-OCT-04'. See Details.}

\item{USER_ID}{e.g. 'REFRESH'. See Details.}

\item{SENSITIVE_IND}{e.g. 'NA'. See Details.}

\item{UPDATE_DATE}{e.g. '17-DEC-13'. See Details.}

\item{PUBLIC_IND}{e.g. 'Y'. See Details.}
}
\description{
Retrieve frs alt name data from fii database
}

