% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fii_frs_alternative_id.R
\name{fii_frs_alternative_id}
\alias{fii_frs_alternative_id}
\title{Retrieve frs alternative id data from fii database}
\usage{
fii_frs_alternative_id(PGM_SYS_ACRNM = NULL, PGM_SYS_ID = NULL,
  ALTERNATIVE_ID = NULL, ALTERNATIVE_ID_TYPE = NULL,
  SOURCE_OF_DATA = NULL, LAST_REPORTED_DATE = NULL, FIPS_CODE = NULL,
  CREATE_DATE = NULL, USER_ID = NULL, SENSITIVE_IND = NULL,
  PUBLIC_IND = NULL)
}
\arguments{
\item{PGM_SYS_ACRNM}{e.g. 'ACRES'. See Details.}

\item{PGM_SYS_ID}{e.g. '10389'. See Details.}

\item{ALTERNATIVE_ID}{e.g. '0407271'. See Details.}

\item{ALTERNATIVE_ID_TYPE}{e.g. 'UNKNOWN'. See Details.}

\item{SOURCE_OF_DATA}{e.g. 'ACRES'. See Details.}

\item{LAST_REPORTED_DATE}{e.g. 'NA'. See Details.}

\item{FIPS_CODE}{e.g. 'NA'. See Details.}

\item{CREATE_DATE}{e.g. '26-FEB-13'. See Details.}

\item{USER_ID}{e.g. 'REFRESH'. See Details.}

\item{SENSITIVE_IND}{e.g. 'N'. See Details.}

\item{PUBLIC_IND}{e.g. 'Y'. See Details.}
}
\description{
Retrieve frs alternative id data from fii database
}
