% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fii_frs_code_description.R
\name{fii_frs_code_description}
\alias{fii_frs_code_description}
\title{Retrieve frs code description data from fii database}
\usage{
fii_frs_code_description(CODE_TYPE = NULL, CODE_VALUE = NULL,
  CODE_DESCRIPTION = NULL)
}
\arguments{
\item{CODE_TYPE}{e.g. 'DATAQUALITYCODE'. See Details.}

\item{CODE_VALUE}{e.g. 'e'. See Details.}

\item{CODE_DESCRIPTION}{e.g. 'ERRONEOUS'. See Details.}
}
\description{
Retrieve frs code description data from fii database
}

