% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fii_frs_contact.R
\name{fii_frs_contact}
\alias{fii_frs_contact}
\title{Retrieve frs contact data from fii database}
\usage{
fii_frs_contact(CONTACT_UIN = NULL, FULL_NAME = NULL, TITLE = NULL,
  DIVISION_NAME = NULL, PHONE_NUMBER = NULL, ALTERNATE_PHONE = NULL,
  FAX_NUMBER = NULL, EMAIL_ADDRESS = NULL, SOURCE_OF_DATA = NULL,
  LAST_REPORTED_DATE = NULL, CREATE_DATE = NULL, UPDATE_DATE = NULL,
  USER_ID = NULL, STD_FULL_NAME = NULL, SENSITIVE_IND = NULL,
  PUBLIC_IND = NULL)
}
\arguments{
\item{CONTACT_UIN}{e.g. '100474089689'. See Details.}

\item{FULL_NAME}{e.g. 'RALPH HOWARD'. See Details.}

\item{TITLE}{e.g. 'NA'. See Details.}

\item{DIVISION_NAME}{e.g. 'NA'. See Details.}

\item{PHONE_NUMBER}{e.g. '4045628829'. See Details.}

\item{ALTERNATE_PHONE}{e.g. 'NA'. See Details.}

\item{FAX_NUMBER}{e.g. 'NA'. See Details.}

\item{EMAIL_ADDRESS}{e.g. 'NA'. See Details.}

\item{SOURCE_OF_DATA}{e.g. 'CERCLIS'. See Details.}

\item{LAST_REPORTED_DATE}{e.g. 'NA'. See Details.}

\item{CREATE_DATE}{e.g. '31-MAY-13'. See Details.}

\item{UPDATE_DATE}{e.g. 'NA'. See Details.}

\item{USER_ID}{e.g. 'REFRESH_CON_INS'. See Details.}

\item{STD_FULL_NAME}{e.g. 'RALPH HOWARD'. See Details.}

\item{SENSITIVE_IND}{e.g. 'N'. See Details.}

\item{PUBLIC_IND}{e.g. 'Y'. See Details.}
}
\description{
Retrieve frs contact data from fii database
}

