% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fii_frs_facility_site.R
\name{fii_frs_facility_site}
\alias{fii_frs_facility_site}
\title{Retrieve frs facility site data from fii database}
\usage{
fii_frs_facility_site(REGISTRY_ID = NULL, PRIMARY_NAME = NULL,
  LOCATION_ADDRESS = NULL, SUPPLEMENTAL_LOCATION = NULL, CITY_NAME = NULL,
  COUNTY_NAME = NULL, FIPS_CODE = NULL, STATE_CODE = NULL,
  STATE_NAME = NULL, COUNTRY_NAME = NULL, POSTAL_CODE = NULL,
  FEDERAL_FACILITY_CODE = NULL, FEDERAL_AGENCY_CODE = NULL,
  TRIBAL_LAND_CODE = NULL, TRIBAL_LAND_NAME = NULL,
  LEGISLATIVE_DIST_NUM = NULL, EPA_REGION_CODE = NULL,
  SITE_TYPE_NAME = NULL, INTEREST_STATUS_CODE = NULL,
  DATA_QUALITY_CODE = NULL, LOCATION_DESCRIPTION = NULL, STD_NAME = NULL,
  STD_LOC_ADDRESS = NULL, STD_HOUSE_NUMBER = NULL, STD_STREET_NAME = NULL,
  STD_BASE_NAME = NULL, STD_PREFIX = NULL, STD_SUFFIX = NULL,
  STD_STYPE_BEFORE = NULL, STD_STYPE_AFTER = NULL, STD_POSTAL_CODE = NULL,
  STD_STATE_CODE = NULL, STD_CITY_NAME = NULL, STD_COUNTY_NAME = NULL,
  STD_COUNTY_FIPS = NULL, STD_COUNTRY = NULL, STD_FULL_ADDRESS = NULL,
  ADDR_TYPE = NULL, LAST_REPORTED_DATE = NULL, CREATE_DATE = NULL,
  UPDATE_DATE = NULL, USER_ID = NULL, PARENT_REGISTRY_ID = NULL,
  SENSITIVE_IND = NULL, REVIEW_FLAG = NULL, REVIEW_REASON = NULL,
  SMALL_BUS_IND = NULL, ENV_JUSTICE_CODE = NULL, PUBLIC_IND = NULL,
  ICIS_IDENTIFIER = NULL, REFRESH_DATE = NULL, STAND_ALONE_FLAG = NULL)
}
\arguments{
\item{REGISTRY_ID}{e.g. '110000307668'. See Details.}

\item{PRIMARY_NAME}{e.g. 'HB FULLER COMPANY HBF PUERTO RICO'. See Details.}

\item{LOCATION_ADDRESS}{e.g. 'CALLE C NUMBER 26 LUCHETTI INDUSTRIAL PARK'. See Details.}

\item{SUPPLEMENTAL_LOCATION}{e.g. 'NA'. See Details.}

\item{CITY_NAME}{e.g. 'BAYAMON'. See Details.}

\item{COUNTY_NAME}{e.g. 'BAYAMON'. See Details.}

\item{FIPS_CODE}{e.g. '72021'. See Details.}

\item{STATE_CODE}{e.g. 'PR'. See Details.}

\item{STATE_NAME}{e.g. 'PUERTO RICO'. See Details.}

\item{COUNTRY_NAME}{e.g. 'USA'. See Details.}

\item{POSTAL_CODE}{e.g. '00961'. See Details.}

\item{FEDERAL_FACILITY_CODE}{e.g. 'N'. See Details.}

\item{FEDERAL_AGENCY_CODE}{e.g. 'NA'. See Details.}

\item{TRIBAL_LAND_CODE}{e.g. 'N'. See Details.}

\item{TRIBAL_LAND_NAME}{e.g. 'NA'. See Details.}

\item{LEGISLATIVE_DIST_NUM}{e.g. '- '. See Details.}

\item{EPA_REGION_CODE}{e.g. '02'. See Details.}

\item{SITE_TYPE_NAME}{e.g. 'STATIONARY'. See Details.}

\item{INTEREST_STATUS_CODE}{e.g. 'A'. See Details.}

\item{DATA_QUALITY_CODE}{e.g. 'V'. See Details.}

\item{LOCATION_DESCRIPTION}{e.g. 'NA'. See Details.}

\item{STD_NAME}{e.g. 'HAMBGR FULLER HBF PR'. See Details.}

\item{STD_LOC_ADDRESS}{e.g. 'CLL C NUMBER 26 LUCHETTI INDUSTRIAL PARK'. See Details.}

\item{STD_HOUSE_NUMBER}{e.g. 'NA'. See Details.}

\item{STD_STREET_NAME}{e.g. 'NA'. See Details.}

\item{STD_BASE_NAME}{e.g. 'NA'. See Details.}

\item{STD_PREFIX}{e.g. 'NA'. See Details.}

\item{STD_SUFFIX}{e.g. 'NA'. See Details.}

\item{STD_STYPE_BEFORE}{e.g. 'NA'. See Details.}

\item{STD_STYPE_AFTER}{e.g. 'NA'. See Details.}

\item{STD_POSTAL_CODE}{e.g. '00739'. See Details.}

\item{STD_STATE_CODE}{e.g. 'PR'. See Details.}

\item{STD_CITY_NAME}{e.g. 'CIDRA'. See Details.}

\item{STD_COUNTY_NAME}{e.g. 'CIDRA MUNICIPIO'. See Details.}

\item{STD_COUNTY_FIPS}{e.g. '72041'. See Details.}

\item{STD_COUNTRY}{e.g. 'PR'. See Details.}

\item{STD_FULL_ADDRESS}{e.g. 'CLL C NUMBER 26 LUCHETTI INDUSTRIAL PARK, CIDRA, PR  00739 PR'. See Details.}

\item{ADDR_TYPE}{e.g. 'REGULAR URBAN'. See Details.}

\item{LAST_REPORTED_DATE}{e.g. 'NA'. See Details.}

\item{CREATE_DATE}{e.g. '01-MAR-00'. See Details.}

\item{UPDATE_DATE}{e.g. '31-DEC-15'. See Details.}

\item{USER_ID}{e.g. 'DSMITH04'. See Details.}

\item{PARENT_REGISTRY_ID}{e.g. 'NA'. See Details.}

\item{SENSITIVE_IND}{e.g. 'N'. See Details.}

\item{REVIEW_FLAG}{e.g. 'NA'. See Details.}

\item{REVIEW_REASON}{e.g. 'NA'. See Details.}

\item{SMALL_BUS_IND}{e.g. 'NA'. See Details.}

\item{ENV_JUSTICE_CODE}{e.g. 'NA'. See Details.}

\item{PUBLIC_IND}{e.g. 'Y'. See Details.}

\item{ICIS_IDENTIFIER}{e.g. 'NA'. See Details.}

\item{REFRESH_DATE}{e.g. 'NA'. See Details.}

\item{STAND_ALONE_FLAG}{e.g. 'N'. See Details.}
}
\description{
Retrieve frs facility site data from fii database
}

