% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fii_frs_interest.R
\name{fii_frs_interest}
\alias{fii_frs_interest}
\title{Retrieve frs interest data from fii database}
\usage{
fii_frs_interest(PGM_SYS_ACRNM = NULL, PGM_SYS_ID = NULL,
  INTEREST_TYPE = NULL, FED_STATE_CODE = NULL, START_DATE = NULL,
  START_DATE_QUALIFIER = NULL, END_DATE = NULL,
  END_DATE_QUALIFIER = NULL, SOURCE_OF_DATA = NULL,
  LAST_REPORTED_DATE = NULL, CREATE_DATE = NULL, UPDATE_DATE = NULL,
  USER_ID = NULL, SENSITIVE_IND = NULL, PUBLIC_IND = NULL,
  ACTIVE_STATUS = NULL)
}
\arguments{
\item{PGM_SYS_ACRNM}{e.g. 'ACES'. See Details.}

\item{PGM_SYS_ID}{e.g. '170001791693'. See Details.}

\item{INTEREST_TYPE}{e.g. 'STATE MASTER'. See Details.}

\item{FED_STATE_CODE}{e.g. 'S'. See Details.}

\item{START_DATE}{e.g. 'NA'. See Details.}

\item{START_DATE_QUALIFIER}{e.g. 'NA'. See Details.}

\item{END_DATE}{e.g. 'NA'. See Details.}

\item{END_DATE_QUALIFIER}{e.g. 'NA'. See Details.}

\item{SOURCE_OF_DATA}{e.g. 'ACES'. See Details.}

\item{LAST_REPORTED_DATE}{e.g. 'NA'. See Details.}

\item{CREATE_DATE}{e.g. '21-DEC-08'. See Details.}

\item{UPDATE_DATE}{e.g. '24-FEB-15'. See Details.}

\item{USER_ID}{e.g. 'REFRESH'. See Details.}

\item{SENSITIVE_IND}{e.g. 'N'. See Details.}

\item{PUBLIC_IND}{e.g. 'Y'. See Details.}

\item{ACTIVE_STATUS}{e.g. 'NA'. See Details.}
}
\description{
Retrieve frs interest data from fii database
}
