% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fii_frs_interest_ref.R
\name{fii_frs_interest_ref}
\alias{fii_frs_interest_ref}
\title{Retrieve frs interest ref data from fii database}
\usage{
fii_frs_interest_ref(INTEREST_TYPE = NULL, INTEREST_DESC = NULL,
  PROGRAM_CATEGORY = NULL, PROGRAM_CATEGORY_DESC = NULL,
  QUERY_FLAG = NULL, PUBLIC_IND = NULL, CREATE_DATE = NULL,
  UPDATE_DATE = NULL)
}
\arguments{
\item{INTEREST_TYPE}{e.g. '401 CERTIFICATION'. See Details.}

\item{INTEREST_DESC}{e.g. 'UNDER SECTION 401 OF THE CLEAN WATER ACT (CWA), STATES AND TRIBES HAVE THE AUTHORITY TO REVIEW AND THEN DENY, CERTIFY, OR CONDITION FEDERAL PERMITS OR LICENSES THAT MIGHT RESULT IN A DISCHARGE TO STATE OR TRIBAL WATERS, INCLUDING WETLANDS, TO ENSURE COMPLIANCE WITH STATE WATER QUALITY STANDARDS.  THE MAJOR FEDERAL LICENSES AND PERMITS SUBJECT TO SECTION 401 ARE SECTION 402 AND 404 PERMITS (IN NONDELEGATED STATES), FEDERAL ENERGY REGULATORY COMMISSION (FERC) HYDROPOWER LICENSES, AND RIVERS AND HARBORS ACT SECTION 9 AND 10 PERMITS.'. See Details.}

\item{PROGRAM_CATEGORY}{e.g. 'WASTE WATER PROGRAMS'. See Details.}

\item{PROGRAM_CATEGORY_DESC}{e.g. 'ENVIRONMENTAL PROGRAMS THAT REGULATE DISCHARGES OF POLLUTANTS TO WATERS OF THE UNITED STATES, AS REQUIRED BY THE CLEAN WATER ACT.'. See Details.}

\item{QUERY_FLAG}{e.g. 'N'. See Details.}

\item{PUBLIC_IND}{e.g. 'Y'. See Details.}

\item{CREATE_DATE}{e.g. '10-MAR-04'. See Details.}

\item{UPDATE_DATE}{e.g. '27-MAR-12'. See Details.}
}
\description{
Retrieve frs interest ref data from fii database
}

