% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fii_frs_mailing_address.R
\name{fii_frs_mailing_address}
\alias{fii_frs_mailing_address}
\title{Retrieve frs mailing address data from fii database}
\usage{
fii_frs_mailing_address(MAIL_UIN = NULL, MAILING_ADDRESS = NULL,
  SUPPLEMENTAL_ADDRESS = NULL, CITY_NAME = NULL, STATE_CODE = NULL,
  STATE_NAME = NULL, POSTAL_CODE = NULL, COUNTRY_NAME = NULL,
  SOURCE_OF_DATA = NULL, LAST_REPORTED_DATE = NULL,
  STD_MAIL_ADDRESS = NULL, STD_SUPPLEMENTAL_ADDRESS = NULL,
  STD_CITY_NAME = NULL, CREATE_DATE = NULL, UPDATE_DATE = NULL,
  USER_ID = NULL, SENSITIVE_IND = NULL, PUBLIC_IND = NULL)
}
\arguments{
\item{MAIL_UIN}{e.g. '000015308560'. See Details.}

\item{MAILING_ADDRESS}{e.g. '141 ROLLINS ROAD'. See Details.}

\item{SUPPLEMENTAL_ADDRESS}{e.g. 'NA'. See Details.}

\item{CITY_NAME}{e.g. 'ROLLINSFORD'. See Details.}

\item{STATE_CODE}{e.g. 'NH'. See Details.}

\item{STATE_NAME}{e.g. 'NEW HAMPSHIRE'. See Details.}

\item{POSTAL_CODE}{e.g. '03869'. See Details.}

\item{COUNTRY_NAME}{e.g. 'NA'. See Details.}

\item{SOURCE_OF_DATA}{e.g. 'AIRS/AFS'. See Details.}

\item{LAST_REPORTED_DATE}{e.g. 'NA'. See Details.}

\item{STD_MAIL_ADDRESS}{e.g. '141 ROLLINS ROAD'. See Details.}

\item{STD_SUPPLEMENTAL_ADDRESS}{e.g. 'NA'. See Details.}

\item{STD_CITY_NAME}{e.g. 'ROLLINSFORD'. See Details.}

\item{CREATE_DATE}{e.g. '03-JUN-02'. See Details.}

\item{UPDATE_DATE}{e.g. '11-MAR-10'. See Details.}

\item{USER_ID}{e.g. 'REFRESH_MAI_UPD'. See Details.}

\item{SENSITIVE_IND}{e.g. 'N'. See Details.}

\item{PUBLIC_IND}{e.g. 'Y'. See Details.}
}
\description{
Retrieve frs mailing address data from fii database
}

