% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fii_frs_naics.R
\name{fii_frs_naics}
\alias{fii_frs_naics}
\title{Retrieve frs naics data from fii database}
\usage{
fii_frs_naics(NAICS_UIN = NULL, PGM_SYS_ACRNM = NULL,
  PGM_SYS_ID = NULL, NAICS_CODE = NULL, PRIMARY_INDICATOR = NULL,
  SOURCE_OF_DATA = NULL, LAST_REPORTED_DATE = NULL,
  CREATE_DATE = NULL, UPDATE_DATE = NULL, USER_ID = NULL,
  SENSITIVE_IND = NULL, PUBLIC_IND = NULL, SUP_INTEREST_ID = NULL)
}
\arguments{
\item{NAICS_UIN}{e.g. '17298'. See Details.}

\item{PGM_SYS_ACRNM}{e.g. 'RCRAINFO'. See Details.}

\item{PGM_SYS_ID}{e.g. 'ALR000031252'. See Details.}

\item{NAICS_CODE}{e.g. '211111'. See Details.}

\item{PRIMARY_INDICATOR}{e.g. 'PRIMARY'. See Details.}

\item{SOURCE_OF_DATA}{e.g. 'RCRAINFO'. See Details.}

\item{LAST_REPORTED_DATE}{e.g. 'NA'. See Details.}

\item{CREATE_DATE}{e.g. '28-DEC-04'. See Details.}

\item{UPDATE_DATE}{e.g. '28-SEP-10'. See Details.}

\item{USER_ID}{e.g. 'REFRESH'. See Details.}

\item{SENSITIVE_IND}{e.g. 'N'. See Details.}

\item{PUBLIC_IND}{e.g. 'Y'. See Details.}

\item{SUP_INTEREST_ID}{e.g. 'NA'. See Details.}
}
\description{
Retrieve frs naics data from fii database
}
