% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fii_frs_organization.R
\name{fii_frs_organization}
\alias{fii_frs_organization}
\title{Retrieve frs organization data from fii database}
\usage{
fii_frs_organization(ORG_UIN = NULL, ORG_NAME = NULL,
  STD_ORG_NAME = NULL, ORG_TYPE = NULL, DUNS_NUMBER = NULL,
  DIVISION_NAME = NULL, PHONE_NUMBER = NULL, ALTERNATE_PHONE = NULL,
  FAX_NUMBER = NULL, EMAIL_ADDRESS = NULL, EIN = NULL,
  STATE_BUSINESS_ID = NULL, SOURCE_OF_DATA = NULL,
  LAST_REPORTED_DATE = NULL, CREATE_DATE = NULL, UPDATE_DATE = NULL,
  SENSITIVE_IND = NULL, USER_ID = NULL, PUBLIC_IND = NULL)
}
\arguments{
\item{ORG_UIN}{e.g. '000014958995'. See Details.}

\item{ORG_NAME}{e.g. 'CALUMET OIL COMPANY'. See Details.}

\item{STD_ORG_NAME}{e.g. 'CALUMET OIL COMPANY'. See Details.}

\item{ORG_TYPE}{e.g. 'NA'. See Details.}

\item{DUNS_NUMBER}{e.g. 'NA'. See Details.}

\item{DIVISION_NAME}{e.g. 'NA'. See Details.}

\item{PHONE_NUMBER}{e.g. '9187493606'. See Details.}

\item{ALTERNATE_PHONE}{e.g. 'NA'. See Details.}

\item{FAX_NUMBER}{e.g. 'NA'. See Details.}

\item{EMAIL_ADDRESS}{e.g. 'NA'. See Details.}

\item{EIN}{e.g. 'NA'. See Details.}

\item{STATE_BUSINESS_ID}{e.g. 'NA'. See Details.}

\item{SOURCE_OF_DATA}{e.g. 'ICIS'. See Details.}

\item{LAST_REPORTED_DATE}{e.g. '11-JAN-05'. See Details.}

\item{CREATE_DATE}{e.g. '02-JUN-02'. See Details.}

\item{UPDATE_DATE}{e.g. '20-JAN-05'. See Details.}

\item{SENSITIVE_IND}{e.g. 'N'. See Details.}

\item{USER_ID}{e.g. 'REFRESH_ORG_UPD'. See Details.}

\item{PUBLIC_IND}{e.g. 'Y'. See Details.}
}
\description{
Retrieve frs organization data from fii database
}
