% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fii_frs_program_facility.R
\name{fii_frs_program_facility}
\alias{fii_frs_program_facility}
\title{Retrieve frs program facility data from fii database}
\usage{
fii_frs_program_facility(PGM_SYS_ACRNM = NULL, PGM_SYS_ID = NULL,
  REGISTRY_ID = NULL, PRIMARY_NAME = NULL, LOCATION_ADDRESS = NULL,
  SUPPLEMENTAL_LOCATION = NULL, CITY_NAME = NULL, COUNTY_NAME = NULL,
  FIPS_CODE = NULL, STATE_CODE = NULL, STATE_NAME = NULL,
  COUNTRY_NAME = NULL, POSTAL_CODE = NULL,
  FEDERAL_FACILITY_CODE = NULL, FEDERAL_AGENCY_CODE = NULL,
  TRIBAL_LAND_CODE = NULL, TRIBAL_LAND_NAME = NULL,
  LEGISLATIVE_DIST_NUM = NULL, HUC_CODE_8 = NULL, HUC_CODE_12 = NULL,
  EPA_REGION_CODE = NULL, SITE_TYPE_NAME = NULL,
  DATA_QUALITY_CODE = NULL, LAST_REPORTED_DATE = NULL,
  STD_NAME = NULL, STD_LOC_ADDRESS = NULL, STD_HOUSE_NUMBER = NULL,
  STD_STREET_NAME = NULL, STD_BASE_NAME = NULL, STD_PREFIX = NULL,
  STD_SUFFIX = NULL, STD_STYPE_BEFORE = NULL, STD_STYPE_AFTER = NULL,
  STD_POSTAL_CODE = NULL, STD_CITY_NAME = NULL,
  STD_COUNTY_NAME = NULL, STD_STATE_CODE = NULL,
  STD_COUNTY_FIPS = NULL, STD_COUNTRY = NULL,
  STD_FULL_ADDRESS = NULL, ADDR_TYPE = NULL, LINK_MTHD = NULL,
  LOCATION_DESCRIPTION = NULL, CREATE_DATE = NULL,
  UPDATE_DATE = NULL, USER_ID = NULL, SENSITIVE_IND = NULL,
  USER_COMMENT = NULL, PUBLIC_IND = NULL, REFRESH_DATE = NULL,
  SMALL_BUS_IND = NULL, ENV_JUSTICE_CODE = NULL,
  PARENT_PGM_SYS_ID = NULL, STAND_ALONE_FLAG = NULL,
  SOURCE_OF_DATA = NULL)
}
\arguments{
\item{PGM_SYS_ACRNM}{e.g. 'ACES'. See Details.}

\item{PGM_SYS_ID}{e.g. '170002007174'. See Details.}

\item{REGISTRY_ID}{e.g. '110056309481'. See Details.}

\item{PRIMARY_NAME}{e.g. 'ILLINI NISSAN'. See Details.}

\item{LOCATION_ADDRESS}{e.g. '615 W MARKETVIEW DR'. See Details.}

\item{SUPPLEMENTAL_LOCATION}{e.g. 'NA'. See Details.}

\item{CITY_NAME}{e.g. 'CHAMPAIGN'. See Details.}

\item{COUNTY_NAME}{e.g. 'CHAMPAIGN'. See Details.}

\item{FIPS_CODE}{e.g. '17019'. See Details.}

\item{STATE_CODE}{e.g. 'IL'. See Details.}

\item{STATE_NAME}{e.g. 'ILLINOIS'. See Details.}

\item{COUNTRY_NAME}{e.g. 'UNITED STATES'. See Details.}

\item{POSTAL_CODE}{e.g. '61820'. See Details.}

\item{FEDERAL_FACILITY_CODE}{e.g. 'NA'. See Details.}

\item{FEDERAL_AGENCY_CODE}{e.g. 'NA'. See Details.}

\item{TRIBAL_LAND_CODE}{e.g. 'NA'. See Details.}

\item{TRIBAL_LAND_NAME}{e.g. 'NA'. See Details.}

\item{LEGISLATIVE_DIST_NUM}{e.g. 'NA'. See Details.}

\item{HUC_CODE_8}{e.g. 'NA'. See Details.}

\item{HUC_CODE_12}{e.g. 'NA'. See Details.}

\item{EPA_REGION_CODE}{e.g. '05'. See Details.}

\item{SITE_TYPE_NAME}{e.g. 'STATIONARY'. See Details.}

\item{DATA_QUALITY_CODE}{e.g. 'V'. See Details.}

\item{LAST_REPORTED_DATE}{e.g. '15-NOV-13'. See Details.}

\item{STD_NAME}{e.g. 'ILLINI NISSAN'. See Details.}

\item{STD_LOC_ADDRESS}{e.g. '615 W MARKETVIEW DR'. See Details.}

\item{STD_HOUSE_NUMBER}{e.g. '615'. See Details.}

\item{STD_STREET_NAME}{e.g. 'W MARKETVIEW DR'. See Details.}

\item{STD_BASE_NAME}{e.g. 'MARKETVIEW'. See Details.}

\item{STD_PREFIX}{e.g. 'W'. See Details.}

\item{STD_SUFFIX}{e.g. 'NA'. See Details.}

\item{STD_STYPE_BEFORE}{e.g. 'NA'. See Details.}

\item{STD_STYPE_AFTER}{e.g. 'DR'. See Details.}

\item{STD_POSTAL_CODE}{e.g. '61822'. See Details.}

\item{STD_CITY_NAME}{e.g. 'CHAMPAIGN'. See Details.}

\item{STD_COUNTY_NAME}{e.g. 'CHAMPAIGN COUNTY'. See Details.}

\item{STD_STATE_CODE}{e.g. 'IL'. See Details.}

\item{STD_COUNTY_FIPS}{e.g. '17019'. See Details.}

\item{STD_COUNTRY}{e.g. 'US'. See Details.}

\item{STD_FULL_ADDRESS}{e.g. '615 W MARKETVIEW DR, CHAMPAIGN, IL  61822 US'. See Details.}

\item{ADDR_TYPE}{e.g. 'NA'. See Details.}

\item{LINK_MTHD}{e.g. 'NA'. See Details.}

\item{LOCATION_DESCRIPTION}{e.g. 'NA'. See Details.}

\item{CREATE_DATE}{e.g. '27-NOV-13'. See Details.}

\item{UPDATE_DATE}{e.g. 'NA'. See Details.}

\item{USER_ID}{e.g. 'REFRESH'. See Details.}

\item{SENSITIVE_IND}{e.g. 'N'. See Details.}

\item{USER_COMMENT}{e.g. 'NA'. See Details.}

\item{PUBLIC_IND}{e.g. 'Y'. See Details.}

\item{REFRESH_DATE}{e.g. '27-NOV-13'. See Details.}

\item{SMALL_BUS_IND}{e.g. 'NA'. See Details.}

\item{ENV_JUSTICE_CODE}{e.g. 'NA'. See Details.}

\item{PARENT_PGM_SYS_ID}{e.g. 'NA'. See Details.}

\item{STAND_ALONE_FLAG}{e.g. 'NA'. See Details.}

\item{SOURCE_OF_DATA}{e.g. 'NA'. See Details.}
}
\description{
Retrieve frs program facility data from fii database
}
