% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fii_frs_sic.R
\name{fii_frs_sic}
\alias{fii_frs_sic}
\title{Retrieve frs sic data from fii database}
\usage{
fii_frs_sic(SIC_UIN = NULL, PGM_SYS_ACRNM = NULL, PGM_SYS_ID = NULL,
  SIC_CODE = NULL, PRIMARY_INDICATOR = NULL, SOURCE_OF_DATA = NULL,
  LAST_REPORTED_DATE = NULL, SENSITIVE_IND = NULL, CREATE_DATE = NULL,
  UPDATE_DATE = NULL, USER_ID = NULL, PUBLIC_IND = NULL,
  SUP_INTEREST_ID = NULL)
}
\arguments{
\item{SIC_UIN}{e.g. '1'. See Details.}

\item{PGM_SYS_ACRNM}{e.g. 'WA-FSIS'. See Details.}

\item{PGM_SYS_ID}{e.g. '29613699'. See Details.}

\item{SIC_CODE}{e.g. '9999'. See Details.}

\item{PRIMARY_INDICATOR}{e.g. 'SECONDARY'. See Details.}

\item{SOURCE_OF_DATA}{e.g. 'WA-FSIS'. See Details.}

\item{LAST_REPORTED_DATE}{e.g. 'NA'. See Details.}

\item{SENSITIVE_IND}{e.g. 'N'. See Details.}

\item{CREATE_DATE}{e.g. '26-SEP-03'. See Details.}

\item{UPDATE_DATE}{e.g. '28-DEC-08'. See Details.}

\item{USER_ID}{e.g. 'REFRESH'. See Details.}

\item{PUBLIC_IND}{e.g. 'Y'. See Details.}

\item{SUP_INTEREST_ID}{e.g. 'NA'. See Details.}
}
\description{
Retrieve frs sic data from fii database
}

