% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fii_frs_supplemental_interest.R
\name{fii_frs_supplemental_interest}
\alias{fii_frs_supplemental_interest}
\title{Retrieve frs supplemental interest data from fii database}
\usage{
fii_frs_supplemental_interest(SUP_INTEREST_ID = NULL, PGM_SYS_ACRNM = NULL,
  PGM_SYS_ID = NULL, SUP_PGM_SYS_ACRNM = NULL, SUP_PGM_SYS_ID = NULL,
  SUP_INTEREST_TYPE = NULL, START_DATE = NULL,
  START_DATE_QUALIFIER = NULL, END_DATE = NULL, END_DATE_QUALIFIER = NULL,
  SOURCE_OF_DATA = NULL, LAST_REPORTED_DATE = NULL,
  REPORTED_SUP_INTEREST_TYPE = NULL, CREATE_DATE = NULL,
  UPDATE_DATE = NULL, USER_ID = NULL, SENSITIVE_IND = NULL,
  PUBLIC_IND = NULL)
}
\arguments{
\item{SUP_INTEREST_ID}{e.g. '301676'. See Details.}

\item{PGM_SYS_ACRNM}{e.g. 'OH-CORE'. See Details.}

\item{PGM_SYS_ID}{e.g. '47847'. See Details.}

\item{SUP_PGM_SYS_ACRNM}{e.g. 'NA'. See Details.}

\item{SUP_PGM_SYS_ID}{e.g. '18-5541-166-2'. See Details.}

\item{SUP_INTEREST_TYPE}{e.g. 'EPCRA'. See Details.}

\item{START_DATE}{e.g. 'NA'. See Details.}

\item{START_DATE_QUALIFIER}{e.g. 'NA'. See Details.}

\item{END_DATE}{e.g. 'NA'. See Details.}

\item{END_DATE_QUALIFIER}{e.g. 'NA'. See Details.}

\item{SOURCE_OF_DATA}{e.g. 'OH-CORE'. See Details.}

\item{LAST_REPORTED_DATE}{e.g. 'NA'. See Details.}

\item{REPORTED_SUP_INTEREST_TYPE}{e.g. 'STATE RTK REPORTER'. See Details.}

\item{CREATE_DATE}{e.g. '17-AUG-01'. See Details.}

\item{UPDATE_DATE}{e.g. '17-OCT-12'. See Details.}

\item{USER_ID}{e.g. 'FRS'. See Details.}

\item{SENSITIVE_IND}{e.g. 'N'. See Details.}

\item{PUBLIC_IND}{e.g. 'Y'. See Details.}
}
\description{
Retrieve frs supplemental interest data from fii database
}

