% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fii_frs_system_ref.R
\name{fii_frs_system_ref}
\alias{fii_frs_system_ref}
\title{Retrieve frs system ref data from fii database}
\usage{
fii_frs_system_ref(PGM_SYS_ACRNM = NULL, PGM_SYS_NAME = NULL,
  PGM_SYS_DESC = NULL, PUBLIC_IND = NULL, FED_STATE_CODE = NULL,
  CREATE_DATE = NULL, UPDATE_DATE = NULL)
}
\arguments{
\item{PGM_SYS_ACRNM}{e.g. 'ACES'. See Details.}

\item{PGM_SYS_NAME}{e.g. 'AGENCY COMPLIANCE AND ENFORCEMENT SYSTEMS'. See Details.}

\item{PGM_SYS_DESC}{e.g. 'The Agency Compliance and Enforcement Systems (ACES) application supports the compliance and enforcement activities that exist primarily within the Illinois Bureaus of Air, Water, and Land, the Division of Legal Counsel, and the Office of Chemical Safety. The intent of the system is to track compliance and enforcement processes and to share the information throughout the agency, the public and with other entities.'. See Details.}

\item{PUBLIC_IND}{e.g. 'Y'. See Details.}

\item{FED_STATE_CODE}{e.g. 'S'. See Details.}

\item{CREATE_DATE}{e.g. 'NA'. See Details.}

\item{UPDATE_DATE}{e.g. 'NA'. See Details.}
}
\description{
Retrieve frs system ref data from fii database
}

