% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fii_frs_tribal_entity_ref.R
\name{fii_frs_tribal_entity_ref}
\alias{fii_frs_tribal_entity_ref}
\title{Retrieve frs tribal entity ref data from fii database}
\usage{
fii_frs_tribal_entity_ref(TRIBAL_LAND_R_CODE = NULL,
  TRIBAL_ENTITY_NAME = NULL, FEDERAL_REGISTER = NULL,
  ORGANIZATION_CODE = NULL)
}
\arguments{
\item{TRIBAL_LAND_R_CODE}{e.g. '-999'. See Details.}

\item{TRIBAL_ENTITY_NAME}{e.g. '-99999'. See Details.}

\item{FEDERAL_REGISTER}{e.g. 'NA'. See Details.}

\item{ORGANIZATION_CODE}{e.g. 'NA'. See Details.}
}
\description{
Retrieve frs tribal entity ref data from fii database
}

