% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fii_frs_tribal_land_ref.R
\name{fii_frs_tribal_land_ref}
\alias{fii_frs_tribal_land_ref}
\title{Retrieve frs tribal land ref data from fii database}
\usage{
fii_frs_tribal_land_ref(TRIBAL_LAND_R_CODE = NULL, TRIBAL_LAND_NAME = NULL,
  USBC_TRIBAL_LAND_CODE = NULL)
}
\arguments{
\item{TRIBAL_LAND_R_CODE}{e.g. '-999'. See Details.}

\item{TRIBAL_LAND_NAME}{e.g. '-99999'. See Details.}

\item{USBC_TRIBAL_LAND_CODE}{e.g. 'NA'. See Details.}
}
\description{
Retrieve frs tribal land ref data from fii database
}

