% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fii_g_info.R
\name{fii_g_info}
\alias{fii_g_info}
\title{Retrieve g info data from fii database}
\usage{
fii_g_info(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  UREA_PRODUCED_METHOD = NULL, UREA_PRODUCED_OTHER_METHOD = NULL,
  CO2_CONSUMED_METHOD = NULL, CO2_CONSUMED_OTHER_METHOD = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000630'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{UREA_PRODUCED_METHOD}{e.g. 'Company records'. See Details.}

\item{UREA_PRODUCED_OTHER_METHOD}{e.g. 'NA'. See Details.}

\item{CO2_CONSUMED_METHOD}{e.g. 'Company records'. See Details.}

\item{CO2_CONSUMED_OTHER_METHOD}{e.g. 'NA'. See Details.}
}
\description{
Retrieve g info data from fii database
}

