% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_aa_fossil_fuel_information.R
\name{ghg_aa_fossil_fuel_information}
\alias{ghg_aa_fossil_fuel_information}
\title{Retrieve aa fossil fuel information data from ghg database}
\usage{
ghg_aa_fossil_fuel_information(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  TIER_3_OPERATING_HOURS_SUB = NULL, FACILITY_NAME = NULL,
  UNIT_NAME = NULL, FUEL_TYPE = NULL, FUEL_TYPE_BLEND = NULL,
  TIER_1_METHOD_EQUATION = NULL, TIER_1_START_DATE = NULL,
  TIER_1_END_DATE = NULL, TIER_1_CO2_EMISSIONS = NULL,
  TIER_1_CH4_EMISSIONS = NULL, TIER_1_N2O_EMISSIONS = NULL,
  TIER_1_CH4_EMISSIONS_CO2E = NULL, TIER_1_N2O_EMISSIONS_CO2E = NULL,
  TIER_2_METHOD_EQUATION = NULL, TIER_2_START_DATE = NULL,
  TIER_2_END_DATE = NULL, TIER_2_HHV_DETERMIN_FREQUENCY = NULL,
  TIER2_HHV_DETERMIN_FREQ_OTH = NULL, TIER_2_CO2_EMISSIONS = NULL,
  TIER_2_CH4_EMISSIONS = NULL, TIER_2_N2O_EMISSIONS = NULL,
  TIER_2_CH4_EMISSIONS_CO2E = NULL, TIER_2_N2O_EMISSIONS_CO2E = NULL,
  TIER_3_METHOD_EQUATION = NULL, TIER_3_START_DATE = NULL,
  TIER_3_END_DATE = NULL, TIER3_CARBON_DETERMIN_FREQUENC = NULL,
  TIER3_CARBON_DETERMIN_NUM_SUB = NULL,
  TIER3_CARBON_DETERMIN_NUM_VALI = NULL, TIER3_MOL_DETERMIN_FREQ = NULL,
  TIER3_MOL_DETERMIN_FREQ_OTHER = NULL, TIER3_MOL_DETERMIN_NUMBER = NULL,
  TIER3_MOL_DETERMIN_NUM_VALID = NULL, TIER_3_CO2_EMISSIONS = NULL,
  TIER_3_CH4_EMISSIONS = NULL, TIER_3_N2O_EMISSIONS = NULL,
  TIER_3_CH4_EMISSIONS_CO2E = NULL, TIER_3_N2O_EMISSIONS_CO2E = NULL,
  PCT_HEAT_INPUT_COMBUSTED = NULL, PCT_HEAT_INPUT_COMBUSTED_IND = NULL,
  PCT_HEAT_INPUT_COMBUSTED_SUB = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000166'. See Details.}

\item{REPORTING_YEAR}{e.g. '2014'. See Details.}

\item{TIER_3_OPERATING_HOURS_SUB}{e.g. 'NA'. See Details.}

\item{FACILITY_NAME}{e.g. 'KapStone Paper and Packaging Corp'. See Details.}

\item{UNIT_NAME}{e.g. '09-PU-004'. See Details.}

\item{FUEL_TYPE}{e.g. 'Natural Gas (Weighted U.S. Average)'. See Details.}

\item{FUEL_TYPE_BLEND}{e.g. 'NA'. See Details.}

\item{TIER_1_METHOD_EQUATION}{e.g. 'NA'. See Details.}

\item{TIER_1_START_DATE}{e.g. 'NA'. See Details.}

\item{TIER_1_END_DATE}{e.g. 'NA'. See Details.}

\item{TIER_1_CO2_EMISSIONS}{e.g. 'NA'. See Details.}

\item{TIER_1_CH4_EMISSIONS}{e.g. 'NA'. See Details.}

\item{TIER_1_N2O_EMISSIONS}{e.g. 'NA'. See Details.}

\item{TIER_1_CH4_EMISSIONS_CO2E}{e.g. 'NA'. See Details.}

\item{TIER_1_N2O_EMISSIONS_CO2E}{e.g. 'NA'. See Details.}

\item{TIER_2_METHOD_EQUATION}{e.g. 'Tier 2 (Equation C-2'. See Details.}

\item{TIER_2_START_DATE}{e.g. '2014-01-01-00:0'. See Details.}

\item{TIER_2_END_DATE}{e.g. '2014-12-31-00:0'. See Details.}

\item{TIER_2_HHV_DETERMIN_FREQUENCY}{e.g. 'Monthly'. See Details.}

\item{TIER2_HHV_DETERMIN_FREQ_OTH}{e.g. 'NA'. See Details.}

\item{TIER_2_CO2_EMISSIONS}{e.g. '26844'. See Details.}

\item{TIER_2_CH4_EMISSIONS}{e.g. '1.37'. See Details.}

\item{TIER_2_N2O_EMISSIONS}{e.g. '0'. See Details.}

\item{TIER_2_CH4_EMISSIONS_CO2E}{e.g. '34.3'. See Details.}

\item{TIER_2_N2O_EMISSIONS_CO2E}{e.g. '0'. See Details.}

\item{TIER_3_METHOD_EQUATION}{e.g. 'NA'. See Details.}

\item{TIER_3_START_DATE}{e.g. 'NA'. See Details.}

\item{TIER_3_END_DATE}{e.g. 'NA'. See Details.}

\item{TIER3_CARBON_DETERMIN_FREQUENC}{e.g. 'NA'. See Details.}

\item{TIER3_CARBON_DETERMIN_NUM_SUB}{e.g. 'NA'. See Details.}

\item{TIER3_CARBON_DETERMIN_NUM_VALI}{e.g. 'NA'. See Details.}

\item{TIER3_MOL_DETERMIN_FREQ}{e.g. 'NA'. See Details.}

\item{TIER3_MOL_DETERMIN_FREQ_OTHER}{e.g. 'NA'. See Details.}

\item{TIER3_MOL_DETERMIN_NUMBER}{e.g. 'NA'. See Details.}

\item{TIER3_MOL_DETERMIN_NUM_VALID}{e.g. 'NA'. See Details.}

\item{TIER_3_CO2_EMISSIONS}{e.g. 'NA'. See Details.}

\item{TIER_3_CH4_EMISSIONS}{e.g. 'NA'. See Details.}

\item{TIER_3_N2O_EMISSIONS}{e.g. 'NA'. See Details.}

\item{TIER_3_CH4_EMISSIONS_CO2E}{e.g. 'NA'. See Details.}

\item{TIER_3_N2O_EMISSIONS_CO2E}{e.g. 'NA'. See Details.}

\item{PCT_HEAT_INPUT_COMBUSTED}{e.g. '0.98'. See Details.}

\item{PCT_HEAT_INPUT_COMBUSTED_IND}{e.g. 'NA'. See Details.}

\item{PCT_HEAT_INPUT_COMBUSTED_SUB}{e.g. 'NA'. See Details.}
}
\description{
Retrieve aa fossil fuel information data from ghg database
}

