% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_aa_fossil_fuel_tier_2_info.R
\name{ghg_aa_fossil_fuel_tier_2_info}
\alias{ghg_aa_fossil_fuel_tier_2_info}
\title{Retrieve aa fossil fuel tier 2 info data from ghg database}
\usage{
ghg_aa_fossil_fuel_tier_2_info(FACILITY_ID = NULL,
  REPORTING_YEAR = NULL, FACILITY_NAME = NULL, UNIT_NAME = NULL,
  TIER_2_MONTH_NAME = NULL, TIER_2_HHV_SUBSTITUTE_IND = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000166'. See Details.}

\item{REPORTING_YEAR}{e.g. '2014'. See Details.}

\item{FACILITY_NAME}{e.g. 'KapStone Paper and Packaging Corp'. See Details.}

\item{UNIT_NAME}{e.g. '08-PU-012'. See Details.}

\item{TIER_2_MONTH_NAME}{e.g. 'September'. See Details.}

\item{TIER_2_HHV_SUBSTITUTE_IND}{e.g. 'N'. See Details.}
}
\description{
Retrieve aa fossil fuel tier 2 info data from ghg database
}
