% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_aa_makeup_chemical_info.R
\name{ghg_aa_makeup_chemical_info}
\alias{ghg_aa_makeup_chemical_info}
\title{Retrieve aa makeup chemical info data from ghg database}
\usage{
ghg_aa_makeup_chemical_info(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  CACO3_SUBSTITUTE_INDICATOR = NULL, MASS_MAKEUP_CHEMICAL_CO2 = NULL,
  NA2CO3_SUBSTITUTE_INDICATOR = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000166'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{CACO3_SUBSTITUTE_INDICATOR}{e.g. 'N'. See Details.}

\item{MASS_MAKEUP_CHEMICAL_CO2}{e.g. '0'. See Details.}

\item{NA2CO3_SUBSTITUTE_INDICATOR}{e.g. 'N'. See Details.}
}
\description{
Retrieve aa makeup chemical info data from ghg database
}

