% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_aa_spent_liquor_information.R
\name{ghg_aa_spent_liquor_information}
\alias{ghg_aa_spent_liquor_information}
\title{Retrieve aa spent liquor information data from ghg database}
\usage{
ghg_aa_spent_liquor_information(FACILITY_ID = NULL,
  REPORTING_YEAR = NULL, FACILITY_NAME = NULL,
  SPENT_LIQUOR_CO2_EMISSIONS = NULL, SPENT_LIQUOR_CH4_EMISSIONS = NULL,
  UNIT_NAME = NULL, UNIT_TYPE = NULL,
  SPENT_LIQUOR_N2O_EMISSIONS = NULL,
  BIOMASS_CH4_EMISSIONS_FACTOR = NULL,
  BIOMASS_N2O_EMISSIONS_FACTOR = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000166'. See Details.}

\item{REPORTING_YEAR}{e.g. '2014'. See Details.}

\item{FACILITY_NAME}{e.g. 'KapStone Paper and Packaging Corp'. See Details.}

\item{SPENT_LIQUOR_CO2_EMISSIONS}{e.g. 'NA'. See Details.}

\item{SPENT_LIQUOR_CH4_EMISSIONS}{e.g. 'NA'. See Details.}

\item{UNIT_NAME}{e.g. '09-PU-004'. See Details.}

\item{UNIT_TYPE}{e.g. 'Pulp Mill Lime Kiln'. See Details.}

\item{SPENT_LIQUOR_N2O_EMISSIONS}{e.g. 'NA'. See Details.}

\item{BIOMASS_CH4_EMISSIONS_FACTOR}{e.g. 'NA'. See Details.}

\item{BIOMASS_N2O_EMISSIONS_FACTOR}{e.g. 'NA'. See Details.}
}
\description{
Retrieve aa spent liquor information data from ghg database
}
