% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_aa_subpart_level_information.R
\name{ghg_aa_subpart_level_information}
\alias{ghg_aa_subpart_level_information}
\title{Retrieve aa subpart level information data from ghg database}
\usage{
ghg_aa_subpart_level_information(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  GHG_NAME = NULL, GHG_QUANTITY = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000166'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{GHG_NAME}{e.g. 'Carbon Dioxide'. See Details.}

\item{GHG_QUANTITY}{e.g. '27668.9'. See Details.}
}
\description{
Retrieve aa subpart level information data from ghg database
}

