% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_c_biogenic_co2_details.R
\name{ghg_c_biogenic_co2_details}
\alias{ghg_c_biogenic_co2_details}
\title{Retrieve c biogenic co2 details data from ghg database}
\usage{
ghg_c_biogenic_co2_details(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, UNIT_NAME = NULL, MSW_SAMPLING_QUARTER = NULL,
  MSW_SAMPLE_ANALYSIS_RESULT = NULL, FOSSIL_BIOMASS_QUARTER = NULL,
  FOSSIL_BIOMASS_ANALYSIS_RESULT = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000036'. See Details.}

\item{REPORTING_YEAR}{e.g. '2014'. See Details.}

\item{FACILITY_NAME}{e.g. 'UNIVERSITY OF CINCINNATI'. See Details.}

\item{UNIT_NAME}{e.g. 'B108'. See Details.}

\item{MSW_SAMPLING_QUARTER}{e.g. 'NA'. See Details.}

\item{MSW_SAMPLE_ANALYSIS_RESULT}{e.g. 'NA'. See Details.}

\item{FOSSIL_BIOMASS_QUARTER}{e.g. 'First Quarter'. See Details.}

\item{FOSSIL_BIOMASS_ANALYSIS_RESULT}{e.g. '0'. See Details.}
}
\description{
Retrieve c biogenic co2 details data from ghg database
}

