% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_c_cems_quarterly_co2.R
\name{ghg_c_cems_quarterly_co2}
\alias{ghg_c_cems_quarterly_co2}
\title{Retrieve c cems quarterly co2 data from ghg database}
\usage{
ghg_c_cems_quarterly_co2(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, UNIT_NAME = NULL,
  CEMS_CO2_EMISSIONS_QUARTER_1 = NULL, CEMS_CO2_EMISSIONS_QUARTER_2 = NULL,
  CEMS_CO2_EMISSIONS_QUARTER_3 = NULL, CEMS_CO2_EMISSIONS_QUARTER_4 = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000156'. See Details.}

\item{REPORTING_YEAR}{e.g. '2014'. See Details.}

\item{FACILITY_NAME}{e.g. 'Mittal Steel USA - Indiana Harbor East'. See Details.}

\item{UNIT_NAME}{e.g. 'CS-5BH_CEMS'. See Details.}

\item{CEMS_CO2_EMISSIONS_QUARTER_1}{e.g. '476897.3292'. See Details.}

\item{CEMS_CO2_EMISSIONS_QUARTER_2}{e.g. '322457.8216'. See Details.}

\item{CEMS_CO2_EMISSIONS_QUARTER_3}{e.g. '357420.8886'. See Details.}

\item{CEMS_CO2_EMISSIONS_QUARTER_4}{e.g. '534324.3595'. See Details.}
}
\description{
Retrieve c cems quarterly co2 data from ghg database
}

