% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_c_configuration_level_info.R
\name{ghg_c_configuration_level_info}
\alias{ghg_c_configuration_level_info}
\title{Retrieve c configuration level info data from ghg database}
\usage{
ghg_c_configuration_level_info(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, UNIT_NAME = NULL, UNIT_TYPE = NULL,
  UNIT_TYPE_OTHER = NULL, IND_UNIT_HEAT_INPUT_CAPACITY = NULL,
  IND_UNIT_HEATINPUTCAPACITY_UOM = NULL,
  IND_UNIT_HEATCAPACITY_UOM_OTHR = NULL, AGGR_HIGH_HEAT_CAPACITY = NULL,
  AGGR_HIGH_HEAT_CAPACITY_UOM = NULL, COMMON_PIPE_HEAT_CAPACITY = NULL,
  COMMON_PIPE_HEAT_CAPACITY_UOM = NULL, COMMON_STACK_NUMBER_OF_UNITS = NULL,
  SORBENT_MOLECULAR_WEIGHT = NULL, SORBENT_RATIO = NULL,
  COMMON_STACK_COMBINED_HEAT = NULL, COMMON_STACK_COMBINED_HEAT_UOM = NULL,
  TIER_123_SORBENT_CO2_EMISSIONS = NULL,
  TIER123_BIOGENIC_CO2_EMISSIONS = NULL,
  TIER_1_2_3_FOSSIL_CO2_EMISSION = NULL, TIER_4_OPERATING_HOURS = NULL,
  TIER_4_OP_HRS_CO2_CONC_SUB = NULL, TIER_4_OP_HRS_STACK_FLOWRT_SUB = NULL,
  TIER_4_OP_HRS_STACK_MOIST_SUB = NULL,
  TIER4_NONBIOGENIC_CO2EMISSION = NULL,
  TIER_4_BIOGENIC_CO2_EMISSIONS = NULL, TIER_4_TOTAL_CO2_EMISSIONS = NULL,
  TIER_4_METHODOLOGY_START_DATE = NULL, TIER_4_METHODOLOGY_END_DATE = NULL,
  TIER4_VFF = NULL, TIER4_V_TOTAL = NULL, TIER4_V_BIO = NULL,
  TIER4_ANNUAL_BIO_CO2_EMISSION = NULL, PART_75_CO2_EMISSIONS_METHOD = NULL,
  PART_75_METHOD_STARTDATE = NULL, PART_75_METHODOLOGY_ENDDATE = NULL,
  PART_75_HEAT_INPUT_METHOD = NULL, PART_75_CO2_METHOD = NULL,
  PART_75_OP_HRS_FUEL_FLOWRATE = NULL, PART_75_OP_HRS_HHV_SUB = NULL,
  PART_75_OP_HRS_CO2_CONC_SUB = NULL, PART_75_OP_HRS_FLOWRATE_SUB = NULL,
  PART_75_OP_HRS_MOISTURE_SUB = NULL, BIOGENIC_CO2_MASS_EMISS_MSW = NULL,
  TOTAL_AMT_SORBENT_USED = NULL, PLANT_CODE = NULL, CP_USEIVTIND = NULL,
  CS_USEIVTIND = NULL, IND_USE_IVT_IND = NULL, AGG_USEIVTIND = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000001'. See Details.}

\item{REPORTING_YEAR}{e.g. '2010'. See Details.}

\item{FACILITY_NAME}{e.g. 'PSE Ferndale Generating Station'. See Details.}

\item{UNIT_NAME}{e.g. 'GT-1A'. See Details.}

\item{UNIT_TYPE}{e.g. 'CCCT (CC (Turbine, combined cycle))'. See Details.}

\item{UNIT_TYPE_OTHER}{e.g. 'NA'. See Details.}

\item{IND_UNIT_HEAT_INPUT_CAPACITY}{e.g. '1019.7'. See Details.}

\item{IND_UNIT_HEATINPUTCAPACITY_UOM}{e.g. 'mmBtu/hr'. See Details.}

\item{IND_UNIT_HEATCAPACITY_UOM_OTHR}{e.g. 'NA'. See Details.}

\item{AGGR_HIGH_HEAT_CAPACITY}{e.g. 'NA'. See Details.}

\item{AGGR_HIGH_HEAT_CAPACITY_UOM}{e.g. 'NA'. See Details.}

\item{COMMON_PIPE_HEAT_CAPACITY}{e.g. 'NA'. See Details.}

\item{COMMON_PIPE_HEAT_CAPACITY_UOM}{e.g. 'NA'. See Details.}

\item{COMMON_STACK_NUMBER_OF_UNITS}{e.g. 'NA'. See Details.}

\item{SORBENT_MOLECULAR_WEIGHT}{e.g. 'NA'. See Details.}

\item{SORBENT_RATIO}{e.g. 'NA'. See Details.}

\item{COMMON_STACK_COMBINED_HEAT}{e.g. 'NA'. See Details.}

\item{COMMON_STACK_COMBINED_HEAT_UOM}{e.g. 'NA'. See Details.}

\item{TIER_123_SORBENT_CO2_EMISSIONS}{e.g. '0'. See Details.}

\item{TIER123_BIOGENIC_CO2_EMISSIONS}{e.g. '0'. See Details.}

\item{TIER_1_2_3_FOSSIL_CO2_EMISSION}{e.g. 'NA'. See Details.}

\item{TIER_4_OPERATING_HOURS}{e.g. 'NA'. See Details.}

\item{TIER_4_OP_HRS_CO2_CONC_SUB}{e.g. 'NA'. See Details.}

\item{TIER_4_OP_HRS_STACK_FLOWRT_SUB}{e.g. 'NA'. See Details.}

\item{TIER_4_OP_HRS_STACK_MOIST_SUB}{e.g. 'NA'. See Details.}

\item{TIER4_NONBIOGENIC_CO2EMISSION}{e.g. 'NA'. See Details.}

\item{TIER_4_BIOGENIC_CO2_EMISSIONS}{e.g. 'NA'. See Details.}

\item{TIER_4_TOTAL_CO2_EMISSIONS}{e.g. 'NA'. See Details.}

\item{TIER_4_METHODOLOGY_START_DATE}{e.g. 'NA'. See Details.}

\item{TIER_4_METHODOLOGY_END_DATE}{e.g. 'NA'. See Details.}

\item{TIER4_VFF}{e.g. 'NA'. See Details.}

\item{TIER4_V_TOTAL}{e.g. 'NA'. See Details.}

\item{TIER4_V_BIO}{e.g. 'NA'. See Details.}

\item{TIER4_ANNUAL_BIO_CO2_EMISSION}{e.g. 'NA'. See Details.}

\item{PART_75_CO2_EMISSIONS_METHOD}{e.g. 'NA'. See Details.}

\item{PART_75_METHOD_STARTDATE}{e.g. 'NA'. See Details.}

\item{PART_75_METHODOLOGY_ENDDATE}{e.g. 'NA'. See Details.}

\item{PART_75_HEAT_INPUT_METHOD}{e.g. 'NA'. See Details.}

\item{PART_75_CO2_METHOD}{e.g. 'NA'. See Details.}

\item{PART_75_OP_HRS_FUEL_FLOWRATE}{e.g. 'NA'. See Details.}

\item{PART_75_OP_HRS_HHV_SUB}{e.g. 'NA'. See Details.}

\item{PART_75_OP_HRS_CO2_CONC_SUB}{e.g. 'NA'. See Details.}

\item{PART_75_OP_HRS_FLOWRATE_SUB}{e.g. 'NA'. See Details.}

\item{PART_75_OP_HRS_MOISTURE_SUB}{e.g. 'NA'. See Details.}

\item{BIOGENIC_CO2_MASS_EMISS_MSW}{e.g. 'NA'. See Details.}

\item{TOTAL_AMT_SORBENT_USED}{e.g. 'NA'. See Details.}

\item{PLANT_CODE}{e.g. 'NA'. See Details.}

\item{CP_USEIVTIND}{e.g. 'NA'. See Details.}

\item{CS_USEIVTIND}{e.g. 'NA'. See Details.}

\item{IND_USE_IVT_IND}{e.g. 'NA'. See Details.}

\item{AGG_USEIVTIND}{e.g. 'NA'. See Details.}
}
\description{
Retrieve c configuration level info data from ghg database
}

