% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_c_fuel_level_information.R
\name{ghg_c_fuel_level_information}
\alias{ghg_c_fuel_level_information}
\title{Retrieve c fuel level information data from ghg database}
\usage{
ghg_c_fuel_level_information(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  T3FREQUENCYOFCCDETERMINATION = NULL,
  T3OTHERFREQFCCDETERMINATION = NULL,
  T3TOTALNUMBERCCDETERMINATIONS = NULL,
  T3NUMBERVALIDCCDETERMINATIONS = NULL,
  T3FREQUENCYOFMWDETERMINATION = NULL,
  T3OTHERFREQOFMWDETERMINATION = NULL,
  T3TOTALNUMBERMWTDETERMINATIONS = NULL,
  T3NUMBERVALIDMWDETERMINATIONS = NULL,
  T3NUMOFOPTHOURSSUBFUELUSAGE = NULL, T4CH4COMBUSTIONEMISSIONS = NULL,
  T4N2OCOMBUSTIONEMISSIONS = NULL, TIER4_CH4_EMISSIONS_CO2E = NULL,
  TIER4_N2O_EMISSIONS_CO2E = NULL, PART_75_CH4_EMISSIONS_CO2E = NULL,
  PART_75_N2O_EMISSIONS_CO2E = NULL, PART_75_ANNUAL_HEAT_INPUT = NULL,
  FACILITY_NAME = NULL, UNIT_NAME = NULL, UNIT_TYPE = NULL,
  FUEL_TYPE = NULL, FUEL_TYPE_OTHER = NULL, FUEL_TYPE_BLEND = NULL,
  TIER_1_METHOD_EQUATION = NULL, TIER1_METHODOLOGY_START_DATE = NULL,
  TIER1_METHODOLOGY_END_DATE = NULL,
  TIER1_CO2_COMBUSTION_EMISSIONS = NULL,
  TIER1_CH4_COMBUSTION_EMISSIONS = NULL,
  TIER1_N2O_COMBUSTION_EMISSIONS = NULL,
  TIER1_CH4_EMISSIONS_CO2E = NULL, TIER1_N2O_EMISSIONS_CO2E = NULL,
  TIER_2_METHOD_EQUATION = NULL, TIER2_METHODOLOGY_START_DATE = NULL,
  TIER2_METHODOLOGY_END_DATE = NULL,
  TIER2_FREQ_HHV_DETERMINATION = NULL,
  TIER2_OTH_FREQ_HHV_DETERMIN = NULL,
  TIER2_CO2_COMBUSTION_EMISSIONS = NULL,
  TIER2_CH4_COMBUSTION_EMISSIONS = NULL,
  TIER2_N2O_COMBUSTION_EMISSIONS = NULL,
  TIER2_CH4_EMISSIONS_CO2E = NULL, TIER2_N2O_EMISSIONS_CO2E = NULL,
  TIER_3_METHOD_EQUATION = NULL, TIER3_METHODOLOGY_START_DATE = NULL,
  TIER3_METHODOLOGY_END_DATE = NULL,
  TIER3_CO2_COMBUSTION_EMISSIONS = NULL,
  TIER3_CH4_COMBUSTION_EMISSIONS = NULL,
  TIER3_N2O_COMBUSTION_EMISSIONS = NULL,
  TIER3_CH4_EMISSIONS_CO2E = NULL, TIER3_N2O_EMISSIONS_CO2E = NULL,
  T3MOLARVOLUMECONSTANTVALUE = NULL, OTHER_OR_BLEND_FUEL_TYPE = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000001'. See Details.}

\item{REPORTING_YEAR}{e.g. '2010'. See Details.}

\item{T3FREQUENCYOFCCDETERMINATION}{e.g. 'NA'. See Details.}

\item{T3OTHERFREQFCCDETERMINATION}{e.g. 'NA'. See Details.}

\item{T3TOTALNUMBERCCDETERMINATIONS}{e.g. 'NA'. See Details.}

\item{T3NUMBERVALIDCCDETERMINATIONS}{e.g. 'NA'. See Details.}

\item{T3FREQUENCYOFMWDETERMINATION}{e.g. 'NA'. See Details.}

\item{T3OTHERFREQOFMWDETERMINATION}{e.g. 'NA'. See Details.}

\item{T3TOTALNUMBERMWTDETERMINATIONS}{e.g. 'NA'. See Details.}

\item{T3NUMBERVALIDMWDETERMINATIONS}{e.g. 'NA'. See Details.}

\item{T3NUMOFOPTHOURSSUBFUELUSAGE}{e.g. 'NA'. See Details.}

\item{T4CH4COMBUSTIONEMISSIONS}{e.g. 'NA'. See Details.}

\item{T4N2OCOMBUSTIONEMISSIONS}{e.g. 'NA'. See Details.}

\item{TIER4_CH4_EMISSIONS_CO2E}{e.g. 'NA'. See Details.}

\item{TIER4_N2O_EMISSIONS_CO2E}{e.g. 'NA'. See Details.}

\item{PART_75_CH4_EMISSIONS_CO2E}{e.g. 'NA'. See Details.}

\item{PART_75_N2O_EMISSIONS_CO2E}{e.g. 'NA'. See Details.}

\item{PART_75_ANNUAL_HEAT_INPUT}{e.g. 'NA'. See Details.}

\item{FACILITY_NAME}{e.g. 'PSE Ferndale Generating Station'. See Details.}

\item{UNIT_NAME}{e.g. 'GT-1A'. See Details.}

\item{UNIT_TYPE}{e.g. 'CCCT (CC (Turbine, combined cycle))'. See Details.}

\item{FUEL_TYPE}{e.g. 'Distillate Fuel Oil No. 2'. See Details.}

\item{FUEL_TYPE_OTHER}{e.g. 'NA'. See Details.}

\item{FUEL_TYPE_BLEND}{e.g. 'NA'. See Details.}

\item{TIER_1_METHOD_EQUATION}{e.g. 'NA'. See Details.}

\item{TIER1_METHODOLOGY_START_DATE}{e.g. 'NA'. See Details.}

\item{TIER1_METHODOLOGY_END_DATE}{e.g. 'NA'. See Details.}

\item{TIER1_CO2_COMBUSTION_EMISSIONS}{e.g. 'NA'. See Details.}

\item{TIER1_CH4_COMBUSTION_EMISSIONS}{e.g. 'NA'. See Details.}

\item{TIER1_N2O_COMBUSTION_EMISSIONS}{e.g. 'NA'. See Details.}

\item{TIER1_CH4_EMISSIONS_CO2E}{e.g. 'NA'. See Details.}

\item{TIER1_N2O_EMISSIONS_CO2E}{e.g. 'NA'. See Details.}

\item{TIER_2_METHOD_EQUATION}{e.g. 'Tier 2 (Equation C-2a)'. See Details.}

\item{TIER2_METHODOLOGY_START_DATE}{e.g. '2010-01-01'. See Details.}

\item{TIER2_METHODOLOGY_END_DATE}{e.g. '2010-12-31'. See Details.}

\item{TIER2_FREQ_HHV_DETERMINATION}{e.g. 'Other (specify)'. See Details.}

\item{TIER2_OTH_FREQ_HHV_DETERMIN}{e.g. 'AP-42 Emission Factor'. See Details.}

\item{TIER2_CO2_COMBUSTION_EMISSIONS}{e.g. '44.1'. See Details.}

\item{TIER2_CH4_COMBUSTION_EMISSIONS}{e.g. '0'. See Details.}

\item{TIER2_N2O_COMBUSTION_EMISSIONS}{e.g. '0'. See Details.}

\item{TIER2_CH4_EMISSIONS_CO2E}{e.g. '0'. See Details.}

\item{TIER2_N2O_EMISSIONS_CO2E}{e.g. '0'. See Details.}

\item{TIER_3_METHOD_EQUATION}{e.g. 'NA'. See Details.}

\item{TIER3_METHODOLOGY_START_DATE}{e.g. 'NA'. See Details.}

\item{TIER3_METHODOLOGY_END_DATE}{e.g. 'NA'. See Details.}

\item{TIER3_CO2_COMBUSTION_EMISSIONS}{e.g. 'NA'. See Details.}

\item{TIER3_CH4_COMBUSTION_EMISSIONS}{e.g. 'NA'. See Details.}

\item{TIER3_N2O_COMBUSTION_EMISSIONS}{e.g. 'NA'. See Details.}

\item{TIER3_CH4_EMISSIONS_CO2E}{e.g. 'NA'. See Details.}

\item{TIER3_N2O_EMISSIONS_CO2E}{e.g. 'NA'. See Details.}

\item{T3MOLARVOLUMECONSTANTVALUE}{e.g. 'NA'. See Details.}

\item{OTHER_OR_BLEND_FUEL_TYPE}{e.g. 'NA'. See Details.}
}
\description{
Retrieve c fuel level information data from ghg database
}
