% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_c_subpart_level_information.R
\name{ghg_c_subpart_level_information}
\alias{ghg_c_subpart_level_information}
\title{Retrieve c subpart level information data from ghg database}
\usage{
ghg_c_subpart_level_information(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, GHG_GAS_NAME = NULL, GHG_QUANTITY = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000001'. See Details.}

\item{REPORTING_YEAR}{e.g. '2010'. See Details.}

\item{FACILITY_NAME}{e.g. 'PSE Ferndale Generating Station'. See Details.}

\item{GHG_GAS_NAME}{e.g. 'Biogenic Carbon dioxide'. See Details.}

\item{GHG_QUANTITY}{e.g. '0'. See Details.}
}
\description{
Retrieve c subpart level information data from ghg database
}

