% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_c_tier2_monthly_hhv.R
\name{ghg_c_tier2_monthly_hhv}
\alias{ghg_c_tier2_monthly_hhv}
\title{Retrieve c tier2 monthly hhv data from ghg database}
\usage{
ghg_c_tier2_monthly_hhv(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, TIER2_NAME = NULL, MONTH_NAME = NULL,
  HIGH_HEAT_VALUE_INDICATOR = NULL, FUEL_TYPE = NULL,
  UNIT_NAME = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000001'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{FACILITY_NAME}{e.g. 'PSE Ferndale Generating Station'. See Details.}

\item{TIER2_NAME}{e.g. 'Tier 2 (Equation C-2a)'. See Details.}

\item{MONTH_NAME}{e.g. 'September'. See Details.}

\item{HIGH_HEAT_VALUE_INDICATOR}{e.g. 'N'. See Details.}

\item{FUEL_TYPE}{e.g. 'Distillate Fuel Oil No. 2'. See Details.}

\item{UNIT_NAME}{e.g. 'GT-1A'. See Details.}
}
\description{
Retrieve c tier2 monthly hhv data from ghg database
}
