% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_d_fuel_level_information.R
\name{ghg_d_fuel_level_information}
\alias{ghg_d_fuel_level_information}
\title{Retrieve d fuel level information data from ghg database}
\usage{
ghg_d_fuel_level_information(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  N2O_EMISSIONS_CO2E = NULL, TOTAL_ANNUAL_HEAT_INPUT = NULL,
  FACILITY_NAME = NULL, UNIT_NAME = NULL, UNIT_TYPE = NULL,
  FUEL_TYPE = NULL, FUEL_TYPE_OTHER = NULL, FUEL_TYPE_BLEND = NULL,
  CH4_EMISSIONS_CO2E = NULL, OTHER_OR_BLEND_FUEL_TYPE = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000001'. See Details.}

\item{REPORTING_YEAR}{e.g. '2013'. See Details.}

\item{N2O_EMISSIONS_CO2E}{e.g. '111'. See Details.}

\item{TOTAL_ANNUAL_HEAT_INPUT}{e.g. '3725813'. See Details.}

\item{FACILITY_NAME}{e.g. 'PSE Ferndale Generating Station'. See Details.}

\item{UNIT_NAME}{e.g. 'CT-1A'. See Details.}

\item{UNIT_TYPE}{e.g. 'Electricity Generator'. See Details.}

\item{FUEL_TYPE}{e.g. 'Natural Gas (Weighted U.S. Average)'. See Details.}

\item{FUEL_TYPE_OTHER}{e.g. 'NA'. See Details.}

\item{FUEL_TYPE_BLEND}{e.g. 'NA'. See Details.}

\item{CH4_EMISSIONS_CO2E}{e.g. '93.1'. See Details.}

\item{OTHER_OR_BLEND_FUEL_TYPE}{e.g. 'NA'. See Details.}
}
\description{
Retrieve d fuel level information data from ghg database
}
