% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_dd_facility_info.R
\name{ghg_dd_facility_info}
\alias{ghg_dd_facility_info}
\title{Retrieve dd facility info data from ghg database}
\usage{
ghg_dd_facility_info(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, GAS_NAME_SF6_OR_PFC = NULL,
  BEGINNING_NAME_PLATE_CAPACITY = NULL, IS_GAS_REPORTED = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000068'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{FACILITY_NAME}{e.g. 'Apex Generating Station'. See Details.}

\item{GAS_NAME_SF6_OR_PFC}{e.g. 'Sulfur hexafluoride'. See Details.}

\item{BEGINNING_NAME_PLATE_CAPACITY}{e.g. '1355'. See Details.}

\item{IS_GAS_REPORTED}{e.g. 'Yes'. See Details.}
}
\description{
Retrieve dd facility info data from ghg database
}
