% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_dd_increase_nameplate.R
\name{ghg_dd_increase_nameplate}
\alias{ghg_dd_increase_nameplate}
\title{Retrieve dd increase nameplate data from ghg database}
\usage{
ghg_dd_increase_nameplate(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, GAS_NAME_SF6_OR_PFC = NULL,
  NEW_EQUIPMENT_NAMEPLATE_CAP = NULL,
  RETIRING_EQUIP_NAMEPLATE_CAP = NULL,
  NET_INCREASE_NAMEPLATE_CAP_RPT = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000068'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{FACILITY_NAME}{e.g. 'Apex Generating Station'. See Details.}

\item{GAS_NAME_SF6_OR_PFC}{e.g. 'Sulfur hexafluoride'. See Details.}

\item{NEW_EQUIPMENT_NAMEPLATE_CAP}{e.g. '0'. See Details.}

\item{RETIRING_EQUIP_NAMEPLATE_CAP}{e.g. '0'. See Details.}

\item{NET_INCREASE_NAMEPLATE_CAP_RPT}{e.g. '0'. See Details.}
}
\description{
Retrieve dd increase nameplate data from ghg database
}
