% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_dd_sf6pfc_acquisitions.R
\name{ghg_dd_sf6pfc_acquisitions}
\alias{ghg_dd_sf6pfc_acquisitions}
\title{Retrieve dd sf6pfc acquisitions data from ghg database}
\usage{
ghg_dd_sf6pfc_acquisitions(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, GAS_NAME_SF6_OR_PFC = NULL,
  GAS_PURCHASED_FROM_EQUIP_MANUF = NULL, GAS_PURCHASED_IN_BULK = NULL,
  GAS_RETURN_AFTER_OFFSITE_RECYC = NULL,
  ACQUISITIONS_OF_GAS_REPORTED = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000068'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{FACILITY_NAME}{e.g. 'Apex Generating Station'. See Details.}

\item{GAS_NAME_SF6_OR_PFC}{e.g. 'Sulfur hexafluoride'. See Details.}

\item{GAS_PURCHASED_FROM_EQUIP_MANUF}{e.g. '0'. See Details.}

\item{GAS_PURCHASED_IN_BULK}{e.g. '0'. See Details.}

\item{GAS_RETURN_AFTER_OFFSITE_RECYC}{e.g. '0'. See Details.}

\item{ACQUISITIONS_OF_GAS_REPORTED}{e.g. '0'. See Details.}
}
\description{
Retrieve dd sf6pfc acquisitions data from ghg database
}

