% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_dd_sf6pfc_decreases.R
\name{ghg_dd_sf6pfc_decreases}
\alias{ghg_dd_sf6pfc_decreases}
\title{Retrieve dd sf6pfc decreases data from ghg database}
\usage{
ghg_dd_sf6pfc_decreases(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, GAS_NAME_SF6_OR_PFC = NULL,
  GAS_STORED_IN_CONTAIN_BEG_YEAR = NULL,
  GAS_STORED_IN_CONTAIN_END_YEAR = NULL,
  DECREASE_IN_INVENTORY_REPORTED = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000068'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{FACILITY_NAME}{e.g. 'Apex Generating Station'. See Details.}

\item{GAS_NAME_SF6_OR_PFC}{e.g. 'Sulfur hexafluoride'. See Details.}

\item{GAS_STORED_IN_CONTAIN_BEG_YEAR}{e.g. '926.5'. See Details.}

\item{GAS_STORED_IN_CONTAIN_END_YEAR}{e.g. '926.5'. See Details.}

\item{DECREASE_IN_INVENTORY_REPORTED}{e.g. '0'. See Details.}
}
\description{
Retrieve dd sf6pfc decreases data from ghg database
}
