% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_dd_sf6pfc_disbursement.R
\name{ghg_dd_sf6pfc_disbursement}
\alias{ghg_dd_sf6pfc_disbursement}
\title{Retrieve dd sf6pfc disbursement data from ghg database}
\usage{
ghg_dd_sf6pfc_disbursement(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  GAS_RETURNED_TO_SUPPLIERS = NULL,
  GAS_CONTAINED_IN_EQUIPMNT_SOLD = NULL, FACILITY_NAME = NULL,
  GAS_NAME_SF6_OR_PFC = NULL, GAS_SENT_OFFSITE_FOR_DESTRUCT = NULL,
  GAS_SENT_OFFSITE_FOR_RECYCLING = NULL,
  DISBURSEMENTS_OF_GAS_REPORTED = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000068'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{GAS_RETURNED_TO_SUPPLIERS}{e.g. '0'. See Details.}

\item{GAS_CONTAINED_IN_EQUIPMNT_SOLD}{e.g. '0'. See Details.}

\item{FACILITY_NAME}{e.g. 'Apex Generating Station'. See Details.}

\item{GAS_NAME_SF6_OR_PFC}{e.g. 'Sulfur hexafluoride'. See Details.}

\item{GAS_SENT_OFFSITE_FOR_DESTRUCT}{e.g. '0'. See Details.}

\item{GAS_SENT_OFFSITE_FOR_RECYCLING}{e.g. '0'. See Details.}

\item{DISBURSEMENTS_OF_GAS_REPORTED}{e.g. '0'. See Details.}
}
\description{
Retrieve dd sf6pfc disbursement data from ghg database
}
