% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_dd_subpart_level_information.R
\name{ghg_dd_subpart_level_information}
\alias{ghg_dd_subpart_level_information}
\title{Retrieve dd subpart level information data from ghg database}
\usage{
ghg_dd_subpart_level_information(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, GHG_NAME = NULL, GHG_QUANTITY = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000068'. See Details.}

\item{REPORTING_YEAR}{e.g. '2012'. See Details.}

\item{FACILITY_NAME}{e.g. 'Apex Generating Station'. See Details.}

\item{GHG_NAME}{e.g. 'Sulfur hexafluoride'. See Details.}

\item{GHG_QUANTITY}{e.g. '0'. See Details.}
}
\description{
Retrieve dd subpart level information data from ghg database
}

