% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_dd_transmission_line.R
\name{ghg_dd_transmission_line}
\alias{ghg_dd_transmission_line}
\title{Retrieve dd transmission line data from ghg database}
\usage{
ghg_dd_transmission_line(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, LINES_CARRY_VOLT_ABOVE35KV_LEN = NULL,
  LINES_CARRY_VOLT_BELOW35KV_LEN = NULL,
  MISS_DATA_USED_LINES_ABOVE35KV = NULL,
  MISS_DATA_USED_LINES_BELOW35KV = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000068'. See Details.}

\item{REPORTING_YEAR}{e.g. '2012'. See Details.}

\item{FACILITY_NAME}{e.g. 'Apex Generating Station'. See Details.}

\item{LINES_CARRY_VOLT_ABOVE35KV_LEN}{e.g. '0'. See Details.}

\item{LINES_CARRY_VOLT_BELOW35KV_LEN}{e.g. '0'. See Details.}

\item{MISS_DATA_USED_LINES_ABOVE35KV}{e.g. 'No'. See Details.}

\item{MISS_DATA_USED_LINES_BELOW35KV}{e.g. 'No'. See Details.}
}
\description{
Retrieve dd transmission line data from ghg database
}

