% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_dd_user_emissions_calc.R
\name{ghg_dd_user_emissions_calc}
\alias{ghg_dd_user_emissions_calc}
\title{Retrieve dd user emissions calc data from ghg database}
\usage{
ghg_dd_user_emissions_calc(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  USE_REMISSIONS_CALCULATED = NULL, FACILITY_NAME = NULL,
  GAS_NAME_SF6_OR_PFC = NULL, ACQUISITIONS_OF_GAS_CALCULATED = NULL,
  DECREASE_IN_GAS_INVENTORY_CALC = NULL,
  DISBURSEMENTS_OF_GAS_CALCULATE = NULL,
  NET_INCREASE_NAMEPLATE_CAP_CAL = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000068'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{USE_REMISSIONS_CALCULATED}{e.g. '0'. See Details.}

\item{FACILITY_NAME}{e.g. 'Apex Generating Station'. See Details.}

\item{GAS_NAME_SF6_OR_PFC}{e.g. 'Sulfur hexafluoride'. See Details.}

\item{ACQUISITIONS_OF_GAS_CALCULATED}{e.g. '0'. See Details.}

\item{DECREASE_IN_GAS_INVENTORY_CALC}{e.g. '0'. See Details.}

\item{DISBURSEMENTS_OF_GAS_CALCULATE}{e.g. '0'. See Details.}

\item{NET_INCREASE_NAMEPLATE_CAP_CAL}{e.g. '0'. See Details.}
}
\description{
Retrieve dd user emissions calc data from ghg database
}

