% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_ee_facility_info.R
\name{ghg_ee_facility_info}
\alias{ghg_ee_facility_info}
\title{Retrieve ee facility info data from ghg database}
\usage{
ghg_ee_facility_info(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  NUMBER_CHLORIDE_PROCESS_LINES = NULL, FACILITY_NAME = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000086'. See Details.}

\item{REPORTING_YEAR}{e.g. '2013'. See Details.}

\item{NUMBER_CHLORIDE_PROCESS_LINES}{e.g. '2'. See Details.}

\item{FACILITY_NAME}{e.g. 'DUPONT JOHNSONVILLE'. See Details.}
}
\description{
Retrieve ee facility info data from ghg database
}
