% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_ee_nocems_monthlydetails.R
\name{ghg_ee_nocems_monthlydetails}
\alias{ghg_ee_nocems_monthlydetails}
\title{Retrieve ee nocems monthlydetails data from ghg database}
\usage{
ghg_ee_nocems_monthlydetails(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  BASIS_FOR_CARB_CONTENT_MONTHLY = NULL, MONTH_NAME = NULL,
  UNIT_TYPE = NULL, UNIT_NAME = NULL, FACILITY_NAME = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000086'. See Details.}

\item{REPORTING_YEAR}{e.g. '2013'. See Details.}

\item{BASIS_FOR_CARB_CONTENT_MONTHLY}{e.g. 'SUPPLIER RECORDS'. See Details.}

\item{MONTH_NAME}{e.g. 'JANUARY'. See Details.}

\item{UNIT_TYPE}{e.g. 'NA'. See Details.}

\item{UNIT_NAME}{e.g. 'JV1'. See Details.}

\item{FACILITY_NAME}{e.g. 'DUPONT JOHNSONVILLE'. See Details.}
}
\description{
Retrieve ee nocems monthlydetails data from ghg database
}
