% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_ee_nocemstio2details.R
\name{ghg_ee_nocemstio2details}
\alias{ghg_ee_nocemstio2details}
\title{Retrieve ee nocemstio2details data from ghg database}
\usage{
ghg_ee_nocemstio2details(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  MTHS_MISDATA_PROCS_USED_CCWAST = NULL,
  ANN_CO2_EMM_PER_PROCESS_LINE = NULL, UNIT_TYPE = NULL,
  UNIT_NAME = NULL, FACILITY_NAME = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000086'. See Details.}

\item{REPORTING_YEAR}{e.g. '2014'. See Details.}

\item{MTHS_MISDATA_PROCS_USED_CCWAST}{e.g. '0'. See Details.}

\item{ANN_CO2_EMM_PER_PROCESS_LINE}{e.g. 'NA'. See Details.}

\item{UNIT_TYPE}{e.g. 'NA'. See Details.}

\item{UNIT_NAME}{e.g. 'JV2'. See Details.}

\item{FACILITY_NAME}{e.g. 'DUPONT JOHNSONVILLE'. See Details.}
}
\description{
Retrieve ee nocemstio2details data from ghg database
}
