% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_ee_subpart_level_information.R
\name{ghg_ee_subpart_level_information}
\alias{ghg_ee_subpart_level_information}
\title{Retrieve ee subpart level information data from ghg database}
\usage{
ghg_ee_subpart_level_information(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  GHG_QUANTITY = NULL, GHG_NAME = NULL, FACILITY_NAME = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000086'. See Details.}

\item{REPORTING_YEAR}{e.g. '2010'. See Details.}

\item{GHG_QUANTITY}{e.g. '278683'. See Details.}

\item{GHG_NAME}{e.g. 'CARBON DIOXIDE'. See Details.}

\item{FACILITY_NAME}{e.g. 'DUPONT JOHNSONVILLE'. See Details.}
}
\description{
Retrieve ee subpart level information data from ghg database
}

