% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_f_smelter_details.R
\name{ghg_f_smelter_details}
\alias{ghg_f_smelter_details}
\title{Retrieve f smelter details data from ghg database}
\usage{
ghg_f_smelter_details(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  DATE_SMELTER_SSCOEFF_MEASURED = NULL, UNIT_NAME = NULL,
  FACILITY_NAME = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000635'. See Details.}

\item{REPORTING_YEAR}{e.g. '2013'. See Details.}

\item{DATE_SMELTER_SSCOEFF_MEASURED}{e.g. '2011-10-28'. See Details.}

\item{UNIT_NAME}{e.g. 'POTLINE 1, 2 AND 5'. See Details.}

\item{FACILITY_NAME}{e.g. 'ALCOA INC WENATCHEE WORKS'. See Details.}
}
\description{
Retrieve f smelter details data from ghg database
}
