% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_ff_degas_qtrly.R
\name{ghg_ff_degas_qtrly}
\alias{ghg_ff_degas_qtrly}
\title{Retrieve ff degas qtrly data from ghg database}
\usage{
ghg_ff_degas_qtrly(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  TIME_SUB_DATA_USED_PRESSURE = NULL,
  TIME_SUB_DATA_USED_MOISTURE = NULL,
  START_DATE_FOR_DEGAS_SYSTEM = NULL, START_DATE_EQUIP_NOT_FUNC = NULL,
  END_DATE_EQUIP_NOT_FUNC = NULL, GAS_SYSTEM_UNIT_ID = NULL,
  QUARTER = NULL, TIME_SUB_DATA_USED_TEMPERATURE = NULL,
  DEGAS_MINING_OPERAT_STOP_DATE = NULL, METHOD_USED_CH4_CONC = NULL,
  QTRLY_CEMS_METHANE_CONCENTRATI = NULL,
  TIME_SUB_DATA_USD_QTR_CEMS_CH4 = NULL,
  QTRLY_METHANE_CONC_WEEKLY = NULL,
  TIME_SUB_DATA_USD_QTR_WKLY_CH4 = NULL, IS_EQ_FF9_REQ = NULL,
  GAS_ORGANIC_CONC_CORRCTN_FACTR = NULL,
  MOISTURE_CONTENT_WKLY_CH4 = NULL, PRESSURE_WKLY_CH4 = NULL,
  TEMPERATURE_USED_WKLY_CH4 = NULL, FACILITY_NAME = NULL,
  TIME_SUBSTITUTED_DATA_FOR_MCF = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1004441'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{TIME_SUB_DATA_USED_PRESSURE}{e.g. 'NA'. See Details.}

\item{TIME_SUB_DATA_USED_MOISTURE}{e.g. 'NA'. See Details.}

\item{START_DATE_FOR_DEGAS_SYSTEM}{e.g. 'NA'. See Details.}

\item{START_DATE_EQUIP_NOT_FUNC}{e.g. 'NA'. See Details.}

\item{END_DATE_EQUIP_NOT_FUNC}{e.g. 'NA'. See Details.}

\item{GAS_SYSTEM_UNIT_ID}{e.g. 'NO DEGASIFICATION GAS COLLECTION '. See Details.}

\item{QUARTER}{e.g. 'NA'. See Details.}

\item{TIME_SUB_DATA_USED_TEMPERATURE}{e.g. 'NA'. See Details.}

\item{DEGAS_MINING_OPERAT_STOP_DATE}{e.g. 'NA'. See Details.}

\item{METHOD_USED_CH4_CONC}{e.g. 'NA'. See Details.}

\item{QTRLY_CEMS_METHANE_CONCENTRATI}{e.g. 'NA'. See Details.}

\item{TIME_SUB_DATA_USD_QTR_CEMS_CH4}{e.g. 'NA'. See Details.}

\item{QTRLY_METHANE_CONC_WEEKLY}{e.g. 'NA'. See Details.}

\item{TIME_SUB_DATA_USD_QTR_WKLY_CH4}{e.g. 'NA'. See Details.}

\item{IS_EQ_FF9_REQ}{e.g. 'NA'. See Details.}

\item{GAS_ORGANIC_CONC_CORRCTN_FACTR}{e.g. 'NA'. See Details.}

\item{MOISTURE_CONTENT_WKLY_CH4}{e.g. 'NA'. See Details.}

\item{PRESSURE_WKLY_CH4}{e.g. 'NA'. See Details.}

\item{TEMPERATURE_USED_WKLY_CH4}{e.g. 'NA'. See Details.}

\item{FACILITY_NAME}{e.g. 'COAL RIVER MINING, LLC'. See Details.}

\item{TIME_SUBSTITUTED_DATA_FOR_MCF}{e.g. 'NA'. See Details.}
}
\description{
Retrieve ff degas qtrly data from ghg database
}
