% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_ff_degas_weekly.R
\name{ghg_ff_degas_weekly}
\alias{ghg_ff_degas_weekly}
\title{Retrieve ff degas weekly data from ghg database}
\usage{
ghg_ff_degas_weekly(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  GAS_SYSTEM_UNIT_ID = NULL, QUARTER = NULL, WEEK = NULL,
  DEGAS_FLOW_RATE_METHOD = NULL, TIME_SUB_DATA_WK_DEGAS_VOL_FLO = NULL,
  START_DATE_MONIT_EQUIP_NOT_FUN = NULL,
  STOP_DATE_MONIT_EQUIP_NOT_FUNC = NULL,
  WKLY_CH4_LIBER_MONITORING_PT = NULL,
  WEEKLY_VOLUMETRIC_FLOW_RATE = NULL,
  WKLY_VOLUMETRIC_FLOW_RATE_UNIT = NULL, FACILITY_NAME = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1003058'. See Details.}

\item{REPORTING_YEAR}{e.g. '2014'. See Details.}

\item{GAS_SYSTEM_UNIT_ID}{e.g. 'DE-GAS #1'. See Details.}

\item{QUARTER}{e.g. '4'. See Details.}

\item{WEEK}{e.g. '6'. See Details.}

\item{DEGAS_FLOW_RATE_METHOD}{e.g. 'MONITORED USING 98.324(C)(2)'. See Details.}

\item{TIME_SUB_DATA_WK_DEGAS_VOL_FLO}{e.g. '0.01'. See Details.}

\item{START_DATE_MONIT_EQUIP_NOT_FUN}{e.g. 'NA'. See Details.}

\item{STOP_DATE_MONIT_EQUIP_NOT_FUNC}{e.g. 'NA'. See Details.}

\item{WKLY_CH4_LIBER_MONITORING_PT}{e.g. '1.29'. See Details.}

\item{WEEKLY_VOLUMETRIC_FLOW_RATE}{e.g. '162'. See Details.}

\item{WKLY_VOLUMETRIC_FLOW_RATE_UNIT}{e.g. 'ACFM'. See Details.}

\item{FACILITY_NAME}{e.g. 'OAKTOWN FUELS MINE NO.1'. See Details.}
}
\description{
Retrieve ff degas weekly data from ghg database
}
