% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_ff_degasification.R
\name{ghg_ff_degasification}
\alias{ghg_ff_degasification}
\title{Retrieve ff degasification data from ghg database}
\usage{
ghg_ff_degasification(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  DEGAS_SYS_ANNUAL_OPERATING_HRS = NULL, GAS_SYSTEM_UNIT_ID = NULL,
  DEGAS_SYSTM_SURFACE_AREA = NULL, FACILITY_NAME = NULL,
  DEGAS_NUM_WELLS = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1001687'. See Details.}

\item{REPORTING_YEAR}{e.g. '2013'. See Details.}

\item{DEGAS_SYS_ANNUAL_OPERATING_HRS}{e.g. '0'. See Details.}

\item{GAS_SYSTEM_UNIT_ID}{e.g. 'NONE'. See Details.}

\item{DEGAS_SYSTM_SURFACE_AREA}{e.g. '0'. See Details.}

\item{FACILITY_NAME}{e.g. 'BLUE DIAMOND MINING MINE 77'. See Details.}

\item{DEGAS_NUM_WELLS}{e.g. '0'. See Details.}
}
\description{
Retrieve ff degasification data from ghg database
}
