% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_ff_destroffsite_backup.R
\name{ghg_ff_destroffsite_backup}
\alias{ghg_ff_destroffsite_backup}
\title{Retrieve ff destroffsite backup data from ghg database}
\usage{
ghg_ff_destroffsite_backup(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  INDV_WEL_SHFT_DSTR_OFFST_TRNS9 = NULL,
  INDV_WEL_SHFT_DSTR_OFST_TRNS10 = NULL, DEVICE_DESCRIPTION = NULL,
  ADDITIONAL_DESCRIPTION = NULL, BACKUP_DESTRUC_DEVICE_PRESENCE = NULL,
  ANNUAL_OP_HRS_DESTRUC_DEVICE = NULL,
  ANNU_OP_HRS_BCKUP1_DEST_DEVICE = NULL,
  ANNU_OP_HRS_BCKUP2_DEST_DEVICE = NULL, DESTRUCTION_DEVICE_ID = NULL,
  FACILITY_NAME = NULL, IS_GAS_DESTROYED_OFFSITE = NULL,
  GAS_DESTR_MINE_TRANS_OFFSTE = NULL,
  DESTR_EFFICIEN_BACKUP_1_DEVICE = NULL,
  DESTR_EFFICIEN_BACKUP_2_DEVICE = NULL,
  DESTR_EFFICIEN_PRIMARY_DEVICE = NULL,
  INDV_WEL_SHFT_DSTR_OFFST_TRNS1 = NULL,
  INDV_WEL_SHFT_DSTR_OFFST_TRNS2 = NULL,
  INDV_WEL_SHFT_DSTR_OFFST_TRNS3 = NULL,
  INDV_WEL_SHFT_DSTR_OFFST_TRNS4 = NULL,
  INDV_WEL_SHFT_DSTR_OFFST_TRNS5 = NULL,
  INDV_WEL_SHFT_DSTR_OFFST_TRNS6 = NULL,
  INDV_WEL_SHFT_DSTR_OFFST_TRNS7 = NULL,
  INDV_WEL_SHFT_DSTR_OFFST_TRNS8 = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1001687'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{INDV_WEL_SHFT_DSTR_OFFST_TRNS9}{e.g. 'NA'. See Details.}

\item{INDV_WEL_SHFT_DSTR_OFST_TRNS10}{e.g. 'NA'. See Details.}

\item{DEVICE_DESCRIPTION}{e.g. 'NA'. See Details.}

\item{ADDITIONAL_DESCRIPTION}{e.g. 'NA'. See Details.}

\item{BACKUP_DESTRUC_DEVICE_PRESENCE}{e.g. 'NA'. See Details.}

\item{ANNUAL_OP_HRS_DESTRUC_DEVICE}{e.g. 'NA'. See Details.}

\item{ANNU_OP_HRS_BCKUP1_DEST_DEVICE}{e.g. 'NA'. See Details.}

\item{ANNU_OP_HRS_BCKUP2_DEST_DEVICE}{e.g. 'NA'. See Details.}

\item{DESTRUCTION_DEVICE_ID}{e.g. 'NONE'. See Details.}

\item{FACILITY_NAME}{e.g. 'BLUE DIAMOND MINING MINE 77'. See Details.}

\item{IS_GAS_DESTROYED_OFFSITE}{e.g. 'NA'. See Details.}

\item{GAS_DESTR_MINE_TRANS_OFFSTE}{e.g. 'NA'. See Details.}

\item{DESTR_EFFICIEN_BACKUP_1_DEVICE}{e.g. 'NA'. See Details.}

\item{DESTR_EFFICIEN_BACKUP_2_DEVICE}{e.g. 'NA'. See Details.}

\item{DESTR_EFFICIEN_PRIMARY_DEVICE}{e.g. 'NA'. See Details.}

\item{INDV_WEL_SHFT_DSTR_OFFST_TRNS1}{e.g. 'NA'. See Details.}

\item{INDV_WEL_SHFT_DSTR_OFFST_TRNS2}{e.g. 'NA'. See Details.}

\item{INDV_WEL_SHFT_DSTR_OFFST_TRNS3}{e.g. 'NA'. See Details.}

\item{INDV_WEL_SHFT_DSTR_OFFST_TRNS4}{e.g. 'NA'. See Details.}

\item{INDV_WEL_SHFT_DSTR_OFFST_TRNS5}{e.g. 'NA'. See Details.}

\item{INDV_WEL_SHFT_DSTR_OFFST_TRNS6}{e.g. 'NA'. See Details.}

\item{INDV_WEL_SHFT_DSTR_OFFST_TRNS7}{e.g. 'NA'. See Details.}

\item{INDV_WEL_SHFT_DSTR_OFFST_TRNS8}{e.g. 'NA'. See Details.}
}
\description{
Retrieve ff destroffsite backup data from ghg database
}
