% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_ff_destroffsite_qtrly.R
\name{ghg_ff_destroffsite_qtrly}
\alias{ghg_ff_destroffsite_qtrly}
\title{Retrieve ff destroffsite qtrly data from ghg database}
\usage{
ghg_ff_destroffsite_qtrly(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, DESTR_OFFSITE_UNIT_ID = NULL, QUARTER = NULL,
  QTRLY_CH4_DESTR_ONSITE = NULL, QTRLY_CH4_TRANS_OFFSITE = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1001687'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{FACILITY_NAME}{e.g. 'BLUE DIAMOND MINING MINE 77'. See Details.}

\item{DESTR_OFFSITE_UNIT_ID}{e.g. 'NONE'. See Details.}

\item{QUARTER}{e.g. '1'. See Details.}

\item{QTRLY_CH4_DESTR_ONSITE}{e.g. 'NA'. See Details.}

\item{QTRLY_CH4_TRANS_OFFSITE}{e.g. 'NA'. See Details.}
}
\description{
Retrieve ff destroffsite qtrly data from ghg database
}

