% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_ff_destroffsite_weekly.R
\name{ghg_ff_destroffsite_weekly}
\alias{ghg_ff_destroffsite_weekly}
\title{Retrieve ff destroffsite weekly data from ghg database}
\usage{
ghg_ff_destroffsite_weekly(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  WEEKLY_DESTRUCTION_DEVICE_ID = NULL,
  TIME_SUB_WKLY_VOL_FLOW_METH_DE = NULL,
  TIME_SUB_WKLY_METH_CON_METH_DE = NULL,
  WKLY_FLW_RT_USD_CALC_METH_DEST = NULL,
  WKLY_VOLUMETRIC_FLOW_RATE_UNIT = NULL,
  WEEKLY_METHANE_CONCENTRATION = NULL, FACILITY_NAME = NULL,
  QUARTER = NULL, WEEK = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1001687'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{WEEKLY_DESTRUCTION_DEVICE_ID}{e.g. 'NONE'. See Details.}

\item{TIME_SUB_WKLY_VOL_FLOW_METH_DE}{e.g. '0'. See Details.}

\item{TIME_SUB_WKLY_METH_CON_METH_DE}{e.g. '0'. See Details.}

\item{WKLY_FLW_RT_USD_CALC_METH_DEST}{e.g. '0'. See Details.}

\item{WKLY_VOLUMETRIC_FLOW_RATE_UNIT}{e.g. 'NA'. See Details.}

\item{WEEKLY_METHANE_CONCENTRATION}{e.g. 'NA'. See Details.}

\item{FACILITY_NAME}{e.g. 'BLUE DIAMOND MINING MINE 77'. See Details.}

\item{QUARTER}{e.g. '1'. See Details.}

\item{WEEK}{e.g. '1'. See Details.}
}
\description{
Retrieve ff destroffsite weekly data from ghg database
}
