% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_ff_subpart_level_information.R
\name{ghg_ff_subpart_level_information}
\alias{ghg_ff_subpart_level_information}
\title{Retrieve ff subpart level information data from ghg database}
\usage{
ghg_ff_subpart_level_information(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, GHG_NAME = NULL, GHG_QUANTITY = NULL,
  MSHA_IDENTIFICATION_NUMBER = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000386'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{FACILITY_NAME}{e.g. 'NO. 65'. See Details.}

\item{GHG_NAME}{e.g. 'METHANE'. See Details.}

\item{GHG_QUANTITY}{e.g. '391.06'. See Details.}

\item{MSHA_IDENTIFICATION_NUMBER}{e.g. 'NA'. See Details.}
}
\description{
Retrieve ff subpart level information data from ghg database
}

