% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_ff_summary_source.R
\name{ghg_ff_summary_source}
\alias{ghg_ff_summary_source}
\title{Retrieve ff summary source data from ghg database}
\usage{
ghg_ff_summary_source(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, QUARTER = NULL,
  QTR_METH_DEST_TRNS_OFFSTE_CALC = NULL,
  QTR_METH_DEST_TRNS_OFFSTE_USER = NULL,
  QTR_METH_LIBER_DEGAS_CALC = NULL, QTR_METH_LIBER_DEGAS_USER = NULL,
  QTR_METH_LIBER_VENT_CALC = NULL, QTR_METH_LIBER_VENT_USER = NULL,
  NUMBER_OF_WELLS = NULL, NUM_OF_SHAFTS = NULL,
  IS_GAS_COLL_SYSTEM = NULL, QUARTER_1_CO2_EMISSIONS = NULL,
  TOT_QRTLY_CH4 = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000386'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{FACILITY_NAME}{e.g. 'NO. 65'. See Details.}

\item{QUARTER}{e.g. 'QUARTER 1 (JAN-MAR)'. See Details.}

\item{QTR_METH_DEST_TRNS_OFFSTE_CALC}{e.g. '0'. See Details.}

\item{QTR_METH_DEST_TRNS_OFFSTE_USER}{e.g. 'N'. See Details.}

\item{QTR_METH_LIBER_DEGAS_CALC}{e.g. '0'. See Details.}

\item{QTR_METH_LIBER_DEGAS_USER}{e.g. 'N'. See Details.}

\item{QTR_METH_LIBER_VENT_CALC}{e.g. '154.19'. See Details.}

\item{QTR_METH_LIBER_VENT_USER}{e.g. 'N'. See Details.}

\item{NUMBER_OF_WELLS}{e.g. 'NA'. See Details.}

\item{NUM_OF_SHAFTS}{e.g. 'NA'. See Details.}

\item{IS_GAS_COLL_SYSTEM}{e.g. 'N'. See Details.}

\item{QUARTER_1_CO2_EMISSIONS}{e.g. '0'. See Details.}

\item{TOT_QRTLY_CH4}{e.g. '154.19'. See Details.}
}
\description{
Retrieve ff summary source data from ghg database
}
