% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_ff_well_and_shaft.R
\name{ghg_ff_well_and_shaft}
\alias{ghg_ff_well_and_shaft}
\title{Retrieve ff well and shaft data from ghg database}
\usage{
ghg_ff_well_and_shaft(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  WELL_SHAFT_DAYS_IN_OPERATION = NULL, WELL_SHAFT_CLOSE_DATE = NULL,
  WELL_SHAFT_START_DATE = NULL, CENTRALIZD_MONITORING_POINT_ID = NULL,
  FACILITY_NAME = NULL, WELL_SHAFT_ID = NULL,
  IS_THIS_WELL_OR_SHAFT = NULL, GAS_SYSTEM_UNIT_ID = NULL,
  DESCRIPTION = NULL, ADDITIONAL_DESCRIPTION = NULL,
  WELL_SHAFT_MONITORED_STATUS = NULL, MON_POINT_NAME = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1001681'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{WELL_SHAFT_DAYS_IN_OPERATION}{e.g. 'NA'. See Details.}

\item{WELL_SHAFT_CLOSE_DATE}{e.g. 'NA'. See Details.}

\item{WELL_SHAFT_START_DATE}{e.g. 'NA'. See Details.}

\item{CENTRALIZD_MONITORING_POINT_ID}{e.g. 'NA'. See Details.}

\item{FACILITY_NAME}{e.g. 'WILLOW LAKE MINE'. See Details.}

\item{WELL_SHAFT_ID}{e.g. 'WILLOW LAKE SOUTH EXHAUST FAN'. See Details.}

\item{IS_THIS_WELL_OR_SHAFT}{e.g. 'SHAFT'. See Details.}

\item{GAS_SYSTEM_UNIT_ID}{e.g. 'NA'. See Details.}

\item{DESCRIPTION}{e.g. 'EXHAUST VENT SHAFT'. See Details.}

\item{ADDITIONAL_DESCRIPTION}{e.g. 'NA'. See Details.}

\item{WELL_SHAFT_MONITORED_STATUS}{e.g. 'MONITORED INDIVIDUALLY'. See Details.}

\item{MON_POINT_NAME}{e.g. 'NA'. See Details.}
}
\description{
Retrieve ff well and shaft data from ghg database
}
