% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_g_non_cems_source_info.R
\name{ghg_g_non_cems_source_info}
\alias{ghg_g_non_cems_source_info}
\title{Retrieve g non cems source info data from ghg database}
\usage{
ghg_g_non_cems_source_info(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  UNIT_NAME = NULL, CO2_EMISSIONS = NULL, MONTH_NAME = NULL,
  FEED_STOCK_QUANTITY_METHD_MONT = NULL,
  FEED_STOCK_QUANTITY_OTHER_METH = NULL,
  FEEDSTOCK_CARBON_CONTENT_BASIS = NULL,
  FEED_STOCK_QUANTITY_SUB_DATA = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000630'. See Details.}

\item{REPORTING_YEAR}{e.g. '2013'. See Details.}

\item{UNIT_NAME}{e.g. 'Kellogg'. See Details.}

\item{CO2_EMISSIONS}{e.g. '644804'. See Details.}

\item{MONTH_NAME}{e.g. 'January'. See Details.}

\item{FEED_STOCK_QUANTITY_METHD_MONT}{e.g. 'Flow meter'. See Details.}

\item{FEED_STOCK_QUANTITY_OTHER_METH}{e.g. 'NA'. See Details.}

\item{FEEDSTOCK_CARBON_CONTENT_BASIS}{e.g. 'Supplier records'. See Details.}

\item{FEED_STOCK_QUANTITY_SUB_DATA}{e.g. 'N'. See Details.}
}
\description{
Retrieve g non cems source info data from ghg database
}

