% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_gg_facility_info.R
\name{ghg_gg_facility_info}
\alias{ghg_gg_facility_info}
\title{Retrieve gg facility info data from ghg database}
\usage{
ghg_gg_facility_info(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  NUMBER_OF_WAELZ_KILNS = NULL, NUM_OF_ELECTRO_THERMIC_FURNACE = NULL,
  FACILITY_NAME = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1004424'. See Details.}

\item{REPORTING_YEAR}{e.g. '2010'. See Details.}

\item{NUMBER_OF_WAELZ_KILNS}{e.g. 'NA'. See Details.}

\item{NUM_OF_ELECTRO_THERMIC_FURNACE}{e.g. 'NA'. See Details.}

\item{FACILITY_NAME}{e.g. 'HORSEHEAD CORP. MONACA SMELTER'. See Details.}
}
\description{
Retrieve gg facility info data from ghg database
}
