% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_gg_subpart_level_information.R
\name{ghg_gg_subpart_level_information}
\alias{ghg_gg_subpart_level_information}
\title{Retrieve gg subpart level information data from ghg database}
\usage{
ghg_gg_subpart_level_information(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  OTHER_GHG_NAME = NULL, GHG_QUANTITY_OVERRIDE_INDICATR = NULL,
  GHG_QUANTITY_UNIT_OF_MEASURE = NULL, GHG_QUANTITY = NULL,
  GHG_NAME = NULL, FACILITY_NAME = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1004424'. See Details.}

\item{REPORTING_YEAR}{e.g. '2010'. See Details.}

\item{OTHER_GHG_NAME}{e.g. 'NA'. See Details.}

\item{GHG_QUANTITY_OVERRIDE_INDICATR}{e.g. 'NA'. See Details.}

\item{GHG_QUANTITY_UNIT_OF_MEASURE}{e.g. 'METRIC TONS'. See Details.}

\item{GHG_QUANTITY}{e.g. '198494'. See Details.}

\item{GHG_NAME}{e.g. 'CARBON DIOXIDE'. See Details.}

\item{FACILITY_NAME}{e.g. 'HORSEHEAD CORP. MONACA SMELTER'. See Details.}
}
\description{
Retrieve gg subpart level information data from ghg database
}

