% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_h_cems_details.R
\name{ghg_h_cems_details}
\alias{ghg_h_cems_details}
\title{Retrieve h cems details data from ghg database}
\usage{
ghg_h_cems_details(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  CEMS_CO2_COMBUSTION_EMISSIONS = NULL,
  TIER4_SUB_OPER_HRS_GAS_FLOW_RT = NULL,
  TIER4_SUB_OPER_HRS_MOIST_CONT = NULL,
  CEMS_METHODOLOGY_START_DATE = NULL, CEMS_METHODOLOGY_END_DATE = NULL,
  CEMS_FUEL = NULL, UNIT_NAMES = NULL,
  CEMS_CO2_EMISSIONS_BY_QUARTER = NULL, FACILITY_NAME = NULL,
  CML_NAME = NULL, CML_CONFIGURATION_TYPE = NULL,
  CEMS_BIOGENIC_CO2_EMISSIONS = NULL,
  CEMS_NON_BIOGENIC_CO2_EMISSION = NULL, CEMS_CO2_EMISSIONS = NULL,
  CEMS_CH4_EMISSIONS = NULL, CEMS_N2O_EMISSIONS = NULL,
  CML_TOTAL_OPERATING_HOURS = NULL, TIER4_SUB_OPER_HRS_CO2_CONC = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000198'. See Details.}

\item{REPORTING_YEAR}{e.g. '2013'. See Details.}

\item{CEMS_CO2_COMBUSTION_EMISSIONS}{e.g. 'NA'. See Details.}

\item{TIER4_SUB_OPER_HRS_GAS_FLOW_RT}{e.g. '0'. See Details.}

\item{TIER4_SUB_OPER_HRS_MOIST_CONT}{e.g. '0'. See Details.}

\item{CEMS_METHODOLOGY_START_DATE}{e.g. '2013-01-01'. See Details.}

\item{CEMS_METHODOLOGY_END_DATE}{e.g. '2013-12-31'. See Details.}

\item{CEMS_FUEL}{e.g. 'natural gas, petroleum coke, and recycled plastics'. See Details.}

\item{UNIT_NAMES}{e.g. 'Lehigh Cement Co. LLC, York, PA 17404'. See Details.}

\item{CEMS_CO2_EMISSIONS_BY_QUARTER}{e.g. 'First Quarter-8061.4 Metric Tons'. See Details.}

\item{FACILITY_NAME}{e.g. 'LEHIGH CEMENT CO LLC'. See Details.}

\item{CML_NAME}{e.g. 'Lehigh Cement Co., York, PA 17404'. See Details.}

\item{CML_CONFIGURATION_TYPE}{e.g. 'Process/stationary combustion units share common stack'. See Details.}

\item{CEMS_BIOGENIC_CO2_EMISSIONS}{e.g. '0'. See Details.}

\item{CEMS_NON_BIOGENIC_CO2_EMISSION}{e.g. '74736.2'. See Details.}

\item{CEMS_CO2_EMISSIONS}{e.g. '74736.2'. See Details.}

\item{CEMS_CH4_EMISSIONS}{e.g. '3.11'. See Details.}

\item{CEMS_N2O_EMISSIONS}{e.g. '0.382'. See Details.}

\item{CML_TOTAL_OPERATING_HOURS}{e.g. '6781'. See Details.}

\item{TIER4_SUB_OPER_HRS_CO2_CONC}{e.g. '360'. See Details.}
}
\description{
Retrieve h cems details data from ghg database
}
