% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_h_cems_info.R
\name{ghg_h_cems_info}
\alias{ghg_h_cems_info}
\title{Retrieve h cems info data from ghg database}
\usage{
ghg_h_cems_info(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, CEMS_UNIT_NAME = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000198'. See Details.}

\item{REPORTING_YEAR}{e.g. '2013'. See Details.}

\item{FACILITY_NAME}{e.g. 'LEHIGH CEMENT CO LLC'. See Details.}

\item{CEMS_UNIT_NAME}{e.g. 'Lehigh Cement Co. LLC, York, PA 17404'. See Details.}
}
\description{
Retrieve h cems info data from ghg database
}
