% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_h_no_cems_month_details.R
\name{ghg_h_no_cems_month_details}
\alias{ghg_h_no_cems_month_details}
\title{Retrieve h no cems month details data from ghg database}
\usage{
ghg_h_no_cems_month_details(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FRAC_NON_CALCINED_CAO = NULL, FRAC_NON_CALCINED_MGO = NULL,
  FRAC_TOT_CAO_MISSININFO = NULL, FRAC_TOT_MGO_CLINKR_MISSININFO = NULL,
  MONTH_NAME = NULL, CLINKER_PRODUCTION_METHOD = NULL,
  OTHER_CLINKER_PRODUCTION_METHD = NULL,
  DET_METHD_NON_CALCINED_CALCIUM = NULL,
  DET_METHD_NON_CALCINED_MAGNESI = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000243'. See Details.}

\item{REPORTING_YEAR}{e.g. '2010'. See Details.}

\item{FRAC_NON_CALCINED_CAO}{e.g. 'N'. See Details.}

\item{FRAC_NON_CALCINED_MGO}{e.g. 'N'. See Details.}

\item{FRAC_TOT_CAO_MISSININFO}{e.g. 'N'. See Details.}

\item{FRAC_TOT_MGO_CLINKR_MISSININFO}{e.g. 'N'. See Details.}

\item{MONTH_NAME}{e.g. 'January'. See Details.}

\item{CLINKER_PRODUCTION_METHOD}{e.g. 'Kiln specific feed-to-clinker ratio'. See Details.}

\item{OTHER_CLINKER_PRODUCTION_METHD}{e.g. 'NA'. See Details.}

\item{DET_METHD_NON_CALCINED_CALCIUM}{e.g. 'Method ASTM114-09'. See Details.}

\item{DET_METHD_NON_CALCINED_MAGNESI}{e.g. 'Method ASTM114-09'. See Details.}
}
\description{
Retrieve h no cems month details data from ghg database
}

