% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_h_no_cems_qrtr_details.R
\name{ghg_h_no_cems_qrtr_details}
\alias{ghg_h_no_cems_qrtr_details}
\title{Retrieve h no cems qrtr details data from ghg database}
\usage{
ghg_h_no_cems_qrtr_details(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, UNIT_NAME = NULL, QUARTERNAME = NULL,
  CKDCO2EMISSIONSFACTOR = NULL, CEMENT_DUST_NOT_RECYCYLED_KILN = NULL,
  MISS_DATA_FRAC_NON_CALCND_CAO = NULL,
  MISS_DATA_FRAC_NON_CALCND_MGO = NULL, DET_METHD_NON_CALCINED_CAO = NULL,
  OTHDETMTHDNONCALCAO = NULL, DET_METHD_NON_CALCINED_MAGNESI = NULL,
  OTHDETMTHDNONCALMGO = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000243'. See Details.}

\item{REPORTING_YEAR}{e.g. '2010'. See Details.}

\item{FACILITY_NAME}{e.g. 'LAFARGE MIDWEST INC JOPPA PLANT'. See Details.}

\item{UNIT_NAME}{e.g. 'GHGK1Process'. See Details.}

\item{QUARTERNAME}{e.g. 'Third Quarter'. See Details.}

\item{CKDCO2EMISSIONSFACTOR}{e.g. 'NA'. See Details.}

\item{CEMENT_DUST_NOT_RECYCYLED_KILN}{e.g. 'N'. See Details.}

\item{MISS_DATA_FRAC_NON_CALCND_CAO}{e.g. 'N'. See Details.}

\item{MISS_DATA_FRAC_NON_CALCND_MGO}{e.g. 'N'. See Details.}

\item{DET_METHD_NON_CALCINED_CAO}{e.g. 'Other'. See Details.}

\item{OTHDETMTHDNONCALCAO}{e.g. 'Method ASTM 114-09'. See Details.}

\item{DET_METHD_NON_CALCINED_MAGNESI}{e.g. 'Other'. See Details.}

\item{OTHDETMTHDNONCALMGO}{e.g. 'Method ASTM 114-09'. See Details.}
}
\description{
Retrieve h no cems qrtr details data from ghg database
}

