% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_h_no_cems_raw_material_details.R
\name{ghg_h_no_cems_raw_material_details}
\alias{ghg_h_no_cems_raw_material_details}
\title{Retrieve h no cems raw material details data from ghg database}
\usage{
ghg_h_no_cems_raw_material_details(FACILITY_ID = NULL,
  REPORTING_YEAR = NULL, FACILITY_NAME = NULL, UNIT_NAME = NULL,
  RAW_KILNFD_CONSUMPTION = NULL, ORG_CARBON_RAWKILNFD = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000243'. See Details.}

\item{REPORTING_YEAR}{e.g. '2010'. See Details.}

\item{FACILITY_NAME}{e.g. 'LAFARGE MIDWEST INC JOPPA PLANT'. See Details.}

\item{UNIT_NAME}{e.g. 'GHGK1Process'. See Details.}

\item{RAW_KILNFD_CONSUMPTION}{e.g. 'NA'. See Details.}

\item{ORG_CARBON_RAWKILNFD}{e.g. 'NA'. See Details.}
}
\description{
Retrieve h no cems raw material details data from ghg database
}

