% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_hh_active_aeration_sys_detls.R
\name{ghg_hh_active_aeration_sys_detls}
\alias{ghg_hh_active_aeration_sys_detls}
\title{Retrieve hh active aeration sys detls data from ghg database}
\usage{
ghg_hh_active_aeration_sys_detls(FACILITY_ID = NULL,
  REPORTING_YEAR = NULL, FACILITY_NAME = NULL,
  AERATION_SYS_BLOWER_CAPACITY = NULL,
  LNDFIL_FRAC_AFFECTED_AERATION = NULL,
  AERATION_SYS_BLOW_HRS_OF_OPRAT = NULL, OTHER_MCF_FACTORS = NULL,
  AERATION_SYS_ADD_DESCRIPTION = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1004073'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{FACILITY_NAME}{e.g. 'MARQUETTE CO-SOLID WASTE LF'. See Details.}

\item{AERATION_SYS_BLOWER_CAPACITY}{e.g. '3085'. See Details.}

\item{LNDFIL_FRAC_AFFECTED_AERATION}{e.g. '0.99'. See Details.}

\item{AERATION_SYS_BLOW_HRS_OF_OPRAT}{e.g. '8760'. See Details.}

\item{OTHER_MCF_FACTORS}{e.g. 'AERATION OCCURS BY UTILIZING INNOVATIVE TECHNOLOGY THAT IS NOT APPLICABLE TO THESE FORMS OR MODELS.'. See Details.}

\item{AERATION_SYS_ADD_DESCRIPTION}{e.g. 'AERATION OCCURS BY UTILIZING INNOVATIVE TECHNOLOGY THAT IS NOT APPLICABLE TO THESE FORMS OR MODELS.'. See Details.}
}
\description{
Retrieve hh active aeration sys detls data from ghg database
}
