% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_hh_ann_waste_disposal_qty.R
\name{ghg_hh_ann_waste_disposal_qty}
\alias{ghg_hh_ann_waste_disposal_qty}
\title{Retrieve hh ann waste disposal qty data from ghg database}
\usage{
ghg_hh_ann_waste_disposal_qty(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  WASTE_DISPOSAL_QTY_RY_BY_MTHD = NULL, FACILITY_NAME = NULL,
  WASTE_DISP_REPORTING_YEAR = NULL, TOTAL_WASTE_DISPOSAL_QTY_RY = NULL,
  METHOD_USED_TO_FIND_QTY = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000010'. See Details.}

\item{REPORTING_YEAR}{e.g. '2014'. See Details.}

\item{WASTE_DISPOSAL_QTY_RY_BY_MTHD}{e.g. '117582'. See Details.}

\item{FACILITY_NAME}{e.g. 'SANTA ROSA CENTRAL LANDFILL'. See Details.}

\item{WASTE_DISP_REPORTING_YEAR}{e.g. '2010'. See Details.}

\item{TOTAL_WASTE_DISPOSAL_QTY_RY}{e.g. '117582'. See Details.}

\item{METHOD_USED_TO_FIND_QTY}{e.g. 'USED SCALES TO WEIGH LOADS BEFORE OFF-LOADING AND EITHER USED SCALES TO WEIGH INDIVIDUAL LOADS AFTER OFF-LOADING OR USED REPRESENTATIVE TARE VEHICLE/CONTAINER WEIGHTS'. See Details.}
}
\description{
Retrieve hh ann waste disposal qty data from ghg database
}

