% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_hh_cover_type_details.R
\name{ghg_hh_cover_type_details}
\alias{ghg_hh_cover_type_details}
\title{Retrieve hh cover type details data from ghg database}
\usage{
ghg_hh_cover_type_details(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, COVER_TYPE_NAME = NULL, SURFACE_AREA = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000010'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{FACILITY_NAME}{e.g. 'SANTA ROSA CENTRAL LANDFILL'. See Details.}

\item{COVER_TYPE_NAME}{e.g. 'OTHER SOIL MIXTURE '. See Details.}

\item{SURFACE_AREA}{e.g. '517998'. See Details.}
}
\description{
Retrieve hh cover type details data from ghg database
}
