% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_hh_gas_collection_mnthly_detls.R
\name{ghg_hh_gas_collection_mnthly_detls}
\alias{ghg_hh_gas_collection_mnthly_detls}
\title{Retrieve hh gas collection mnthly detls data from ghg database}
\usage{
ghg_hh_gas_collection_mnthly_detls(FACILITY_ID = NULL,
  REPORTING_YEAR = NULL, FACILITY_NAME = NULL, MONTH_NAME = NULL,
  AVERAGE_MONTHLY_TEMPERATURE = NULL, AVERAGE_MONTHLY_PRESSURE = NULL,
  AVG_MNTHLY_MOISTURE_CONTENT = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000366'. See Details.}

\item{REPORTING_YEAR}{e.g. '2012'. See Details.}

\item{FACILITY_NAME}{e.g. 'MEAD VALLEY DISPOSAL SITE LANDFILL'. See Details.}

\item{MONTH_NAME}{e.g. 'MAY'. See Details.}

\item{AVERAGE_MONTHLY_TEMPERATURE}{e.g. 'NA'. See Details.}

\item{AVERAGE_MONTHLY_PRESSURE}{e.g. '1.00787'. See Details.}

\item{AVG_MNTHLY_MOISTURE_CONTENT}{e.g. 'NA'. See Details.}
}
\description{
Retrieve hh gas collection mnthly detls data from ghg database
}

