% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_hh_gas_collection_system_detls.R
\name{ghg_hh_gas_collection_system_detls}
\alias{ghg_hh_gas_collection_system_detls}
\title{Retrieve hh gas collection system detls data from ghg database}
\usage{
ghg_hh_gas_collection_system_detls(FACILITY_ID = NULL,
  REPORTING_YEAR = NULL, FACILITY_NAME = NULL,
  ANNUAL_LANDFILL_GAS_FLOW = NULL, ANNL_AVG_METHANE_CONCENTRATION = NULL,
  ANNL_AVG_METHANE_CONC_DAYS_SUB = NULL,
  ANNL_AVG_METHANE_CONC_WK_SUB = NULL,
  GAS_COLLECTION_SYS_MANUFACTURE = NULL,
  GAS_COLLECTION_SYSTEM_CAPACITY = NULL, NUMBER_OF_WELLS = NULL,
  EQUATION_HH5_RESULT = NULL, EQUATION_HH7_RESULT = NULL,
  EQUATION_HH6_RESULT = NULL, EQUATION_HH8_RESULT = NULL,
  CH4_GEN_INPUT_TO_EQUATION_HH6 = NULL, CH4_DESTRUCTION_LOCATION = NULL,
  GCS_ANNUAL_OPERATING_HOURS = NULL, ANNUAL_QTY_RECOVERD_CH4_EQ_HH4 = NULL,
  EST_GAS_COLLECTION_EFFICIENCY = NULL,
  LNDFIL_GAS_FLOW_MESR_WET_BASIS = NULL, CH4_CONC_MEASURED_WET_BASIS = NULL,
  BACKUP_DESTRUCTION_DEVICE = NULL, BACKUP_DESTR_DEVICE_ANN_OP_HRS = NULL,
  DESTRUCTION_EFFICIENCY_PCT = NULL, PRIMARY_DESTR_DEVICE_ANN_OP_HR = NULL,
  IS_PRESSURE_INCORPORATED = NULL, IS_TEMPERATURE_INCORPORATED = NULL,
  ANNL_AVG_METHANE_CONC_MONS_SUB = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000010'. See Details.}

\item{REPORTING_YEAR}{e.g. '2014'. See Details.}

\item{FACILITY_NAME}{e.g. 'SANTA ROSA CENTRAL LANDFILL'. See Details.}

\item{ANNUAL_LANDFILL_GAS_FLOW}{e.g. '7144560'. See Details.}

\item{ANNL_AVG_METHANE_CONCENTRATION}{e.g. '55.9'. See Details.}

\item{ANNL_AVG_METHANE_CONC_DAYS_SUB}{e.g. '61'. See Details.}

\item{ANNL_AVG_METHANE_CONC_WK_SUB}{e.g. 'NA'. See Details.}

\item{GAS_COLLECTION_SYS_MANUFACTURE}{e.g. 'C B %26 I'. See Details.}

\item{GAS_COLLECTION_SYSTEM_CAPACITY}{e.g. '750'. See Details.}

\item{NUMBER_OF_WELLS}{e.g. '30'. See Details.}

\item{EQUATION_HH5_RESULT}{e.g. '4627.51'. See Details.}

\item{EQUATION_HH7_RESULT}{e.g. '9650.83'. See Details.}

\item{EQUATION_HH6_RESULT}{e.g. '1510.76'. See Details.}

\item{EQUATION_HH8_RESULT}{e.g. '6534.09'. See Details.}

\item{CH4_GEN_INPUT_TO_EQUATION_HH6}{e.g. 'EQUATION HH-1'. See Details.}

\item{CH4_DESTRUCTION_LOCATION}{e.g. 'ON-SITE'. See Details.}

\item{GCS_ANNUAL_OPERATING_HOURS}{e.g. 'NA'. See Details.}

\item{ANNUAL_QTY_RECOVERD_CH4_EQ_HH4}{e.g. '504.7'. See Details.}

\item{EST_GAS_COLLECTION_EFFICIENCY}{e.g. '0.3'. See Details.}

\item{LNDFIL_GAS_FLOW_MESR_WET_BASIS}{e.g. 'Y'. See Details.}

\item{CH4_CONC_MEASURED_WET_BASIS}{e.g. 'Y'. See Details.}

\item{BACKUP_DESTRUCTION_DEVICE}{e.g. 'NA'. See Details.}

\item{BACKUP_DESTR_DEVICE_ANN_OP_HRS}{e.g. 'NA'. See Details.}

\item{DESTRUCTION_EFFICIENCY_PCT}{e.g. 'NA'. See Details.}

\item{PRIMARY_DESTR_DEVICE_ANN_OP_HR}{e.g. 'NA'. See Details.}

\item{IS_PRESSURE_INCORPORATED}{e.g. 'Y'. See Details.}

\item{IS_TEMPERATURE_INCORPORATED}{e.g. 'Y'. See Details.}

\item{ANNL_AVG_METHANE_CONC_MONS_SUB}{e.g. 'NA'. See Details.}
}
\description{
Retrieve hh gas collection system detls data from ghg database
}

