% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_hh_hist_waste_qty_method.R
\name{ghg_hh_hist_waste_qty_method}
\alias{ghg_hh_hist_waste_qty_method}
\title{Retrieve hh hist waste qty method data from ghg database}
\usage{
ghg_hh_hist_waste_qty_method(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, SCALES_USED_FOR_HIST_WSTE_QTY = NULL,
  TIPPING_RCPT_USED_HIST_WST_QTY = NULL,
  TIPNG_RCP_HIST_WSTE_QTY_STR_YR = NULL,
  TIPNG_RCP_HIST_WSTE_QTY_END_YR = NULL,
  HIST_WASTE_QUANTITY_EST_METHOD = NULL,
  HIST_WSTE_QTY_EST_MTHD_STRT_YR = NULL,
  HIST_WSTE_QTY_EST_MTHD_END_YR = NULL,
  HIST_WSTE_QTY_EST_MTHD_RSN_USE = NULL, LANDFILL_CAPACITY_IN_MTHD_3 = NULL,
  WKG_CAP_HIST_WSTE_QTY_START_YR = NULL,
  WKG_CAP_HIST_WSTE_QTY_END_YR = NULL, IS_WORKING_CAPACITY_USED_IND = NULL,
  YEAR_FOR_METHOD_2 = NULL, EST_POP_SERVD_BY_LANDFL_MTHD_2 = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000010'. See Details.}

\item{REPORTING_YEAR}{e.g. '2013'. See Details.}

\item{FACILITY_NAME}{e.g. 'SANTA ROSA CENTRAL LANDFILL'. See Details.}

\item{SCALES_USED_FOR_HIST_WSTE_QTY}{e.g. 'Y'. See Details.}

\item{TIPPING_RCPT_USED_HIST_WST_QTY}{e.g. 'N'. See Details.}

\item{TIPNG_RCP_HIST_WSTE_QTY_STR_YR}{e.g. 'NA'. See Details.}

\item{TIPNG_RCP_HIST_WSTE_QTY_END_YR}{e.g. 'NA'. See Details.}

\item{HIST_WASTE_QUANTITY_EST_METHOD}{e.g. 'METHOD #1:  ASSUME ALL PRIOR YEAR’S WASTE DISPOSAL QUANTITIES ARE THE SAME AS THE WASTE QUANTITY IN '. See Details.}

\item{HIST_WSTE_QTY_EST_MTHD_STRT_YR}{e.g. '1978'. See Details.}

\item{HIST_WSTE_QTY_EST_MTHD_END_YR}{e.g. '1997'. See Details.}

\item{HIST_WSTE_QTY_EST_MTHD_RSN_USE}{e.g. 'BEST FIT FOR AVAILABLE INFORMATION'. See Details.}

\item{LANDFILL_CAPACITY_IN_MTHD_3}{e.g. 'NA'. See Details.}

\item{WKG_CAP_HIST_WSTE_QTY_START_YR}{e.g. 'NA'. See Details.}

\item{WKG_CAP_HIST_WSTE_QTY_END_YR}{e.g. 'NA'. See Details.}

\item{IS_WORKING_CAPACITY_USED_IND}{e.g. 'N'. See Details.}

\item{YEAR_FOR_METHOD_2}{e.g. 'NA'. See Details.}

\item{EST_POP_SERVD_BY_LANDFL_MTHD_2}{e.g. 'NA'. See Details.}
}
\description{
Retrieve hh hist waste qty method data from ghg database
}

