% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_hh_hist_yr_waste_qty.R
\name{ghg_hh_hist_yr_waste_qty}
\alias{ghg_hh_hist_yr_waste_qty}
\title{Retrieve hh hist yr waste qty data from ghg database}
\usage{
ghg_hh_hist_yr_waste_qty(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, HIST_WASTE_DISPOSAL_YEAR = NULL,
  HISTYR_ANN_WSTE_DISPOSAL_QTY = NULL, HISTYR_ANN_WSTE_QTY_BY_MTHD = NULL,
  HISTYR_ANN_WSTE_QTY_MTHD_USED = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000010'. See Details.}

\item{REPORTING_YEAR}{e.g. '2013'. See Details.}

\item{FACILITY_NAME}{e.g. 'SANTA ROSA CENTRAL LANDFILL'. See Details.}

\item{HIST_WASTE_DISPOSAL_YEAR}{e.g. '1997'. See Details.}

\item{HISTYR_ANN_WSTE_DISPOSAL_QTY}{e.g. '46586'. See Details.}

\item{HISTYR_ANN_WSTE_QTY_BY_MTHD}{e.g. 'NA'. See Details.}

\item{HISTYR_ANN_WSTE_QTY_MTHD_USED}{e.g. 'USED WORKING CAPACITY FOR EACH VEHICLE/CONTAINER'. See Details.}
}
\description{
Retrieve hh hist yr waste qty data from ghg database
}

