% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_hh_hist_yr_waste_qty_detl.R
\name{ghg_hh_hist_yr_waste_qty_detl}
\alias{ghg_hh_hist_yr_waste_qty_detl}
\title{Retrieve hh hist yr waste qty detl data from ghg database}
\usage{
ghg_hh_hist_yr_waste_qty_detl(FACILITY_ID = NULL,
  REPORTING_YEAR = NULL, HISTYR_PERCENT_BY_WEIGHT = NULL,
  FACILITY_NAME = NULL, HIST_WASTE_DISPOSAL_YEAR = NULL,
  HISTYR_DECAY_RATE = NULL, HISTYR_FRACTION_OF_DOC_DISS = NULL,
  HISTYR_DOC_VAL = NULL, HISTYR_IS_MCF_VALUE_DEFAULT_IN = NULL,
  HISTYR_METHANE_FRACT_DET_MTHD = NULL,
  HISTYR_ANN_WST_QTY_SUB_DATA_US = NULL,
  HISTYR_ANNUAL_MCF_VALUE = NULL, HISTYR_METHANE_FRACTION = NULL,
  HISTYR_ANN_WST_QTY_NUM_DAY_SUB = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000010'. See Details.}

\item{REPORTING_YEAR}{e.g. '2013'. See Details.}

\item{HISTYR_PERCENT_BY_WEIGHT}{e.g. '1'. See Details.}

\item{FACILITY_NAME}{e.g. 'SANTA ROSA CENTRAL LANDFILL'. See Details.}

\item{HIST_WASTE_DISPOSAL_YEAR}{e.g. '2009'. See Details.}

\item{HISTYR_DECAY_RATE}{e.g. '0.057'. See Details.}

\item{HISTYR_FRACTION_OF_DOC_DISS}{e.g. '0.5'. See Details.}

\item{HISTYR_DOC_VAL}{e.g. '0.2'. See Details.}

\item{HISTYR_IS_MCF_VALUE_DEFAULT_IN}{e.g. 'NA'. See Details.}

\item{HISTYR_METHANE_FRACT_DET_MTHD}{e.g. 'NA'. See Details.}

\item{HISTYR_ANN_WST_QTY_SUB_DATA_US}{e.g. 'N'. See Details.}

\item{HISTYR_ANNUAL_MCF_VALUE}{e.g. 'NA'. See Details.}

\item{HISTYR_METHANE_FRACTION}{e.g. 'NA'. See Details.}

\item{HISTYR_ANN_WST_QTY_NUM_DAY_SUB}{e.g. 'NA'. See Details.}
}
\description{
Retrieve hh hist yr waste qty detl data from ghg database
}
