% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_hh_landfill_info.R
\name{ghg_hh_landfill_info}
\alias{ghg_hh_landfill_info}
\title{Retrieve hh landfill info data from ghg database}
\usage{
ghg_hh_landfill_info(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  ESTIMATED_YR_OF_LNDFIL_CLOSURE = NULL,
  FIRST_YR_LNDFIL_ACCEPTED_WASTE = NULL,
  LAST_YR_LNDFIL_ACCEPTED_WASTE = NULL,
  LEACHATE_RECIRCULATION_FREQ = NULL, IS_LANDFILL_OPEN = NULL,
  DOES_LNDFIL_HAVE_GAS_CLCT = NULL, LEACHATE_RECIRCULATION_USED = NULL,
  PASSIVE_VENTS_OR_FLARES = NULL, LANDFILL_CAPACITY = NULL,
  SCALES_PRESENT = NULL, LNDFIL_SURFACE_CONTAINING_WSTE = NULL,
  ANNUAL_MODELED_CH4_GENERATION = NULL,
  FIRST_YEAR_EMISSIONS_REPORTING = NULL, FACILITY_NAME = NULL,
  ANNUAL_MCF_VALUE = NULL, IS_MCF_VALUE_DEFAULT_INDICATOR = NULL,
  METHANE_FRAC_DETER_ANNUAL_VALU = NULL,
  METHANE_FRACTION_DETER_METHOD = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000010'. See Details.}

\item{REPORTING_YEAR}{e.g. '2014'. See Details.}

\item{ESTIMATED_YR_OF_LNDFIL_CLOSURE}{e.g. '2034'. See Details.}

\item{FIRST_YR_LNDFIL_ACCEPTED_WASTE}{e.g. '1978'. See Details.}

\item{LAST_YR_LNDFIL_ACCEPTED_WASTE}{e.g. 'NA'. See Details.}

\item{LEACHATE_RECIRCULATION_FREQ}{e.g. 'NOT USED FOR THE PAST 10 YEARS'. See Details.}

\item{IS_LANDFILL_OPEN}{e.g. 'Y'. See Details.}

\item{DOES_LNDFIL_HAVE_GAS_CLCT}{e.g. 'Y'. See Details.}

\item{LEACHATE_RECIRCULATION_USED}{e.g. 'N'. See Details.}

\item{PASSIVE_VENTS_OR_FLARES}{e.g. 'Y'. See Details.}

\item{LANDFILL_CAPACITY}{e.g. '3169311'. See Details.}

\item{SCALES_PRESENT}{e.g. 'Y'. See Details.}

\item{LNDFIL_SURFACE_CONTAINING_WSTE}{e.g. '314445'. See Details.}

\item{ANNUAL_MODELED_CH4_GENERATION}{e.g. '5141.68'. See Details.}

\item{FIRST_YEAR_EMISSIONS_REPORTING}{e.g. '2010'. See Details.}

\item{FACILITY_NAME}{e.g. 'SANTA ROSA CENTRAL LANDFILL'. See Details.}

\item{ANNUAL_MCF_VALUE}{e.g. '1'. See Details.}

\item{IS_MCF_VALUE_DEFAULT_INDICATOR}{e.g. 'N'. See Details.}

\item{METHANE_FRAC_DETER_ANNUAL_VALU}{e.g. '0.5'. See Details.}

\item{METHANE_FRACTION_DETER_METHOD}{e.g. 'DEFAULT'. See Details.}
}
\description{
Retrieve hh landfill info data from ghg database
}
