% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_hh_subpart_level_information.R
\name{ghg_hh_subpart_level_information}
\alias{ghg_hh_subpart_level_information}
\title{Retrieve hh subpart level information data from ghg database}
\usage{
ghg_hh_subpart_level_information(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, GHG_NAME = NULL, GHG_QUANTITY = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000010'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{FACILITY_NAME}{e.g. 'SANTA ROSA CENTRAL LANDFILL'. See Details.}

\item{GHG_NAME}{e.g. 'METHANE'. See Details.}

\item{GHG_QUANTITY}{e.g. '3925.29'. See Details.}
}
\description{
Retrieve hh subpart level information data from ghg database
}

