% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_hh_waste_qty_details.R
\name{ghg_hh_waste_qty_details}
\alias{ghg_hh_waste_qty_details}
\title{Retrieve hh waste qty details data from ghg database}
\usage{
ghg_hh_waste_qty_details(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  DECAY_RATE = NULL, METHANE_FRACT_ANNL_VAL = NULL,
  FRACTION_OF_DOC_DISSIMILATED = NULL, FACILITY_NAME = NULL,
  YEAR_WASTE_DISPOSED = NULL, ANNL_WST_QTY_SUB_DATA_USE = NULL,
  ANNL_WST_QTY_NUM_DAYS_SUB = NULL, METHANE_FRACT_DETERM_MTHD = NULL,
  IS_MCF_VALUE_DEFAULT_IND = NULL, ANNUAL_MCF_VALUE = NULL,
  PERCENT_BY_WEIGHT = NULL, DEGRADABLE_ORGANIC_CARBON_VAL = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000010'. See Details.}

\item{REPORTING_YEAR}{e.g. '2010'. See Details.}

\item{DECAY_RATE}{e.g. '0.05'. See Details.}

\item{METHANE_FRACT_ANNL_VAL}{e.g. '0.5'. See Details.}

\item{FRACTION_OF_DOC_DISSIMILATED}{e.g. '0.5'. See Details.}

\item{FACILITY_NAME}{e.g. 'SANTA ROSA CENTRAL LANDFILL'. See Details.}

\item{YEAR_WASTE_DISPOSED}{e.g. '2012'. See Details.}

\item{ANNL_WST_QTY_SUB_DATA_USE}{e.g. 'N'. See Details.}

\item{ANNL_WST_QTY_NUM_DAYS_SUB}{e.g. 'NA'. See Details.}

\item{METHANE_FRACT_DETERM_MTHD}{e.g. 'DEFAULT'. See Details.}

\item{IS_MCF_VALUE_DEFAULT_IND}{e.g. 'N'. See Details.}

\item{ANNUAL_MCF_VALUE}{e.g. '1'. See Details.}

\item{PERCENT_BY_WEIGHT}{e.g. '0.79'. See Details.}

\item{DEGRADABLE_ORGANIC_CARBON_VAL}{e.g. '0.31'. See Details.}
}
\description{
Retrieve hh waste qty details data from ghg database
}
