% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_i_abatement_info.R
\name{ghg_i_abatement_info}
\alias{ghg_i_abatement_info}
\title{Retrieve i abatement info data from ghg database}
\usage{
ghg_i_abatement_info(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  ABATEMENT_SYSTEMS_IN_USE = NULL, FACILITY_NAME = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000354'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{ABATEMENT_SYSTEMS_IN_USE}{e.g. 'YES'. See Details.}

\item{FACILITY_NAME}{e.g. 'CYPRESS SEMICONDUCTOR MINNESOTA INC'. See Details.}
}
\description{
Retrieve i abatement info data from ghg database
}

