% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_i_facility_details.R
\name{ghg_i_facility_details}
\alias{ghg_i_facility_details}
\title{Retrieve i facility details data from ghg database}
\usage{
ghg_i_facility_details(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FAC_MAN_GREA_THAN_300ML_WAFERS = NULL,
  FAC_MANUFACTURE_300ML_WAFERS = NULL, FAC_MANUFACTURE_200ML_WAFERS = NULL,
  FAC_MANUFACTURE_150ML_WAFERS = NULL, FAC_MANUFACTURE_MEMS_PVS_LCDS = NULL,
  FAC_MANUFACTURE_SEMICONDUCTORS = NULL, FACILITY_NAME = NULL,
  APPORTIONING_MODEL_COMPARISON = NULL, VERIFICATION_START_DATE = NULL,
  VERIFICATION_END_DATE = NULL, CERT_OF_GASES_SELECTED = NULL,
  FAC_MAN_LESS_THAN_150ML_WAFERS = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000354'. See Details.}

\item{REPORTING_YEAR}{e.g. '2012'. See Details.}

\item{FAC_MAN_GREA_THAN_300ML_WAFERS}{e.g. 'NO'. See Details.}

\item{FAC_MANUFACTURE_300ML_WAFERS}{e.g. 'NO'. See Details.}

\item{FAC_MANUFACTURE_200ML_WAFERS}{e.g. 'YES'. See Details.}

\item{FAC_MANUFACTURE_150ML_WAFERS}{e.g. 'NO'. See Details.}

\item{FAC_MANUFACTURE_MEMS_PVS_LCDS}{e.g. 'NO'. See Details.}

\item{FAC_MANUFACTURE_SEMICONDUCTORS}{e.g. 'YES'. See Details.}

\item{FACILITY_NAME}{e.g. 'CYPRESS SEMICONDUCTOR MINNESOTA INC'. See Details.}

\item{APPORTIONING_MODEL_COMPARISON}{e.g. '0.2'. See Details.}

\item{VERIFICATION_START_DATE}{e.g. '01-JAN-12'. See Details.}

\item{VERIFICATION_END_DATE}{e.g. '31-DEC-12'. See Details.}

\item{CERT_OF_GASES_SELECTED}{e.g. 'CERTIFIED'. See Details.}

\item{FAC_MAN_LESS_THAN_150ML_WAFERS}{e.g. 'NA'. See Details.}
}
\description{
Retrieve i facility details data from ghg database
}

