% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_i_fhtf_info.R
\name{ghg_i_fhtf_info}
\alias{ghg_i_fhtf_info}
\title{Retrieve i fhtf info data from ghg database}
\usage{
ghg_i_fhtf_info(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FHTF_EMISSION_CALC_METHOD = NULL, ANNUAL_FHTF_EMISSIONS = NULL,
  OTHER_FHTF_CAS_NUMBER = NULL, OTHER_FHTF_NAME = NULL,
  FACILITY_NAME = NULL, FHTF_NAME = NULL, FHTF_CAS_NUMBER = NULL,
  OTHER_FHTF_FORMULA = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000354'. See Details.}

\item{REPORTING_YEAR}{e.g. '2013'. See Details.}

\item{FHTF_EMISSION_CALC_METHOD}{e.g. 'EQUATION I-16 (MASS BALANCE)'. See Details.}

\item{ANNUAL_FHTF_EMISSIONS}{e.g. '0.4547'. See Details.}

\item{OTHER_FHTF_CAS_NUMBER}{e.g. '163702-06-5, 163702-05-4'. See Details.}

\item{OTHER_FHTF_NAME}{e.g. 'HFE-7200'. See Details.}

\item{FACILITY_NAME}{e.g. 'CYPRESS SEMICONDUCTOR MINNESOTA INC'. See Details.}

\item{FHTF_NAME}{e.g. 'OTHER F-HTF (SPECIFY)'. See Details.}

\item{FHTF_CAS_NUMBER}{e.g. '163702-06-5, 163702-05-4'. See Details.}

\item{OTHER_FHTF_FORMULA}{e.g. 'C4F9OC2H5'. See Details.}
}
\description{
Retrieve i fhtf info data from ghg database
}
