% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_i_pv_mems_lcd_info.R
\name{ghg_i_pv_mems_lcd_info}
\alias{ghg_i_pv_mems_lcd_info}
\title{Retrieve i pv mems lcd info data from ghg database}
\usage{
ghg_i_pv_mems_lcd_info(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  OTHER_FGHG_CAS_NUMBER = NULL, OTHER_FGHG_FORMULA = NULL,
  PROCESS_TYPE = NULL, ANNUAL_FGHG_EMISSIONS = NULL,
  FACILITY_NAME = NULL, FGHG_NAME = NULL, FGHG_CAS_NUMBER = NULL,
  OTHER_FGHG_NAME = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1001199'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{OTHER_FGHG_CAS_NUMBER}{e.g. 'NA'. See Details.}

\item{OTHER_FGHG_FORMULA}{e.g. 'NA'. See Details.}

\item{PROCESS_TYPE}{e.g. 'PLASMA ETCHING'. See Details.}

\item{ANNUAL_FGHG_EMISSIONS}{e.g. '0.00034'. See Details.}

\item{FACILITY_NAME}{e.g. 'EASTMAN KODAK - KODAK PARK'. See Details.}

\item{FGHG_NAME}{e.g. 'HFC-23'. See Details.}

\item{FGHG_CAS_NUMBER}{e.g. '75-46-7'. See Details.}

\item{OTHER_FGHG_NAME}{e.g. 'NA'. See Details.}
}
\description{
Retrieve i pv mems lcd info data from ghg database
}
