% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_ii_biogas_rec_proc.R
\name{ghg_ii_biogas_rec_proc}
\alias{ghg_ii_biogas_rec_proc}
\title{Retrieve ii biogas rec proc data from ghg database}
\usage{
ghg_ii_biogas_rec_proc(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, ANAEROBIC_PROCESS_ID = NULL,
  IS_CH4_CONC_MONITORED_WEEKLY = NULL, WEEK = NULL,
  IS_MISS_DATA_USED_BIOGAS_FLOW = NULL,
  IS_MISS_DATA_USED_METHANE_CONC = NULL, WEEK_AVG_BIOGAS_TEMPERATURE = NULL,
  IS_BIOGAS_TEMPERATURE_INCORP = NULL, WEEK_AVG_BIOGAS_PRESSURE = NULL,
  IS_BIOGAS_PRESSURE_INCORP = NULL, IS_MOISTURE_CONTENT_INCORP = NULL,
  WEEK_AVG_BIOGAS_MOISTURE_CONT = NULL, IS_BIOGAS_FLOW_WET_DRY_BASIS = NULL,
  IS_BIOGAS_METHAN_WETDRY_BASIS = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000035'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{FACILITY_NAME}{e.g. 'ANDERSONS MARATHON ETHANOL LLC'. See Details.}

\item{ANAEROBIC_PROCESS_ID}{e.g. 'Greenville Methanator System'. See Details.}

\item{IS_CH4_CONC_MONITORED_WEEKLY}{e.g. 'Yes'. See Details.}

\item{WEEK}{e.g. '1'. See Details.}

\item{IS_MISS_DATA_USED_BIOGAS_FLOW}{e.g. 'No'. See Details.}

\item{IS_MISS_DATA_USED_METHANE_CONC}{e.g. 'Yes'. See Details.}

\item{WEEK_AVG_BIOGAS_TEMPERATURE}{e.g. 'NA'. See Details.}

\item{IS_BIOGAS_TEMPERATURE_INCORP}{e.g. 'Yes'. See Details.}

\item{WEEK_AVG_BIOGAS_PRESSURE}{e.g. 'NA'. See Details.}

\item{IS_BIOGAS_PRESSURE_INCORP}{e.g. 'Yes'. See Details.}

\item{IS_MOISTURE_CONTENT_INCORP}{e.g. 'Yes'. See Details.}

\item{WEEK_AVG_BIOGAS_MOISTURE_CONT}{e.g. 'NA'. See Details.}

\item{IS_BIOGAS_FLOW_WET_DRY_BASIS}{e.g. 'NA'. See Details.}

\item{IS_BIOGAS_METHAN_WETDRY_BASIS}{e.g. 'NA'. See Details.}
}
\description{
Retrieve ii biogas rec proc data from ghg database
}

