% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_ii_ch4_gen_process.R
\name{ghg_ii_ch4_gen_process}
\alias{ghg_ii_ch4_gen_process}
\title{Retrieve ii ch4 gen process data from ghg database}
\usage{
ghg_ii_ch4_gen_process(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, ANAEROBIC_PROCESS_ID = NULL, WEEK = NULL,
  DOES_FAC_MEASURE_COD_BOD_CONC = NULL, WEEKLY_AVG_COD_BOD_CONC = NULL,
  IS_MISS_DATA_USED_COD_BOD_CONC = NULL,
  WEEKLY_VOLUME_OF_WASTEWATER = NULL,
  METHANE_PRODUCTION_POTENTIAL = NULL,
  IS_MISS_DATA_USED_VOL_WSTWATER = NULL,
  METHANE_CONVERSION_FACTOR = NULL,
  ANNUAL_MASS_METHANE_GEN_BOD5 = NULL,
  ANNUAL_MASS_METHANE_GEN_COD = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000035'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{FACILITY_NAME}{e.g. 'ANDERSONS MARATHON ETHANOL LLC'. See Details.}

\item{ANAEROBIC_PROCESS_ID}{e.g. 'Greenville Methanator System'. See Details.}

\item{WEEK}{e.g. '1'. See Details.}

\item{DOES_FAC_MEASURE_COD_BOD_CONC}{e.g. 'COD'. See Details.}

\item{WEEKLY_AVG_COD_BOD_CONC}{e.g. '3.67'. See Details.}

\item{IS_MISS_DATA_USED_COD_BOD_CONC}{e.g. 'No'. See Details.}

\item{WEEKLY_VOLUME_OF_WASTEWATER}{e.g. '3038.65'. See Details.}

\item{METHANE_PRODUCTION_POTENTIAL}{e.g. '0.25'. See Details.}

\item{IS_MISS_DATA_USED_VOL_WSTWATER}{e.g. 'NA'. See Details.}

\item{METHANE_CONVERSION_FACTOR}{e.g. '0.8'. See Details.}

\item{ANNUAL_MASS_METHANE_GEN_BOD5}{e.g. 'NA'. See Details.}

\item{ANNUAL_MASS_METHANE_GEN_COD}{e.g. '2.2303691'. See Details.}
}
\description{
Retrieve ii ch4 gen process data from ghg database
}
