% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_ii_equ_ii1_or_ii2.R
\name{ghg_ii_equ_ii1_or_ii2}
\alias{ghg_ii_equ_ii1_or_ii2}
\title{Retrieve ii equ ii1 or ii2 data from ghg database}
\usage{
ghg_ii_equ_ii1_or_ii2(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, ANAEROBIC_PROCESS_ID = NULL,
  ANNUAL_MASS_METHANE_GEN_RPTVAL = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000035'. See Details.}

\item{REPORTING_YEAR}{e.g. '2012'. See Details.}

\item{FACILITY_NAME}{e.g. 'ANDERSONS MARATHON ETHANOL LLC'. See Details.}

\item{ANAEROBIC_PROCESS_ID}{e.g. 'Greenville Methanator System'. See Details.}

\item{ANNUAL_MASS_METHANE_GEN_RPTVAL}{e.g. '190.7'. See Details.}
}
\description{
Retrieve ii equ ii1 or ii2 data from ghg database
}
