% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_ii_equ_ii4_input.R
\name{ghg_ii_equ_ii4_input}
\alias{ghg_ii_equ_ii4_input}
\title{Retrieve ii equ ii4 input data from ghg database}
\usage{
ghg_ii_equ_ii4_input(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, ANAEROBIC_PROCESS_ID = NULL,
  ANNUAL_QTY_METHANE_RECOV_RPT = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000035'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{FACILITY_NAME}{e.g. 'ANDERSONS MARATHON ETHANOL LLC'. See Details.}

\item{ANAEROBIC_PROCESS_ID}{e.g. 'Greenville Methanator System'. See Details.}

\item{ANNUAL_QTY_METHANE_RECOV_RPT}{e.g. '256.63'. See Details.}
}
\description{
Retrieve ii equ ii4 input data from ghg database
}
