% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_ii_equation_ii3.R
\name{ghg_ii_equation_ii3}
\alias{ghg_ii_equation_ii3}
\title{Retrieve ii equation ii3 data from ghg database}
\usage{
ghg_ii_equation_ii3(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, ANAEROBIC_PROCESS_ID = NULL,
  ANNUAL_MASS_METHANE_EMISSIONS = NULL,
  ANNUAL_MASS_METHANE_GENERATED = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000038'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{FACILITY_NAME}{e.g. 'TYSON FRESH MEATS'. See Details.}

\item{ANAEROBIC_PROCESS_ID}{e.g. 'Uncovered Anaerobic Lagoons'. See Details.}

\item{ANNUAL_MASS_METHANE_EMISSIONS}{e.g. '1488.82'. See Details.}

\item{ANNUAL_MASS_METHANE_GENERATED}{e.g. '1488.82'. See Details.}
}
\description{
Retrieve ii equation ii3 data from ghg database
}

