% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_ii_equation_ii6.R
\name{ghg_ii_equation_ii6}
\alias{ghg_ii_equation_ii6}
\title{Retrieve ii equation ii6 data from ghg database}
\usage{
ghg_ii_equation_ii6(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, ANAEROBIC_PROCESS_ID = NULL,
  ANAEROBIC_PROCESS_LEAKAGE = NULL,
  ANNUAL_MASS_METHANE_EMISSIONS = NULL,
  ANNUAL_METHANE_RECOVERED = NULL,
  ANNUALOPHOURSBACKUPDESTRDEVICE = NULL,
  ANNUAL_OP_HOURS_PRIMARY_DEVICE = NULL,
  FRACT_OP_HOURS_PRIMARY_DEVICE = NULL,
  FRACT_OP_HOURS_SECOND_DEVICE = NULL,
  IS_BACKUP_DESTR_DEVICE_PRESENT = NULL,
  METHANE_COLLECTION_EFFICIENCY = NULL,
  METHANE_DESTRUCTION_LOCATION = NULL,
  PRIMARY_DEVICE_DESTRUCTION_EFF = NULL,
  SECOND_DEVICE_DESTRUCTION_EFF = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000035'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{FACILITY_NAME}{e.g. 'ANDERSONS MARATHON ETHANOL LLC'. See Details.}

\item{ANAEROBIC_PROCESS_ID}{e.g. 'Greenville Methanator System'. See Details.}

\item{ANAEROBIC_PROCESS_LEAKAGE}{e.g. '2.5922222'. See Details.}

\item{ANNUAL_MASS_METHANE_EMISSIONS}{e.g. '11.74'. See Details.}

\item{ANNUAL_METHANE_RECOVERED}{e.g. '256.63'. See Details.}

\item{ANNUALOPHOURSBACKUPDESTRDEVICE}{e.g. '35'. See Details.}

\item{ANNUAL_OP_HOURS_PRIMARY_DEVICE}{e.g. '8585'. See Details.}

\item{FRACT_OP_HOURS_PRIMARY_DEVICE}{e.g. '0.9800228'. See Details.}

\item{FRACT_OP_HOURS_SECOND_DEVICE}{e.g. '0.0039954'. See Details.}

\item{IS_BACKUP_DESTR_DEVICE_PRESENT}{e.g. 'Yes'. See Details.}

\item{METHANE_COLLECTION_EFFICIENCY}{e.g. '0.99'. See Details.}

\item{METHANE_DESTRUCTION_LOCATION}{e.g. 'On-Site'. See Details.}

\item{PRIMARY_DEVICE_DESTRUCTION_EFF}{e.g. '0.98'. See Details.}

\item{SECOND_DEVICE_DESTRUCTION_EFF}{e.g. '0.98'. See Details.}
}
\description{
Retrieve ii equation ii6 data from ghg database
}
