% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_ii_equation_ii7.R
\name{ghg_ii_equation_ii7}
\alias{ghg_ii_equation_ii7}
\title{Retrieve ii equation ii7 data from ghg database}
\usage{
ghg_ii_equation_ii7(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, ANAEROBIC_PROCESS_ID = NULL,
  ANNUAL_MASS_METHANE_EMITTED = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000035'. See Details.}

\item{REPORTING_YEAR}{e.g. '2012'. See Details.}

\item{FACILITY_NAME}{e.g. 'ANDERSONS MARATHON ETHANOL LLC'. See Details.}

\item{ANAEROBIC_PROCESS_ID}{e.g. 'The Andersons Marathon Ethanol - Greenville'. See Details.}

\item{ANNUAL_MASS_METHANE_EMITTED}{e.g. '8.93'. See Details.}
}
\description{
Retrieve ii equation ii7 data from ghg database
}
