% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_ii_process_details.R
\name{ghg_ii_process_details}
\alias{ghg_ii_process_details}
\title{Retrieve ii process details data from ghg database}
\usage{
ghg_ii_process_details(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, ANAEROBIC_PROCESS_ID = NULL,
  IS_BIOGAS_RECOVERED = NULL, DOES_FAC_MEASURE_COD_BOD_CONC = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000035'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{FACILITY_NAME}{e.g. 'ANDERSONS MARATHON ETHANOL LLC'. See Details.}

\item{ANAEROBIC_PROCESS_ID}{e.g. 'Greenville Methanator System'. See Details.}

\item{IS_BIOGAS_RECOVERED}{e.g. 'Yes'. See Details.}

\item{DOES_FAC_MEASURE_COD_BOD_CONC}{e.g. 'COD'. See Details.}
}
\description{
Retrieve ii process details data from ghg database
}
