% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_k_info.R
\name{ghg_k_info}
\alias{ghg_k_info}
\title{Retrieve k info data from ghg database}
\usage{
ghg_k_info(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  TOTALNUMELECTARCFURNAC = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1002969'. See Details.}

\item{REPORTING_YEAR}{e.g. '2010'. See Details.}

\item{TOTALNUMELECTARCFURNAC}{e.g. '2'. See Details.}
}
\description{
Retrieve k info data from ghg database
}

