% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_k_non_cems_source_info.R
\name{ghg_k_non_cems_source_info}
\alias{ghg_k_non_cems_source_info}
\title{Retrieve k non cems source info data from ghg database}
\usage{
ghg_k_non_cems_source_info(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  UNIT_NAME = NULL, CH4_EMISSIONS = NULL, CO2_EMISSIONS = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1002969'. See Details.}

\item{REPORTING_YEAR}{e.g. '2010'. See Details.}

\item{UNIT_NAME}{e.g. 'Furnace #6'. See Details.}

\item{CH4_EMISSIONS}{e.g. '11.13'. See Details.}

\item{CO2_EMISSIONS}{e.g. '52087'. See Details.}
}
\description{
Retrieve k non cems source info data from ghg database
}

