% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_mm_subpart_level_information.R
\name{ghg_mm_subpart_level_information}
\alias{ghg_mm_subpart_level_information}
\title{Retrieve mm subpart level information data from ghg database}
\usage{
ghg_mm_subpart_level_information(FACILITY_ID = NULL,
  REPORTING_YEAR = NULL, FACILITY_NAME = NULL, CO2_QUANTITY = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000025'. See Details.}

\item{REPORTING_YEAR}{e.g. '2012'. See Details.}

\item{FACILITY_NAME}{e.g. 'PLACID REFINING CO LLC - PORT ALLEN REFINERY'. See Details.}

\item{CO2_QUANTITY}{e.g. '8542528.7'. See Details.}
}
\description{
Retrieve mm subpart level information data from ghg database
}
