% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_n_cems_details.R
\name{ghg_n_cems_details}
\alias{ghg_n_cems_details}
\title{Retrieve n cems details data from ghg database}
\usage{
ghg_n_cems_details(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  CML_NAME = NULL, CML_DESCRIPTION = NULL, CML_CONFIGURATION_TYPE = NULL,
  CEMS_BIOGENIC_CO2_EMISSIONS = NULL, CEMS_NON_BIOGENIC_CO2_EMISSION = NULL,
  CEMS_ANNUAL_CO2_EMISSIONS = NULL, CEMS_CH4_EMISSIONS = NULL,
  CEMS_N2O_EMISSIONS = NULL, CML_TOTAL_OPERATING_HOURS = NULL,
  TIER4_SUB_OPER_HRS_CO2_CONC = NULL, TIER4_SUB_OPER_HRS_GAS_FLOW_RT = NULL,
  TIER4_SUB_OPER_HRS_MOIST_CONT = NULL,
  CEMS_CO2_COMBUSTION_EMISSIONS = NULL, CEMS_METHODOLOGY_START_DATE = NULL,
  CEMS_METHODOLOGY_END_DATE = NULL, CEMS_FUEL = NULL,
  CEMS_QUARTER_NAME = NULL, CEMS_CO2_EMISSIONS_BY_QUARTER = NULL,
  CEMS_GLASS_FURNACE_NAME = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1004085'. See Details.}

\item{REPORTING_YEAR}{e.g. '2012'. See Details.}

\item{CML_NAME}{e.g. 'ES-67 Glass Melting Furnace Poll Control'. See Details.}

\item{CML_DESCRIPTION}{e.g. 'Idled in December 2008; but remains covered by Title V Air Permit'. See Details.}

\item{CML_CONFIGURATION_TYPE}{e.g. 'Single process/process unit exhausts to dedicated stack'. See Details.}

\item{CEMS_BIOGENIC_CO2_EMISSIONS}{e.g. '0'. See Details.}

\item{CEMS_NON_BIOGENIC_CO2_EMISSION}{e.g. '0'. See Details.}

\item{CEMS_ANNUAL_CO2_EMISSIONS}{e.g. '0'. See Details.}

\item{CEMS_CH4_EMISSIONS}{e.g. '0'. See Details.}

\item{CEMS_N2O_EMISSIONS}{e.g. '0'. See Details.}

\item{CML_TOTAL_OPERATING_HOURS}{e.g. '0'. See Details.}

\item{TIER4_SUB_OPER_HRS_CO2_CONC}{e.g. '0'. See Details.}

\item{TIER4_SUB_OPER_HRS_GAS_FLOW_RT}{e.g. '0'. See Details.}

\item{TIER4_SUB_OPER_HRS_MOIST_CONT}{e.g. '0'. See Details.}

\item{CEMS_CO2_COMBUSTION_EMISSIONS}{e.g. 'NA'. See Details.}

\item{CEMS_METHODOLOGY_START_DATE}{e.g. '2012-01-01'. See Details.}

\item{CEMS_METHODOLOGY_END_DATE}{e.g. '2012-12-31'. See Details.}

\item{CEMS_FUEL}{e.g. 'Natural Gas Fired with Propane Backup'. See Details.}

\item{CEMS_QUARTER_NAME}{e.g. 'Third Quarter'. See Details.}

\item{CEMS_CO2_EMISSIONS_BY_QUARTER}{e.g. '0'. See Details.}

\item{CEMS_GLASS_FURNACE_NAME}{e.g. 'ES67 Glass Melting Furnace #2'. See Details.}
}
\description{
Retrieve n cems details data from ghg database
}

