% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_n_cems_info.R
\name{ghg_n_cems_info}
\alias{ghg_n_cems_info}
\title{Retrieve n cems info data from ghg database}
\usage{
ghg_n_cems_info(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  CEMS_UNIT_NAME = NULL, CEMS_UNIT_TYPE = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1004085'. See Details.}

\item{REPORTING_YEAR}{e.g. '2010'. See Details.}

\item{CEMS_UNIT_NAME}{e.g. 'ES67 Glass Melting Furnace #2'. See Details.}

\item{CEMS_UNIT_TYPE}{e.g. 'Continuous Glass Melting Furnace'. See Details.}
}
\description{
Retrieve n cems info data from ghg database
}
