% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_n_info.R
\name{ghg_n_info}
\alias{ghg_n_info}
\title{Retrieve n info data from ghg database}
\usage{
ghg_n_info(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  TOTAL_NUMBER_OF_FURNACES = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000002'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{TOTAL_NUMBER_OF_FURNACES}{e.g. '2'. See Details.}
}
\description{
Retrieve n info data from ghg database
}

