% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_n_non_cems_source_info.R
\name{ghg_n_non_cems_source_info}
\alias{ghg_n_non_cems_source_info}
\title{Retrieve n non cems source info data from ghg database}
\usage{
ghg_n_non_cems_source_info(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  UNIT_TYPE = NULL, CO2_EMISSIONS = NULL,
  TIMES_MISS_DATA_PROCE_USED_RAW = NULL,
  TIMES_MISS_DATA_PRO_USED_FRAC = NULL,
  CALCINATION_FRACTION_METHOD = NULL,
  OTHR_CALC_FRAC_METHD_DESCRIP = NULL, MASS_FRACTION_TEST_DATE = NULL,
  MASS_FRACTION_TEST_METHOD = NULL, UNIT_NAME = NULL,
  OTHER_TEST_METHOD = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000002'. See Details.}

\item{REPORTING_YEAR}{e.g. '2010'. See Details.}

\item{UNIT_TYPE}{e.g. 'Continuous Glass Melting Furnace'. See Details.}

\item{CO2_EMISSIONS}{e.g. '11956.7'. See Details.}

\item{TIMES_MISS_DATA_PROCE_USED_RAW}{e.g. '0'. See Details.}

\item{TIMES_MISS_DATA_PRO_USED_FRAC}{e.g. '0'. See Details.}

\item{CALCINATION_FRACTION_METHOD}{e.g. 'Default value (1.0)'. See Details.}

\item{OTHR_CALC_FRAC_METHD_DESCRIP}{e.g. 'NA'. See Details.}

\item{MASS_FRACTION_TEST_DATE}{e.g. '2010-12-31'. See Details.}

\item{MASS_FRACTION_TEST_METHOD}{e.g. 'Default Method per 98.143(c)'. See Details.}

\item{UNIT_NAME}{e.g. 'Furnace #2'. See Details.}

\item{OTHER_TEST_METHOD}{e.g. 'NA'. See Details.}
}
\description{
Retrieve n non cems source info data from ghg database
}
