% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_nn_ldc_details.R
\name{ghg_nn_ldc_details}
\alias{ghg_nn_ldc_details}
\title{Retrieve nn ldc details data from ghg database}
\usage{
ghg_nn_ldc_details(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  GAS_DELIV_OTHR_LDCS_DWNS_PIPE = NULL, OTHINDSTRYSTDFORVOL = NULL,
  LDC_NN1_CO2 = NULL, LDC_NN2_CO2 = NULL, NN3_CO2 = NULL,
  NN4_CO2 = NULL, NN5_CO2 = NULL, NN3_EF_INDUSTRY_STNDRD = NULL,
  NN3_EF_OTHER_INDUSTRY_STNDRD = NULL, NN4_EF_INDUSTRY_STNDRD = NULL,
  NN4_EF_OTHER_INDUSTRY_STNDRD = NULL, NN5_EF_INDUSTRY_STNDRD = NULL,
  NN5_EF_OTHER_INDUSTRY_STNDRD = NULL,
  LDC_NN1_HHV_INDUSTRY_STNDRD = NULL,
  LDC_NN1_HHV_OTH_INDUSTRY_STD = NULL,
  LDC_NN1_EF_INDUSTRY_STNDRD = NULL,
  LDC_NN1_EF_OTH_INDUSTRY_STD = NULL,
  LDC_NN2_EF_INDUSTRY_STNDRD = NULL,
  LDC_NN2_EF_OTH_INDUSTRY_STD = NULL, FACILITY_NAME = NULL,
  CITY_GATE_VOLUME_INDUSTRY_STND = NULL, GAS_STORED = NULL,
  GASBP_CITYGATE_NO_DAYS_SUBDATA = NULL,
  LDC_NN5B_CO2_MASS_QUANTITY = NULL, LDC_NN5A_CO2_MASS_QUANTITY = NULL,
  ANNVOLGASRMVDFRMSTRG = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000008'. See Details.}

\item{REPORTING_YEAR}{e.g. '2010'. See Details.}

\item{GAS_DELIV_OTHR_LDCS_DWNS_PIPE}{e.g. '0'. See Details.}

\item{OTHINDSTRYSTDFORVOL}{e.g. 'NA'. See Details.}

\item{LDC_NN1_CO2}{e.g. 'NA'. See Details.}

\item{LDC_NN2_CO2}{e.g. '391676.1'. See Details.}

\item{NN3_CO2}{e.g. '0'. See Details.}

\item{NN4_CO2}{e.g. '86036'. See Details.}

\item{NN5_CO2}{e.g. '0'. See Details.}

\item{NN3_EF_INDUSTRY_STNDRD}{e.g. 'NA'. See Details.}

\item{NN3_EF_OTHER_INDUSTRY_STNDRD}{e.g. 'NA'. See Details.}

\item{NN4_EF_INDUSTRY_STNDRD}{e.g. 'NA'. See Details.}

\item{NN4_EF_OTHER_INDUSTRY_STNDRD}{e.g. 'NA'. See Details.}

\item{NN5_EF_INDUSTRY_STNDRD}{e.g. 'NA'. See Details.}

\item{NN5_EF_OTHER_INDUSTRY_STNDRD}{e.g. 'NA'. See Details.}

\item{LDC_NN1_HHV_INDUSTRY_STNDRD}{e.g. 'NA'. See Details.}

\item{LDC_NN1_HHV_OTH_INDUSTRY_STD}{e.g. 'NA'. See Details.}

\item{LDC_NN1_EF_INDUSTRY_STNDRD}{e.g. 'NA'. See Details.}

\item{LDC_NN1_EF_OTH_INDUSTRY_STD}{e.g. 'NA'. See Details.}

\item{LDC_NN2_EF_INDUSTRY_STNDRD}{e.g. 'NA'. See Details.}

\item{LDC_NN2_EF_OTH_INDUSTRY_STD}{e.g. 'NA'. See Details.}

\item{FACILITY_NAME}{e.g. 'St. Lawrence Gas Co., Inc.'. See Details.}

\item{CITY_GATE_VOLUME_INDUSTRY_STND}{e.g. 'Industry standard practices'. See Details.}

\item{GAS_STORED}{e.g. '0'. See Details.}

\item{GASBP_CITYGATE_NO_DAYS_SUBDATA}{e.g. 'NA'. See Details.}

\item{LDC_NN5B_CO2_MASS_QUANTITY}{e.g. 'NA'. See Details.}

\item{LDC_NN5A_CO2_MASS_QUANTITY}{e.g. 'NA'. See Details.}

\item{ANNVOLGASRMVDFRMSTRG}{e.g. 'NA'. See Details.}
}
\description{
Retrieve nn ldc details data from ghg database
}
