% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_nn_ldc_nat_gas_deliveries.R
\name{ghg_nn_ldc_nat_gas_deliveries}
\alias{ghg_nn_ldc_nat_gas_deliveries}
\title{Retrieve nn ldc nat gas deliveries data from ghg database}
\usage{
ghg_nn_ldc_nat_gas_deliveries(FACILITY_ID = NULL,
  REPORTING_YEAR = NULL, FACILITY_NAME = NULL,
  END_USER_CATEGORY = NULL, VOLUME_DELIVERED = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000008'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{FACILITY_NAME}{e.g. 'St. Lawrence Gas Co., Inc.'. See Details.}

\item{END_USER_CATEGORY}{e.g. 'Residential consumers'. See Details.}

\item{VOLUME_DELIVERED}{e.g. '1470567'. See Details.}
}
\description{
Retrieve nn ldc nat gas deliveries data from ghg database
}
