% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_p_cems_details.R
\name{ghg_p_cems_details}
\alias{ghg_p_cems_details}
\title{Retrieve p cems details data from ghg database}
\usage{
ghg_p_cems_details(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  CEMS_CH4_EMISSIONS = NULL, CEMS_N2O_EMISSIONS = NULL,
  CML_TOT_NUM_OF_OPR_HRS = NULL, TIER4_SUB_OPR_HRS_CO2_CONC = NULL,
  TIER_4_SUB_OPR_HRS_GAS_FLOW_RT = NULL,
  TIER4_SUB_OPR_HRS_MOIST_CONT = NULL, CEMS_CO2_COMBUSTION_EMISSIONS = NULL,
  CEMS_METHODOLOGY_START_DATE = NULL, CEMS_METHODOLOGY_END_DATE = NULL,
  CEMS_FUEL = NULL, CEMS_QUARTER_NAME = NULL,
  CEMS_CO2_EMISSIONS_BY_QUARTER = NULL, CML_NAME = NULL,
  CML_DESCRIPTION = NULL, TIER_4_UNIT = NULL,
  CML_CONFIGURATION_TYPE = NULL, CEMS_BIOGENIC_CO2_EMISSIONS = NULL,
  CEMS_NON_BIOGENIC_CO2_EMISS = NULL, CEMS_ANNUAL_CO2_EMISSIONS = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1002122'. See Details.}

\item{REPORTING_YEAR}{e.g. '2014'. See Details.}

\item{CEMS_CH4_EMISSIONS}{e.g. '0.32'. See Details.}

\item{CEMS_N2O_EMISSIONS}{e.g. '0.032'. See Details.}

\item{CML_TOT_NUM_OF_OPR_HRS}{e.g. '8744'. See Details.}

\item{TIER4_SUB_OPR_HRS_CO2_CONC}{e.g. '194'. See Details.}

\item{TIER_4_SUB_OPR_HRS_GAS_FLOW_RT}{e.g. '194'. See Details.}

\item{TIER4_SUB_OPR_HRS_MOIST_CONT}{e.g. '0'. See Details.}

\item{CEMS_CO2_COMBUSTION_EMISSIONS}{e.g. 'NA'. See Details.}

\item{CEMS_METHODOLOGY_START_DATE}{e.g. '01-JAN-14'. See Details.}

\item{CEMS_METHODOLOGY_END_DATE}{e.g. '31-DEC-14'. See Details.}

\item{CEMS_FUEL}{e.g. 'Process/stationary combustion units share a common stack was chosen to allow for the reporting of CH'. See Details.}

\item{CEMS_QUARTER_NAME}{e.g. 'Fourth Quarter'. See Details.}

\item{CEMS_CO2_EMISSIONS_BY_QUARTER}{e.g. '50312.8'. See Details.}

\item{CML_NAME}{e.g. 'APCI Tesoro Martinez CEM'. See Details.}

\item{CML_DESCRIPTION}{e.g. 'NA'. See Details.}

\item{TIER_4_UNIT}{e.g. 'APCI Tesoro Martinez'. See Details.}

\item{CML_CONFIGURATION_TYPE}{e.g. 'Process/stationary combustion units share common stack'. See Details.}

\item{CEMS_BIOGENIC_CO2_EMISSIONS}{e.g. '0'. See Details.}

\item{CEMS_NON_BIOGENIC_CO2_EMISS}{e.g. '255049.1'. See Details.}

\item{CEMS_ANNUAL_CO2_EMISSIONS}{e.g. '255049.1'. See Details.}
}
\description{
Retrieve p cems details data from ghg database
}

