% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_p_unit_info.R
\name{ghg_p_unit_info}
\alias{ghg_p_unit_info}
\title{Retrieve p unit info data from ghg database}
\usage{
ghg_p_unit_info(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  UNIT_NAME = NULL, UNIT_TYPE = NULL, CO2_EMISSIONS = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000043'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{UNIT_NAME}{e.g. 'Hydrogen Unit 4'. See Details.}

\item{UNIT_TYPE}{e.g. 'Hydrogen production process unit'. See Details.}

\item{CO2_EMISSIONS}{e.g. '467777.78'. See Details.}
}
\description{
Retrieve p unit info data from ghg database
}
