% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_p_unit_info_cems.R
\name{ghg_p_unit_info_cems}
\alias{ghg_p_unit_info_cems}
\title{Retrieve p unit info cems data from ghg database}
\usage{
ghg_p_unit_info_cems(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  CEMS_UNIT_NAME = NULL, CEMS_UNIT_TYPE = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1002122'. See Details.}

\item{REPORTING_YEAR}{e.g. '2010'. See Details.}

\item{CEMS_UNIT_NAME}{e.g. 'APCI Tesoro Martinez'. See Details.}

\item{CEMS_UNIT_TYPE}{e.g. 'Hydrogen production process unit'. See Details.}
}
\description{
Retrieve p unit info cems data from ghg database
}
