% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_pp_calc_method_data_quality.R
\name{ghg_pp_calc_method_data_quality}
\alias{ghg_pp_calc_method_data_quality}
\title{Retrieve pp calc method data quality data from ghg database}
\usage{
ghg_pp_calc_method_data_quality(FACILITY_ID = NULL,
  REPORTING_YEAR = NULL, FACILITY_NAME = NULL,
  CALCULATION_METHODOLOGY = NULL,
  SUBSTITUTE_MASSVOLUME_DATA_USE = NULL,
  SUB_CONCENTRATION_DATA_USE = NULL,
  SUBSTITUTE_DENSITY_DATA_USE = NULL, IMPORTER_EXPORTER_TYPE = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000035'. See Details.}

\item{REPORTING_YEAR}{e.g. '2013'. See Details.}

\item{FACILITY_NAME}{e.g. 'ANDERSONS MARATHON ETHANOL LLC'. See Details.}

\item{CALCULATION_METHODOLOGY}{e.g. 'Flow Meters 1'. See Details.}

\item{SUBSTITUTE_MASSVOLUME_DATA_USE}{e.g. 'NA'. See Details.}

\item{SUB_CONCENTRATION_DATA_USE}{e.g. 'NA'. See Details.}

\item{SUBSTITUTE_DENSITY_DATA_USE}{e.g. 'NA'. See Details.}

\item{IMPORTER_EXPORTER_TYPE}{e.g. 'NA'. See Details.}
}
\description{
Retrieve pp calc method data quality data from ghg database
}
