% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_pp_co2_end_use_details.R
\name{ghg_pp_co2_end_use_details}
\alias{ghg_pp_co2_end_use_details}
\title{Retrieve pp co2 end use details data from ghg database}
\usage{
ghg_pp_co2_end_use_details(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, END_USE_FOR_CO2 = NULL,
  END_USE_APP_IMPORTEXPORTSTATUS = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000035'. See Details.}

\item{REPORTING_YEAR}{e.g. '2013'. See Details.}

\item{FACILITY_NAME}{e.g. 'ANDERSONS MARATHON ETHANOL LLC'. See Details.}

\item{END_USE_FOR_CO2}{e.g. 'Food and beverage'. See Details.}

\item{END_USE_APP_IMPORTEXPORTSTATUS}{e.g. 'All'. See Details.}
}
\description{
Retrieve pp co2 end use details data from ghg database
}
