% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_pp_co2_flow_measure_equipment.R
\name{ghg_pp_co2_flow_measure_equipment}
\alias{ghg_pp_co2_flow_measure_equipment}
\title{Retrieve pp co2 flow measure equipment data from ghg database}
\usage{
ghg_pp_co2_flow_measure_equipment(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  EQUIPMENT_CALIBRATION_STANDARD = NULL, FACILITY_NAME = NULL,
  MEASUREMENT_EQUIPMENT_TYPE = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000035'. See Details.}

\item{REPORTING_YEAR}{e.g. '2013'. See Details.}

\item{EQUIPMENT_CALIBRATION_STANDARD}{e.g. 'Manufacturer'. See Details.}

\item{FACILITY_NAME}{e.g. 'ANDERSONS MARATHON ETHANOL LLC'. See Details.}

\item{MEASUREMENT_EQUIPMENT_TYPE}{e.g. 'Mass flow meter'. See Details.}
}
\description{
Retrieve pp co2 flow measure equipment data from ghg database
}

