% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_pp_mass_flow_measure_details.R
\name{ghg_pp_mass_flow_measure_details}
\alias{ghg_pp_mass_flow_measure_details}
\title{Retrieve pp mass flow measure details data from ghg database}
\usage{
ghg_pp_mass_flow_measure_details(FACILITY_ID = NULL,
  REPORTING_YEAR = NULL, FACILITY_NAME = NULL,
  STREAM_OR_METER_NAME = NULL, QUARTER_IDENTIFICATION = NULL,
  CONCENTRATION_MEASURE_STANDARD = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000035'. See Details.}

\item{REPORTING_YEAR}{e.g. '2013'. See Details.}

\item{FACILITY_NAME}{e.g. 'ANDERSONS MARATHON ETHANOL LLC'. See Details.}

\item{STREAM_OR_METER_NAME}{e.g. 'FL-3805'. See Details.}

\item{QUARTER_IDENTIFICATION}{e.g. 'First Quarter'. See Details.}

\item{CONCENTRATION_MEASURE_STANDARD}{e.g. 'Industry Standard'. See Details.}
}
\description{
Retrieve pp mass flow measure details data from ghg database
}
