% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_pp_meter_and_equipment_details.R
\name{ghg_pp_meter_and_equipment_details}
\alias{ghg_pp_meter_and_equipment_details}
\title{Retrieve pp meter and equipment details data from ghg database}
\usage{
ghg_pp_meter_and_equipment_details(FACILITY_ID = NULL,
  REPORTING_YEAR = NULL, METER_NAME = NULL, METER_TYPE = NULL,
  FACILITY_NAME = NULL, BIOMASS_PERCENTAGE = NULL,
  UPSTREAM_DEHYDRATION_EQUIPMENT = NULL,
  DOWNSTREAM_DEHYDRATION_EQUIP = NULL,
  UPSTREAM_COMPRESSION_EQUIPMENT = NULL,
  DOWNSTREAM_COMPRESSION_EQUIP = NULL, OTHER_EQUIPMENT_UPSTREAM = NULL,
  OTHER_EQUIPMENT_DOWNSTREAM = NULL, UPSTREAM_CAPTURE_EQUIPMENT = NULL,
  DOWNSTREAM_CAPTURE_EQUIPMENT = NULL, DESC_OTHR_PROCESS_EQUIP = NULL,
  MASSORVOLUME = NULL, METERSEGREGATIONLOCATION = NULL,
  METERIMPORTEXPORT = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000035'. See Details.}

\item{REPORTING_YEAR}{e.g. '2013'. See Details.}

\item{METER_NAME}{e.g. 'FL-3805'. See Details.}

\item{METER_TYPE}{e.g. 'CO2 Flow Meter'. See Details.}

\item{FACILITY_NAME}{e.g. 'ANDERSONS MARATHON ETHANOL LLC'. See Details.}

\item{BIOMASS_PERCENTAGE}{e.g. '100'. See Details.}

\item{UPSTREAM_DEHYDRATION_EQUIPMENT}{e.g. 'N'. See Details.}

\item{DOWNSTREAM_DEHYDRATION_EQUIP}{e.g. 'N'. See Details.}

\item{UPSTREAM_COMPRESSION_EQUIPMENT}{e.g. 'N'. See Details.}

\item{DOWNSTREAM_COMPRESSION_EQUIP}{e.g. 'N'. See Details.}

\item{OTHER_EQUIPMENT_UPSTREAM}{e.g. 'N'. See Details.}

\item{OTHER_EQUIPMENT_DOWNSTREAM}{e.g. 'N'. See Details.}

\item{UPSTREAM_CAPTURE_EQUIPMENT}{e.g. 'NA'. See Details.}

\item{DOWNSTREAM_CAPTURE_EQUIPMENT}{e.g. 'NA'. See Details.}

\item{DESC_OTHR_PROCESS_EQUIP}{e.g. 'NA'. See Details.}

\item{MASSORVOLUME}{e.g. 'NA'. See Details.}

\item{METERSEGREGATIONLOCATION}{e.g. 'NA'. See Details.}

\item{METERIMPORTEXPORT}{e.g. 'NA'. See Details.}
}
\description{
Retrieve pp meter and equipment details data from ghg database
}

