% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_pp_subpart_level_information.R
\name{ghg_pp_subpart_level_information}
\alias{ghg_pp_subpart_level_information}
\title{Retrieve pp subpart level information data from ghg database}
\usage{
ghg_pp_subpart_level_information(FACILITY_ID = NULL,
  REPORTING_YEAR = NULL, FACILITY_NAME = NULL, GHG_NAME = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000035'. See Details.}

\item{REPORTING_YEAR}{e.g. '2013'. See Details.}

\item{FACILITY_NAME}{e.g. 'ANDERSONS MARATHON ETHANOL LLC'. See Details.}

\item{GHG_NAME}{e.g. 'Carbon Dioxide'. See Details.}
}
\description{
Retrieve pp subpart level information data from ghg database
}
