% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_pp_volume_flow_measure_details.R
\name{ghg_pp_volume_flow_measure_details}
\alias{ghg_pp_volume_flow_measure_details}
\title{Retrieve pp volume flow measure details data from ghg database}
\usage{
ghg_pp_volume_flow_measure_details(FACILITY_ID = NULL,
  REPORTING_YEAR = NULL, FACILITY_NAME = NULL,
  STREAM_OR_METER_NAME = NULL, QUARTER_IDENTIFICATION = NULL,
  CONCENTRATION_MEASURE_STANDARD = NULL,
  DENSITY_MEASUREMENT_STANDARD = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000195'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{FACILITY_NAME}{e.g. 'DRI Dock Dehydration Facility'. See Details.}

\item{STREAM_OR_METER_NAME}{e.g. 'DRI Dock 5401'. See Details.}

\item{QUARTER_IDENTIFICATION}{e.g. 'First Quarter'. See Details.}

\item{CONCENTRATION_MEASURE_STANDARD}{e.g. 'AGA8 method, 1992 (detailed) and AGA NX-19 method from the AGA3 report.'. See Details.}

\item{DENSITY_MEASUREMENT_STANDARD}{e.g. 'AGA8 method, 1992 (detailed) and AGA NX-19 method from the AGA3 report.'. See Details.}
}
\description{
Retrieve pp volume flow measure details data from ghg database
}
