% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_pub_dim_facility.R
\name{ghg_pub_dim_facility}
\alias{ghg_pub_dim_facility}
\title{Retrieve pub dim facility data from ghg database}
\usage{
ghg_pub_dim_facility(FACILITY_ID = NULL, LATITUDE = NULL,
  LONGITUDE = NULL, CITY = NULL, STATE = NULL, ZIP = NULL,
  COUNTY_FIPS = NULL, COUNTY = NULL, ADDRESS1 = NULL, ADDRESS2 = NULL,
  FACILITY_NAME = NULL, STATE_NAME = NULL, NAICS_CODE = NULL,
  YEAR = NULL, BAMM_USED_DESC = NULL, EMISSION_CLASSIFICATION_CODE = NULL,
  PROGRAM_NAME = NULL, PROGRAM_SYS_ID = NULL, FRS_ID = NULL,
  CEMS_USED = NULL, CO2_CAPTURED = NULL, REPORTED_SUBPARTS = NULL,
  BAMM_APPROVED = NULL, EMITTED_CO2_SUPPLIED = NULL,
  TRIBAL_LAND_ID = NULL, EGGRT_FACILITY_ID = NULL, PARENT_COMPANY = NULL,
  REPORTED_INDUSTRY_TYPES = NULL, FACILITY_TYPES = NULL,
  SUBMISSION_ID = NULL, UU_RD_EXEMPT = NULL, REPORTING_STATUS = NULL,
  PROCESS_STATIONARY_CML = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000001'. See Details.}

\item{LATITUDE}{e.g. '48.82667'. See Details.}

\item{LONGITUDE}{e.g. '-122.71583'. See Details.}

\item{CITY}{e.g. 'FERNDALE'. See Details.}

\item{STATE}{e.g. 'WA'. See Details.}

\item{ZIP}{e.g. '98248'. See Details.}

\item{COUNTY_FIPS}{e.g. '53073'. See Details.}

\item{COUNTY}{e.g. 'WHATCOM'. See Details.}

\item{ADDRESS1}{e.g. '5105 LAKE TERRELL ROAD'. See Details.}

\item{ADDRESS2}{e.g. 'NA'. See Details.}

\item{FACILITY_NAME}{e.g. 'PSE Ferndale Generating Station'. See Details.}

\item{STATE_NAME}{e.g. 'Washington'. See Details.}

\item{NAICS_CODE}{e.g. '221112'. See Details.}

\item{YEAR}{e.g. '2010'. See Details.}

\item{BAMM_USED_DESC}{e.g. 'NA'. See Details.}

\item{EMISSION_CLASSIFICATION_CODE}{e.g. 'CU_ONLY'. See Details.}

\item{PROGRAM_NAME}{e.g. 'FRS'. See Details.}

\item{PROGRAM_SYS_ID}{e.g. '110000490166'. See Details.}

\item{FRS_ID}{e.g. '110000490166'. See Details.}

\item{CEMS_USED}{e.g. 'NA'. See Details.}

\item{CO2_CAPTURED}{e.g. 'NA'. See Details.}

\item{REPORTED_SUBPARTS}{e.g. 'C'. See Details.}

\item{BAMM_APPROVED}{e.g. 'NA'. See Details.}

\item{EMITTED_CO2_SUPPLIED}{e.g. 'NA'. See Details.}

\item{TRIBAL_LAND_ID}{e.g. 'NA'. See Details.}

\item{EGGRT_FACILITY_ID}{e.g. 'NA'. See Details.}

\item{PARENT_COMPANY}{e.g. 'Tenaska Energy, Inc. and Tenaska Energy Holdings, LLC (11.667%);Diamond Generating Corporation (14.00002%);Empeco IV, LLC and USPF II Ferndale Holdings, LLC (74.33298%);'. See Details.}

\item{REPORTED_INDUSTRY_TYPES}{e.g. 'C'. See Details.}

\item{FACILITY_TYPES}{e.g. 'Direct Emitter'. See Details.}

\item{SUBMISSION_ID}{e.g. '5752'. See Details.}

\item{UU_RD_EXEMPT}{e.g. 'NA'. See Details.}

\item{REPORTING_STATUS}{e.g. 'NA'. See Details.}

\item{PROCESS_STATIONARY_CML}{e.g. 'NA'. See Details.}
}
\description{
Retrieve pub dim facility data from ghg database
}

