% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_q_cems_details.R
\name{ghg_q_cems_details}
\alias{ghg_q_cems_details}
\title{Retrieve q cems details data from ghg database}
\usage{
ghg_q_cems_details(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  CML_NAME = NULL, CML_CONFIGURATION_TYPE = NULL,
  CEMS_BIOGENIC_CO2_EMISSIONS = NULL,
  CEMS_NON_BIOGENIC_CO2_EMISSION = NULL,
  CEMS_ANNUAL_CO2_EMISSIONS = NULL, CEMS_CH4_EMISSIONS = NULL,
  CEMS_N2O_EMISSIONS = NULL, CML_TOTAL_OPERATING_HOURS = NULL,
  TIER4_SUB_OPER_HRS_CO2_CONC = NULL,
  TIER4_SUB_OPER_HRS_GAS_FLOW_RT = NULL,
  TIER4_SUB_OPER_HRS_MOIST_CONT = NULL,
  CEMS_CO2_COMBUSTION_EMISSIONS = NULL,
  CEMS_METHODOLOGY_START_DATE = NULL, CEMS_METHODOLOGY_END_DATE = NULL,
  CEMS_FUEL = NULL, CEMS_QUARTER_NAME = NULL,
  CEMS_CO2_EMISSIONS_BY_QUARTER = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000156'. See Details.}

\item{REPORTING_YEAR}{e.g. '2013'. See Details.}

\item{CML_NAME}{e.g. 'UTC_Sinter Plant_CEMS'. See Details.}

\item{CML_CONFIGURATION_TYPE}{e.g. 'Single process/process unit exhausts to dedicated stack'. See Details.}

\item{CEMS_BIOGENIC_CO2_EMISSIONS}{e.g. '0'. See Details.}

\item{CEMS_NON_BIOGENIC_CO2_EMISSION}{e.g. '150219'. See Details.}

\item{CEMS_ANNUAL_CO2_EMISSIONS}{e.g. '150219'. See Details.}

\item{CEMS_CH4_EMISSIONS}{e.g. 'NA'. See Details.}

\item{CEMS_N2O_EMISSIONS}{e.g. 'NA'. See Details.}

\item{CML_TOTAL_OPERATING_HOURS}{e.g. '6029'. See Details.}

\item{TIER4_SUB_OPER_HRS_CO2_CONC}{e.g. '2'. See Details.}

\item{TIER4_SUB_OPER_HRS_GAS_FLOW_RT}{e.g. '2'. See Details.}

\item{TIER4_SUB_OPER_HRS_MOIST_CONT}{e.g. '2'. See Details.}

\item{CEMS_CO2_COMBUSTION_EMISSIONS}{e.g. 'NA'. See Details.}

\item{CEMS_METHODOLOGY_START_DATE}{e.g. '2013-01-01'. See Details.}

\item{CEMS_METHODOLOGY_END_DATE}{e.g. '2013-12-31'. See Details.}

\item{CEMS_FUEL}{e.g. 'NA'. See Details.}

\item{CEMS_QUARTER_NAME}{e.g. 'First Quarter'. See Details.}

\item{CEMS_CO2_EMISSIONS_BY_QUARTER}{e.g. '48759'. See Details.}
}
\description{
Retrieve q cems details data from ghg database
}
