% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_q_cems_info.R
\name{ghg_q_cems_info}
\alias{ghg_q_cems_info}
\title{Retrieve q cems info data from ghg database}
\usage{
ghg_q_cems_info(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  CEMS_UNIT_NAME = NULL, CEMS_UNIT_TYPE = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000156'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{CEMS_UNIT_NAME}{e.g. 'UTC_Sinter Plant'. See Details.}

\item{CEMS_UNIT_TYPE}{e.g. 'Sinter Process'. See Details.}
}
\description{
Retrieve q cems info data from ghg database
}

