% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_q_flare_information.R
\name{ghg_q_flare_information}
\alias{ghg_q_flare_information}
\title{Retrieve q flare information data from ghg database}
\usage{
ghg_q_flare_information(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  C_FROM_CH4_FRACTION_BASIS = NULL, FLARE_TYPE = NULL,
  FLARE_TYPE_OTHER = NULL, FLARE_SERVICE = NULL,
  FLARE_SERVICE_OTHER = NULL, C_FROM_CH4_FRACTION_BASIS_OTHR = NULL,
  UNIT_NAME = NULL, UNIT_TYPE = NULL, CO2_CALCULATION_METHOD = NULL,
  FLARE_CO2_EMISSIONS = NULL, CO2EMMVALUOM = NULL,
  FLARE_CH4_EMISSIONS = NULL, CH4EMMVALUOM = NULL,
  FLARE_N2O_EMISSIONS = NULL, N2OEMMVALUOM = NULL,
  Y1A_MEASUREMENT_PERIOD = NULL, Y1A_VOL_FLARE_GAS = NULL,
  Y1AEQVOLOFFLAREGASUOM = NULL, Y1A_VOL_FLARE_GAS_METHOD = NULL,
  Y1A_MOL_WT = NULL, Y1AEQANNAVGMOLWTUOM = NULL,
  Y1A_MOL_WT_METHOD_OTHER = NULL, Y1A_C_CONTENT = NULL,
  Y1AEQANNAVGCCONTENTUOM = NULL, Y1A_C_CONTENT_METHOD_OTHER = NULL,
  Y1B_MEASUREMENT_PERIOD = NULL, Y1B_VOL_FLARE_GAS = NULL,
  Y1BANNVOLFLAREGASUOM = NULL, Y1B_VOL_FLARE_GAS_METHOD = NULL,
  Y1B_CO2_CONCENTRATION = NULL, Y1BANNAVGCO2CONCUOM = NULL,
  Y1B_CO2_CONCENTRATION_METHOD = NULL,
  Y1B_CO2_CONCENTRA_METHOD_OTHER = NULL,
  Y1B_NUMBER_NONCO2_C_COMPOUNDS = NULL, Y2_MEASUREMENT_PERIOD = NULL,
  Y2_VOL_FLARE_GAS = NULL, Y2ANNVOLFLAREGASUOM = NULL,
  Y2_VOL_FLARE_GAS_METHOD = NULL, Y2_HHV = NULL, Y2ANNHHVALUOM = NULL,
  Y2_HHV_METHOD = NULL, Y2_HHV_METHOD_OTHER = NULL,
  Y2_VOL_FLARE_GAS_CONDITIONS = NULL, Y2_HHV_CONDITIONS = NULL,
  Y3_NUMBER_SSM_EVENTS = NULL, Y1A_MOL_WT_METHOD = NULL,
  Y1A_C_CONTENT_METHOD = NULL, Y1B_CONC_NON_CO2_C_COMP_METHD = NULL,
  Y1BANNAVGCONCMTHD = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000156'. See Details.}

\item{REPORTING_YEAR}{e.g. '2013'. See Details.}

\item{C_FROM_CH4_FRACTION_BASIS}{e.g. 'NA'. See Details.}

\item{FLARE_TYPE}{e.g. 'Unassisted'. See Details.}

\item{FLARE_TYPE_OTHER}{e.g. 'NA'. See Details.}

\item{FLARE_SERVICE}{e.g. 'Unit flare'. See Details.}

\item{FLARE_SERVICE_OTHER}{e.g. 'NA'. See Details.}

\item{C_FROM_CH4_FRACTION_BASIS_OTHR}{e.g. 'NA'. See Details.}

\item{UNIT_NAME}{e.g. 'SRC_IH7_BFG Flare 1_195'. See Details.}

\item{UNIT_TYPE}{e.g. 'Flare'. See Details.}

\item{CO2_CALCULATION_METHOD}{e.g. '98.253(b)(1)(iii) - Equation Y-3 Start-up, Shutdown, Malfunction'. See Details.}

\item{FLARE_CO2_EMISSIONS}{e.g. '124224.4'. See Details.}

\item{CO2EMMVALUOM}{e.g. 'Metric Tons'. See Details.}

\item{FLARE_CH4_EMISSIONS}{e.g. '0.01'. See Details.}

\item{CH4EMMVALUOM}{e.g. 'Metric Tons'. See Details.}

\item{FLARE_N2O_EMISSIONS}{e.g. '0.042'. See Details.}

\item{N2OEMMVALUOM}{e.g. 'Metric Tons'. See Details.}

\item{Y1A_MEASUREMENT_PERIOD}{e.g. 'NA'. See Details.}

\item{Y1A_VOL_FLARE_GAS}{e.g. 'NA'. See Details.}

\item{Y1AEQVOLOFFLAREGASUOM}{e.g. 'NA'. See Details.}

\item{Y1A_VOL_FLARE_GAS_METHOD}{e.g. 'NA'. See Details.}

\item{Y1A_MOL_WT}{e.g. 'NA'. See Details.}

\item{Y1AEQANNAVGMOLWTUOM}{e.g. 'NA'. See Details.}

\item{Y1A_MOL_WT_METHOD_OTHER}{e.g. 'NA'. See Details.}

\item{Y1A_C_CONTENT}{e.g. 'NA'. See Details.}

\item{Y1AEQANNAVGCCONTENTUOM}{e.g. 'NA'. See Details.}

\item{Y1A_C_CONTENT_METHOD_OTHER}{e.g. 'NA'. See Details.}

\item{Y1B_MEASUREMENT_PERIOD}{e.g. 'NA'. See Details.}

\item{Y1B_VOL_FLARE_GAS}{e.g. 'NA'. See Details.}

\item{Y1BANNVOLFLAREGASUOM}{e.g. 'NA'. See Details.}

\item{Y1B_VOL_FLARE_GAS_METHOD}{e.g. 'NA'. See Details.}

\item{Y1B_CO2_CONCENTRATION}{e.g. 'NA'. See Details.}

\item{Y1BANNAVGCO2CONCUOM}{e.g. 'NA'. See Details.}

\item{Y1B_CO2_CONCENTRATION_METHOD}{e.g. 'NA'. See Details.}

\item{Y1B_CO2_CONCENTRA_METHOD_OTHER}{e.g. 'NA'. See Details.}

\item{Y1B_NUMBER_NONCO2_C_COMPOUNDS}{e.g. 'NA'. See Details.}

\item{Y2_MEASUREMENT_PERIOD}{e.g. 'NA'. See Details.}

\item{Y2_VOL_FLARE_GAS}{e.g. 'NA'. See Details.}

\item{Y2ANNVOLFLAREGASUOM}{e.g. 'NA'. See Details.}

\item{Y2_VOL_FLARE_GAS_METHOD}{e.g. 'NA'. See Details.}

\item{Y2_HHV}{e.g. 'NA'. See Details.}

\item{Y2ANNHHVALUOM}{e.g. 'NA'. See Details.}

\item{Y2_HHV_METHOD}{e.g. 'NA'. See Details.}

\item{Y2_HHV_METHOD_OTHER}{e.g. 'NA'. See Details.}

\item{Y2_VOL_FLARE_GAS_CONDITIONS}{e.g. 'NA'. See Details.}

\item{Y2_HHV_CONDITIONS}{e.g. 'NA'. See Details.}

\item{Y3_NUMBER_SSM_EVENTS}{e.g. '0'. See Details.}

\item{Y1A_MOL_WT_METHOD}{e.g. 'NA'. See Details.}

\item{Y1A_C_CONTENT_METHOD}{e.g. 'NA'. See Details.}

\item{Y1B_CONC_NON_CO2_C_COMP_METHD}{e.g. 'NA'. See Details.}

\item{Y1BANNAVGCONCMTHD}{e.g. 'NA'. See Details.}
}
\description{
Retrieve q flare information data from ghg database
}

