% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_q_subpart_level_information.R
\name{ghg_q_subpart_level_information}
\alias{ghg_q_subpart_level_information}
\title{Retrieve q subpart level information data from ghg database}
\usage{
ghg_q_subpart_level_information(FACILITY_ID = NULL,
  REPORTING_YEAR = NULL, GHG_NAME = NULL, GHG_QUANTITY = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000029'. See Details.}

\item{REPORTING_YEAR}{e.g. '2010'. See Details.}

\item{GHG_NAME}{e.g. 'Biogenic Carbon dioxide'. See Details.}

\item{GHG_QUANTITY}{e.g. '0'. See Details.}
}
\description{
Retrieve q subpart level information data from ghg database
}
