% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_q_unit_details.R
\name{ghg_q_unit_details}
\alias{ghg_q_unit_details}
\title{Retrieve q unit details data from ghg database}
\usage{
ghg_q_unit_details(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  Q5_CO2_EMISSIONS = NULL, Q6_CO2_EMISSIONS = NULL,
  Q7_CO2_EMISSIONS = NULL, SS_EF_CALC_CO2 = NULL,
  CP_EF_CALC_CO2 = NULL, NUM_MOS_CC_MISS_DATA = NULL,
  CARBON_CONTENT__DET_MTHD = NULL, DESC_OTHR_CC_BASIS = NULL,
  NUM_MOS_MISS_DATA_METHD = NULL, UNIT_NAME = NULL, UNIT_TYPE = NULL,
  BASIS_MOLTEN_STEEL_CC = NULL, Q1_CO2_EMISSIONS = NULL,
  Q2_CO2_EMISSIONS = NULL, Q3_CO2_EMISSIONS = NULL,
  Q4_CO2_EMISSIONS = NULL, HOURLY_AVG_CO2_EMISSION_RATE = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000029'. See Details.}

\item{REPORTING_YEAR}{e.g. '2010'. See Details.}

\item{Q5_CO2_EMISSIONS}{e.g. 'NA'. See Details.}

\item{Q6_CO2_EMISSIONS}{e.g. 'NA'. See Details.}

\item{Q7_CO2_EMISSIONS}{e.g. 'NA'. See Details.}

\item{SS_EF_CALC_CO2}{e.g. '56615.9'. See Details.}

\item{CP_EF_CALC_CO2}{e.g. 'NA'. See Details.}

\item{NUM_MOS_CC_MISS_DATA}{e.g. 'NA'. See Details.}

\item{CARBON_CONTENT__DET_MTHD}{e.g. 'NA'. See Details.}

\item{DESC_OTHR_CC_BASIS}{e.g. 'NA'. See Details.}

\item{NUM_MOS_MISS_DATA_METHD}{e.g. 'NA'. See Details.}

\item{UNIT_NAME}{e.g. 'Melt Shop EAF'. See Details.}

\item{UNIT_TYPE}{e.g. 'Electric Arc Furnace (EAF)'. See Details.}

\item{BASIS_MOLTEN_STEEL_CC}{e.g. 'NA'. See Details.}

\item{Q1_CO2_EMISSIONS}{e.g. 'NA'. See Details.}

\item{Q2_CO2_EMISSIONS}{e.g. 'NA'. See Details.}

\item{Q3_CO2_EMISSIONS}{e.g. 'NA'. See Details.}

\item{Q4_CO2_EMISSIONS}{e.g. 'NA'. See Details.}

\item{HOURLY_AVG_CO2_EMISSION_RATE}{e.g. 'NA'. See Details.}
}
\description{
Retrieve q unit details data from ghg database
}
