% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_r_facility_info.R
\name{ghg_r_facility_info}
\alias{ghg_r_facility_info}
\title{Retrieve r facility info data from ghg database}
\usage{
ghg_r_facility_info(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, IS_CEMS_USED = NULL,
  NUM_SMELTING_FURNACES_CEMS = NULL,
  NUM_SMELTING_FURNACES_NOCEMS = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000601'. See Details.}

\item{REPORTING_YEAR}{e.g. '2010'. See Details.}

\item{FACILITY_NAME}{e.g. 'EXIDE CORP READING SMELTER DIV'. See Details.}

\item{IS_CEMS_USED}{e.g. 'No'. See Details.}

\item{NUM_SMELTING_FURNACES_CEMS}{e.g. 'NA'. See Details.}

\item{NUM_SMELTING_FURNACES_NOCEMS}{e.g. '4'. See Details.}
}
\description{
Retrieve r facility info data from ghg database
}
