% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_r_feedstock_info.R
\name{ghg_r_feedstock_info}
\alias{ghg_r_feedstock_info}
\title{Retrieve r feedstock info data from ghg database}
\usage{
ghg_r_feedstock_info(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, FURNACE_ASSOC_WITH_FEEDSTOCK = NULL,
  SUBSTITUTE_DATA_PROCEDURE = NULL, OTHER_SUBS_DATA_PROCEDURE = NULL,
  NUM_MONTHS_MASS_MATERIAL = NULL, METHOD_USED_CARBON_CONTENT = NULL,
  NUM_MONTHS_CARBON_MATERIAL = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000601'. See Details.}

\item{REPORTING_YEAR}{e.g. '2012'. See Details.}

\item{FACILITY_NAME}{e.g. 'EXIDE CORP READING SMELTER DIV'. See Details.}

\item{FURNACE_ASSOC_WITH_FEEDSTOCK}{e.g. '122'. See Details.}

\item{SUBSTITUTE_DATA_PROCEDURE}{e.g. 'Other (specify)'. See Details.}

\item{OTHER_SUBS_DATA_PROCEDURE}{e.g. 'Use of months with good data to correlate with months with missing data.  Correlated with production'. See Details.}

\item{NUM_MONTHS_MASS_MATERIAL}{e.g. '0'. See Details.}

\item{METHOD_USED_CARBON_CONTENT}{e.g. 'ASTM E1941-04'. See Details.}

\item{NUM_MONTHS_CARBON_MATERIAL}{e.g. '0'. See Details.}
}
\description{
Retrieve r feedstock info data from ghg database
}

