% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_r_smelting_furnace_info.R
\name{ghg_r_smelting_furnace_info}
\alias{ghg_r_smelting_furnace_info}
\title{Retrieve r smelting furnace info data from ghg database}
\usage{
ghg_r_smelting_furnace_info(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, FURNACE_ID = NULL,
  ANNUAL_UNROUNDED_CO2_EMIS = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000601'. See Details.}

\item{REPORTING_YEAR}{e.g. '2013'. See Details.}

\item{FACILITY_NAME}{e.g. 'EXIDE CORP READING SMELTER DIV'. See Details.}

\item{FURNACE_ID}{e.g. '122'. See Details.}

\item{ANNUAL_UNROUNDED_CO2_EMIS}{e.g. '1317.7'. See Details.}
}
\description{
Retrieve r smelting furnace info data from ghg database
}

