% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_s_lime_byproduct_details.R
\name{ghg_s_lime_byproduct_details}
\alias{ghg_s_lime_byproduct_details}
\title{Retrieve s lime byproduct details data from ghg database}
\usage{
ghg_s_lime_byproduct_details(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  MONTH_NAME = NULL, CAO_CHEMICAL_COMPO_ANALY = NULL,
  CAO_CHEMICAL_COMPO_ANALY_METHD = NULL,
  CAO_CHEMICAL_COMPO_METHD_OTHR = NULL, MGO_CHEMICAL_COMPO_ANALY = NULL,
  MGO_CHEMICAL_COMPO_ANALY_METHD = NULL,
  MGO_CHEMICAL_COMPO_METHD_OTHR = NULL, MASS_SOLD_ANALYSIS_METHOD = NULL,
  MASS_SOLD_ANALYSIS_METHOD_OTHR = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000019'. See Details.}

\item{REPORTING_YEAR}{e.g. '2014'. See Details.}

\item{MONTH_NAME}{e.g. 'NA'. See Details.}

\item{CAO_CHEMICAL_COMPO_ANALY}{e.g. 'NA'. See Details.}

\item{CAO_CHEMICAL_COMPO_ANALY_METHD}{e.g. 'NA'. See Details.}

\item{CAO_CHEMICAL_COMPO_METHD_OTHR}{e.g. 'NA'. See Details.}

\item{MGO_CHEMICAL_COMPO_ANALY}{e.g. 'NA'. See Details.}

\item{MGO_CHEMICAL_COMPO_ANALY_METHD}{e.g. 'NA'. See Details.}

\item{MGO_CHEMICAL_COMPO_METHD_OTHR}{e.g. 'NA'. See Details.}

\item{MASS_SOLD_ANALYSIS_METHOD}{e.g. 'NA'. See Details.}

\item{MASS_SOLD_ANALYSIS_METHOD_OTHR}{e.g. 'NA'. See Details.}
}
\description{
Retrieve s lime byproduct details data from ghg database
}

