% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_s_lime_product_details.R
\name{ghg_s_lime_product_details}
\alias{ghg_s_lime_product_details}
\title{Retrieve s lime product details data from ghg database}
\usage{
ghg_s_lime_product_details(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  MASS_SOLD_ANALYSIS_METHOD_OTHR = NULL,
  MASS_PRODUCED_ANALYSIS_METHOD = NULL,
  MASS_PRODUCED_ANLS_METHD_OTHR = NULL, MASS_SOLD_ANALYSIS = NULL,
  IS_CO2_USED_ONSITE_INDICATOR = NULL, AMOUNT_CO2_USED_ONSITE_METHOD = NULL,
  AMT_CO2_USED_ONSITE_METHD_OTHR = NULL, MONTH_NAME = NULL,
  CAO_CHEMICAL_COMPO_ANALY = NULL, CAO_CHEMICAL_COMPO_ANALY_METHD = NULL,
  CAO_CHEMICAL_COMPO_METHD_OTHR = NULL, MGO_CHEMICAL_COMPO_ANALY = NULL,
  MGO_CHEMICAL_COMPO_ANALY_METHD = NULL,
  MGO_CHEMICAL_COMPO_METHD_OTHR = NULL, MASS_SOLD_ANALYSIS_METHOD = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000019'. See Details.}

\item{REPORTING_YEAR}{e.g. '2014'. See Details.}

\item{MASS_SOLD_ANALYSIS_METHOD_OTHR}{e.g. 'None Sold'. See Details.}

\item{MASS_PRODUCED_ANALYSIS_METHOD}{e.g. 'Other'. See Details.}

\item{MASS_PRODUCED_ANLS_METHD_OTHR}{e.g. 'Per October 2011 EPA Memo to Doug Emerson American Crystal Sugar'. See Details.}

\item{MASS_SOLD_ANALYSIS}{e.g. 'N'. See Details.}

\item{IS_CO2_USED_ONSITE_INDICATOR}{e.g. 'Y'. See Details.}

\item{AMOUNT_CO2_USED_ONSITE_METHOD}{e.g. 'Other method'. See Details.}

\item{AMT_CO2_USED_ONSITE_METHD_OTHR}{e.g. 'Industry process review of conversion back to calcium carbonate suggest 100% recovery (Beet Sugar Te'. See Details.}

\item{MONTH_NAME}{e.g. 'January'. See Details.}

\item{CAO_CHEMICAL_COMPO_ANALY}{e.g. 'N'. See Details.}

\item{CAO_CHEMICAL_COMPO_ANALY_METHD}{e.g. 'ASTM C25-06'. See Details.}

\item{CAO_CHEMICAL_COMPO_METHD_OTHR}{e.g. 'NA'. See Details.}

\item{MGO_CHEMICAL_COMPO_ANALY}{e.g. 'N'. See Details.}

\item{MGO_CHEMICAL_COMPO_ANALY_METHD}{e.g. 'ASTM C25-06'. See Details.}

\item{MGO_CHEMICAL_COMPO_METHD_OTHR}{e.g. 'NA'. See Details.}

\item{MASS_SOLD_ANALYSIS_METHOD}{e.g. 'Other'. See Details.}
}
\description{
Retrieve s lime product details data from ghg database
}

