% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_ss_eq_ss5.R
\name{ghg_ss_eq_ss5}
\alias{ghg_ss_eq_ss5}
\title{Retrieve ss eq ss5 data from ghg database}
\usage{
ghg_ss_eq_ss5(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, TOTAL_EMIT_PERFLUOROCYCLOPROPA = NULL,
  TOTAL_EMIT_FLUOROCYCLOBUTANE = NULL,
  TOTAL_EMIT_PERFLUOROETHANE = NULL,
  TOTAL_EMIT_PERFLUOROMETHANE = NULL,
  TOTAL_EMIT_PERFLUOROPROPANE = NULL,
  TOTAL_EMIT_PERFLUOROBUTANE = NULL,
  TOTAL_EMIT_PERFLUOROPENTANE = NULL,
  TOTAL_EMIT_PERFLUOROHEXANE = NULL, TOTAL_EMITTED_PFC_9_1_18 = NULL,
  TOTAL_EMITTED_SF6 = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000039'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{FACILITY_NAME}{e.g. 'Alstom Grid Inc'. See Details.}

\item{TOTAL_EMIT_PERFLUOROCYCLOPROPA}{e.g. 'NA'. See Details.}

\item{TOTAL_EMIT_FLUOROCYCLOBUTANE}{e.g. '0'. See Details.}

\item{TOTAL_EMIT_PERFLUOROETHANE}{e.g. 'NA'. See Details.}

\item{TOTAL_EMIT_PERFLUOROMETHANE}{e.g. 'NA'. See Details.}

\item{TOTAL_EMIT_PERFLUOROPROPANE}{e.g. 'NA'. See Details.}

\item{TOTAL_EMIT_PERFLUOROBUTANE}{e.g. 'NA'. See Details.}

\item{TOTAL_EMIT_PERFLUOROPENTANE}{e.g. 'NA'. See Details.}

\item{TOTAL_EMIT_PERFLUOROHEXANE}{e.g. 'NA'. See Details.}

\item{TOTAL_EMITTED_PFC_9_1_18}{e.g. 'NA'. See Details.}

\item{TOTAL_EMITTED_SF6}{e.g. 'NA'. See Details.}
}
\description{
Retrieve ss eq ss5 data from ghg database
}
