% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_ss_ghg_acquisition.R
\name{ghg_ss_ghg_acquisition}
\alias{ghg_ss_ghg_acquisition}
\title{Retrieve ss ghg acquisition data from ghg database}
\usage{
ghg_ss_ghg_acquisition(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, GAS_NAME = NULL,
  ACQUISITION_CALCULATED = NULL, ACQUISITION_OVERRIDE = NULL,
  LBS_PURCHASED_BULK = NULL, LBS_RETURNED_EQUIPMENT_USERS = NULL,
  LBS_RETURNED_OFFSITE_RECYCLING = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000039'. See Details.}

\item{REPORTING_YEAR}{e.g. '2013'. See Details.}

\item{FACILITY_NAME}{e.g. 'Alstom Grid Inc'. See Details.}

\item{GAS_NAME}{e.g. 'Sulfur hexafluoride'. See Details.}

\item{ACQUISITION_CALCULATED}{e.g. 'NA'. See Details.}

\item{ACQUISITION_OVERRIDE}{e.g. 'NA'. See Details.}

\item{LBS_PURCHASED_BULK}{e.g. 'NA'. See Details.}

\item{LBS_RETURNED_EQUIPMENT_USERS}{e.g. 'NA'. See Details.}

\item{LBS_RETURNED_OFFSITE_RECYCLING}{e.g. '0'. See Details.}
}
\description{
Retrieve ss ghg acquisition data from ghg database
}
