% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_ss_ghg_disbursement.R
\name{ghg_ss_ghg_disbursement}
\alias{ghg_ss_ghg_disbursement}
\title{Retrieve ss ghg disbursement data from ghg database}
\usage{
ghg_ss_ghg_disbursement(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, GAS_NAME = NULL, DISBURSEMENT_CALCULATED = NULL,
  DISBURSEMENT_OVERRIDE = NULL, LBS_CONTAINED_NEW_EQUIP_DELIV = NULL,
  LBS_DELIV_EQUIP_IN_CONTAINERS = NULL, LBS_OFFSITE_RECYCLING = NULL,
  LBS_RETURNED_TO_SUPPLIERS = NULL, LBS_SENT_OFFSITE_DESTRUCTION = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000039'. See Details.}

\item{REPORTING_YEAR}{e.g. '2013'. See Details.}

\item{FACILITY_NAME}{e.g. 'Alstom Grid Inc'. See Details.}

\item{GAS_NAME}{e.g. 'Sulfur hexafluoride'. See Details.}

\item{DISBURSEMENT_CALCULATED}{e.g. '79730.33'. See Details.}

\item{DISBURSEMENT_OVERRIDE}{e.g. 'N'. See Details.}

\item{LBS_CONTAINED_NEW_EQUIP_DELIV}{e.g. '9650.5'. See Details.}

\item{LBS_DELIV_EQUIP_IN_CONTAINERS}{e.g. '69349'. See Details.}

\item{LBS_OFFSITE_RECYCLING}{e.g. '0'. See Details.}

\item{LBS_RETURNED_TO_SUPPLIERS}{e.g. '730.83'. See Details.}

\item{LBS_SENT_OFFSITE_DESTRUCTION}{e.g. '0'. See Details.}
}
\description{
Retrieve ss ghg disbursement data from ghg database
}

