% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_ss_ghg_information.R
\name{ghg_ss_ghg_information}
\alias{ghg_ss_ghg_information}
\title{Retrieve ss ghg information data from ghg database}
\usage{
ghg_ss_ghg_information(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, GAS_NAME = NULL,
  GHG_EMISSIONS_UNROUNDED = NULL, TOTAL_EMISSIONS_UNROUNDED = NULL,
  MFG_EMISSIONS_UNROUNDED = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000039'. See Details.}

\item{REPORTING_YEAR}{e.g. '2013'. See Details.}

\item{FACILITY_NAME}{e.g. 'Alstom Grid Inc'. See Details.}

\item{GAS_NAME}{e.g. 'Sulfur hexafluoride'. See Details.}

\item{GHG_EMISSIONS_UNROUNDED}{e.g. '2625.64'. See Details.}

\item{TOTAL_EMISSIONS_UNROUNDED}{e.g. '0'. See Details.}

\item{MFG_EMISSIONS_UNROUNDED}{e.g. '2625.64'. See Details.}
}
\description{
Retrieve ss ghg information data from ghg database
}
