% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_ss_install_emissions.R
\name{ghg_ss_install_emissions}
\alias{ghg_ss_install_emissions}
\title{Retrieve ss install emissions data from ghg database}
\usage{
ghg_ss_install_emissions(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, GAS_NAME = NULL, TOTAL_EMISSIONS_RPT_ROUND = NULL,
  TOTAL_EMISSIONS_RPT_ROUND_OVR = NULL, TOTAL_MASS_CHARGE_EQUIPMENT = NULL,
  TOTAL_MASS_FILL_EQUIPMENT = NULL, TOTAL_NAMEPLATE_CAPACITY = NULL,
  CHARGE_EQ_MISSING_DATA_METHOD = NULL,
  OFFSITE_FILL_MISSING_DATA_MTHD = NULL, NC_MISSING_DATA_METHOD = NULL,
  CHARGE_EQ_MISSING_DATA = NULL, CHARGE_EQ_MISSING_DATA_REASON = NULL,
  OFFSITE_FILL_MISSING_DATA = NULL, OFFSITE_FILL_MISSING_DATA_REAS = NULL,
  NC_MISSING_DATA = NULL, NC_MISSING_DATA_REASON = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000039'. See Details.}

\item{REPORTING_YEAR}{e.g. '2013'. See Details.}

\item{FACILITY_NAME}{e.g. 'Alstom Grid Inc'. See Details.}

\item{GAS_NAME}{e.g. 'PFC-4-1-12 (Perfluoropentane)'. See Details.}

\item{TOTAL_EMISSIONS_RPT_ROUND}{e.g. '0'. See Details.}

\item{TOTAL_EMISSIONS_RPT_ROUND_OVR}{e.g. 'N'. See Details.}

\item{TOTAL_MASS_CHARGE_EQUIPMENT}{e.g. 'NA'. See Details.}

\item{TOTAL_MASS_FILL_EQUIPMENT}{e.g. 'NA'. See Details.}

\item{TOTAL_NAMEPLATE_CAPACITY}{e.g. 'NA'. See Details.}

\item{CHARGE_EQ_MISSING_DATA_METHOD}{e.g. 'NA'. See Details.}

\item{OFFSITE_FILL_MISSING_DATA_MTHD}{e.g. 'NA'. See Details.}

\item{NC_MISSING_DATA_METHOD}{e.g. 'NA'. See Details.}

\item{CHARGE_EQ_MISSING_DATA}{e.g. 'NA'. See Details.}

\item{CHARGE_EQ_MISSING_DATA_REASON}{e.g. 'NA'. See Details.}

\item{OFFSITE_FILL_MISSING_DATA}{e.g. 'NA'. See Details.}

\item{OFFSITE_FILL_MISSING_DATA_REAS}{e.g. 'NA'. See Details.}

\item{NC_MISSING_DATA}{e.g. 'NA'. See Details.}

\item{NC_MISSING_DATA_REASON}{e.g. 'NA'. See Details.}
}
\description{
Retrieve ss install emissions data from ghg database
}

