% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_ss_mass_deliv_new_eq.R
\name{ghg_ss_mass_deliv_new_eq}
\alias{ghg_ss_mass_deliv_new_eq}
\title{Retrieve ss mass deliv new eq data from ghg database}
\usage{
ghg_ss_mass_deliv_new_eq(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  NUM_OF_SAMPLES_FLUOROPENTANE = NULL, NUM_OF_SAMPLES_FLUOROPROPANE = NULL,
  NUM_OF_SAMPLESSULFURHEXAFLUORI = NULL,
  UPPER_BOUND_PERFLUORODECALIN = NULL, LOWER_BOUND_PERFLUORODECALIN = NULL,
  MAKE_MODEL_FLUOROBUTANE = NULL, MAKE_MODEL_FLUOROCYCLOBUTANE = NULL,
  MAKE_MODEL_FLUOROCYCLOPROPANE = NULL, MAKE_MODEL_FLUORODECALIN = NULL,
  MAKE_MODEL_FLUOROETHANE = NULL, MAKE_MODEL_FLUOROHEXANE = NULL,
  MAKE_MODEL_FLUOROMETHANE = NULL, MAKE_MODEL_FLUOROPENTANE = NULL,
  MAKE_MODELFLUOROPROPANE = NULL, MEAN_VALUE_MASS_PERFLUOROBUTAN = NULL,
  MEAN_VALUE_MASS_FLUOROCYCLOBUT = NULL,
  MEAN_VALUE_MASS_FLUOROCYCLOPRO = NULL,
  MEAN_VALUE_MASS_FLUORODECALIN = NULL, MEAN_VALUE_MASS_FLUOROETHANE = NULL,
  MEAN_VALUE_MASS_FLUOROHEXANE = NULL, MEAN_VALUE_MASS_FLUOROMETHANE = NULL,
  MEAN_VALUE_MASS_FLUOROPENTANE = NULL,
  MEAN_VALUE_MASS_FLUOROPROPANE = NULL,
  MEAN_VALUE_MASS_SULFURHEXAFLUO = NULL, NUM_OF_SAMPLES_FLUOROBUTANE = NULL,
  NUM_OF_SAMPLES_FLUOROCYCLOBUTA = NULL,
  NUM_OF_SAMPLES_FLUOROCYCLOPROP = NULL,
  NUM_OF_SAMPLES_FLUORODECALIN = NULL, NUM_OF_SAMPLES_FLUOROETHANE = NULL,
  NUM_OF_SAMPLES_FLUOROHEXANE = NULL, NUM_OF_SAMPLES_FLUOROMETHANE = NULL,
  FACILITY_NAME = NULL, MAKE_MODEL_GROUP_CONDITION_ID = NULL,
  METHOD_SPECIFIED = NULL, UPPER_BOUND_SULFURHEXAFLUORIDE = NULL,
  LOWER_BOUND_SULFURHEXAFLUORIDE = NULL,
  UPPER_BOUND_PERFLUOROMETHANE = NULL, LOWER_BOUND_PERFLUOROMETHANE = NULL,
  UPPER_BOUND_PERFLUOROETHANE = NULL, LOWER_BOUND_PERFLUOROETHANE = NULL,
  UPPER_BOUND_PERFLUOROPROPANE = NULL, LOWER_BOUND_PERFLUOROPROPANE = NULL,
  UPPERBOUNDC_C3F6 = NULL, LOWER_BOUNDC_C3F6 = NULL,
  UPPER_BOUND_PERFLUOROBUTANE = NULL, LOWER_BOUND_PERFLUOROBUTANE = NULL,
  UPPER_BOUND_PERFLUOROCYCLOBUTA = NULL,
  LOWER_BOUND_PERFLUOROCYCLOBUTA = NULL,
  UPPER_BOUND_PERFLUOROPENTANE = NULL, LOWER_BOUND_PERFLUOROPENTANE = NULL,
  UPPER_BOUND_PERFLUOROHEXANE = NULL, LOWER_BOUND_PERFLUOROHEXANE = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000039'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{NUM_OF_SAMPLES_FLUOROPENTANE}{e.g. 'NA'. See Details.}

\item{NUM_OF_SAMPLES_FLUOROPROPANE}{e.g. 'NA'. See Details.}

\item{NUM_OF_SAMPLESSULFURHEXAFLUORI}{e.g. 'NA'. See Details.}

\item{UPPER_BOUND_PERFLUORODECALIN}{e.g. 'NA'. See Details.}

\item{LOWER_BOUND_PERFLUORODECALIN}{e.g. 'NA'. See Details.}

\item{MAKE_MODEL_FLUOROBUTANE}{e.g. 'NA'. See Details.}

\item{MAKE_MODEL_FLUOROCYCLOBUTANE}{e.g. 'NA'. See Details.}

\item{MAKE_MODEL_FLUOROCYCLOPROPANE}{e.g. 'NA'. See Details.}

\item{MAKE_MODEL_FLUORODECALIN}{e.g. 'NA'. See Details.}

\item{MAKE_MODEL_FLUOROETHANE}{e.g. 'NA'. See Details.}

\item{MAKE_MODEL_FLUOROHEXANE}{e.g. 'NA'. See Details.}

\item{MAKE_MODEL_FLUOROMETHANE}{e.g. 'NA'. See Details.}

\item{MAKE_MODEL_FLUOROPENTANE}{e.g. 'NA'. See Details.}

\item{MAKE_MODELFLUOROPROPANE}{e.g. 'NA'. See Details.}

\item{MEAN_VALUE_MASS_PERFLUOROBUTAN}{e.g. 'NA'. See Details.}

\item{MEAN_VALUE_MASS_FLUOROCYCLOBUT}{e.g. 'NA'. See Details.}

\item{MEAN_VALUE_MASS_FLUOROCYCLOPRO}{e.g. 'NA'. See Details.}

\item{MEAN_VALUE_MASS_FLUORODECALIN}{e.g. 'NA'. See Details.}

\item{MEAN_VALUE_MASS_FLUOROETHANE}{e.g. 'NA'. See Details.}

\item{MEAN_VALUE_MASS_FLUOROHEXANE}{e.g. 'NA'. See Details.}

\item{MEAN_VALUE_MASS_FLUOROMETHANE}{e.g. 'NA'. See Details.}

\item{MEAN_VALUE_MASS_FLUOROPENTANE}{e.g. 'NA'. See Details.}

\item{MEAN_VALUE_MASS_FLUOROPROPANE}{e.g. 'NA'. See Details.}

\item{MEAN_VALUE_MASS_SULFURHEXAFLUO}{e.g. 'NA'. See Details.}

\item{NUM_OF_SAMPLES_FLUOROBUTANE}{e.g. 'NA'. See Details.}

\item{NUM_OF_SAMPLES_FLUOROCYCLOBUTA}{e.g. 'NA'. See Details.}

\item{NUM_OF_SAMPLES_FLUOROCYCLOPROP}{e.g. 'NA'. See Details.}

\item{NUM_OF_SAMPLES_FLUORODECALIN}{e.g. 'NA'. See Details.}

\item{NUM_OF_SAMPLES_FLUOROETHANE}{e.g. 'NA'. See Details.}

\item{NUM_OF_SAMPLES_FLUOROHEXANE}{e.g. 'NA'. See Details.}

\item{NUM_OF_SAMPLES_FLUOROMETHANE}{e.g. 'NA'. See Details.}

\item{FACILITY_NAME}{e.g. 'Alstom Grid Inc'. See Details.}

\item{MAKE_MODEL_GROUP_CONDITION_ID}{e.g. 'DT12'. See Details.}

\item{METHOD_SPECIFIED}{e.g. 'Yes'. See Details.}

\item{UPPER_BOUND_SULFURHEXAFLUORIDE}{e.g. 'NA'. See Details.}

\item{LOWER_BOUND_SULFURHEXAFLUORIDE}{e.g. 'NA'. See Details.}

\item{UPPER_BOUND_PERFLUOROMETHANE}{e.g. 'NA'. See Details.}

\item{LOWER_BOUND_PERFLUOROMETHANE}{e.g. 'NA'. See Details.}

\item{UPPER_BOUND_PERFLUOROETHANE}{e.g. 'NA'. See Details.}

\item{LOWER_BOUND_PERFLUOROETHANE}{e.g. 'NA'. See Details.}

\item{UPPER_BOUND_PERFLUOROPROPANE}{e.g. 'NA'. See Details.}

\item{LOWER_BOUND_PERFLUOROPROPANE}{e.g. 'NA'. See Details.}

\item{UPPERBOUNDC_C3F6}{e.g. 'NA'. See Details.}

\item{LOWER_BOUNDC_C3F6}{e.g. 'NA'. See Details.}

\item{UPPER_BOUND_PERFLUOROBUTANE}{e.g. 'NA'. See Details.}

\item{LOWER_BOUND_PERFLUOROBUTANE}{e.g. 'NA'. See Details.}

\item{UPPER_BOUND_PERFLUOROCYCLOBUTA}{e.g. 'NA'. See Details.}

\item{LOWER_BOUND_PERFLUOROCYCLOBUTA}{e.g. 'NA'. See Details.}

\item{UPPER_BOUND_PERFLUOROPENTANE}{e.g. 'NA'. See Details.}

\item{LOWER_BOUND_PERFLUOROPENTANE}{e.g. 'NA'. See Details.}

\item{UPPER_BOUND_PERFLUOROHEXANE}{e.g. 'NA'. See Details.}

\item{LOWER_BOUND_PERFLUOROHEXANE}{e.g. 'NA'. See Details.}
}
\description{
Retrieve ss mass deliv new eq data from ghg database
}

