% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_ss_msng_dta_emission.R
\name{ghg_ss_msng_dta_emission}
\alias{ghg_ss_msng_dta_emission}
\title{Retrieve ss msng dta emission data from ghg database}
\usage{
ghg_ss_msng_dta_emission(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  MTHD_END_YR_TO_ESTIMATE_DATA = NULL,
  MTHD_EQPT_NAMEPLATE_CAP_TO_EST = NULL,
  MTHD_EQPT_USER_TO_ESTIMAT_DATA = NULL,
  MTHD_NEW_EQUIP_TO_ESTIMAT_DATA = NULL,
  MTHD_OFFSITE_DESTRCTION_TO_EST = NULL,
  MTHD_OFFSITE_RECYC_TO_EST_DATA = NULL,
  MTHD_RECYCLE_TO_ESTIMATE_DATA = NULL,
  MTHD_RETURND_EQUIP_TO_EST_DATA = NULL,
  MTHD_RETURNED_SUPP_TO_EST_DATA = NULL, MISSING_DATA_BEGIN_YEAR = NULL,
  MISSING_DATA_BULK_PURCHASE = NULL, MISSING_DATA_END_YEAR = NULL,
  MISS_DATA_EQPT_NAMEPLATE_CAPCT = NULL, MISSING_DATA_EQUIPMENT_USER = NULL,
  MISSING_DATA_NEW_EQUIPMENT = NULL, MISSING_DATA_OFFSITE_DESTRUCTI = NULL,
  MISSING_DATA_OFFSITE_RECYCLE = NULL, MISSING_DATA_RECYCLE = NULL,
  MISSING_DATA_RETURNED_EQUIPMEN = NULL,
  MISSING_DATA_RETURNED_SUPPLIER = NULL,
  MISSING_DATA_BEGIN_YEAR_REASON = NULL,
  MISS_DATA_BULK_PURCHASE_REASON = NULL,
  MISSING_DATA_END_YEAR_REASON = NULL,
  MISS_DATA_NAMEPLATE_CAPCTY_REA = NULL,
  MISS_DATA_EQUIPMENTUSERREASON = NULL,
  MISSING_DATA_NEW_EQUIP_REASON = NULL,
  MISS_DATA_OFFSITE_DESTR_REASON = NULL,
  MISS_DATA_OFFSITE_RECY_REASON = NULL, MISSING_DATA_RECYCLE_REASON = NULL,
  MISSING_DATA_RETURNED_EQUIP = NULL, MISSING_DATA_RETURN_SUPPLIERS = NULL,
  FACILITY_NAME = NULL, GAS_NAME = NULL,
  MTHD_BEGIN_YR_TO_ESTIMATE_DATA = NULL, MTHD_BULK_PURCH_TO_EST_DATA = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000039'. See Details.}

\item{REPORTING_YEAR}{e.g. '2013'. See Details.}

\item{MTHD_END_YR_TO_ESTIMATE_DATA}{e.g. 'NA'. See Details.}

\item{MTHD_EQPT_NAMEPLATE_CAP_TO_EST}{e.g. 'NA'. See Details.}

\item{MTHD_EQPT_USER_TO_ESTIMAT_DATA}{e.g. 'NA'. See Details.}

\item{MTHD_NEW_EQUIP_TO_ESTIMAT_DATA}{e.g. 'NA'. See Details.}

\item{MTHD_OFFSITE_DESTRCTION_TO_EST}{e.g. 'NA'. See Details.}

\item{MTHD_OFFSITE_RECYC_TO_EST_DATA}{e.g. 'NA'. See Details.}

\item{MTHD_RECYCLE_TO_ESTIMATE_DATA}{e.g. 'NA'. See Details.}

\item{MTHD_RETURND_EQUIP_TO_EST_DATA}{e.g. 'NA'. See Details.}

\item{MTHD_RETURNED_SUPP_TO_EST_DATA}{e.g. 'NA'. See Details.}

\item{MISSING_DATA_BEGIN_YEAR}{e.g. 'No'. See Details.}

\item{MISSING_DATA_BULK_PURCHASE}{e.g. 'No'. See Details.}

\item{MISSING_DATA_END_YEAR}{e.g. 'No'. See Details.}

\item{MISS_DATA_EQPT_NAMEPLATE_CAPCT}{e.g. 'No'. See Details.}

\item{MISSING_DATA_EQUIPMENT_USER}{e.g. 'No'. See Details.}

\item{MISSING_DATA_NEW_EQUIPMENT}{e.g. 'No'. See Details.}

\item{MISSING_DATA_OFFSITE_DESTRUCTI}{e.g. 'No'. See Details.}

\item{MISSING_DATA_OFFSITE_RECYCLE}{e.g. 'No'. See Details.}

\item{MISSING_DATA_RECYCLE}{e.g. 'No'. See Details.}

\item{MISSING_DATA_RETURNED_EQUIPMEN}{e.g. 'No'. See Details.}

\item{MISSING_DATA_RETURNED_SUPPLIER}{e.g. 'No'. See Details.}

\item{MISSING_DATA_BEGIN_YEAR_REASON}{e.g. 'NA'. See Details.}

\item{MISS_DATA_BULK_PURCHASE_REASON}{e.g. 'NA'. See Details.}

\item{MISSING_DATA_END_YEAR_REASON}{e.g. 'NA'. See Details.}

\item{MISS_DATA_NAMEPLATE_CAPCTY_REA}{e.g. 'NA'. See Details.}

\item{MISS_DATA_EQUIPMENTUSERREASON}{e.g. 'NA'. See Details.}

\item{MISSING_DATA_NEW_EQUIP_REASON}{e.g. 'NA'. See Details.}

\item{MISS_DATA_OFFSITE_DESTR_REASON}{e.g. 'NA'. See Details.}

\item{MISS_DATA_OFFSITE_RECY_REASON}{e.g. 'NA'. See Details.}

\item{MISSING_DATA_RECYCLE_REASON}{e.g. 'NA'. See Details.}

\item{MISSING_DATA_RETURNED_EQUIP}{e.g. 'NA'. See Details.}

\item{MISSING_DATA_RETURN_SUPPLIERS}{e.g. 'NA'. See Details.}

\item{FACILITY_NAME}{e.g. 'Alstom Grid Inc'. See Details.}

\item{GAS_NAME}{e.g. 'Sulfur hexafluoride'. See Details.}

\item{MTHD_BEGIN_YR_TO_ESTIMATE_DATA}{e.g. 'NA'. See Details.}

\item{MTHD_BULK_PURCH_TO_EST_DATA}{e.g. 'NA'. See Details.}
}
\description{
Retrieve ss msng dta emission data from ghg database
}

