% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_ss_msng_dta_mak_modl.R
\name{ghg_ss_msng_dta_mak_modl}
\alias{ghg_ss_msng_dta_mak_modl}
\title{Retrieve ss msng dta mak modl data from ghg database}
\usage{
ghg_ss_msng_dta_mak_modl(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, MAKE_MODEL_GROUP_CONDITION_ID = NULL,
  METHOD_SPECIFIED = NULL, MTHD_LOWER_BOUND_FLUOROBUTANE = NULL,
  MTHD_LOWER_BND_FLUOCYCLOBUTANE = NULL,
  MTHD_LOWER_BND_FLUOCYCLOPROPAN = NULL,
  MTHD_LOWER_BOUND_FLUORODECALIN = NULL,
  MTHD_LOWER_BOUND_FLUOROETHANE = NULL,
  MTHD_LOWER_BOUND_FLUOROHEXANE = NULL,
  MTHD_LOWER_BOUND_FLUOROMETHANE = NULL,
  MTHD_LOWER_BOUND_FLUOROPENTANE = NULL,
  MISS_UPPER_BOUND_FLUOROBUTANE = NULL,
  MISS_UPPER_BND_FLUOROBUTANE_RE = NULL,
  MISS_UPPR_BND_FLUOCYCLOBUTANE = NULL,
  MISS_UPPER_BND_FLUOCYCBUTAN_RE = NULL,
  MTHD_MEAN_VAL_FLUOROMETHANE = NULL,
  MTHD_MEAN_VAL_FLUOROPENTANE = NULL,
  MTHD_MEAN_VAL_FLUOROPROPANE = NULL,
  MTHD_MEAN_VAL_SULFURHEXAFLUORI = NULL,
  MTHD_NUM_SAMP_FLUOROBUTANE = NULL,
  MTHD_NUM_SAMP_FLUOROCYCLOBUTAN = NULL,
  MTHD_NUM_SAMP_FLUOROCYCLOPROPA = NULL,
  MTHD_NUM_SAMP_FLUORODECALIN = NULL,
  MTHD_NUM_SAMP_FLUOROETHANE = NULL, MTHD_NUM_SAMP_FLUOROHEXANE = NULL,
  MTHD_NUM_SAMP_FLUOROMETHANE = NULL,
  MTHD_NUM_SAMP_FLUOROPENTANE = NULL,
  MTHD_NUM_SAMP_FLUOROPROPANE = NULL,
  MTHD_NUM_SAMP_SULFURHEXAFLUORI = NULL,
  MTHD_UPPER_BOUND_FLUOROBUTANE = NULL,
  MTHD_UPPER_BND_FLUOCYCLOBUTANE = NULL,
  MTHD_UPPER_BND_FLUOCYCLOPROPAN = NULL,
  MTHD_UPPER_BOUND_FLUORODECALIN = NULL,
  MTHD_UPPER_BOUND_FLUOROETHANE = NULL,
  MTHD_UPPER_BOUND_FLUOROHEXANE = NULL,
  MTHD_UPPER_BOUND_FLUOROMETHANE = NULL,
  MTHD_UPPER_BOUND_FLUOROPENTANE = NULL,
  MTHD_UPPER_BOUND_FLUOROPROPANE = NULL,
  MTHD_UPPER_BND_SULFURHEXAFLUOR = NULL,
  MISS_LOWER_BOUND_FLUOROBUTANE = NULL,
  MISS_LOWER_BND_FLUOROBUTANE_RE = NULL,
  MISS_LOWER_BOUND_FLUOCYCLOBUTA = NULL,
  MISS_LOWER_BND_FLUOCYCBUTAN_RE = NULL,
  MISS_LOWER_BND_FLUOCYCLOPROPAN = NULL,
  MISS_LOWER_BND_FLUOCYCPROPA_RE = NULL,
  MISS_LOWER_BOUND_FLUORODECALIN = NULL,
  MISS_LOWER_BND_FLUODECALIN_RE = NULL,
  MISS_LOWER_BND_FLUOROETHANE = NULL,
  MISS_LOWER_BND_FLUOROETHANE_RE = NULL,
  MISS_LOWER_BOUNDFLUOROHEXANE = NULL,
  MISS_LOWER_BND_FLUOROHEXANE_RE = NULL,
  MISS_LOWER_BOUND_FLUOROMETHANE = NULL,
  MISS_LOWER_BND_FLUOROMETHAN_RE = NULL,
  MISS_LOWER_BOUND_FLUOROPENTANE = NULL,
  MISS_NUM_SAMP_FLUOROBUTANE_RE = NULL,
  MISS_NUM_SAMP_FLUOROCYCLOBUTAN = NULL,
  MISS_NUM_SAMP_FLUOROCYCBUTA_RE = NULL,
  MISS_NUM_SAMP_FLUOROCYCLOPROPA = NULL,
  MISS_NUM_SAMP_FLUOCYCPROPAN_RE = NULL,
  MISS_NUM_SAMP_FLUORODECALIN = NULL,
  MISS_NUM_SAMP_FLUORODECALIN_RE = NULL,
  MISS_NUM_SAMP_FLUOROETHANE = NULL,
  MISS_NUM_SAMP_FLUOROETHANE_RE = NULL,
  MISS_NUM_SAMP_FLUOROHEXANE = NULL,
  MISS_NUM_SAMP_FLUOROHEXANE_RE = NULL,
  MISS_NUM_SAMP_FLUOROMETHANE = NULL,
  MISS_NUM_SAMP_FLUOROMETHANE_RE = NULL,
  MISS_NUM_SAMP_FLUOROPENTANE = NULL,
  MISS_NUM_SAMP_FLUOROPENTANE_RE = NULL,
  MISS_NUM_SAMP_FLUOROPROPANE = NULL,
  MISS_NUM_SAMP_FLUOROPROPANE_RE = NULL,
  MISS_NUM_SAMP_SULFURHEXAFLUORI = NULL,
  MISS_NUM_SAMP_SULFURHEXAFLU_RE = NULL,
  MISS_UPPR_BOUND_FLUOCYCPROPANE = NULL,
  MISS_UPPER_BOUND_FLUORODECALIN = NULL,
  MISS_LOWER_BND_FLUOROPENTAN_RE = NULL,
  MISS_LOWER_BOUND_FLUOROPROPANE = NULL,
  MISS_LOWER_BND_FLUOROPROPAN_RE = NULL,
  MISS_LOWER_BND_SULFURHEXAFLUO = NULL,
  MISS_LOWER_BND_SULFURHEXAFL_RE = NULL,
  MISS_MEAN_VAL_FLUOROBUTANE = NULL,
  MISS_MEAN_VAL_FLUOROBUTANE_RE = NULL,
  MISS_MEAN_VAL_FLUOROCYCLOBUTAN = NULL,
  MISS_MEAN_VAL_FLUOROCYCBUTA_RE = NULL,
  MISS_MEAN_VAL_FLUOROCYCLOPROPA = NULL,
  MISS_MEAN_VAL_FLUOCYCPROPAN_RE = NULL,
  MISS_MEAN_VAL_FLUORODECALIN = NULL,
  MISS_MEAN_VAL_FLUORODECALIN_RE = NULL,
  MISS_MEAN_VAL_FLUOROETHANE = NULL,
  MISS_MEAN_VAL_FLUOROETHANE_RE = NULL,
  MISS_MEAN_VAL_FLUOROHEXANE = NULL,
  MISS_MEAN_VAL_FLUOROHEXANE_RE = NULL,
  MISS_MEAN_VAL_FLUOROMETHANE = NULL,
  MISS_MEAN_VAL_FLUOROMETHANE_RE = NULL,
  MISS_MEAN_VAL_FLUOROPENTANE = NULL,
  MISS_MEAN_VAL_FLUOROPENTANE_RE = NULL,
  MISS_MEAN_VAL_FLUOROPROPANE = NULL,
  MISS_MEAN_VAL_FLUOROPROPANE_RE = NULL,
  MISS_MEAN_VAL_SULFURHEXAFLUORI = NULL,
  MISS_MEAN_VAL_SULFURHEXAFLU_RE = NULL,
  MISS_NUM_SAMP_FLUOROBUTANE = NULL,
  MISS_UPPER_BND_FLUOCYCPROPA_RE = NULL,
  MISS_UPPER_BOUND_FLUOROETHANE = NULL,
  MTHD_LOWER_BOUND_FLUOROPROPANE = NULL,
  MTHD_LOWER_BND_SULFURHEXAFLUOR = NULL,
  MTHD_MEAN_VAL_FLUOROBUTANE = NULL,
  MTHD_MEAN_VA_LFLUOROCYCLOBUTAN = NULL,
  MTHD_MEAN_VAL_FLUOROCYCLOPROPA = NULL,
  MTHD_MEAN_VAL_FLUORODECALIN = NULL,
  MTHD_MEAN_VAL_FLUOROETHANE = NULL, MTHD_MEAN_VAL_FLUOROHEXANE = NULL,
  MISS_UPPER_BND_FLUODECALIN_RE = NULL,
  MISS_UPPR_BOUND_FLUOROETHANERE = NULL,
  MISS_UPPER_BOUND_FLUOROHEXANE = NULL,
  MISS_UPPR_BOUND_FLUOROHEXAN_RE = NULL,
  MISS_UPPER_BOUND_FLUOROMETHANE = NULL,
  MISS_UPPR_BOUND_FLUOROMETHANRE = NULL,
  MISS_UPPR_BOUND_FLUOROPENTANE = NULL,
  MISS_UPPR_BOUND_FLUOROPENTANRE = NULL,
  MISS_UPPER_BOUND_FLUOROPROPANE = NULL,
  MISS_UPPR_BND_PERFLUOROPROP_RE = NULL,
  MISS_UPPR_BOUND_SULFURHEXAFLUO = NULL,
  MISS_UPPR_BOUND_SULFURHEXAF_RE = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000039'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{FACILITY_NAME}{e.g. 'Alstom Grid Inc'. See Details.}

\item{MAKE_MODEL_GROUP_CONDITION_ID}{e.g. 'DT12'. See Details.}

\item{METHOD_SPECIFIED}{e.g. 'Yes'. See Details.}

\item{MTHD_LOWER_BOUND_FLUOROBUTANE}{e.g. 'NA'. See Details.}

\item{MTHD_LOWER_BND_FLUOCYCLOBUTANE}{e.g. 'NA'. See Details.}

\item{MTHD_LOWER_BND_FLUOCYCLOPROPAN}{e.g. 'NA'. See Details.}

\item{MTHD_LOWER_BOUND_FLUORODECALIN}{e.g. 'NA'. See Details.}

\item{MTHD_LOWER_BOUND_FLUOROETHANE}{e.g. 'NA'. See Details.}

\item{MTHD_LOWER_BOUND_FLUOROHEXANE}{e.g. 'NA'. See Details.}

\item{MTHD_LOWER_BOUND_FLUOROMETHANE}{e.g. 'NA'. See Details.}

\item{MTHD_LOWER_BOUND_FLUOROPENTANE}{e.g. 'NA'. See Details.}

\item{MISS_UPPER_BOUND_FLUOROBUTANE}{e.g. 'NA'. See Details.}

\item{MISS_UPPER_BND_FLUOROBUTANE_RE}{e.g. 'NA'. See Details.}

\item{MISS_UPPR_BND_FLUOCYCLOBUTANE}{e.g. 'NA'. See Details.}

\item{MISS_UPPER_BND_FLUOCYCBUTAN_RE}{e.g. 'NA'. See Details.}

\item{MTHD_MEAN_VAL_FLUOROMETHANE}{e.g. 'NA'. See Details.}

\item{MTHD_MEAN_VAL_FLUOROPENTANE}{e.g. 'NA'. See Details.}

\item{MTHD_MEAN_VAL_FLUOROPROPANE}{e.g. 'NA'. See Details.}

\item{MTHD_MEAN_VAL_SULFURHEXAFLUORI}{e.g. 'NA'. See Details.}

\item{MTHD_NUM_SAMP_FLUOROBUTANE}{e.g. 'NA'. See Details.}

\item{MTHD_NUM_SAMP_FLUOROCYCLOBUTAN}{e.g. 'NA'. See Details.}

\item{MTHD_NUM_SAMP_FLUOROCYCLOPROPA}{e.g. 'NA'. See Details.}

\item{MTHD_NUM_SAMP_FLUORODECALIN}{e.g. 'NA'. See Details.}

\item{MTHD_NUM_SAMP_FLUOROETHANE}{e.g. 'NA'. See Details.}

\item{MTHD_NUM_SAMP_FLUOROHEXANE}{e.g. 'NA'. See Details.}

\item{MTHD_NUM_SAMP_FLUOROMETHANE}{e.g. 'NA'. See Details.}

\item{MTHD_NUM_SAMP_FLUOROPENTANE}{e.g. 'NA'. See Details.}

\item{MTHD_NUM_SAMP_FLUOROPROPANE}{e.g. 'NA'. See Details.}

\item{MTHD_NUM_SAMP_SULFURHEXAFLUORI}{e.g. 'NA'. See Details.}

\item{MTHD_UPPER_BOUND_FLUOROBUTANE}{e.g. 'NA'. See Details.}

\item{MTHD_UPPER_BND_FLUOCYCLOBUTANE}{e.g. 'NA'. See Details.}

\item{MTHD_UPPER_BND_FLUOCYCLOPROPAN}{e.g. 'NA'. See Details.}

\item{MTHD_UPPER_BOUND_FLUORODECALIN}{e.g. 'NA'. See Details.}

\item{MTHD_UPPER_BOUND_FLUOROETHANE}{e.g. 'NA'. See Details.}

\item{MTHD_UPPER_BOUND_FLUOROHEXANE}{e.g. 'NA'. See Details.}

\item{MTHD_UPPER_BOUND_FLUOROMETHANE}{e.g. 'NA'. See Details.}

\item{MTHD_UPPER_BOUND_FLUOROPENTANE}{e.g. 'NA'. See Details.}

\item{MTHD_UPPER_BOUND_FLUOROPROPANE}{e.g. 'NA'. See Details.}

\item{MTHD_UPPER_BND_SULFURHEXAFLUOR}{e.g. 'NA'. See Details.}

\item{MISS_LOWER_BOUND_FLUOROBUTANE}{e.g. 'NA'. See Details.}

\item{MISS_LOWER_BND_FLUOROBUTANE_RE}{e.g. 'NA'. See Details.}

\item{MISS_LOWER_BOUND_FLUOCYCLOBUTA}{e.g. 'NA'. See Details.}

\item{MISS_LOWER_BND_FLUOCYCBUTAN_RE}{e.g. 'NA'. See Details.}

\item{MISS_LOWER_BND_FLUOCYCLOPROPAN}{e.g. 'NA'. See Details.}

\item{MISS_LOWER_BND_FLUOCYCPROPA_RE}{e.g. 'NA'. See Details.}

\item{MISS_LOWER_BOUND_FLUORODECALIN}{e.g. 'NA'. See Details.}

\item{MISS_LOWER_BND_FLUODECALIN_RE}{e.g. 'NA'. See Details.}

\item{MISS_LOWER_BND_FLUOROETHANE}{e.g. 'NA'. See Details.}

\item{MISS_LOWER_BND_FLUOROETHANE_RE}{e.g. 'NA'. See Details.}

\item{MISS_LOWER_BOUNDFLUOROHEXANE}{e.g. 'NA'. See Details.}

\item{MISS_LOWER_BND_FLUOROHEXANE_RE}{e.g. 'NA'. See Details.}

\item{MISS_LOWER_BOUND_FLUOROMETHANE}{e.g. 'NA'. See Details.}

\item{MISS_LOWER_BND_FLUOROMETHAN_RE}{e.g. 'NA'. See Details.}

\item{MISS_LOWER_BOUND_FLUOROPENTANE}{e.g. 'NA'. See Details.}

\item{MISS_NUM_SAMP_FLUOROBUTANE_RE}{e.g. 'NA'. See Details.}

\item{MISS_NUM_SAMP_FLUOROCYCLOBUTAN}{e.g. 'NA'. See Details.}

\item{MISS_NUM_SAMP_FLUOROCYCBUTA_RE}{e.g. 'NA'. See Details.}

\item{MISS_NUM_SAMP_FLUOROCYCLOPROPA}{e.g. 'NA'. See Details.}

\item{MISS_NUM_SAMP_FLUOCYCPROPAN_RE}{e.g. 'NA'. See Details.}

\item{MISS_NUM_SAMP_FLUORODECALIN}{e.g. 'NA'. See Details.}

\item{MISS_NUM_SAMP_FLUORODECALIN_RE}{e.g. 'NA'. See Details.}

\item{MISS_NUM_SAMP_FLUOROETHANE}{e.g. 'NA'. See Details.}

\item{MISS_NUM_SAMP_FLUOROETHANE_RE}{e.g. 'NA'. See Details.}

\item{MISS_NUM_SAMP_FLUOROHEXANE}{e.g. 'NA'. See Details.}

\item{MISS_NUM_SAMP_FLUOROHEXANE_RE}{e.g. 'NA'. See Details.}

\item{MISS_NUM_SAMP_FLUOROMETHANE}{e.g. 'NA'. See Details.}

\item{MISS_NUM_SAMP_FLUOROMETHANE_RE}{e.g. 'NA'. See Details.}

\item{MISS_NUM_SAMP_FLUOROPENTANE}{e.g. 'NA'. See Details.}

\item{MISS_NUM_SAMP_FLUOROPENTANE_RE}{e.g. 'NA'. See Details.}

\item{MISS_NUM_SAMP_FLUOROPROPANE}{e.g. 'NA'. See Details.}

\item{MISS_NUM_SAMP_FLUOROPROPANE_RE}{e.g. 'NA'. See Details.}

\item{MISS_NUM_SAMP_SULFURHEXAFLUORI}{e.g. 'NA'. See Details.}

\item{MISS_NUM_SAMP_SULFURHEXAFLU_RE}{e.g. 'NA'. See Details.}

\item{MISS_UPPR_BOUND_FLUOCYCPROPANE}{e.g. 'NA'. See Details.}

\item{MISS_UPPER_BOUND_FLUORODECALIN}{e.g. 'NA'. See Details.}

\item{MISS_LOWER_BND_FLUOROPENTAN_RE}{e.g. 'NA'. See Details.}

\item{MISS_LOWER_BOUND_FLUOROPROPANE}{e.g. 'NA'. See Details.}

\item{MISS_LOWER_BND_FLUOROPROPAN_RE}{e.g. 'NA'. See Details.}

\item{MISS_LOWER_BND_SULFURHEXAFLUO}{e.g. 'NA'. See Details.}

\item{MISS_LOWER_BND_SULFURHEXAFL_RE}{e.g. 'NA'. See Details.}

\item{MISS_MEAN_VAL_FLUOROBUTANE}{e.g. 'NA'. See Details.}

\item{MISS_MEAN_VAL_FLUOROBUTANE_RE}{e.g. 'NA'. See Details.}

\item{MISS_MEAN_VAL_FLUOROCYCLOBUTAN}{e.g. 'NA'. See Details.}

\item{MISS_MEAN_VAL_FLUOROCYCBUTA_RE}{e.g. 'NA'. See Details.}

\item{MISS_MEAN_VAL_FLUOROCYCLOPROPA}{e.g. 'NA'. See Details.}

\item{MISS_MEAN_VAL_FLUOCYCPROPAN_RE}{e.g. 'NA'. See Details.}

\item{MISS_MEAN_VAL_FLUORODECALIN}{e.g. 'NA'. See Details.}

\item{MISS_MEAN_VAL_FLUORODECALIN_RE}{e.g. 'NA'. See Details.}

\item{MISS_MEAN_VAL_FLUOROETHANE}{e.g. 'NA'. See Details.}

\item{MISS_MEAN_VAL_FLUOROETHANE_RE}{e.g. 'NA'. See Details.}

\item{MISS_MEAN_VAL_FLUOROHEXANE}{e.g. 'NA'. See Details.}

\item{MISS_MEAN_VAL_FLUOROHEXANE_RE}{e.g. 'NA'. See Details.}

\item{MISS_MEAN_VAL_FLUOROMETHANE}{e.g. 'NA'. See Details.}

\item{MISS_MEAN_VAL_FLUOROMETHANE_RE}{e.g. 'NA'. See Details.}

\item{MISS_MEAN_VAL_FLUOROPENTANE}{e.g. 'NA'. See Details.}

\item{MISS_MEAN_VAL_FLUOROPENTANE_RE}{e.g. 'NA'. See Details.}

\item{MISS_MEAN_VAL_FLUOROPROPANE}{e.g. 'NA'. See Details.}

\item{MISS_MEAN_VAL_FLUOROPROPANE_RE}{e.g. 'NA'. See Details.}

\item{MISS_MEAN_VAL_SULFURHEXAFLUORI}{e.g. 'NA'. See Details.}

\item{MISS_MEAN_VAL_SULFURHEXAFLU_RE}{e.g. 'NA'. See Details.}

\item{MISS_NUM_SAMP_FLUOROBUTANE}{e.g. 'NA'. See Details.}

\item{MISS_UPPER_BND_FLUOCYCPROPA_RE}{e.g. 'NA'. See Details.}

\item{MISS_UPPER_BOUND_FLUOROETHANE}{e.g. 'NA'. See Details.}

\item{MTHD_LOWER_BOUND_FLUOROPROPANE}{e.g. 'NA'. See Details.}

\item{MTHD_LOWER_BND_SULFURHEXAFLUOR}{e.g. 'NA'. See Details.}

\item{MTHD_MEAN_VAL_FLUOROBUTANE}{e.g. 'NA'. See Details.}

\item{MTHD_MEAN_VA_LFLUOROCYCLOBUTAN}{e.g. 'NA'. See Details.}

\item{MTHD_MEAN_VAL_FLUOROCYCLOPROPA}{e.g. 'NA'. See Details.}

\item{MTHD_MEAN_VAL_FLUORODECALIN}{e.g. 'NA'. See Details.}

\item{MTHD_MEAN_VAL_FLUOROETHANE}{e.g. 'NA'. See Details.}

\item{MTHD_MEAN_VAL_FLUOROHEXANE}{e.g. 'NA'. See Details.}

\item{MISS_UPPER_BND_FLUODECALIN_RE}{e.g. 'NA'. See Details.}

\item{MISS_UPPR_BOUND_FLUOROETHANERE}{e.g. 'NA'. See Details.}

\item{MISS_UPPER_BOUND_FLUOROHEXANE}{e.g. 'NA'. See Details.}

\item{MISS_UPPR_BOUND_FLUOROHEXAN_RE}{e.g. 'NA'. See Details.}

\item{MISS_UPPER_BOUND_FLUOROMETHANE}{e.g. 'NA'. See Details.}

\item{MISS_UPPR_BOUND_FLUOROMETHANRE}{e.g. 'NA'. See Details.}

\item{MISS_UPPR_BOUND_FLUOROPENTANE}{e.g. 'NA'. See Details.}

\item{MISS_UPPR_BOUND_FLUOROPENTANRE}{e.g. 'NA'. See Details.}

\item{MISS_UPPER_BOUND_FLUOROPROPANE}{e.g. 'NA'. See Details.}

\item{MISS_UPPR_BND_PERFLUOROPROP_RE}{e.g. 'NA'. See Details.}

\item{MISS_UPPR_BOUND_SULFURHEXAFLUO}{e.g. 'NA'. See Details.}

\item{MISS_UPPR_BOUND_SULFURHEXAF_RE}{e.g. 'NA'. See Details.}
}
\description{
Retrieve ss msng dta mak modl data from ghg database
}
