% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_ss_subpart_level_information.R
\name{ghg_ss_subpart_level_information}
\alias{ghg_ss_subpart_level_information}
\title{Retrieve ss subpart level information data from ghg database}
\usage{
ghg_ss_subpart_level_information(FACILITY_ID = NULL,
  REPORTING_YEAR = NULL, FACILITY_NAME = NULL, GAS_NAME = NULL,
  OTHER_GHG_NAME = NULL, GHG_QUANTITY = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000039'. See Details.}

\item{REPORTING_YEAR}{e.g. '2013'. See Details.}

\item{FACILITY_NAME}{e.g. 'Alstom Grid Inc'. See Details.}

\item{GAS_NAME}{e.g. 'PFC-116 (Perfluoroethane)'. See Details.}

\item{OTHER_GHG_NAME}{e.g. 'NA'. See Details.}

\item{GHG_QUANTITY}{e.g. '0'. See Details.}
}
\description{
Retrieve ss subpart level information data from ghg database
}
