% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_ss_summary.R
\name{ghg_ss_summary}
\alias{ghg_ss_summary}
\title{Retrieve ss summary data from ghg database}
\usage{
ghg_ss_summary(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, GAS_NAME = NULL, IS_EQUATION_SS5_USED = NULL,
  IS_ESTIMATED_GHG = NULL, IS_GHG_MASS_DISBURSED = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000039'. See Details.}

\item{REPORTING_YEAR}{e.g. '2013'. See Details.}

\item{FACILITY_NAME}{e.g. 'Alstom Grid Inc'. See Details.}

\item{GAS_NAME}{e.g. 'Sulfur hexafluoride'. See Details.}

\item{IS_EQUATION_SS5_USED}{e.g. 'Yes'. See Details.}

\item{IS_ESTIMATED_GHG}{e.g. 'Yes'. See Details.}

\item{IS_GHG_MASS_DISBURSED}{e.g. 'No'. See Details.}
}
\description{
Retrieve ss summary data from ghg database
}

