% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_ss_user_emissions.R
\name{ghg_ss_user_emissions}
\alias{ghg_ss_user_emissions}
\title{Retrieve ss user emissions data from ghg database}
\usage{
ghg_ss_user_emissions(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, GAS_NAME = NULL, LBS_ACQUISITION = NULL,
  LBS_DECREASE_INVENTORY = NULL, LBS_DISBURSEMENTS = NULL,
  LBS_USER_EMISSION = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000039'. See Details.}

\item{REPORTING_YEAR}{e.g. '2013'. See Details.}

\item{FACILITY_NAME}{e.g. 'Alstom Grid Inc'. See Details.}

\item{GAS_NAME}{e.g. 'Sulfur hexafluoride'. See Details.}

\item{LBS_ACQUISITION}{e.g. '80415.5'. See Details.}

\item{LBS_DECREASE_INVENTORY}{e.g. '1940.47'. See Details.}

\item{LBS_DISBURSEMENTS}{e.g. '79730.33'. See Details.}

\item{LBS_USER_EMISSION}{e.g. '2625.64'. See Details.}
}
\description{
Retrieve ss user emissions data from ghg database
}
