% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_t_cover_or_carrier_gas.R
\name{ghg_t_cover_or_carrier_gas}
\alias{ghg_t_cover_or_carrier_gas}
\title{Retrieve t cover or carrier gas data from ghg database}
\usage{
ghg_t_cover_or_carrier_gas(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, MISS_DATA_QUANTITY_EMISSION = NULL,
  TOTAL_ANNUAL_EMISSION = NULL, WERE_MISSING_DATA_USED = NULL,
  MISSING_DATA_METHOD = NULL, TIME_PERIOD_USED = NULL,
  TIME_PERIOD_USED_SCALE = NULL, PROTECTION_TECHNOLOGIES_DESC = NULL,
  COVER_GROUP_OF_FLUORINATED_GHG = NULL, OTHER_CARRIER_GAS_NAME = NULL,
  CARRIER_GROUP_OF_FLUORINTD_GHG = NULL,
  OTHER_CARRIER_CAS_NUMBER = NULL, OTHER_CARRIER_GAS_FORMULA = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000030'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{FACILITY_NAME}{e.g. 'US MAGNESIUM LLC'. See Details.}

\item{MISS_DATA_QUANTITY_EMISSION}{e.g. 'NA'. See Details.}

\item{TOTAL_ANNUAL_EMISSION}{e.g. '22.4936364'. See Details.}

\item{WERE_MISSING_DATA_USED}{e.g. 'NO'. See Details.}

\item{MISSING_DATA_METHOD}{e.g. 'NA'. See Details.}

\item{TIME_PERIOD_USED}{e.g. 'NA'. See Details.}

\item{TIME_PERIOD_USED_SCALE}{e.g. 'NA'. See Details.}

\item{PROTECTION_TECHNOLOGIES_DESC}{e.g. 'NA'. See Details.}

\item{COVER_GROUP_OF_FLUORINATED_GHG}{e.g. 'NA'. See Details.}

\item{OTHER_CARRIER_GAS_NAME}{e.g. 'NA'. See Details.}

\item{CARRIER_GROUP_OF_FLUORINTD_GHG}{e.g. 'NA'. See Details.}

\item{OTHER_CARRIER_CAS_NUMBER}{e.g. 'NA'. See Details.}

\item{OTHER_CARRIER_GAS_FORMULA}{e.g. 'NA'. See Details.}
}
\description{
Retrieve t cover or carrier gas data from ghg database
}
