% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_t_production_process.R
\name{ghg_t_production_process}
\alias{ghg_t_production_process}
\title{Retrieve t production process data from ghg database}
\usage{
ghg_t_production_process(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, PRODUCT_PROCESS_TYPE_OTHER = NULL,
  PRODUCT_PROCESS_TYPE = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000030'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{FACILITY_NAME}{e.g. 'US MAGNESIUM LLC'. See Details.}

\item{PRODUCT_PROCESS_TYPE_OTHER}{e.g. 'NA'. See Details.}

\item{PRODUCT_PROCESS_TYPE}{e.g. 'PRIMARY PRODUCTION'. See Details.}
}
\description{
Retrieve t production process data from ghg database
}

