% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_t_subpart_level_information.R
\name{ghg_t_subpart_level_information}
\alias{ghg_t_subpart_level_information}
\title{Retrieve t subpart level information data from ghg database}
\usage{
ghg_t_subpart_level_information(FACILITY_ID = NULL,
  REPORTING_YEAR = NULL, GHG_QUANTITY = NULL, FACILITY_NAME = NULL,
  GHG_NAME = NULL, OTHER_GREENHOUSE_GAS_NAME = NULL,
  OTHER_GAS_GHG_GROUP = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000030'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{GHG_QUANTITY}{e.g. '398.2'. See Details.}

\item{FACILITY_NAME}{e.g. 'US MAGNESIUM LLC'. See Details.}

\item{GHG_NAME}{e.g. 'OTHER GHG COVER GAS'. See Details.}

\item{OTHER_GREENHOUSE_GAS_NAME}{e.g. 'CARBON DIOXIDE'. See Details.}

\item{OTHER_GAS_GHG_GROUP}{e.g. 'NA'. See Details.}
}
\description{
Retrieve t subpart level information data from ghg database
}
