% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_tt_hist_waste_method.R
\name{ghg_tt_hist_waste_method}
\alias{ghg_tt_hist_waste_method}
\title{Retrieve tt hist waste method data from ghg database}
\usage{
ghg_tt_hist_waste_method(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  DECAY_RATE = NULL, IS_BASED_ON_TOTAL_VOLIT_SOLIDS = NULL,
  IS_DEFAULT_VAL_FROM_TABLE_TT1 = NULL,
  ANAEROB_60DAY_BIODEG_TEST_USED = NULL, WASTEDISPOSEDQUANTITY = NULL,
  WASTESTREAMDOCVALUE = NULL, FACILITY_NAME = NULL,
  WASTE_STREAM_NAME = NULL, HISTORICAL_WASTE_REPORTINGYEAR = NULL,
  USE_MTHD1_FOR_HIST_WASTE_QUAN = NULL,
  USE_MTHD2_FOR_HIST_WASTE_QUAN = NULL,
  USE_MTHD3_FOR_HIST_WASTE_QUAN = NULL,
  USE_MTHD4_FOR_HIST_WASTE_QUAN = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000093'. See Details.}

\item{REPORTING_YEAR}{e.g. '2014'. See Details.}

\item{DECAY_RATE}{e.g. '0.06'. See Details.}

\item{IS_BASED_ON_TOTAL_VOLIT_SOLIDS}{e.g. 'N'. See Details.}

\item{IS_DEFAULT_VAL_FROM_TABLE_TT1}{e.g. 'Y'. See Details.}

\item{ANAEROB_60DAY_BIODEG_TEST_USED}{e.g. 'N'. See Details.}

\item{WASTEDISPOSEDQUANTITY}{e.g. '5365'. See Details.}

\item{WASTESTREAMDOCVALUE}{e.g. '0.2'. See Details.}

\item{FACILITY_NAME}{e.g. 'MOMENTIVE PERFORMANCE MATERIALS'. See Details.}

\item{WASTE_STREAM_NAME}{e.g. 'LANDFILL 1'. See Details.}

\item{HISTORICAL_WASTE_REPORTINGYEAR}{e.g. '1982'. See Details.}

\item{USE_MTHD1_FOR_HIST_WASTE_QUAN}{e.g. 'N'. See Details.}

\item{USE_MTHD2_FOR_HIST_WASTE_QUAN}{e.g. 'N'. See Details.}

\item{USE_MTHD3_FOR_HIST_WASTE_QUAN}{e.g. 'N'. See Details.}

\item{USE_MTHD4_FOR_HIST_WASTE_QUAN}{e.g. 'Y'. See Details.}
}
\description{
Retrieve tt hist waste method data from ghg database
}

