% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_tt_landfill_details.R
\name{ghg_tt_landfill_details}
\alias{ghg_tt_landfill_details}
\title{Retrieve tt landfill details data from ghg database}
\usage{
ghg_tt_landfill_details(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, LANDFILL_OPEN = NULL,
  LANDFILL_GAS_COLL_SYSTEM_INDIC = NULL,
  LEACHATE_RECIRCULATION_INDICA = NULL,
  LEACHATE_RECIRCULATION_FREQ = NULL,
  IS_MCF_VALUE_DEFAULT_INDICATOR = NULL,
  PASSIVE_VENT_FLARE_INDICATOR = NULL, COVER_TYPE_NAME = NULL,
  LF_SURFACEAREA_BY_COVERTYPE = NULL,
  ESTIMATED_YEAR_OF_LF_CLOSURE = NULL,
  LAST_YEAR_LF_RECEIVED_WASTE = NULL, MCF_VALUE = NULL,
  AERATION_BLOWER_CAPACITY = NULL,
  LANDFILL_FRAC_AFFECT_AERATION = NULL,
  AERATION_BLOWER_OPERATING_HRS = NULL,
  FACTORS_USED_AS_BASIS_FOR_MCF = NULL,
  AERATION_SYSTEM_DESCRIPTION = NULL,
  ANNL_METHANE_EMISSIONS_EQ_TT6 = NULL, NUMBER_OF_WASTE_STREAMS = NULL,
  FRACTION_METHANE_IN_LF_GAS = NULL,
  FRCT_METHANE_LF_GAS_DTRM_IND = NULL,
  STARTINGYEARFORACCEPTINGWASTE = NULL,
  DOESFACKNOWWHENLANDFILLOPENED = NULL,
  ISWASTEQUANTITYDATAAVAILIND = NULL, LANDFILLCAPACITY = NULL,
  LANDFILLFIRSTRECEIVEDWASTE = NULL,
  PRIORYRWASTEDISPDATAFIRSTAVAIL = NULL, YEARLASTRECEIVEDWASTE = NULL,
  ANNUALMODELEDCH4GENTT1 = NULL, ANNUAL_VOL_LANDFILL_GAS_FLOW = NULL,
  ANNUAL_AVERAGE_METHANE_CONCEN = NULL,
  ON_SITE_DESTRUCTION_INDICATOR = NULL,
  GAS_COLL_SYSTEM_MANUFACTURER = NULL, GAS_COLL_SYSTEM_CAPACITY = NULL,
  NUMBER_OF_WELLS = NULL, MEAS_METHANE_GEN_EQ_HH7 = NULL,
  BASIS_FOR_INPUT_METH_GEN_VALUE = NULL,
  METHANE_EMISSION_EQ_HH6 = NULL, METHANE_EMISSION_EQ_HH8 = NULL,
  GAS_COLL_SYSTEM_ANN_OPER_HRS = NULL,
  ANNUAL_QTY_RECOV_METHANE_HH4 = NULL,
  EST_GAS_COLLECTION_EFFICIENCY = NULL,
  IS_LFG_FLOW_WET_BASIS_INDICATR = NULL,
  IS_CH4_CONC_WET_BASIS_INDICATR = NULL,
  IS_PRESSURE_INCORP_INDICATOR = NULL,
  IS_TEMPERATURE_INCORP_INDICATR = NULL,
  MODELED_METHANE_GEN_EQ_HH5 = NULL,
  BACKUP_DEVICE_ANNUAL_OP_HOURS = NULL,
  DESTRUCTION_EFFICIENCY_PERCENT = NULL,
  BACKUP_DESTRUCTION_DEVICE = NULL,
  PRIMARY_DEVICE_ANNUAL_OP_HOURS = NULL,
  ANN_AVG_METHANE_CONC_MONTHS = NULL, ANN_AVG_METHANE_CONC_DAYS = NULL,
  ANN_VOL_LFGFLOW_DAYS = NULL, METHANE_OXIDATION_FRACTION_HH6 = NULL,
  METHANE_OXIDATION_FRACTION_HH7 = NULL,
  METHANE_OXIDATION_FRACTION_HH8 = NULL,
  METHANE_OXIDATION_FRACTION_TT6 = NULL,
  NGMETH_OXIDATION_FRACTION_TT6 = NULL,
  CUMULATIVE_WASTE_QUANTITY = NULL,
  DOES_FAC_KNW_WHN_LANDFILL_OPEN = NULL,
  IS_WASTE_IN_PLACE_QUAN_KNW_IND = NULL,
  LANDFILL_DESIGN_CAPACITY = NULL,
  LANDFILL_RECVD_WASTE_YEAR_LAST = NULL, LANDFILL_YEAR_OPEN = NULL,
  QUANTITY_AVAILABLE_YEARS = NULL, WASTE_IN_PLACE_QUANTITY = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000093'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{FACILITY_NAME}{e.g. 'MOMENTIVE PERFORMANCE MATERIALS'. See Details.}

\item{LANDFILL_OPEN}{e.g. 'N'. See Details.}

\item{LANDFILL_GAS_COLL_SYSTEM_INDIC}{e.g. 'N'. See Details.}

\item{LEACHATE_RECIRCULATION_INDICA}{e.g. 'N'. See Details.}

\item{LEACHATE_RECIRCULATION_FREQ}{e.g. 'NOT USED FOR THE PAST TEN (10) YEARS'. See Details.}

\item{IS_MCF_VALUE_DEFAULT_INDICATOR}{e.g. 'N'. See Details.}

\item{PASSIVE_VENT_FLARE_INDICATOR}{e.g. 'Y'. See Details.}

\item{COVER_TYPE_NAME}{e.g. 'SAND COVER'. See Details.}

\item{LF_SURFACEAREA_BY_COVERTYPE}{e.g. '4550'. See Details.}

\item{ESTIMATED_YEAR_OF_LF_CLOSURE}{e.g. 'NA'. See Details.}

\item{LAST_YEAR_LF_RECEIVED_WASTE}{e.g. '1989'. See Details.}

\item{MCF_VALUE}{e.g. 'NA'. See Details.}

\item{AERATION_BLOWER_CAPACITY}{e.g. 'NA'. See Details.}

\item{LANDFILL_FRAC_AFFECT_AERATION}{e.g. 'NA'. See Details.}

\item{AERATION_BLOWER_OPERATING_HRS}{e.g. 'NA'. See Details.}

\item{FACTORS_USED_AS_BASIS_FOR_MCF}{e.g. 'NA'. See Details.}

\item{AERATION_SYSTEM_DESCRIPTION}{e.g. 'NA'. See Details.}

\item{ANNL_METHANE_EMISSIONS_EQ_TT6}{e.g. '57.64'. See Details.}

\item{NUMBER_OF_WASTE_STREAMS}{e.g. '1'. See Details.}

\item{FRACTION_METHANE_IN_LF_GAS}{e.g. 'NA'. See Details.}

\item{FRCT_METHANE_LF_GAS_DTRM_IND}{e.g. 'DEFA'. See Details.}

\item{STARTINGYEARFORACCEPTINGWASTE}{e.g. '1972'. See Details.}

\item{DOESFACKNOWWHENLANDFILLOPENED}{e.g. 'NA'. See Details.}

\item{ISWASTEQUANTITYDATAAVAILIND}{e.g. 'NA'. See Details.}

\item{LANDFILLCAPACITY}{e.g. 'NA'. See Details.}

\item{LANDFILLFIRSTRECEIVEDWASTE}{e.g. 'NA'. See Details.}

\item{PRIORYRWASTEDISPDATAFIRSTAVAIL}{e.g. 'NA'. See Details.}

\item{YEARLASTRECEIVEDWASTE}{e.g. 'NA'. See Details.}

\item{ANNUALMODELEDCH4GENTT1}{e.g. '64.05'. See Details.}

\item{ANNUAL_VOL_LANDFILL_GAS_FLOW}{e.g. 'NA'. See Details.}

\item{ANNUAL_AVERAGE_METHANE_CONCEN}{e.g. 'NA'. See Details.}

\item{ON_SITE_DESTRUCTION_INDICATOR}{e.g. 'NA'. See Details.}

\item{GAS_COLL_SYSTEM_MANUFACTURER}{e.g. 'NA'. See Details.}

\item{GAS_COLL_SYSTEM_CAPACITY}{e.g. 'NA'. See Details.}

\item{NUMBER_OF_WELLS}{e.g. 'NA'. See Details.}

\item{MEAS_METHANE_GEN_EQ_HH7}{e.g. 'NA'. See Details.}

\item{BASIS_FOR_INPUT_METH_GEN_VALUE}{e.g. 'NA'. See Details.}

\item{METHANE_EMISSION_EQ_HH6}{e.g. 'NA'. See Details.}

\item{METHANE_EMISSION_EQ_HH8}{e.g. 'NA'. See Details.}

\item{GAS_COLL_SYSTEM_ANN_OPER_HRS}{e.g. 'NA'. See Details.}

\item{ANNUAL_QTY_RECOV_METHANE_HH4}{e.g. 'NA'. See Details.}

\item{EST_GAS_COLLECTION_EFFICIENCY}{e.g. 'NA'. See Details.}

\item{IS_LFG_FLOW_WET_BASIS_INDICATR}{e.g. 'NA'. See Details.}

\item{IS_CH4_CONC_WET_BASIS_INDICATR}{e.g. 'NA'. See Details.}

\item{IS_PRESSURE_INCORP_INDICATOR}{e.g. 'NA'. See Details.}

\item{IS_TEMPERATURE_INCORP_INDICATR}{e.g. 'NA'. See Details.}

\item{MODELED_METHANE_GEN_EQ_HH5}{e.g. 'NA'. See Details.}

\item{BACKUP_DEVICE_ANNUAL_OP_HOURS}{e.g. 'NA'. See Details.}

\item{DESTRUCTION_EFFICIENCY_PERCENT}{e.g. 'NA'. See Details.}

\item{BACKUP_DESTRUCTION_DEVICE}{e.g. 'NA'. See Details.}

\item{PRIMARY_DEVICE_ANNUAL_OP_HOURS}{e.g. 'NA'. See Details.}

\item{ANN_AVG_METHANE_CONC_MONTHS}{e.g. 'NA'. See Details.}

\item{ANN_AVG_METHANE_CONC_DAYS}{e.g. 'NA'. See Details.}

\item{ANN_VOL_LFGFLOW_DAYS}{e.g. 'NA'. See Details.}

\item{METHANE_OXIDATION_FRACTION_HH6}{e.g. 'NA'. See Details.}

\item{METHANE_OXIDATION_FRACTION_HH7}{e.g. 'NA'. See Details.}

\item{METHANE_OXIDATION_FRACTION_HH8}{e.g. 'NA'. See Details.}

\item{METHANE_OXIDATION_FRACTION_TT6}{e.g. 'NA'. See Details.}

\item{NGMETH_OXIDATION_FRACTION_TT6}{e.g. 'NA'. See Details.}

\item{CUMULATIVE_WASTE_QUANTITY}{e.g. 'NA'. See Details.}

\item{DOES_FAC_KNW_WHN_LANDFILL_OPEN}{e.g. 'NA'. See Details.}

\item{IS_WASTE_IN_PLACE_QUAN_KNW_IND}{e.g. 'NA'. See Details.}

\item{LANDFILL_DESIGN_CAPACITY}{e.g. 'NA'. See Details.}

\item{LANDFILL_RECVD_WASTE_YEAR_LAST}{e.g. 'NA'. See Details.}

\item{LANDFILL_YEAR_OPEN}{e.g. 'NA'. See Details.}

\item{QUANTITY_AVAILABLE_YEARS}{e.g. 'NA'. See Details.}

\item{WASTE_IN_PLACE_QUANTITY}{e.g. 'NA'. See Details.}
}
\description{
Retrieve tt landfill details data from ghg database
}
