% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_tt_lf_gas_coll_details.R
\name{ghg_tt_lf_gas_coll_details}
\alias{ghg_tt_lf_gas_coll_details}
\title{Retrieve tt lf gas coll details data from ghg database}
\usage{
ghg_tt_lf_gas_coll_details(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, MONTH_NAME = NULL, AVERAGE_TEMPERATURE = NULL,
  AVERAGE_PRESSURE = NULL, AVERAGE_MOISTURE_CONTENT = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1007453'. See Details.}

\item{REPORTING_YEAR}{e.g. '2014'. See Details.}

\item{FACILITY_NAME}{e.g. 'GEORGIA-PACIFIC CONSUMER PRODUCTS LP - SAVANNAH RIVER MILL'. See Details.}

\item{MONTH_NAME}{e.g. 'JANUARY'. See Details.}

\item{AVERAGE_TEMPERATURE}{e.g. 'NA'. See Details.}

\item{AVERAGE_PRESSURE}{e.g. 'NA'. See Details.}

\item{AVERAGE_MOISTURE_CONTENT}{e.g. '0.023'. See Details.}
}
\description{
Retrieve tt lf gas coll details data from ghg database
}

