% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_tt_waste_depth_details.R
\name{ghg_tt_waste_depth_details}
\alias{ghg_tt_waste_depth_details}
\title{Retrieve tt waste depth details data from ghg database}
\usage{
ghg_tt_waste_depth_details(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, AREA_TYPE = NULL, SURFACE_AREA = NULL,
  WASTE_DEPTH = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1007453'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{FACILITY_NAME}{e.g. 'GEORGIA-PACIFIC CONSUMER PRODUCTS LP - SAVANNAH RIVER MILL'. See Details.}

\item{AREA_TYPE}{e.g. 'A1'. See Details.}

\item{SURFACE_AREA}{e.g. '437384'. See Details.}

\item{WASTE_DEPTH}{e.g. '0'. See Details.}
}
\description{
Retrieve tt waste depth details data from ghg database
}
