% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_tt_wastestream_detls.R
\name{ghg_tt_wastestream_detls}
\alias{ghg_tt_wastestream_detls}
\title{Retrieve tt wastestream detls data from ghg database}
\usage{
ghg_tt_wastestream_detls(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, WASTE_STREAM_NAME = NULL,
  WASTE_STREAM_DESCRIPTION = NULL, WASTE_STREAM_TYPE = NULL,
  ANAEROBIC_BIODEG_TEST_METHOD = NULL,
  OTHER_ANAEROB_BIODEG_TEST_METH = NULL,
  RANGE_YRS_DISP_PRD_DATA_USED = NULL,
  WASTE_STREAM_FIRST_YR_IN_LF = NULL,
  WASTE_STREAM_LAST_YR_IN_LF = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000093'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{FACILITY_NAME}{e.g. 'MOMENTIVE PERFORMANCE MATERIALS'. See Details.}

\item{WASTE_STREAM_NAME}{e.g. 'LANDFILL 1'. See Details.}

\item{WASTE_STREAM_DESCRIPTION}{e.g. 'DEWATERED SOLIDS FROM THE WASTEWATER TREATMENT PLANT, SPENT SILICON-COPPER SLURRY, INCINERATOR RESID'. See Details.}

\item{WASTE_STREAM_TYPE}{e.g. 'OTHER INDUSTRIAL SOLID WASTE'. See Details.}

\item{ANAEROBIC_BIODEG_TEST_METHOD}{e.g. 'NA'. See Details.}

\item{OTHER_ANAEROB_BIODEG_TEST_METH}{e.g. 'NA'. See Details.}

\item{RANGE_YRS_DISP_PRD_DATA_USED}{e.g. 'NA'. See Details.}

\item{WASTE_STREAM_FIRST_YR_IN_LF}{e.g. '1972'. See Details.}

\item{WASTE_STREAM_LAST_YR_IN_LF}{e.g. '1989'. See Details.}
}
\description{
Retrieve tt wastestream detls data from ghg database
}
