% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_u_calcination_frac_methods.R
\name{ghg_u_calcination_frac_methods}
\alias{ghg_u_calcination_frac_methods}
\title{Retrieve u calcination frac methods data from ghg database}
\usage{
ghg_u_calcination_frac_methods(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  OTHER_CARBONATE_MASS_MEAS_MTHD = NULL,
  OTHER_CALCINAT_FRAC_DETRM_MTHD = NULL, SUB_MONTHLY_MASS_CONSUMED = NULL,
  CARBONATE_MASS_MEASR_MTHD = NULL, CALCINATION_FRAC_DETRM_MTHD = NULL,
  FACILITY_NAME = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1001072'. See Details.}

\item{REPORTING_YEAR}{e.g. '2013'. See Details.}

\item{OTHER_CARBONATE_MASS_MEAS_MTHD}{e.g. 'NA'. See Details.}

\item{OTHER_CALCINAT_FRAC_DETRM_MTHD}{e.g. 'NA'. See Details.}

\item{SUB_MONTHLY_MASS_CONSUMED}{e.g. '0'. See Details.}

\item{CARBONATE_MASS_MEASR_MTHD}{e.g. 'PURCHASE RECORDS'. See Details.}

\item{CALCINATION_FRAC_DETRM_MTHD}{e.g. 'DEFAULT OF 1.0'. See Details.}

\item{FACILITY_NAME}{e.g. 'OCCIDENTAL CHEMICAL DALLAS SILICATE PLANT'. See Details.}
}
\description{
Retrieve u calcination frac methods data from ghg database
}

