% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_u_info.R
\name{ghg_u_info}
\alias{ghg_u_info}
\title{Retrieve u info data from ghg database}
\usage{
ghg_u_info(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  EMISSIONS_CALC_METHOD = NULL, FACILITY_NAME = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1001072'. See Details.}

\item{REPORTING_YEAR}{e.g. '2010'. See Details.}

\item{EMISSIONS_CALC_METHOD}{e.g. 'NA'. See Details.}

\item{FACILITY_NAME}{e.g. 'OCCIDENTAL CHEMICAL DALLAS SILICATE PLANT'. See Details.}
}
\description{
Retrieve u info data from ghg database
}

