% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_u_mass_methods.R
\name{ghg_u_mass_methods}
\alias{ghg_u_mass_methods}
\title{Retrieve u mass methods data from ghg database}
\usage{
ghg_u_mass_methods(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  OTHR_CARBONATE_MASS_MEASR_MTHD = NULL,
  SUB_MTHLY_MASS_CARBONAT_OUTPUT = NULL,
  SUB_MTHLY_MASS_CARBONAT_INPUT = NULL, CARBONATE_MASS_MEASR_MTHD = NULL,
  FACILITY_NAME = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1001755'. See Details.}

\item{REPORTING_YEAR}{e.g. '2013'. See Details.}

\item{OTHR_CARBONATE_MASS_MEASR_MTHD}{e.g. 'NA'. See Details.}

\item{SUB_MTHLY_MASS_CARBONAT_OUTPUT}{e.g. '0'. See Details.}

\item{SUB_MTHLY_MASS_CARBONAT_INPUT}{e.g. '0'. See Details.}

\item{CARBONATE_MASS_MEASR_MTHD}{e.g. 'PURCHASE RECORDS'. See Details.}

\item{FACILITY_NAME}{e.g. 'DEFIANCE CASTING OPERATIONS (0320010001)'. See Details.}
}
\description{
Retrieve u mass methods data from ghg database
}

