% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_u_subpart_level_information.R
\name{ghg_u_subpart_level_information}
\alias{ghg_u_subpart_level_information}
\title{Retrieve u subpart level information data from ghg database}
\usage{
ghg_u_subpart_level_information(FACILITY_ID = NULL,
  REPORTING_YEAR = NULL, GHG_QUANTITY = NULL, GHG_NAME = NULL,
  FACILITY_NAME = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1001072'. See Details.}

\item{REPORTING_YEAR}{e.g. '2010'. See Details.}

\item{GHG_QUANTITY}{e.g. '5825.1'. See Details.}

\item{GHG_NAME}{e.g. 'CARBON DIOXIDE'. See Details.}

\item{FACILITY_NAME}{e.g. 'OCCIDENTAL CHEMICAL DALLAS SILICATE PLANT'. See Details.}
}
\description{
Retrieve u subpart level information data from ghg database
}
