% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_uu_inj_co2_facility.R
\name{ghg_uu_inj_co2_facility}
\alias{ghg_uu_inj_co2_facility}
\title{Retrieve uu inj co2 facility data from ghg database}
\usage{
ghg_uu_inj_co2_facility(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  CO2_SOURCES = NULL, FACILITY_NAME = NULL,
  CO2_SRC_SPEC_ANTHROPOGENIC_SRC = NULL,
  PURPOSE_OF_CO2_INJECTION = NULL,
  PURPOSE_OF_CO2_INJECTION_OTHER = NULL,
  PURP_OF_CO2_INJEC_END_DATE = NULL,
  PURP_OF_CO2_INJEC_START_DATE = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1001052'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{CO2_SOURCES}{e.g. 'ELECTRIC GENERATING UNIT'. See Details.}

\item{FACILITY_NAME}{e.g. 'MOUNTAINEER (1301)'. See Details.}

\item{CO2_SRC_SPEC_ANTHROPOGENIC_SRC}{e.g. 'NA'. See Details.}

\item{PURPOSE_OF_CO2_INJECTION}{e.g. 'NA'. See Details.}

\item{PURPOSE_OF_CO2_INJECTION_OTHER}{e.g. 'NA'. See Details.}

\item{PURP_OF_CO2_INJEC_END_DATE}{e.g. 'NA'. See Details.}

\item{PURP_OF_CO2_INJEC_START_DATE}{e.g. 'NA'. See Details.}
}
\description{
Retrieve uu inj co2 facility data from ghg database
}
