% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_uu_inj_co2_unit.R
\name{ghg_uu_inj_co2_unit}
\alias{ghg_uu_inj_co2_unit}
\title{Retrieve uu inj co2 unit data from ghg database}
\usage{
ghg_uu_inj_co2_unit(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  MASS_CONT_Q1_QTY_SUB_PROCD_USE = NULL,
  MASS_CONT_Q3_QUANTITY_METHOD = NULL,
  MASS_CONT_Q3_QTY_METHOD_OTHER = NULL,
  MASS_CONT_Q4_CO2_CONC_METHOD = NULL,
  MASS_CONT_Q4_CO2_CON_MTHD_OTHR = NULL,
  MASS_CONT_Q4_QTY_METHOD_OTHER = NULL,
  MASS_FLOW_Q1_CO2_CONC_METHOD = NULL,
  MASS_FLOW_Q1_CO2_CON_MTHD_OTHR = NULL,
  MASS_FLOW_Q1_CO2_SUB_PROCD_USE = NULL,
  MASS_FLOW_Q1_QTY_REDELVRD_MTHD = NULL,
  MASS_FLOW_Q1_QUANTITY_METHOD = NULL,
  MASS_FLOW_Q1_QTY_METHOD_OTHER = NULL,
  MASS_FLOW_Q2_CO2_CONC_METHOD = NULL,
  MASS_CONT_Q4_CO2_SUB_PROCD_USE = NULL,
  MASS_CONT_Q4_QTY_REDELVRD_SUB = NULL,
  MASS_CONT_Q4_QUANTITY_METHOD = NULL,
  MASS_CONT_Q3_CO2_SUB_PROCD_USE = NULL,
  MASS_CONT_Q3_QTY_REDELVRD_SUB = NULL,
  MASS_CONT_Q2_CO2_CONC_METHOD = NULL,
  MASS_CONT_Q2_CO2_CON_MTHD_OTHR = NULL,
  MASS_CONT_Q2_CO2_SUB_PROCD_USE = NULL,
  MASS_CONT_Q2_QTY_REDELVRD_SUB = NULL,
  MASS_CONT_Q2_QUANTITY_METHOD = NULL,
  MASS_CONT_Q2_QTY_METHOD_OTHER = NULL,
  MASS_CONT_Q3_CO2_CONC_METHOD = NULL,
  MASS_CONT_Q3_CO2_CON_MTHD_OTHR = NULL,
  VOL_FLOW_Q2_CO2_CONC_MTHD_OTHR = NULL,
  VOL_FLOW_Q2_CO2_SUB_PROCD_USED = NULL,
  VOL_FLOW_Q2_QTY_REDELVRD_MTHD = NULL,
  VOL_FLOW_Q2_QUANTITY_METHOD = NULL,
  VOL_FLOW_Q2_QTY_METHOD_OTHER = NULL,
  VOL_FLOW_Q3_CO2_CONC_METHOD = NULL,
  VOL_FLOW_Q3_CO2_CONC_MTHD_OTHR = NULL,
  VOL_FLOW_Q3_CO2_SUB_PROCD_USED = NULL,
  VOL_FLOW_Q3_QTY_REDELVD_METHOD = NULL,
  MASS_CONT_Q1_QTY_METHOD_OTHER = NULL,
  VOL_FLOW_Q3_QUANTITY_METHOD = NULL,
  VOL_FLOW_Q3_QTY_METHOD_OTHER = NULL,
  VOL_FLOW_Q4_CO2_CONC_METHOD = NULL,
  VOL_FLOW_Q4_CO2_CONC_MTHD_OTHR = NULL,
  VOL_FLOW_Q4_CO2_SUB_PROCD_USED = NULL,
  VOL_FLOW_Q4_QTY_REDELVD_METHOD = NULL,
  VOL_FLOW_Q4_QUANTITY_METHOD = NULL,
  VOL_FLOW_Q4_QTY_METHOD_OTHER = NULL,
  MASS_CONT_Q2_QTY_SUB_PROCD_USE = NULL,
  MASS_CONT_Q3_QTY_SUB_PROCD_USE = NULL,
  MASS_CONT_Q4_QTY_SUB_PROCD_USE = NULL,
  MASS_FLW_Q1_QTY_REDV_MTHD_OTH = NULL,
  MASS_FLW_Q1_QTY_REDV_SUB_PRCS = NULL,
  MASS_FLOW_Q1_QTY_SUB_PROCD_USE = NULL,
  MASS_FLW_Q2_QTY_REDV_MTHD_OTH = NULL,
  MASS_FLW_Q2_QTY_REDV_SUB_PRCS = NULL,
  MASS_FLW_Q3_QTY_REDV_MTHD_OTH = NULL,
  MASS_FLOW_Q2_QTY_SUB_PROCD_USE = NULL,
  MASS_FLW_Q3_QTY_REDV_SUB_PRCS = NULL,
  MASS_FLOW_Q3_QTY_SUB_PROCD_USE = NULL,
  MASS_FLW_Q4_QTY_REDV_MTHD_OTH = NULL,
  MASS_FLW_Q4_QTY_REDV_SUB_PRCS = NULL,
  MASS_FLOW_Q4_QTY_SUB_PROCD_USE = NULL,
  VOL_CONT_Q1_QTY_SUB_PROCD_USE = NULL,
  VOL_CONT_Q2_QTY_SUB_PROCD_USE = NULL,
  VOL_CONT_Q3_QTY_SUB_PROCD_USE = NULL,
  VOL_CONT_Q4_QTY_SUB_PROCD_USE = NULL,
  VOL_FLOW_Q1_QTY_REDEL_MTHD_OTH = NULL,
  VOL_FLW_Q1_QTY_REDV_SUB_PR_USE = NULL,
  VOL_FLOW_Q1_QTY_SUB_PROCD = NULL,
  VOL_FLOW_Q2_QTY_REDEL_MTHD_OTH = NULL,
  VOL_FLW_Q2_QTY_REDV_SUB_PR_USE = NULL,
  VOL_FLOW_Q2_QTY_SUB_PROCD = NULL,
  VOL_FLOW_Q3_CO2_RED_PROCD_USED = NULL,
  VOL_FLOW_Q3_QTY_REDEL_MTHD_OTH = NULL,
  VOL_FLW_Q3_QTY_REDV_SUB_PR_USE = NULL,
  VOL_FLOW_Q4_QTY_SUB_PROCD = NULL,
  VOL_FLOW_Q4_QTY_REDEL_MTHD_OTH = NULL,
  VOL_FLW_Q4_QTY_REDV_SUB_PR_USE = NULL,
  MASS_FLOW_Q2_CO2_CON_MTHD_OTHR = NULL,
  MASS_FLOW_Q2_CO2_SUB_PROCD_USE = NULL,
  MASS_FLOW_Q2_QTY_REDELVRD_MTHD = NULL,
  MASS_FLOW_Q2_QUANTITY_METHOD = NULL,
  MASS_FLOW_Q2_QTY_METHOD_OTHER = NULL,
  MASS_FLOW_Q3_CO2_CONC_METHOD = NULL,
  MASS_FLW_Q3_CO2_CONC_MTHD_OTH = NULL,
  MASS_FLOW_Q3_CO2_SUB_PROCD_USE = NULL,
  MASS_FLOW_Q3_QTY_REDELVRD_MTHD = NULL,
  MASS_FLOW_Q3_QUANTITY_METHOD = NULL,
  MASS_FLOW_Q3_QTY_METHOD_OTHER = NULL,
  MASS_FLOW_Q4_CO2_CONC_METHOD = NULL,
  MASS_FLOW_Q4_CO2_CON_MTHD_OTHR = NULL,
  MASS_FLOW_Q4_CO2_SUB_PROCD_USE = NULL,
  MASS_FLOW_Q4_QTY_REDELVRD_MTHD = NULL,
  MASS_FLOW_Q4_QUANTITY_METHOD = NULL,
  MASS_FLOW_Q4_QTY_METHOD_OTHER = NULL, UNIT_NAME = NULL,
  UNIT_TYPE = NULL, VOL_CONT_Q1_CO2_CONC_METHOD = NULL,
  VOL_CONT_Q1_CO2_CONC_MTHD_OTH = NULL,
  VOL_CONT_Q1_CO2_SUB_PROCD_USED = NULL,
  VOL_CONT_Q1_QTY_REDELVRD_SUB = NULL,
  VOL_CONT_Q1_QUANTITY_METHOD = NULL,
  VOL_CONT_Q1_QTY_METHOD_OTHER = NULL,
  VOL_CONT_Q2_CO2_CONC_METHOD = NULL,
  VOL_CONT_Q2_CO2_CONC_MTHD_OTH = NULL,
  VOL_CONT_Q2_CO2_SUB_PROCD_USED = NULL,
  VOL_CONT_Q2_QTY_REDELVRD_SUB = NULL,
  VOL_CONT_Q2_QUANTITY_METHOD = NULL,
  VOL_CONT_Q2_QTY_METHOD_OTHER = NULL,
  VOL_CONT_Q3_CO2_CONC_METHOD = NULL,
  VOL_CONT_Q3_CO2_CONC_MTHD_OTHR = NULL,
  VOL_CONT_Q3_CO2_SUB_PROCD_USED = NULL,
  VOL_CONT_Q3_QTY_REDELVRD_SUB = NULL,
  VOL_CONT_Q3_QUANTITY_METHOD = NULL,
  VOL_CONT_Q3_QTY_METHOD_OTHER = NULL,
  VOL_CONT_Q4_CO2_CONC_METHOD = NULL,
  VOL_CONT_Q4_CO2_CONC_MTHD_OTHR = NULL,
  VOL_CONT_Q4_CO2_SUB_PROCD_USED = NULL,
  VOL_CONT_Q4_QTY_REDELVRD_SUB = NULL,
  VOL_CONT_Q4_QUANTITY_METHOD = NULL,
  VOL_CONT_Q4_QTY_METHOD_OTHER = NULL,
  VOL_FLOW_Q1_CO2_CONC_METHOD = NULL,
  VOL_FLOW_Q1_CO2_CONC_MTHD_OTHR = NULL,
  VOL_FLOW_Q1_CO2_SUB_PROCD_USED = NULL,
  VOL_FLOW_Q1_QTY_REDELVRD_MTHD = NULL,
  VOL_FLOW_Q1_QUANTITY_METHOD = NULL,
  VOL_FLOW_Q1_QTY_METHOD_OTHER = NULL,
  VOL_FLOW_Q2_CO2_CONC_METHOD = NULL, FACILITY_NAME = NULL,
  MASS_CONT_Q1_CO2_CONC_METHOD = NULL,
  MASS_CONT_Q1_CO2_CON_MTHD_OTHR = NULL,
  MASS_CONT_Q1_CO2_SUB_PROCD_USE = NULL,
  MASS_CONT_Q1_QTY_REDELVRD_SUB = NULL,
  MASS_CONT_Q1_QUANTITY_METHOD = NULL,
  VOL_CONT_Q2_QTY_RED_MTHD_OTHR = NULL,
  VOL_CONT_Q3_QTY_REDEL_METHOD = NULL,
  VOL_CONT_Q3_QTY_RED_MTHD_OTHR = NULL,
  VOL_CONT_Q4_QTY_RED_MTHD = NULL,
  VOL_CONT_Q4_QTY_RED_MTHD_OTHR = NULL,
  MASS_CONT_Q1_QTY_REDEL_MTHD = NULL,
  MASS_CONT_Q1_QTY_RED_MTHD_OTH = NULL,
  MASS_CONT_Q2_QTY_RED_METHOD = NULL,
  MASS_CONT_Q2_QTY_RED_MTHD_OTH = NULL,
  MASS_CONT_Q3_QTY_REDEL_METHOD = NULL,
  MASS_CONT_Q3_QTY_RED_MTHD_OTH = NULL,
  MASS_CONT_Q4_QTY_REDEL_METHOD = NULL,
  MASS_CONT_Q4_QTY_RED_MTHD_OTH = NULL,
  VOL_CONT_Q1_QTY_REDEL_METHOD = NULL,
  VOL_CONT_Q1_QTY_RED_MTHD_OTHR = NULL,
  VOL_CONT_Q2_QTY_REDEL_METHOD = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1001052'. See Details.}

\item{REPORTING_YEAR}{e.g. '2014'. See Details.}

\item{MASS_CONT_Q1_QTY_SUB_PROCD_USE}{e.g. 'NA'. See Details.}

\item{MASS_CONT_Q3_QUANTITY_METHOD}{e.g. 'NA'. See Details.}

\item{MASS_CONT_Q3_QTY_METHOD_OTHER}{e.g. 'NA'. See Details.}

\item{MASS_CONT_Q4_CO2_CONC_METHOD}{e.g. 'NA'. See Details.}

\item{MASS_CONT_Q4_CO2_CON_MTHD_OTHR}{e.g. 'NA'. See Details.}

\item{MASS_CONT_Q4_QTY_METHOD_OTHER}{e.g. 'NA'. See Details.}

\item{MASS_FLOW_Q1_CO2_CONC_METHOD}{e.g. 'OTHER (SPECIFY)'. See Details.}

\item{MASS_FLOW_Q1_CO2_CON_MTHD_OTHR}{e.g. 'NA'. See Details.}

\item{MASS_FLOW_Q1_CO2_SUB_PROCD_USE}{e.g. 'N'. See Details.}

\item{MASS_FLOW_Q1_QTY_REDELVRD_MTHD}{e.g. 'OTHER (SPECIFY)'. See Details.}

\item{MASS_FLOW_Q1_QUANTITY_METHOD}{e.g. 'OTHER (SPECIFY)'. See Details.}

\item{MASS_FLOW_Q1_QTY_METHOD_OTHER}{e.g. 'NA'. See Details.}

\item{MASS_FLOW_Q2_CO2_CONC_METHOD}{e.g. 'OTHER (SPECIFY)'. See Details.}

\item{MASS_CONT_Q4_CO2_SUB_PROCD_USE}{e.g. 'NA'. See Details.}

\item{MASS_CONT_Q4_QTY_REDELVRD_SUB}{e.g. 'NA'. See Details.}

\item{MASS_CONT_Q4_QUANTITY_METHOD}{e.g. 'NA'. See Details.}

\item{MASS_CONT_Q3_CO2_SUB_PROCD_USE}{e.g. 'NA'. See Details.}

\item{MASS_CONT_Q3_QTY_REDELVRD_SUB}{e.g. 'NA'. See Details.}

\item{MASS_CONT_Q2_CO2_CONC_METHOD}{e.g. 'NA'. See Details.}

\item{MASS_CONT_Q2_CO2_CON_MTHD_OTHR}{e.g. 'NA'. See Details.}

\item{MASS_CONT_Q2_CO2_SUB_PROCD_USE}{e.g. 'NA'. See Details.}

\item{MASS_CONT_Q2_QTY_REDELVRD_SUB}{e.g. 'NA'. See Details.}

\item{MASS_CONT_Q2_QUANTITY_METHOD}{e.g. 'NA'. See Details.}

\item{MASS_CONT_Q2_QTY_METHOD_OTHER}{e.g. 'NA'. See Details.}

\item{MASS_CONT_Q3_CO2_CONC_METHOD}{e.g. 'NA'. See Details.}

\item{MASS_CONT_Q3_CO2_CON_MTHD_OTHR}{e.g. 'NA'. See Details.}

\item{VOL_FLOW_Q2_CO2_CONC_MTHD_OTHR}{e.g. 'NA'. See Details.}

\item{VOL_FLOW_Q2_CO2_SUB_PROCD_USED}{e.g. 'NA'. See Details.}

\item{VOL_FLOW_Q2_QTY_REDELVRD_MTHD}{e.g. 'NA'. See Details.}

\item{VOL_FLOW_Q2_QUANTITY_METHOD}{e.g. 'NA'. See Details.}

\item{VOL_FLOW_Q2_QTY_METHOD_OTHER}{e.g. 'NA'. See Details.}

\item{VOL_FLOW_Q3_CO2_CONC_METHOD}{e.g. 'NA'. See Details.}

\item{VOL_FLOW_Q3_CO2_CONC_MTHD_OTHR}{e.g. 'NA'. See Details.}

\item{VOL_FLOW_Q3_CO2_SUB_PROCD_USED}{e.g. 'NA'. See Details.}

\item{VOL_FLOW_Q3_QTY_REDELVD_METHOD}{e.g. 'NA'. See Details.}

\item{MASS_CONT_Q1_QTY_METHOD_OTHER}{e.g. 'NA'. See Details.}

\item{VOL_FLOW_Q3_QUANTITY_METHOD}{e.g. 'NA'. See Details.}

\item{VOL_FLOW_Q3_QTY_METHOD_OTHER}{e.g. 'NA'. See Details.}

\item{VOL_FLOW_Q4_CO2_CONC_METHOD}{e.g. 'NA'. See Details.}

\item{VOL_FLOW_Q4_CO2_CONC_MTHD_OTHR}{e.g. 'NA'. See Details.}

\item{VOL_FLOW_Q4_CO2_SUB_PROCD_USED}{e.g. 'NA'. See Details.}

\item{VOL_FLOW_Q4_QTY_REDELVD_METHOD}{e.g. 'NA'. See Details.}

\item{VOL_FLOW_Q4_QUANTITY_METHOD}{e.g. 'NA'. See Details.}

\item{VOL_FLOW_Q4_QTY_METHOD_OTHER}{e.g. 'NA'. See Details.}

\item{MASS_CONT_Q2_QTY_SUB_PROCD_USE}{e.g. 'NA'. See Details.}

\item{MASS_CONT_Q3_QTY_SUB_PROCD_USE}{e.g. 'NA'. See Details.}

\item{MASS_CONT_Q4_QTY_SUB_PROCD_USE}{e.g. 'NA'. See Details.}

\item{MASS_FLW_Q1_QTY_REDV_MTHD_OTH}{e.g. 'NA'. See Details.}

\item{MASS_FLW_Q1_QTY_REDV_SUB_PRCS}{e.g. '0'. See Details.}

\item{MASS_FLOW_Q1_QTY_SUB_PROCD_USE}{e.g. '0'. See Details.}

\item{MASS_FLW_Q2_QTY_REDV_MTHD_OTH}{e.g. 'NA'. See Details.}

\item{MASS_FLW_Q2_QTY_REDV_SUB_PRCS}{e.g. '0'. See Details.}

\item{MASS_FLW_Q3_QTY_REDV_MTHD_OTH}{e.g. 'NA'. See Details.}

\item{MASS_FLOW_Q2_QTY_SUB_PROCD_USE}{e.g. '0'. See Details.}

\item{MASS_FLW_Q3_QTY_REDV_SUB_PRCS}{e.g. '0'. See Details.}

\item{MASS_FLOW_Q3_QTY_SUB_PROCD_USE}{e.g. '0'. See Details.}

\item{MASS_FLW_Q4_QTY_REDV_MTHD_OTH}{e.g. 'NA'. See Details.}

\item{MASS_FLW_Q4_QTY_REDV_SUB_PRCS}{e.g. '0'. See Details.}

\item{MASS_FLOW_Q4_QTY_SUB_PROCD_USE}{e.g. '0'. See Details.}

\item{VOL_CONT_Q1_QTY_SUB_PROCD_USE}{e.g. 'NA'. See Details.}

\item{VOL_CONT_Q2_QTY_SUB_PROCD_USE}{e.g. 'NA'. See Details.}

\item{VOL_CONT_Q3_QTY_SUB_PROCD_USE}{e.g. 'NA'. See Details.}

\item{VOL_CONT_Q4_QTY_SUB_PROCD_USE}{e.g. 'NA'. See Details.}

\item{VOL_FLOW_Q1_QTY_REDEL_MTHD_OTH}{e.g. 'NA'. See Details.}

\item{VOL_FLW_Q1_QTY_REDV_SUB_PR_USE}{e.g. 'NA'. See Details.}

\item{VOL_FLOW_Q1_QTY_SUB_PROCD}{e.g. 'NA'. See Details.}

\item{VOL_FLOW_Q2_QTY_REDEL_MTHD_OTH}{e.g. 'NA'. See Details.}

\item{VOL_FLW_Q2_QTY_REDV_SUB_PR_USE}{e.g. 'NA'. See Details.}

\item{VOL_FLOW_Q2_QTY_SUB_PROCD}{e.g. 'NA'. See Details.}

\item{VOL_FLOW_Q3_CO2_RED_PROCD_USED}{e.g. 'NA'. See Details.}

\item{VOL_FLOW_Q3_QTY_REDEL_MTHD_OTH}{e.g. 'NA'. See Details.}

\item{VOL_FLW_Q3_QTY_REDV_SUB_PR_USE}{e.g. 'NA'. See Details.}

\item{VOL_FLOW_Q4_QTY_SUB_PROCD}{e.g. 'NA'. See Details.}

\item{VOL_FLOW_Q4_QTY_REDEL_MTHD_OTH}{e.g. 'NA'. See Details.}

\item{VOL_FLW_Q4_QTY_REDV_SUB_PR_USE}{e.g. 'NA'. See Details.}

\item{MASS_FLOW_Q2_CO2_CON_MTHD_OTHR}{e.g. 'NA'. See Details.}

\item{MASS_FLOW_Q2_CO2_SUB_PROCD_USE}{e.g. 'N'. See Details.}

\item{MASS_FLOW_Q2_QTY_REDELVRD_MTHD}{e.g. 'OTHER (SPECIFY)'. See Details.}

\item{MASS_FLOW_Q2_QUANTITY_METHOD}{e.g. 'OTHER (SPECIFY)'. See Details.}

\item{MASS_FLOW_Q2_QTY_METHOD_OTHER}{e.g. 'NA'. See Details.}

\item{MASS_FLOW_Q3_CO2_CONC_METHOD}{e.g. 'OTHER (SPECIFY)'. See Details.}

\item{MASS_FLW_Q3_CO2_CONC_MTHD_OTH}{e.g. 'NA'. See Details.}

\item{MASS_FLOW_Q3_CO2_SUB_PROCD_USE}{e.g. 'N'. See Details.}

\item{MASS_FLOW_Q3_QTY_REDELVRD_MTHD}{e.g. 'OTHER (SPECIFY)'. See Details.}

\item{MASS_FLOW_Q3_QUANTITY_METHOD}{e.g. 'OTHER (SPECIFY)'. See Details.}

\item{MASS_FLOW_Q3_QTY_METHOD_OTHER}{e.g. 'NA'. See Details.}

\item{MASS_FLOW_Q4_CO2_CONC_METHOD}{e.g. 'OTHER (SPECIFY)'. See Details.}

\item{MASS_FLOW_Q4_CO2_CON_MTHD_OTHR}{e.g. 'NA'. See Details.}

\item{MASS_FLOW_Q4_CO2_SUB_PROCD_USE}{e.g. 'N'. See Details.}

\item{MASS_FLOW_Q4_QTY_REDELVRD_MTHD}{e.g. 'OTHER (SPECIFY)'. See Details.}

\item{MASS_FLOW_Q4_QUANTITY_METHOD}{e.g. 'OTHER (SPECIFY)'. See Details.}

\item{MASS_FLOW_Q4_QTY_METHOD_OTHER}{e.g. 'NA'. See Details.}

\item{UNIT_NAME}{e.g. 'FIC-0981 TO COMPRESSOR'. See Details.}

\item{UNIT_TYPE}{e.g. 'FLOW METER'. See Details.}

\item{VOL_CONT_Q1_CO2_CONC_METHOD}{e.g. 'NA'. See Details.}

\item{VOL_CONT_Q1_CO2_CONC_MTHD_OTH}{e.g. 'NA'. See Details.}

\item{VOL_CONT_Q1_CO2_SUB_PROCD_USED}{e.g. 'NA'. See Details.}

\item{VOL_CONT_Q1_QTY_REDELVRD_SUB}{e.g. 'NA'. See Details.}

\item{VOL_CONT_Q1_QUANTITY_METHOD}{e.g. 'NA'. See Details.}

\item{VOL_CONT_Q1_QTY_METHOD_OTHER}{e.g. 'NA'. See Details.}

\item{VOL_CONT_Q2_CO2_CONC_METHOD}{e.g. 'NA'. See Details.}

\item{VOL_CONT_Q2_CO2_CONC_MTHD_OTH}{e.g. 'NA'. See Details.}

\item{VOL_CONT_Q2_CO2_SUB_PROCD_USED}{e.g. 'NA'. See Details.}

\item{VOL_CONT_Q2_QTY_REDELVRD_SUB}{e.g. 'NA'. See Details.}

\item{VOL_CONT_Q2_QUANTITY_METHOD}{e.g. 'NA'. See Details.}

\item{VOL_CONT_Q2_QTY_METHOD_OTHER}{e.g. 'NA'. See Details.}

\item{VOL_CONT_Q3_CO2_CONC_METHOD}{e.g. 'NA'. See Details.}

\item{VOL_CONT_Q3_CO2_CONC_MTHD_OTHR}{e.g. 'NA'. See Details.}

\item{VOL_CONT_Q3_CO2_SUB_PROCD_USED}{e.g. 'NA'. See Details.}

\item{VOL_CONT_Q3_QTY_REDELVRD_SUB}{e.g. 'NA'. See Details.}

\item{VOL_CONT_Q3_QUANTITY_METHOD}{e.g. 'NA'. See Details.}

\item{VOL_CONT_Q3_QTY_METHOD_OTHER}{e.g. 'NA'. See Details.}

\item{VOL_CONT_Q4_CO2_CONC_METHOD}{e.g. 'NA'. See Details.}

\item{VOL_CONT_Q4_CO2_CONC_MTHD_OTHR}{e.g. 'NA'. See Details.}

\item{VOL_CONT_Q4_CO2_SUB_PROCD_USED}{e.g. 'NA'. See Details.}

\item{VOL_CONT_Q4_QTY_REDELVRD_SUB}{e.g. 'NA'. See Details.}

\item{VOL_CONT_Q4_QUANTITY_METHOD}{e.g. 'NA'. See Details.}

\item{VOL_CONT_Q4_QTY_METHOD_OTHER}{e.g. 'NA'. See Details.}

\item{VOL_FLOW_Q1_CO2_CONC_METHOD}{e.g. 'NA'. See Details.}

\item{VOL_FLOW_Q1_CO2_CONC_MTHD_OTHR}{e.g. 'NA'. See Details.}

\item{VOL_FLOW_Q1_CO2_SUB_PROCD_USED}{e.g. 'NA'. See Details.}

\item{VOL_FLOW_Q1_QTY_REDELVRD_MTHD}{e.g. 'NA'. See Details.}

\item{VOL_FLOW_Q1_QUANTITY_METHOD}{e.g. 'NA'. See Details.}

\item{VOL_FLOW_Q1_QTY_METHOD_OTHER}{e.g. 'NA'. See Details.}

\item{VOL_FLOW_Q2_CO2_CONC_METHOD}{e.g. 'NA'. See Details.}

\item{FACILITY_NAME}{e.g. 'MOUNTAINEER (1301)'. See Details.}

\item{MASS_CONT_Q1_CO2_CONC_METHOD}{e.g. 'NA'. See Details.}

\item{MASS_CONT_Q1_CO2_CON_MTHD_OTHR}{e.g. 'NA'. See Details.}

\item{MASS_CONT_Q1_CO2_SUB_PROCD_USE}{e.g. 'NA'. See Details.}

\item{MASS_CONT_Q1_QTY_REDELVRD_SUB}{e.g. 'NA'. See Details.}

\item{MASS_CONT_Q1_QUANTITY_METHOD}{e.g. 'NA'. See Details.}

\item{VOL_CONT_Q2_QTY_RED_MTHD_OTHR}{e.g. 'NA'. See Details.}

\item{VOL_CONT_Q3_QTY_REDEL_METHOD}{e.g. 'NA'. See Details.}

\item{VOL_CONT_Q3_QTY_RED_MTHD_OTHR}{e.g. 'NA'. See Details.}

\item{VOL_CONT_Q4_QTY_RED_MTHD}{e.g. 'NA'. See Details.}

\item{VOL_CONT_Q4_QTY_RED_MTHD_OTHR}{e.g. 'NA'. See Details.}

\item{MASS_CONT_Q1_QTY_REDEL_MTHD}{e.g. 'NA'. See Details.}

\item{MASS_CONT_Q1_QTY_RED_MTHD_OTH}{e.g. 'NA'. See Details.}

\item{MASS_CONT_Q2_QTY_RED_METHOD}{e.g. 'NA'. See Details.}

\item{MASS_CONT_Q2_QTY_RED_MTHD_OTH}{e.g. 'NA'. See Details.}

\item{MASS_CONT_Q3_QTY_REDEL_METHOD}{e.g. 'NA'. See Details.}

\item{MASS_CONT_Q3_QTY_RED_MTHD_OTH}{e.g. 'NA'. See Details.}

\item{MASS_CONT_Q4_QTY_REDEL_METHOD}{e.g. 'NA'. See Details.}

\item{MASS_CONT_Q4_QTY_RED_MTHD_OTH}{e.g. 'NA'. See Details.}

\item{VOL_CONT_Q1_QTY_REDEL_METHOD}{e.g. 'NA'. See Details.}

\item{VOL_CONT_Q1_QTY_RED_MTHD_OTHR}{e.g. 'NA'. See Details.}

\item{VOL_CONT_Q2_QTY_REDEL_METHOD}{e.g. 'NA'. See Details.}
}
\description{
Retrieve uu inj co2 unit data from ghg database
}
