% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_v_facility_info.R
\name{ghg_v_facility_info}
\alias{ghg_v_facility_info}
\title{Retrieve v facility info data from ghg database}
\usage{
ghg_v_facility_info(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, NUM_NITRIC_ACID_TRAINS = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1001781'. See Details.}

\item{REPORTING_YEAR}{e.g. '2012'. See Details.}

\item{FACILITY_NAME}{e.g. 'INVISTA S.a r.l.'. See Details.}

\item{NUM_NITRIC_ACID_TRAINS}{e.g. '1'. See Details.}
}
\description{
Retrieve v facility info data from ghg database
}

