% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_v_nitric_acid_train.R
\name{ghg_v_nitric_acid_train}
\alias{ghg_v_nitric_acid_train}
\title{Retrieve v nitric acid train data from ghg database}
\usage{
ghg_v_nitric_acid_train(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, NITRIC_ACID_TRAIN_ID = NULL,
  NUM_MONTH_MISS_DATA_PROCS_USED = NULL,
  IS_ADMIN_APPROV_ALT_METHD_USED = NULL, ALT_METHOD_NAME = NULL,
  ALT_METHOD_DESC = NULL, ALT_METHOD_REQUEST_DATE = NULL,
  ALT_METHOD_APPROVAL_DATE = NULL,
  NUM_PERFOMANCE_TESTS_REPEATED = NULL, PERFORMANCE_TEST_ID_1 = NULL,
  PERFORMANCE_TEST_METHOD_1 = NULL, PERFORMANCE_TEST_ID_2 = NULL,
  PERFORMANCE_TEST_METHOD_2 = NULL, PERFORMANCE_TEST_ID_3 = NULL,
  PERFORMANCE_TEST_METHOD_3 = NULL,
  DOES_NITRIC_ACID_TRAIN_EXHAUST = NULL, IS_NSCR_USED = NULL,
  OTH_ABATEMENT_TECH_USED = NULL, OTH_ABATEMENT_TECH_DESC = NULL,
  ANNUAL_UNROUNDED_N2O_EMISS = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1001781'. See Details.}

\item{REPORTING_YEAR}{e.g. '2013'. See Details.}

\item{FACILITY_NAME}{e.g. 'INVISTA S.a r.l.'. See Details.}

\item{NITRIC_ACID_TRAIN_ID}{e.g. 'AOP-2'. See Details.}

\item{NUM_MONTH_MISS_DATA_PROCS_USED}{e.g. '2'. See Details.}

\item{IS_ADMIN_APPROV_ALT_METHD_USED}{e.g. 'No'. See Details.}

\item{ALT_METHOD_NAME}{e.g. 'NA'. See Details.}

\item{ALT_METHOD_DESC}{e.g. 'NA'. See Details.}

\item{ALT_METHOD_REQUEST_DATE}{e.g. 'NA'. See Details.}

\item{ALT_METHOD_APPROVAL_DATE}{e.g. 'NA'. See Details.}

\item{NUM_PERFOMANCE_TESTS_REPEATED}{e.g. '0'. See Details.}

\item{PERFORMANCE_TEST_ID_1}{e.g. '7/16/2013 & 7/17/2013 AOP GHG Emission Factor Test'. See Details.}

\item{PERFORMANCE_TEST_METHOD_1}{e.g. 'EPA Method 320'. See Details.}

\item{PERFORMANCE_TEST_ID_2}{e.g. 'NA'. See Details.}

\item{PERFORMANCE_TEST_METHOD_2}{e.g. 'NA'. See Details.}

\item{PERFORMANCE_TEST_ID_3}{e.g. 'NA'. See Details.}

\item{PERFORMANCE_TEST_METHOD_3}{e.g. 'NA'. See Details.}

\item{DOES_NITRIC_ACID_TRAIN_EXHAUST}{e.g. 'No'. See Details.}

\item{IS_NSCR_USED}{e.g. 'NA'. See Details.}

\item{OTH_ABATEMENT_TECH_USED}{e.g. 'NA'. See Details.}

\item{OTH_ABATEMENT_TECH_DESC}{e.g. 'NA'. See Details.}

\item{ANNUAL_UNROUNDED_N2O_EMISS}{e.g. '1808'. See Details.}
}
\description{
Retrieve v nitric acid train data from ghg database
}
