% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_v_subpart_level_information.R
\name{ghg_v_subpart_level_information}
\alias{ghg_v_subpart_level_information}
\title{Retrieve v subpart level information data from ghg database}
\usage{
ghg_v_subpart_level_information(FACILITY_ID = NULL,
  REPORTING_YEAR = NULL, FACILITY_NAME = NULL, GHG_NAME = NULL,
  GHG_QUANTITY = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1001781'. See Details.}

\item{REPORTING_YEAR}{e.g. '2013'. See Details.}

\item{FACILITY_NAME}{e.g. 'INVISTA S.a r.l.'. See Details.}

\item{GHG_NAME}{e.g. 'Nitrous Oxide'. See Details.}

\item{GHG_QUANTITY}{e.g. '1808'. See Details.}
}
\description{
Retrieve v subpart level information data from ghg database
}
