% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_w_acid_gas_removal_unit.R
\name{ghg_w_acid_gas_removal_unit}
\alias{ghg_w_acid_gas_removal_unit}
\title{Retrieve w acid gas removal unit data from ghg database}
\usage{
ghg_w_acid_gas_removal_unit(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, INDUSTRY_SEGMENT = NULL,
  ACID_GAS_REMOVAL_CO2_EMISSIONS = NULL,
  ACID_GAS_REMOVAL_CH4_EMM_CO2E = NULL,
  ACID_GAS_REMOVAL_N2O_EMM_CO2E = NULL,
  ACID_GAS_REMOVAL_GHG_EMM_CO2E = NULL, DOES_FAC_HAVE_AGR_UNIT = NULL,
  BAMM_INDICATOR_ACID_GAS = NULL, SUB_INDICATOR_ACID_GAS = NULL,
  BAMM_DESCRIPTION_ACID_GAS = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000064'. See Details.}

\item{REPORTING_YEAR}{e.g. '2014'. See Details.}

\item{FACILITY_NAME}{e.g. 'LONE STAR GAS LIQUIDS'. See Details.}

\item{INDUSTRY_SEGMENT}{e.g. 'ONSHORE NATURAL GAS PROCESSING [98.230(A)(3)]'. See Details.}

\item{ACID_GAS_REMOVAL_CO2_EMISSIONS}{e.g. '0'. See Details.}

\item{ACID_GAS_REMOVAL_CH4_EMM_CO2E}{e.g. 'NA'. See Details.}

\item{ACID_GAS_REMOVAL_N2O_EMM_CO2E}{e.g. 'NA'. See Details.}

\item{ACID_GAS_REMOVAL_GHG_EMM_CO2E}{e.g. '0'. See Details.}

\item{DOES_FAC_HAVE_AGR_UNIT}{e.g. 'NO'. See Details.}

\item{BAMM_INDICATOR_ACID_GAS}{e.g. 'NO'. See Details.}

\item{SUB_INDICATOR_ACID_GAS}{e.g. 'NA'. See Details.}

\item{BAMM_DESCRIPTION_ACID_GAS}{e.g. 'NA'. See Details.}
}
\description{
Retrieve w acid gas removal unit data from ghg database
}
