% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_w_atmospheric_tanks.R
\name{ghg_w_atmospheric_tanks}
\alias{ghg_w_atmospheric_tanks}
\title{Retrieve w atmospheric tanks data from ghg database}
\usage{
ghg_w_atmospheric_tanks(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, INDUSTRY_SEGMENT = NULL,
  ATMOS_TANKS_TOTAL_CO2_EMM = NULL,
  ATMOS_TANKS_TOTAL_CH4_EMM_CO2E = NULL,
  ATMOS_TANKS_TOTAL_N2O_EMM_CO2E = NULL,
  ATMOS_TANKS_TOTAL_GHG_EMM_CO2E = NULL,
  DID_FAC_HAVE_GAS_FRM_PRODU_OIL = NULL,
  BAMM_INDICATOR_ATMOS_TANKS = NULL,
  BAMM_DESCRIPTION_ATMOS_TANKS = NULL,
  SUB_INDICATOR_ATMOS_TANKS = NULL,
  CAL_METH_FIVE_PROD_FRAC_VAPREC = NULL,
  CALC_METH_FIVE_PROD_FRAC_FLRNG = NULL,
  CALC_METH_FIVE_TANKS_WELLPADS = NULL,
  IMPROP_FUNC_DUMP_WELLHD_SEP_CT = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000355'. See Details.}

\item{REPORTING_YEAR}{e.g. '2014'. See Details.}

\item{FACILITY_NAME}{e.g. '#540 BONANZA CREEK ENERGY - DENVER BASIN'. See Details.}

\item{INDUSTRY_SEGMENT}{e.g. 'ONSHORE PETROLEUM AND NATURAL GAS PRODUCTION [98.230(A)(2)]'. See Details.}

\item{ATMOS_TANKS_TOTAL_CO2_EMM}{e.g. '4730.8'. See Details.}

\item{ATMOS_TANKS_TOTAL_CH4_EMM_CO2E}{e.g. '1204.8'. See Details.}

\item{ATMOS_TANKS_TOTAL_N2O_EMM_CO2E}{e.g. '2.4'. See Details.}

\item{ATMOS_TANKS_TOTAL_GHG_EMM_CO2E}{e.g. '5938'. See Details.}

\item{DID_FAC_HAVE_GAS_FRM_PRODU_OIL}{e.g. 'YES'. See Details.}

\item{BAMM_INDICATOR_ATMOS_TANKS}{e.g. 'NO'. See Details.}

\item{BAMM_DESCRIPTION_ATMOS_TANKS}{e.g. 'NA'. See Details.}

\item{SUB_INDICATOR_ATMOS_TANKS}{e.g. 'NO'. See Details.}

\item{CAL_METH_FIVE_PROD_FRAC_VAPREC}{e.g. '0.001'. See Details.}

\item{CALC_METH_FIVE_PROD_FRAC_FLRNG}{e.g. '0.9989'. See Details.}

\item{CALC_METH_FIVE_TANKS_WELLPADS}{e.g. '474'. See Details.}

\item{IMPROP_FUNC_DUMP_WELLHD_SEP_CT}{e.g. '0'. See Details.}
}
\description{
Retrieve w atmospheric tanks data from ghg database
}
