% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_w_blowdown_vent_stacks.R
\name{ghg_w_blowdown_vent_stacks}
\alias{ghg_w_blowdown_vent_stacks}
\title{Retrieve w blowdown vent stacks data from ghg database}
\usage{
ghg_w_blowdown_vent_stacks(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  TOT_BLOWDOWN_CH4_CO2E = NULL, TOT_BLOWDOWN_CO2_EMISSIONS = NULL,
  TOTAL_NUMBER_OF_BLOWDOWNS = NULL, SUB_INDICATOR_BLWDWN_VENT_STK = NULL,
  BAMM_DESC_BLWDWN_VENT_STACKS = NULL,
  BAMM_INDICATOR_BLWDWN_VENT_STK = NULL,
  DOES_FAC_BLOWDOWN_VENT_STACKS = NULL,
  BLDN_VENT_ST_TOT_GHG_EMM_CO2E = NULL,
  BLDN_VENT_ST_TOT_N2O_EMM_CO2E = NULL,
  BLDN_VENT_ST_TOT_CH4_EMM_CO2E = NULL, FACILITY_NAME = NULL,
  INDUSTRY_SEGMENT = NULL, BLWDWN_VENT_ST_TOTAL_CO2_EMM = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000064'. See Details.}

\item{REPORTING_YEAR}{e.g. '2013'. See Details.}

\item{TOT_BLOWDOWN_CH4_CO2E}{e.g. 'NA'. See Details.}

\item{TOT_BLOWDOWN_CO2_EMISSIONS}{e.g. 'NA'. See Details.}

\item{TOTAL_NUMBER_OF_BLOWDOWNS}{e.g. 'NA'. See Details.}

\item{SUB_INDICATOR_BLWDWN_VENT_STK}{e.g. 'NA'. See Details.}

\item{BAMM_DESC_BLWDWN_VENT_STACKS}{e.g. 'NA'. See Details.}

\item{BAMM_INDICATOR_BLWDWN_VENT_STK}{e.g. 'NO'. See Details.}

\item{DOES_FAC_BLOWDOWN_VENT_STACKS}{e.g. 'NO'. See Details.}

\item{BLDN_VENT_ST_TOT_GHG_EMM_CO2E}{e.g. '0'. See Details.}

\item{BLDN_VENT_ST_TOT_N2O_EMM_CO2E}{e.g. 'NA'. See Details.}

\item{BLDN_VENT_ST_TOT_CH4_EMM_CO2E}{e.g. '0'. See Details.}

\item{FACILITY_NAME}{e.g. 'LONE STAR GAS LIQUIDS'. See Details.}

\item{INDUSTRY_SEGMENT}{e.g. 'ONSHORE NATURAL GAS PROCESSING [98.230(A)(3)]'. See Details.}

\item{BLWDWN_VENT_ST_TOTAL_CO2_EMM}{e.g. '0'. See Details.}
}
\description{
Retrieve w blowdown vent stacks data from ghg database
}

