% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_w_centrifugal_cmps_onshore.R
\name{ghg_w_centrifugal_cmps_onshore}
\alias{ghg_w_centrifugal_cmps_onshore}
\title{Retrieve w centrifugal cmps onshore data from ghg database}
\usage{
ghg_w_centrifugal_cmps_onshore(FACILITY_ID = NULL,
  REPORTING_YEAR = NULL, CENF_CMP_ONSH_TOT_CH4_EM_CO2E = NULL,
  INDUSTRY_SEGMENT = NULL, CENF_CMPS_ONSHR_TOT_CO2_EMM = NULL,
  FACILITY_NAME = NULL, CENF_CMPS_GRTOT_N2O_EMM_CO2E = NULL,
  CENF_CMPS_GRTOT_GHG_EMM_CO2E = NULL,
  DOES_FAC_HAVE_CENTG_COMPRESSOR = NULL,
  BAMM_INDICATOR_COMPRESSORS = NULL,
  BAMM_DESCRIPTION_COMPRESSORS = NULL,
  SUB_IND_CENTRIFUGAL_COMPRESSOR = NULL,
  CENF_CMPS_GRTOTAL_CO2_EMM = NULL,
  CENF_CMPS_GRTOT_CH4_EMM_CO2E = NULL, COMPRESSOR_COUNT = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000064'. See Details.}

\item{REPORTING_YEAR}{e.g. '2013'. See Details.}

\item{CENF_CMP_ONSH_TOT_CH4_EM_CO2E}{e.g. 'NA'. See Details.}

\item{INDUSTRY_SEGMENT}{e.g. 'ONSHORE NATURAL GAS PROCESSING [98.230(A)(3)]'. See Details.}

\item{CENF_CMPS_ONSHR_TOT_CO2_EMM}{e.g. 'NA'. See Details.}

\item{FACILITY_NAME}{e.g. 'LONE STAR GAS LIQUIDS'. See Details.}

\item{CENF_CMPS_GRTOT_N2O_EMM_CO2E}{e.g. '0'. See Details.}

\item{CENF_CMPS_GRTOT_GHG_EMM_CO2E}{e.g. '0'. See Details.}

\item{DOES_FAC_HAVE_CENTG_COMPRESSOR}{e.g. 'YES'. See Details.}

\item{BAMM_INDICATOR_COMPRESSORS}{e.g. 'NO'. See Details.}

\item{BAMM_DESCRIPTION_COMPRESSORS}{e.g. 'NA'. See Details.}

\item{SUB_IND_CENTRIFUGAL_COMPRESSOR}{e.g. 'NO'. See Details.}

\item{CENF_CMPS_GRTOTAL_CO2_EMM}{e.g. '0'. See Details.}

\item{CENF_CMPS_GRTOT_CH4_EMM_CO2E}{e.g. '0'. See Details.}

\item{COMPRESSOR_COUNT}{e.g. 'NA'. See Details.}
}
\description{
Retrieve w centrifugal cmps onshore data from ghg database
}
