% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_w_combustion.R
\name{ghg_w_combustion}
\alias{ghg_w_combustion}
\title{Retrieve w combustion data from ghg database}
\usage{
ghg_w_combustion(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  COMBUSTION_TOTAL_N2O_EMM_CO2E = NULL,
  COMBUSTION_TOTAL_CH4_EMM_CO2E = NULL, FACILITY_NAME = NULL,
  COMBUSTION_TOTAL_GHG_EMM_CO2E = NULL,
  COMBUSTION_TOTAL_CO2_EMISSIONS = NULL, BAMMDESCRIPTION = NULL,
  INDUSTRY_SEGMENT = NULL, SUBSTITUTEDATAINDICATOR = NULL,
  DOES_FACILITY_HAVE_COMBUS_EMIS = NULL, BAMMINDICATOR = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000011'. See Details.}

\item{REPORTING_YEAR}{e.g. '2013'. See Details.}

\item{COMBUSTION_TOTAL_N2O_EMM_CO2E}{e.g. '0'. See Details.}

\item{COMBUSTION_TOTAL_CH4_EMM_CO2E}{e.g. '0'. See Details.}

\item{FACILITY_NAME}{e.g. 'CENTRAL HUDSON GAS AND ELECTRIC CORPORATION'. See Details.}

\item{COMBUSTION_TOTAL_GHG_EMM_CO2E}{e.g. '0'. See Details.}

\item{COMBUSTION_TOTAL_CO2_EMISSIONS}{e.g. '0'. See Details.}

\item{BAMMDESCRIPTION}{e.g. 'NA'. See Details.}

\item{INDUSTRY_SEGMENT}{e.g. 'NATURAL GAS DISTRIBUTION [98.230(A)(8)]'. See Details.}

\item{SUBSTITUTEDATAINDICATOR}{e.g. 'NO'. See Details.}

\item{DOES_FACILITY_HAVE_COMBUS_EMIS}{e.g. 'NO'. See Details.}

\item{BAMMINDICATOR}{e.g. 'NO'. See Details.}
}
\description{
Retrieve w combustion data from ghg database
}
