% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_w_dehydrators.R
\name{ghg_w_dehydrators}
\alias{ghg_w_dehydrators}
\title{Retrieve w dehydrators data from ghg database}
\usage{
ghg_w_dehydrators(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, INDUSTRY_SEGMENT = NULL,
  SUB_INDICATOR_DEHYDRATORS = NULL, BAMM_INDICATOR_DEHYDRATORS = NULL,
  BAMM_DESCRIPTION_DEHYDRATORS = NULL,
  DEHYDRATORS_CO2_EMISSIONS = NULL, DEHYDRATORS_CH4_EMISSIONS = NULL,
  DEHYDRATORS_N2O_EMISSIONS = NULL, DEHYDRATORS_GHG_EMISSIONS = NULL,
  FAC_HV_ACIDGAS_REMOVAL_UTS = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000064'. See Details.}

\item{REPORTING_YEAR}{e.g. '2014'. See Details.}

\item{FACILITY_NAME}{e.g. 'LONE STAR GAS LIQUIDS'. See Details.}

\item{INDUSTRY_SEGMENT}{e.g. 'ONSHORE NATURAL GAS PROCESSING [98.230(A)(3)]'. See Details.}

\item{SUB_INDICATOR_DEHYDRATORS}{e.g. 'NA'. See Details.}

\item{BAMM_INDICATOR_DEHYDRATORS}{e.g. 'NO'. See Details.}

\item{BAMM_DESCRIPTION_DEHYDRATORS}{e.g. 'NA'. See Details.}

\item{DEHYDRATORS_CO2_EMISSIONS}{e.g. '0'. See Details.}

\item{DEHYDRATORS_CH4_EMISSIONS}{e.g. '0'. See Details.}

\item{DEHYDRATORS_N2O_EMISSIONS}{e.g. '0'. See Details.}

\item{DEHYDRATORS_GHG_EMISSIONS}{e.g. '0'. See Details.}

\item{FAC_HV_ACIDGAS_REMOVAL_UTS}{e.g. 'NO'. See Details.}
}
\description{
Retrieve w dehydrators data from ghg database
}
