% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_w_enh_oil_rec_inj_pump_bd.R
\name{ghg_w_enh_oil_rec_inj_pump_bd}
\alias{ghg_w_enh_oil_rec_inj_pump_bd}
\title{Retrieve w enh oil rec inj pump bd data from ghg database}
\usage{
ghg_w_enh_oil_rec_inj_pump_bd(FACILITY_ID = NULL,
  REPORTING_YEAR = NULL, DOES_FAC_EOR_INJ_PUMP_BLOWDOWN = NULL,
  EOR_INJ_PUMP_BD_GHG_EMM_CO2E = NULL,
  EOR_INJ_PUMP_BD_N2O_EMM_CO2E = NULL,
  EOR_INJ_PUMP_BD_CH4_EMM_CO2E = NULL,
  EOR_INJ_PUMP_BD_TOTAL_CO2_EMM = NULL, INDUSTRY_SEGMENT = NULL,
  SUB_IND_ENH_OIL_REC_INJ_PUM_BD = NULL, FACILITY_NAME = NULL,
  BAM_IND_ENH_OIL_REC_INJ_PUM_BD = NULL,
  BAM_DES_ENH_OIL_REC_INJ_PUM_BD = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000355'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{DOES_FAC_EOR_INJ_PUMP_BLOWDOWN}{e.g. 'NO'. See Details.}

\item{EOR_INJ_PUMP_BD_GHG_EMM_CO2E}{e.g. '0'. See Details.}

\item{EOR_INJ_PUMP_BD_N2O_EMM_CO2E}{e.g. 'NA'. See Details.}

\item{EOR_INJ_PUMP_BD_CH4_EMM_CO2E}{e.g. 'NA'. See Details.}

\item{EOR_INJ_PUMP_BD_TOTAL_CO2_EMM}{e.g. '0'. See Details.}

\item{INDUSTRY_SEGMENT}{e.g. 'ONSHORE PETROLEUM AND NATURAL GAS PRODUCTION [98.230(A)(2)]'. See Details.}

\item{SUB_IND_ENH_OIL_REC_INJ_PUM_BD}{e.g. 'NA'. See Details.}

\item{FACILITY_NAME}{e.g. '#540 BONANZA CREEK ENERGY - DENVER BASIN'. See Details.}

\item{BAM_IND_ENH_OIL_REC_INJ_PUM_BD}{e.g. 'NA'. See Details.}

\item{BAM_DES_ENH_OIL_REC_INJ_PUM_BD}{e.g. 'NA'. See Details.}
}
\description{
Retrieve w enh oil rec inj pump bd data from ghg database
}
