% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_w_eor_hydrocarbon_liq.R
\name{ghg_w_eor_hydrocarbon_liq}
\alias{ghg_w_eor_hydrocarbon_liq}
\title{Retrieve w eor hydrocarbon liq data from ghg database}
\usage{
ghg_w_eor_hydrocarbon_liq(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  DOES_FAC_HAVE_EOR_LIQ_DISSOLVE = NULL,
  EOR_LIQ_DIS_TOTAL_GHG_EMM_CO2E = NULL,
  EOR_LIQ_DIS_TOTAL_N2O_EMM_CO2E = NULL,
  EOR_LIQ_DIS_TOTAL_CH4_EMM_CO2E = NULL,
  EOR_LIQ_DIS_TOTAL_CO2_EMM = NULL, INDUSTRY_SEGMENT = NULL,
  FACILITY_NAME = NULL, BAMM_INDICATOR_HYDROCARBON_LIQ = NULL,
  SUB_INDICATOR_HYDROCARBON_LIQ = NULL,
  BAMM_DESC_HYDROCARBON_LIQ = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000355'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{DOES_FAC_HAVE_EOR_LIQ_DISSOLVE}{e.g. 'NO'. See Details.}

\item{EOR_LIQ_DIS_TOTAL_GHG_EMM_CO2E}{e.g. '0'. See Details.}

\item{EOR_LIQ_DIS_TOTAL_N2O_EMM_CO2E}{e.g. 'NA'. See Details.}

\item{EOR_LIQ_DIS_TOTAL_CH4_EMM_CO2E}{e.g. 'NA'. See Details.}

\item{EOR_LIQ_DIS_TOTAL_CO2_EMM}{e.g. '0'. See Details.}

\item{INDUSTRY_SEGMENT}{e.g. 'ONSHORE PETROLEUM AND NATURAL GAS PRODUCTION [98.230(A)(2)]'. See Details.}

\item{FACILITY_NAME}{e.g. '#540 BONANZA CREEK ENERGY - DENVER BASIN'. See Details.}

\item{BAMM_INDICATOR_HYDROCARBON_LIQ}{e.g. 'NA'. See Details.}

\item{SUB_INDICATOR_HYDROCARBON_LIQ}{e.g. 'NA'. See Details.}

\item{BAMM_DESC_HYDROCARBON_LIQ}{e.g. 'NA'. See Details.}
}
\description{
Retrieve w eor hydrocarbon liq data from ghg database
}
