% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_w_flare_stacks_details.R
\name{ghg_w_flare_stacks_details}
\alias{ghg_w_flare_stacks_details}
\title{Retrieve w flare stacks details data from ghg database}
\usage{
ghg_w_flare_stacks_details(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  INDUSTRY_SEGMENT = NULL, SUB_INDICATOR_FLARE_STACKS = NULL,
  FLARE_STACKS_CH4_EMISSIONS = NULL, FLARE_STACKS_N2O_EMISSIONS = NULL,
  FLARE_STACKS_TOT_GHG_EMM_CO2E = NULL, DOES_FAC_HAVE_FLARE_STACKS = NULL,
  BAMM_IND_FLARE_STK_DETAILS = NULL, BAMM_DESCRIPTION_FLARE_STACKS = NULL,
  FLARE_STACKS_CO2_EMISSIONS = NULL, FACILITY_NAME = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000064'. See Details.}

\item{REPORTING_YEAR}{e.g. '2014'. See Details.}

\item{INDUSTRY_SEGMENT}{e.g. 'ONSHORE NATURAL GAS PROCESSING [98.230(A)(3)]'. See Details.}

\item{SUB_INDICATOR_FLARE_STACKS}{e.g. 'NA'. See Details.}

\item{FLARE_STACKS_CH4_EMISSIONS}{e.g. '0'. See Details.}

\item{FLARE_STACKS_N2O_EMISSIONS}{e.g. '0'. See Details.}

\item{FLARE_STACKS_TOT_GHG_EMM_CO2E}{e.g. '0'. See Details.}

\item{DOES_FAC_HAVE_FLARE_STACKS}{e.g. 'NO'. See Details.}

\item{BAMM_IND_FLARE_STK_DETAILS}{e.g. 'NO'. See Details.}

\item{BAMM_DESCRIPTION_FLARE_STACKS}{e.g. 'NA'. See Details.}

\item{FLARE_STACKS_CO2_EMISSIONS}{e.g. '0'. See Details.}

\item{FACILITY_NAME}{e.g. 'LONE STAR GAS LIQUIDS'. See Details.}
}
\description{
Retrieve w flare stacks details data from ghg database
}

