% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_w_liquids_unloading.R
\name{ghg_w_liquids_unloading}
\alias{ghg_w_liquids_unloading}
\title{Retrieve w liquids unloading data from ghg database}
\usage{
ghg_w_liquids_unloading(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  LIQUID_UNLOAD_TOT_N2O_EMM_CO2E = NULL,
  LIQUIDS_UNLOADING_TOT_CO2_EMM = NULL, INDUSTRY_SEGMENT = NULL,
  FACILITY_NAME = NULL, LIQUID_UNLOAD_TOT_CH4_EMM_CO2E = NULL,
  SUB_IND_LIQUIDS_UNLOADING = NULL, BAMM_DESC_LIQUIDS_UNLOADING = NULL,
  BAMM_IND_LIQUIDS_UNLOADING = NULL, LIQUID_UNLOAD_TOT_GHG_EMM_CO2E = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000355'. See Details.}

\item{REPORTING_YEAR}{e.g. '2014'. See Details.}

\item{LIQUID_UNLOAD_TOT_N2O_EMM_CO2E}{e.g. 'NA'. See Details.}

\item{LIQUIDS_UNLOADING_TOT_CO2_EMM}{e.g. '10'. See Details.}

\item{INDUSTRY_SEGMENT}{e.g. 'ONSHORE PETROLEUM AND NATURAL GAS PRODUCTION [98.230(A)(2)]'. See Details.}

\item{FACILITY_NAME}{e.g. '#540 BONANZA CREEK ENERGY - DENVER BASIN'. See Details.}

\item{LIQUID_UNLOAD_TOT_CH4_EMM_CO2E}{e.g. '2176.7'. See Details.}

\item{SUB_IND_LIQUIDS_UNLOADING}{e.g. 'NO'. See Details.}

\item{BAMM_DESC_LIQUIDS_UNLOADING}{e.g. 'NA'. See Details.}

\item{BAMM_IND_LIQUIDS_UNLOADING}{e.g. 'NO'. See Details.}

\item{LIQUID_UNLOAD_TOT_GHG_EMM_CO2E}{e.g. '2186.7'. See Details.}
}
\description{
Retrieve w liquids unloading data from ghg database
}

