% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_w_local_dist_companies_details.R
\name{ghg_w_local_dist_companies_details}
\alias{ghg_w_local_dist_companies_details}
\title{Retrieve w local dist companies details data from ghg database}
\usage{
ghg_w_local_dist_companies_details(FACILITY_ID = NULL,
  REPORTING_YEAR = NULL, BAMM_DESC_LOCAL_DIST_CMPNY = NULL,
  BAMM_IND_LOCAL_DIST_CMPNY = NULL,
  ABOVE_GRADE_TRANSFER_STATIONS = NULL,
  YEARS_TRANSFER_STATIONS_MONIT = NULL,
  STATIONS_MONIT_IN_CAL_YR = NULL,
  BELOW_GRADE_TRANSFER_STATIONS = NULL,
  ABOVE_GRADE_METERING_STATIONS = NULL,
  BELOW_GRADE_METERING_STATIONS = NULL,
  ANL_ABV_GRD_TRSFER_STN_CO2_EMM = NULL,
  ANL_ABV_GRD_TR_STN_CH4_EMMCO2E = NULL,
  ANL_BEL_GRDE_TRSFR_STN_CO2_EMM = NULL,
  ANL_BEL_GRD_TR_STN_CH4_EM_CO2E = NULL,
  ANL_ABV_GRDE_MTR_STN_CO2_EMM = NULL,
  ANL_BEL_GRDE_MTR_STNS_CO2_EMM = NULL,
  ANL_BEL_GRD_MT_STN_CH4_EMMCO2E = NULL,
  ANNUAL_DIST_MAINS_CO2_EMM = NULL,
  ANNUAL_DIST_MAINS_CH4_EMM_CO2E = NULL,
  ANNUAL_DIST_SERVICES_CO2_EMM = NULL,
  ANL_DIST_SERVICES_CH4_EMM_CO2E = NULL, FACILITY_NAME = NULL,
  INDUSTRY_SEGMENT = NULL, LOCAL_DIST_CMPNY_CO2_EMISSIONS = NULL,
  LOCAL_DIST_CMPNY_CH4_EMISSIONS = NULL,
  LOCAL_DIST_CMPNY_GHG_EMISSIONS = NULL,
  SUB_INDICATOR_LOCAL_DIST_CMPNY = NULL,
  LOCAL_DIST_CMPNY_N2O_EMISSIONS = NULL,
  ANNUAL_CH4_ABV_GRD_MTRNG_STAT = NULL,
  MILES_OF_UNPR_STEEL_DIST_MAINS = NULL, PR_STEEL_DIST_SERVICES = NULL,
  COPPER_DIST_SERVICES = NULL, MILES_OF_CAST_IRON_DIST_MAINS = NULL,
  MILES_OF_PLSTIC_DIST_MAINS = NULL,
  MILES_OF_PROT_STEEL_DIST_MAINS = NULL,
  MTR_REGULATOR_LEAK_FACTOR = NULL,
  PLASTIC_DISTRIBUTION_SERVICES = NULL, UNPRSTEELDISTSERVICES = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000011'. See Details.}

\item{REPORTING_YEAR}{e.g. '2012'. See Details.}

\item{BAMM_DESC_LOCAL_DIST_CMPNY}{e.g. 'NOT AVAILABLE'. See Details.}

\item{BAMM_IND_LOCAL_DIST_CMPNY}{e.g. 'NO'. See Details.}

\item{ABOVE_GRADE_TRANSFER_STATIONS}{e.g. '4'. See Details.}

\item{YEARS_TRANSFER_STATIONS_MONIT}{e.g. '2'. See Details.}

\item{STATIONS_MONIT_IN_CAL_YR}{e.g. '151'. See Details.}

\item{BELOW_GRADE_TRANSFER_STATIONS}{e.g. '7'. See Details.}

\item{ABOVE_GRADE_METERING_STATIONS}{e.g. '155'. See Details.}

\item{BELOW_GRADE_METERING_STATIONS}{e.g. '0'. See Details.}

\item{ANL_ABV_GRD_TRSFER_STN_CO2_EMM}{e.g. '0'. See Details.}

\item{ANL_ABV_GRD_TR_STN_CH4_EMMCO2E}{e.g. '46.5'. See Details.}

\item{ANL_BEL_GRDE_TRSFR_STN_CO2_EMM}{e.g. '0'. See Details.}

\item{ANL_BEL_GRD_TR_STN_CH4_EM_CO2E}{e.g. '0'. See Details.}

\item{ANL_ABV_GRDE_MTR_STN_CO2_EMM}{e.g. '0'. See Details.}

\item{ANL_BEL_GRDE_MTR_STNS_CO2_EMM}{e.g. '0'. See Details.}

\item{ANL_BEL_GRD_MT_STN_CH4_EMMCO2E}{e.g. '0'. See Details.}

\item{ANNUAL_DIST_MAINS_CO2_EMM}{e.g. '0'. See Details.}

\item{ANNUAL_DIST_MAINS_CH4_EMM_CO2E}{e.g. '18236'. See Details.}

\item{ANNUAL_DIST_SERVICES_CO2_EMM}{e.g. '0'. See Details.}

\item{ANL_DIST_SERVICES_CH4_EMM_CO2E}{e.g. '10936'. See Details.}

\item{FACILITY_NAME}{e.g. 'CENTRAL HUDSON GAS AND ELECTRIC CORPORATION'. See Details.}

\item{INDUSTRY_SEGMENT}{e.g. 'NATURAL GAS DISTRIBUTION [98.230(A)(8)]'. See Details.}

\item{LOCAL_DIST_CMPNY_CO2_EMISSIONS}{e.g. '0'. See Details.}

\item{LOCAL_DIST_CMPNY_CH4_EMISSIONS}{e.g. '29220.8'. See Details.}

\item{LOCAL_DIST_CMPNY_GHG_EMISSIONS}{e.g. '29220.8'. See Details.}

\item{SUB_INDICATOR_LOCAL_DIST_CMPNY}{e.g. 'NO'. See Details.}

\item{LOCAL_DIST_CMPNY_N2O_EMISSIONS}{e.g. 'NA'. See Details.}

\item{ANNUAL_CH4_ABV_GRD_MTRNG_STAT}{e.g. '48.8'. See Details.}

\item{MILES_OF_UNPR_STEEL_DIST_MAINS}{e.g. '156'. See Details.}

\item{PR_STEEL_DIST_SERVICES}{e.g. '7708'. See Details.}

\item{COPPER_DIST_SERVICES}{e.g. '78'. See Details.}

\item{MILES_OF_CAST_IRON_DIST_MAINS}{e.g. '88'. See Details.}

\item{MILES_OF_PLSTIC_DIST_MAINS}{e.g. '607'. See Details.}

\item{MILES_OF_PROT_STEEL_DIST_MAINS}{e.g. '342'. See Details.}

\item{MTR_REGULATOR_LEAK_FACTOR}{e.g. '0.05'. See Details.}

\item{PLASTIC_DISTRIBUTION_SERVICES}{e.g. '33750'. See Details.}

\item{UNPRSTEELDISTSERVICES}{e.g. '15467'. See Details.}
}
\description{
Retrieve w local dist companies details data from ghg database
}
