% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_w_other_emissions_equip_leaks.R
\name{ghg_w_other_emissions_equip_leaks}
\alias{ghg_w_other_emissions_equip_leaks}
\title{Retrieve w other emissions equip leaks data from ghg database}
\usage{
ghg_w_other_emissions_equip_leaks(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  BAMM_DESCRIPTION_EQUIP_LEAKS = NULL, BAMM_INDICATOR_EQUIP_LEAKS = NULL,
  EQUIP_LEAKS_TOTAL_GHG_EMM_CO2E = NULL,
  EQUIP_LEAKS_N2O_EMISSIONS_CO2E = NULL,
  EQUIP_LEAKS_CH4_EMISSIONS_CO2E = NULL, SUB_INDICATOR_EQUIP_LEAKS = NULL,
  INDUSTRY_SEGMENT = NULL, FACILITY_NAME = NULL,
  EQUIP_LEAKS_CO2_EMISSIONS = NULL, DOES_FAC_HAVE_EQUIPMENT_LEAKS = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000011'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{BAMM_DESCRIPTION_EQUIP_LEAKS}{e.g. 'NO'. See Details.}

\item{BAMM_INDICATOR_EQUIP_LEAKS}{e.g. 'NO'. See Details.}

\item{EQUIP_LEAKS_TOTAL_GHG_EMM_CO2E}{e.g. '0'. See Details.}

\item{EQUIP_LEAKS_N2O_EMISSIONS_CO2E}{e.g. 'NA'. See Details.}

\item{EQUIP_LEAKS_CH4_EMISSIONS_CO2E}{e.g. '0'. See Details.}

\item{SUB_INDICATOR_EQUIP_LEAKS}{e.g. 'NO'. See Details.}

\item{INDUSTRY_SEGMENT}{e.g. 'NATURAL GAS DISTRIBUTION [98.230(A)(8)]'. See Details.}

\item{FACILITY_NAME}{e.g. 'CENTRAL HUDSON GAS AND ELECTRIC CORPORATION'. See Details.}

\item{EQUIP_LEAKS_CO2_EMISSIONS}{e.g. '0'. See Details.}

\item{DOES_FAC_HAVE_EQUIPMENT_LEAKS}{e.g. 'YES'. See Details.}
}
\description{
Retrieve w other emissions equip leaks data from ghg database
}

