% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_w_pneumatic_devices.R
\name{ghg_w_pneumatic_devices}
\alias{ghg_w_pneumatic_devices}
\title{Retrieve w pneumatic devices data from ghg database}
\usage{
ghg_w_pneumatic_devices(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  PNEUMATIC_DEVICES_TOT_CO2_EMM = NULL, SUB_IND_PNEUMATIC_DEVICES = NULL,
  FACILITY_NAME = NULL, PNEUMATIC_DEV_TOT_CH4_EMM_CO2E = NULL,
  PNEUM_DEVICES_TOT_N2O_EMM_CO2E = NULL,
  PNEUM_DEVICES_TOT_GHG_EMM_CO2E = NULL,
  DOES_FAC_HAVE_HIGH_BLEED_DEVIC = NULL,
  DOES_FAC_HAVE_INTR_BLEED_DEVIC = NULL,
  DOES_FAC_HAVE_LOW_BLEED_DEVICE = NULL, BAMM_IND_PNEUMATIC_DEVICES = NULL,
  BAMM_DESC_PNEUMATIC_DEVICES = NULL, INDUSTRY_SEGMENT = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000179'. See Details.}

\item{REPORTING_YEAR}{e.g. '2012'. See Details.}

\item{PNEUMATIC_DEVICES_TOT_CO2_EMM}{e.g. '0.3'. See Details.}

\item{SUB_IND_PNEUMATIC_DEVICES}{e.g. 'NO'. See Details.}

\item{FACILITY_NAME}{e.g. 'BERNVILLE STATION'. See Details.}

\item{PNEUMATIC_DEV_TOT_CH4_EMM_CO2E}{e.g. '188.5'. See Details.}

\item{PNEUM_DEVICES_TOT_N2O_EMM_CO2E}{e.g. 'NA'. See Details.}

\item{PNEUM_DEVICES_TOT_GHG_EMM_CO2E}{e.g. '188.8'. See Details.}

\item{DOES_FAC_HAVE_HIGH_BLEED_DEVIC}{e.g. 'YES'. See Details.}

\item{DOES_FAC_HAVE_INTR_BLEED_DEVIC}{e.g. 'NO'. See Details.}

\item{DOES_FAC_HAVE_LOW_BLEED_DEVICE}{e.g. 'NO'. See Details.}

\item{BAMM_IND_PNEUMATIC_DEVICES}{e.g. 'YES'. See Details.}

\item{BAMM_DESC_PNEUMATIC_DEVICES}{e.g. 'BAMM NEEDED TO OBTAIN TOTAL COUNT OF HIGH AND LOW BLEED DEVICES'. See Details.}

\item{INDUSTRY_SEGMENT}{e.g. 'ONSHORE NATURAL GAS TRANSMISSION COMPRESSION [98.230(A)(4)]'. See Details.}
}
\description{
Retrieve w pneumatic devices data from ghg database
}

