% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_w_pneumatic_pumps.R
\name{ghg_w_pneumatic_pumps}
\alias{ghg_w_pneumatic_pumps}
\title{Retrieve w pneumatic pumps data from ghg database}
\usage{
ghg_w_pneumatic_pumps(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  SUB_INDICATOR_PNEUMATIC_PUMPS = NULL,
  BAMM_DESC_PNEUMATIC_PUMPS = NULL,
  BAMM_INDICATOR_PNEUMATIC_PUMPS = NULL,
  DOES_FAC_HAVE_PNEUMATIC_PUMPS = NULL,
  PNEUM_PUMP_TOTAL_GHG_EMM_CO2E = NULL,
  PNEUM_PUMP_TOTAL_N2O_EMM_CO2E = NULL,
  PNEUM_PUMP_TOTAL_CH4_EMM_CO2E = NULL,
  PNEUM_PUMP_TOTAL_CO2_EMISSIONS = NULL, FACILITY_NAME = NULL,
  INDUSTRY_SEGMENT = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000355'. See Details.}

\item{REPORTING_YEAR}{e.g. '2012'. See Details.}

\item{SUB_INDICATOR_PNEUMATIC_PUMPS}{e.g. 'NO'. See Details.}

\item{BAMM_DESC_PNEUMATIC_PUMPS}{e.g. 'NA'. See Details.}

\item{BAMM_INDICATOR_PNEUMATIC_PUMPS}{e.g. 'NO'. See Details.}

\item{DOES_FAC_HAVE_PNEUMATIC_PUMPS}{e.g. 'YES'. See Details.}

\item{PNEUM_PUMP_TOTAL_GHG_EMM_CO2E}{e.g. '706.9'. See Details.}

\item{PNEUM_PUMP_TOTAL_N2O_EMM_CO2E}{e.g. 'NA'. See Details.}

\item{PNEUM_PUMP_TOTAL_CH4_EMM_CO2E}{e.g. '704'. See Details.}

\item{PNEUM_PUMP_TOTAL_CO2_EMISSIONS}{e.g. '2.9'. See Details.}

\item{FACILITY_NAME}{e.g. '#540 BONANZA CREEK ENERGY - DENVER BASIN'. See Details.}

\item{INDUSTRY_SEGMENT}{e.g. 'ONSHORE PETROLEUM AND NATURAL GAS PRODUCTION [98.230(A)(2)]'. See Details.}
}
\description{
Retrieve w pneumatic pumps data from ghg database
}
