% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_w_sub_basin.R
\name{ghg_w_sub_basin}
\alias{ghg_w_sub_basin}
\title{Retrieve w sub basin data from ghg database}
\usage{
ghg_w_sub_basin(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  INDUSTRY_SEGMENT = NULL, FACILITY_NAME = NULL,
  SUB_BASIN_IDENTIFIER = NULL, AVERAGE_PRESSURE = NULL,
  GAS_TO_OIL_RATIO = NULL, API_GRAVITY = NULL,
  SUB_BASIN_FORMATION_TYPE = NULL, SUB_BASIN_COUNTY = NULL,
  SELECTED_BASIN = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000355'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{INDUSTRY_SEGMENT}{e.g. 'ONSHORE PETROLEUM AND NATURAL GAS PRODUCTION [98.230(A)(2)]'. See Details.}

\item{FACILITY_NAME}{e.g. '#540 BONANZA CREEK ENERGY - DENVER BASIN'. See Details.}

\item{SUB_BASIN_IDENTIFIER}{e.g. '540 - WELD, CO (123) - OTHER TIGHT RESERVOIR ROCK'. See Details.}

\item{AVERAGE_PRESSURE}{e.g. 'NA'. See Details.}

\item{GAS_TO_OIL_RATIO}{e.g. 'NA'. See Details.}

\item{API_GRAVITY}{e.g. 'NA'. See Details.}

\item{SUB_BASIN_FORMATION_TYPE}{e.g. 'OTHER TIGHT RESERVOIR ROCK'. See Details.}

\item{SUB_BASIN_COUNTY}{e.g. 'WELD, CO (123)'. See Details.}

\item{SELECTED_BASIN}{e.g. '540 - DENVER BASIN'. See Details.}
}
\description{
Retrieve w sub basin data from ghg database
}

