% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_w_subpart_level_information.R
\name{ghg_w_subpart_level_information}
\alias{ghg_w_subpart_level_information}
\title{Retrieve w subpart level information data from ghg database}
\usage{
ghg_w_subpart_level_information(FACILITY_ID = NULL,
  REPORTING_YEAR = NULL, GHG_NAME = NULL, FACILITY_NAME = NULL,
  GHG_QUANTITY = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000011'. See Details.}

\item{REPORTING_YEAR}{e.g. '2012'. See Details.}

\item{GHG_NAME}{e.g. 'NITROUS OXIDE'. See Details.}

\item{FACILITY_NAME}{e.g. 'CENTRAL HUDSON GAS AND ELECTRIC CORPORATION'. See Details.}

\item{GHG_QUANTITY}{e.g. '0'. See Details.}
}
\description{
Retrieve w subpart level information data from ghg database
}
