% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_w_transmission_tanks.R
\name{ghg_w_transmission_tanks}
\alias{ghg_w_transmission_tanks}
\title{Retrieve w transmission tanks data from ghg database}
\usage{
ghg_w_transmission_tanks(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, INDUSTRY_SEGMENT = NULL,
  TRANS_TANKS_TOTAL_CO2_EMM = NULL, TRANS_TANKS_TOTAL_CH4_EMM_CO2E = NULL,
  TRANS_TANKS_TOTAL_N2O_EMM_CO2E = NULL,
  TRANS_TANKS_TOTAL_GHG_EMM_CO2E = NULL,
  DID_FAC_VENT_STACK_EMISSIONS = NULL, BAMM_IND_TRANSMISSION_TANKS = NULL,
  BAMM_DESC_TRANSMISSION_TANKS = NULL, SUB_IND_TRANSMISSION_TANKS = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000179'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{FACILITY_NAME}{e.g. 'BERNVILLE STATION'. See Details.}

\item{INDUSTRY_SEGMENT}{e.g. 'ONSHORE NATURAL GAS TRANSMISSION COMPRESSION [98.230(A)(4)]'. See Details.}

\item{TRANS_TANKS_TOTAL_CO2_EMM}{e.g. '0'. See Details.}

\item{TRANS_TANKS_TOTAL_CH4_EMM_CO2E}{e.g. '0'. See Details.}

\item{TRANS_TANKS_TOTAL_N2O_EMM_CO2E}{e.g. 'NA'. See Details.}

\item{TRANS_TANKS_TOTAL_GHG_EMM_CO2E}{e.g. '0'. See Details.}

\item{DID_FAC_VENT_STACK_EMISSIONS}{e.g. 'YES'. See Details.}

\item{BAMM_IND_TRANSMISSION_TANKS}{e.g. 'NO'. See Details.}

\item{BAMM_DESC_TRANSMISSION_TANKS}{e.g. 'NA'. See Details.}

\item{SUB_IND_TRANSMISSION_TANKS}{e.g. 'NO'. See Details.}
}
\description{
Retrieve w transmission tanks data from ghg database
}

