% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_w_well_completion_hydraulic.R
\name{ghg_w_well_completion_hydraulic}
\alias{ghg_w_well_completion_hydraulic}
\title{Retrieve w well completion hydraulic data from ghg database}
\usage{
ghg_w_well_completion_hydraulic(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, INDUSTRY_SEGMENT = NULL,
  SUB_BASIN_IDENTIFIER = NULL, WORKOVERS_VENTING_CH4_EMM_CO2E = NULL,
  WORKOVERS_FLARING_CO2_EMM = NULL, WORKOVERS_FLARING_CH4_EMM_CO2E = NULL,
  WORKOVERS_FLARING_N2O_EMM_CO2E = NULL, SEPARATING_COMPLETIONS = NULL,
  SEPARATING_WORKOVERS = NULL, WELL_TYPE = NULL,
  COMPLETION_EQUATION = NULL, WORKOVER_EQUATION = NULL,
  TOTAL_COMPLETIONS = NULL, TOTAL_FLARE_GAS_WORKOVERS = NULL,
  TOTAL_VENT_GAS_WORKOVERS = NULL, COMPLETIONS_VENTING_CO2_EMM = NULL,
  COMPLETIONS_VENT_CH4_EMM_CO2E = NULL, COMPLETIONS_FLARING_CO2_EMM = NULL,
  COMPLETIONS_FLRNG_CH4_EMM_CO2E = NULL,
  COMPLETIONS_FLRNG_N2O_EMM_CO2E = NULL, WORKOVERS_VENTING_CO2_EMM = NULL,
  WORKOVER_BACKFLOW_FLOW_RATE = NULL, WORKOVERS_NMBR_OF_BKFLW_DAYS = NULL,
  COMPLETION_GAS_RECVRD_TO_SALES = NULL, COMPLETIONS_NMBR_BKFLW_DAYS = NULL,
  TWO_YEAR_DELAY_INDICATOR = NULL, WORKOVER_GAS_RECVRD_TO_SALES = NULL,
  API_WELL_NUMBERS = NULL, COMPLETION_BACKFLOW_FLOW_RATE = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000355'. See Details.}

\item{REPORTING_YEAR}{e.g. '2012'. See Details.}

\item{FACILITY_NAME}{e.g. '#540 BONANZA CREEK ENERGY - DENVER BASIN'. See Details.}

\item{INDUSTRY_SEGMENT}{e.g. 'ONSHORE PETROLEUM AND NATURAL GAS PRODUCTION [98.230(A)(2)]'. See Details.}

\item{SUB_BASIN_IDENTIFIER}{e.g. '540 - WELD, CO (123) - OTHER TIGHT RESERVOIR ROCK'. See Details.}

\item{WORKOVERS_VENTING_CH4_EMM_CO2E}{e.g. '0'. See Details.}

\item{WORKOVERS_FLARING_CO2_EMM}{e.g. '0'. See Details.}

\item{WORKOVERS_FLARING_CH4_EMM_CO2E}{e.g. '0'. See Details.}

\item{WORKOVERS_FLARING_N2O_EMM_CO2E}{e.g. '0'. See Details.}

\item{SEPARATING_COMPLETIONS}{e.g. '23'. See Details.}

\item{SEPARATING_WORKOVERS}{e.g. '0'. See Details.}

\item{WELL_TYPE}{e.g. 'HORIZONTAL'. See Details.}

\item{COMPLETION_EQUATION}{e.g. 'EQUATION W-10B'. See Details.}

\item{WORKOVER_EQUATION}{e.g. 'EQUATION W-10B'. See Details.}

\item{TOTAL_COMPLETIONS}{e.g. '25'. See Details.}

\item{TOTAL_FLARE_GAS_WORKOVERS}{e.g. '0'. See Details.}

\item{TOTAL_VENT_GAS_WORKOVERS}{e.g. '0'. See Details.}

\item{COMPLETIONS_VENTING_CO2_EMM}{e.g. '0'. See Details.}

\item{COMPLETIONS_VENT_CH4_EMM_CO2E}{e.g. '0'. See Details.}

\item{COMPLETIONS_FLARING_CO2_EMM}{e.g. '16921'. See Details.}

\item{COMPLETIONS_FLRNG_CH4_EMM_CO2E}{e.g. '1340.5'. See Details.}

\item{COMPLETIONS_FLRNG_N2O_EMM_CO2E}{e.g. '8.9'. See Details.}

\item{WORKOVERS_VENTING_CO2_EMM}{e.g. '0'. See Details.}

\item{WORKOVER_BACKFLOW_FLOW_RATE}{e.g. '0'. See Details.}

\item{WORKOVERS_NMBR_OF_BKFLW_DAYS}{e.g. '0'. See Details.}

\item{COMPLETION_GAS_RECVRD_TO_SALES}{e.g. '145731000'. See Details.}

\item{COMPLETIONS_NMBR_BKFLW_DAYS}{e.g. '0'. See Details.}

\item{TWO_YEAR_DELAY_INDICATOR}{e.g. 'No'. See Details.}

\item{WORKOVER_GAS_RECVRD_TO_SALES}{e.g. '0'. See Details.}

\item{API_WELL_NUMBERS}{e.g. 'NA'. See Details.}

\item{COMPLETION_BACKFLOW_FLOW_RATE}{e.g. '0'. See Details.}
}
\description{
Retrieve w well completion hydraulic data from ghg database
}

