% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_w_well_testing.R
\name{ghg_w_well_testing}
\alias{ghg_w_well_testing}
\title{Retrieve w well testing data from ghg database}
\usage{
ghg_w_well_testing(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  SUB_INDICATOR_WELL_TESTING = NULL, BAMM_DESCRIPTION_WELL_TESTING = NULL,
  BAMM_INDICATOR_WELL_TESTING = NULL, DID_FAC_PERFORM_WELL_TESTING = NULL,
  WELL_TEST_TOTAL_GHG_EMM_CO2E = NULL, WELL_TEST_TOTAL_N2O_EMM_CO2E = NULL,
  WELL_TEST_TOTAL_CO2_EMISSIONS = NULL, INDUSTRY_SEGMENT = NULL,
  FACILITY_NAME = NULL, WELL_TEST_TOTAL_CH4_EMM_CO2E = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000355'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{SUB_INDICATOR_WELL_TESTING}{e.g. 'NA'. See Details.}

\item{BAMM_DESCRIPTION_WELL_TESTING}{e.g. 'NA'. See Details.}

\item{BAMM_INDICATOR_WELL_TESTING}{e.g. 'NA'. See Details.}

\item{DID_FAC_PERFORM_WELL_TESTING}{e.g. 'NO'. See Details.}

\item{WELL_TEST_TOTAL_GHG_EMM_CO2E}{e.g. '0'. See Details.}

\item{WELL_TEST_TOTAL_N2O_EMM_CO2E}{e.g. '0'. See Details.}

\item{WELL_TEST_TOTAL_CO2_EMISSIONS}{e.g. '0'. See Details.}

\item{INDUSTRY_SEGMENT}{e.g. 'ONSHORE PETROLEUM AND NATURAL GAS PRODUCTION [98.230(A)(2)]'. See Details.}

\item{FACILITY_NAME}{e.g. '#540 BONANZA CREEK ENERGY - DENVER BASIN'. See Details.}

\item{WELL_TEST_TOTAL_CH4_EMM_CO2E}{e.g. '0'. See Details.}
}
\description{
Retrieve w well testing data from ghg database
}

