% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_w_wellcompl_without_hydraulic.R
\name{ghg_w_wellcompl_without_hydraulic}
\alias{ghg_w_wellcompl_without_hydraulic}
\title{Retrieve w wellcompl without hydraulic data from ghg database}
\usage{
ghg_w_wellcompl_without_hydraulic(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  COMPLT_WO_HYD_FLAR_N2O_EMMCO2E = NULL, FACILITY_NAME = NULL,
  INDUSTRY_SEGMENT = NULL, COMPLT_WO_HYD_FLAR_CH4_EMMCO2E = NULL,
  TOTAL_GAS_VENTING_DAYS = NULL, COMPLT_WO_HYD_VENTING_CO2_EMM = NULL,
  COMPL_WO_HYD_VENT_CH4_EMM_CO2E = NULL,
  COMPLT_WO_HYD_FLARING_CO2_EMM = NULL, SUB_BASIN_IDENTIFIER = NULL,
  NUMBER_OF_COMPLETIONS = NULL, NUMBER_OF_WRKOVRS_FLR_OR_VNT = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000355'. See Details.}

\item{REPORTING_YEAR}{e.g. '2012'. See Details.}

\item{COMPLT_WO_HYD_FLAR_N2O_EMMCO2E}{e.g. '0'. See Details.}

\item{FACILITY_NAME}{e.g. '#540 BONANZA CREEK ENERGY - DENVER BASIN'. See Details.}

\item{INDUSTRY_SEGMENT}{e.g. 'ONSHORE PETROLEUM AND NATURAL GAS PRODUCTION [98.230(A)(2)]'. See Details.}

\item{COMPLT_WO_HYD_FLAR_CH4_EMMCO2E}{e.g. '0'. See Details.}

\item{TOTAL_GAS_VENTING_DAYS}{e.g. '0'. See Details.}

\item{COMPLT_WO_HYD_VENTING_CO2_EMM}{e.g. '0'. See Details.}

\item{COMPL_WO_HYD_VENT_CH4_EMM_CO2E}{e.g. '6.3'. See Details.}

\item{COMPLT_WO_HYD_FLARING_CO2_EMM}{e.g. '0'. See Details.}

\item{SUB_BASIN_IDENTIFIER}{e.g. '540 - WELD, CO (123) - OTHER TIGHT RESERVOIR ROCK'. See Details.}

\item{NUMBER_OF_COMPLETIONS}{e.g. '7'. See Details.}

\item{NUMBER_OF_WRKOVRS_FLR_OR_VNT}{e.g. '7'. See Details.}
}
\description{
Retrieve w wellcompl without hydraulic data from ghg database
}

