% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_x_cems_details.R
\name{ghg_x_cems_details}
\alias{ghg_x_cems_details}
\title{Retrieve x cems details data from ghg database}
\usage{
ghg_x_cems_details(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  CEMS_CO2_COMBUSTION_EMISSIONS = NULL, CML_NAME = NULL,
  CML_DESCRIPTION = NULL, CML_CONFIGURATION_TYPE = NULL,
  CEMS_BIOGENIC_CO2_EMISSIONS = NULL,
  CEMS_NONBIOGENIC_CO2_EMISSIONS = NULL, CEMS_CO2_EMISSIONS = NULL,
  CEMS_CH4_EMISSIONS = NULL, CEMS_N2O_EMISSIONS = NULL,
  TIER4_SUB_OPER_HRS_CO2_CONC = NULL,
  TIER4_SUB_OPER_HRS_GAS_FLOW_RT = NULL,
  TIER4_SUB_OPER_HRS_MOIST_CONT = NULL,
  CEMS_METHODOLOGY_START_DATE = NULL, CEMS_METHODOLOGY_END_DATE = NULL,
  FUEL_TYPE = NULL, CEMS_QUARTER_NAME = NULL,
  CEMS_CO2_EMISSIONS_BY_QUARTER = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1004900'. See Details.}

\item{REPORTING_YEAR}{e.g. '2012'. See Details.}

\item{CEMS_CO2_COMBUSTION_EMISSIONS}{e.g. 'NA'. See Details.}

\item{CML_NAME}{e.g. 'Chlorine Destruct Tower'. See Details.}

\item{CML_DESCRIPTION}{e.g. 'NA'. See Details.}

\item{CML_CONFIGURATION_TYPE}{e.g. 'Process/stationary combustion units share common stack'. See Details.}

\item{CEMS_BIOGENIC_CO2_EMISSIONS}{e.g. '0'. See Details.}

\item{CEMS_NONBIOGENIC_CO2_EMISSIONS}{e.g. '31334'. See Details.}

\item{CEMS_CO2_EMISSIONS}{e.g. '31334'. See Details.}

\item{CEMS_CH4_EMISSIONS}{e.g. '1.59'. See Details.}

\item{CEMS_N2O_EMISSIONS}{e.g. '0.319'. See Details.}

\item{TIER4_SUB_OPER_HRS_CO2_CONC}{e.g. '0'. See Details.}

\item{TIER4_SUB_OPER_HRS_GAS_FLOW_RT}{e.g. '0'. See Details.}

\item{TIER4_SUB_OPER_HRS_MOIST_CONT}{e.g. '0'. See Details.}

\item{CEMS_METHODOLOGY_START_DATE}{e.g. '01-JAN-12'. See Details.}

\item{CEMS_METHODOLOGY_END_DATE}{e.g. '31-DEC-12'. See Details.}

\item{FUEL_TYPE}{e.g. 'Natural Gas and Process Vent Gas'. See Details.}

\item{CEMS_QUARTER_NAME}{e.g. 'First Quarter'. See Details.}

\item{CEMS_CO2_EMISSIONS_BY_QUARTER}{e.g. '7011'. See Details.}
}
\description{
Retrieve x cems details data from ghg database
}
