% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_x_cems_info.R
\name{ghg_x_cems_info}
\alias{ghg_x_cems_info}
\title{Retrieve x cems info data from ghg database}
\usage{
ghg_x_cems_info(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  CEMS_UNIT_NAME = NULL, CEMS_UNIT_TYPE = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1004900'. See Details.}

\item{REPORTING_YEAR}{e.g. '2014'. See Details.}

\item{CEMS_UNIT_NAME}{e.g. 'VCM-E Plant'. See Details.}

\item{CEMS_UNIT_TYPE}{e.g. 'Petrochemical process unit'. See Details.}
}
\description{
Retrieve x cems info data from ghg database
}

