% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_x_combustion_ethylene_unit.R
\name{ghg_x_combustion_ethylene_unit}
\alias{ghg_x_combustion_ethylene_unit}
\title{Retrieve x combustion ethylene unit data from ghg database}
\usage{
ghg_x_combustion_ethylene_unit(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  OFF_GAS_FRACTION = NULL, UNIT_NAME_COMBUSTION_ETHYLENE = NULL,
  UNIT_TYPE_COMBUSTION_ETHYLENE = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000342'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{OFF_GAS_FRACTION}{e.g. '0.1209'. See Details.}

\item{UNIT_NAME_COMBUSTION_ETHYLENE}{e.g. 'Ethylene Unit'. See Details.}

\item{UNIT_TYPE_COMBUSTION_ETHYLENE}{e.g. 'Petrochemical process unit'. See Details.}
}
\description{
Retrieve x combustion ethylene unit data from ghg database
}

