% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_x_flare_information.R
\name{ghg_x_flare_information}
\alias{ghg_x_flare_information}
\title{Retrieve x flare information data from ghg database}
\usage{
ghg_x_flare_information(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  Y1A_MOL_WT_METHOD_OTHER = NULL, Y1A_C_CONTENT = NULL,
  Y1A_C_CONTENT_METHOD_OTHER = NULL, Y1B_MEASUREMENT_PERIOD = NULL,
  Y1B_VOL_FLARE_GAS = NULL, Y1B_VOL_FLARE_GAS_METHOD = NULL,
  Y1B_CO2_CONCENTRATION = NULL, Y1B_CO2_CONCENTRATION_METHOD = NULL,
  Y1B_CO2_CONC_METHOD_OTHER = NULL,
  Y1B_NUMBER_NONCO2_C_COMPOUNDS = NULL, Y2_MEASUREMENT_PERIOD = NULL,
  Y2_VOL_FLARE_GAS = NULL, Y2_VOL_FLARE_GAS_METHOD = NULL,
  Y2_HHV = NULL, Y2_HHV_METHOD = NULL, Y2_HHV_METHOD_OTHER = NULL,
  Y2_VOL_FLARE_GAS_CONDITIONS = NULL, Y2_HHV_CONDITIONS = NULL,
  Y3_NUMBER_SSM_EVENTS = NULL, FLARE_TYPE = NULL,
  FLARE_TYPE_OTHER = NULL, FLARE_SERVICE = NULL,
  FLARE_SERVICE_OTHER = NULL, C_FROM_CH4_FRACTION_BASIS = NULL,
  C_FROM_CH4_FRACTION_BASIS_OTHR = NULL, UNIT_NAME = NULL,
  UNIT_TYPE = NULL, CO2_CALCULATION_METHOD = NULL,
  FLARE_CO2_EMISSIONS = NULL, FLARE_CH4_EMISSIONS = NULL,
  FLARE_N2O_EMISSIONS = NULL, Y1A_MEASUREMENT_PERIOD = NULL,
  Y1A_VOL_FLARE_GAS = NULL, Y1A_VOL_FLARE_GAS_METHOD = NULL,
  Y1A_MOL_WT = NULL, Y1A_C_CONTENT_METHOD = NULL,
  Y1A_MOL_WT_METHOD = NULL, Y1B_CONC_NON_CO2_C_COMP_METHD = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000342'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{Y1A_MOL_WT_METHOD_OTHER}{e.g. 'NA'. See Details.}

\item{Y1A_C_CONTENT}{e.g. '0.324'. See Details.}

\item{Y1A_C_CONTENT_METHOD_OTHER}{e.g. 'NA'. See Details.}

\item{Y1B_MEASUREMENT_PERIOD}{e.g. 'NA'. See Details.}

\item{Y1B_VOL_FLARE_GAS}{e.g. 'NA'. See Details.}

\item{Y1B_VOL_FLARE_GAS_METHOD}{e.g. 'NA'. See Details.}

\item{Y1B_CO2_CONCENTRATION}{e.g. 'NA'. See Details.}

\item{Y1B_CO2_CONCENTRATION_METHOD}{e.g. 'NA'. See Details.}

\item{Y1B_CO2_CONC_METHOD_OTHER}{e.g. 'NA'. See Details.}

\item{Y1B_NUMBER_NONCO2_C_COMPOUNDS}{e.g. 'NA'. See Details.}

\item{Y2_MEASUREMENT_PERIOD}{e.g. 'NA'. See Details.}

\item{Y2_VOL_FLARE_GAS}{e.g. 'NA'. See Details.}

\item{Y2_VOL_FLARE_GAS_METHOD}{e.g. 'NA'. See Details.}

\item{Y2_HHV}{e.g. 'NA'. See Details.}

\item{Y2_HHV_METHOD}{e.g. 'NA'. See Details.}

\item{Y2_HHV_METHOD_OTHER}{e.g. 'NA'. See Details.}

\item{Y2_VOL_FLARE_GAS_CONDITIONS}{e.g. 'NA'. See Details.}

\item{Y2_HHV_CONDITIONS}{e.g. 'NA'. See Details.}

\item{Y3_NUMBER_SSM_EVENTS}{e.g. 'NA'. See Details.}

\item{FLARE_TYPE}{e.g. 'Steam assisted'. See Details.}

\item{FLARE_TYPE_OTHER}{e.g. 'NA'. See Details.}

\item{FLARE_SERVICE}{e.g. 'Unit flare'. See Details.}

\item{FLARE_SERVICE_OTHER}{e.g. 'NA'. See Details.}

\item{C_FROM_CH4_FRACTION_BASIS}{e.g. 'Chromatographic analysis: manufacturer's instructions'. See Details.}

\item{C_FROM_CH4_FRACTION_BASIS_OTHR}{e.g. 'NA'. See Details.}

\item{UNIT_NAME}{e.g. 'Ethylene Plant Flare 11-MEU-801'. See Details.}

\item{UNIT_TYPE}{e.g. 'Flare'. See Details.}

\item{CO2_CALCULATION_METHOD}{e.g. '98.253(b)(1)(ii)(A) - Equation Y-1a Gas Composition Monitored'. See Details.}

\item{FLARE_CO2_EMISSIONS}{e.g. '4650'. See Details.}

\item{FLARE_CH4_EMISSIONS}{e.g. '23.8'. See Details.}

\item{FLARE_N2O_EMISSIONS}{e.g. '0.047'. See Details.}

\item{Y1A_MEASUREMENT_PERIOD}{e.g. 'Daily'. See Details.}

\item{Y1A_VOL_FLARE_GAS}{e.g. '124856305'. See Details.}

\item{Y1A_VOL_FLARE_GAS_METHOD}{e.g. 'GE Panametrics ultrasonic flow meters calibrated annualy.  Calibrated by the manufacturer.'. See Details.}

\item{Y1A_MOL_WT}{e.g. '20.68'. See Details.}

\item{Y1A_C_CONTENT_METHOD}{e.g. 'Chromatographic analysis: manufacturer's instructions'. See Details.}

\item{Y1A_MOL_WT_METHOD}{e.g. 'ASTM D1945-03,Chromatographic analysis: manufacturer's instructions'. See Details.}

\item{Y1B_CONC_NON_CO2_C_COMP_METHD}{e.g. ' '. See Details.}
}
\description{
Retrieve x flare information data from ghg database
}
