% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_x_process_unit_details.R
\name{ghg_x_process_unit_details}
\alias{ghg_x_process_unit_details}
\title{Retrieve x process unit details data from ghg database}
\usage{
ghg_x_process_unit_details(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  UNIT_NAME_PROCESS = NULL, PROCESS_UNIT_CO2_FRACTION = NULL,
  PROCESS_UNIT_CH4_FRACTION = NULL, PROCESS_UNIT_N2O_FRACTION = NULL,
  CARBON_DIOXIDE_FRACTION = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1004900'. See Details.}

\item{REPORTING_YEAR}{e.g. '2010'. See Details.}

\item{UNIT_NAME_PROCESS}{e.g. 'CS-1'. See Details.}

\item{PROCESS_UNIT_CO2_FRACTION}{e.g. '1'. See Details.}

\item{PROCESS_UNIT_CH4_FRACTION}{e.g. '1'. See Details.}

\item{PROCESS_UNIT_N2O_FRACTION}{e.g. '1'. See Details.}

\item{CARBON_DIOXIDE_FRACTION}{e.g. 'NA'. See Details.}
}
\description{
Retrieve x process unit details data from ghg database
}
