% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_x_subpart_level_information.R
\name{ghg_x_subpart_level_information}
\alias{ghg_x_subpart_level_information}
\title{Retrieve x subpart level information data from ghg database}
\usage{
ghg_x_subpart_level_information(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  GHG_NAME = NULL, GHG_QUANTITY = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000232'. See Details.}

\item{REPORTING_YEAR}{e.g. '2013'. See Details.}

\item{GHG_NAME}{e.g. 'Biogenic Carbon dioxide'. See Details.}

\item{GHG_QUANTITY}{e.g. '0'. See Details.}
}
\description{
Retrieve x subpart level information data from ghg database
}

