% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_x_t4cemsdetails.R
\name{ghg_x_t4cemsdetails}
\alias{ghg_x_t4cemsdetails}
\title{Retrieve x t4cemsdetails data from ghg database}
\usage{
ghg_x_t4cemsdetails(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  CML_TOTAL_OPERATING_HOURS = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1004900'. See Details.}

\item{REPORTING_YEAR}{e.g. '2012'. See Details.}

\item{CML_TOTAL_OPERATING_HOURS}{e.g. '8270'. See Details.}
}
\description{
Retrieve x t4cemsdetails data from ghg database
}

