% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_y_asphaltblowing_info.R
\name{ghg_y_asphaltblowing_info}
\alias{ghg_y_asphaltblowing_info}
\title{Retrieve y asphaltblowing info data from ghg database}
\usage{
ghg_y_asphaltblowing_info(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, UNIT_NAME = NULL, UNIT_TYPE = NULL,
  CONTROL_DEVICE_TYPE = NULL, CONTROL_DEVICE_TYPE_OTHER = NULL,
  C_EMISSION_FACTOR_BASIS_Y16B = NULL,
  C_EMISS_FACTOR_BASIS_Y16B_OTHR = NULL, VOL_FLOW_DISCHARGED_IS_SUB = NULL,
  VOLUMETRIC_FLOW_DISCHARGED = NULL, VOL_FLOW_DISCHARGED_NUM_SUB = NULL,
  VOL_FLOW_DISCHARGED_METHOD = NULL, NUM_OF_VENTING_EVNTS_IS_SUB = NULL,
  NUMBER_OF_VENTING_EVENTS = NULL, NUM_OF_VENTING_EVNTS_NUM_SUB = NULL,
  VENTING_TIME = NULL, CH4_MOLE_FRACTION = NULL,
  CH4_MOLE_FRAC_NUM_SUBSTITUTED = NULL, CH4_MOLE_FRACTION_METHOD = NULL,
  CO2_MOLE_FRAC_IS_SUBSTITUTED = NULL, CO2_MOLE_FRACTION = NULL,
  CO2_MOLE_FRAC_NUM_SUBSTITUTED = NULL, CO2_MOLE_FRACTION_METHOD = NULL,
  N2O_MOLE_FRAC_IS_SUBSTITUTED = NULL, N2O_MOLE_FRACTION = NULL,
  N2O_MOLE_FRAC_NUM_SUBSTITUTED = NULL,
  CH4_MOLE_FRACTION_METHOD_OTHER = NULL,
  CO2_MOLE_FRACTION_METHOD_OTHER = NULL,
  N2O_MOLE_FRACTION_METHOD_OTHER = NULL,
  VOL_FLOW_DISCHARGED_MTHD_OTHER = NULL, CO2_CALCULATION_METHOD = NULL,
  MASS_CO2 = NULL, CH4_EMISSIONS = NULL,
  CO2_EMISSION_FACTOR_BASIS_Y14 = NULL,
  CO2_EMIS_FACTOR_BASIS_Y14_OTHR = NULL,
  CH4_EMISSION_FACTOR_BASIS_Y14 = NULL,
  CH4_EMM_FACTOR_BASIS_Y14_OTHER = NULL,
  CH4_EMISSION_FACTOR_BASIS_Y17 = NULL,
  CH4_EMISS_FACTOR_BAS_Y17_OTHR = NULL, C_EMISSION_FACTOR_BASIS_Y16A = NULL,
  C_EMISS_FACTOR_BASIS_Y16A_OTHR = NULL, CO2_EMM_FACTOR_BASIS_Y16B = NULL,
  CO2_EMISS_FACTOR_BAS_Y16B_OTHR = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1002261'. See Details.}

\item{REPORTING_YEAR}{e.g. '2010'. See Details.}

\item{FACILITY_NAME}{e.g. 'TRICOR REFINING LLC'. See Details.}

\item{UNIT_NAME}{e.g. 'ABA Plant'. See Details.}

\item{UNIT_TYPE}{e.g. 'Asphalt Blowing Unit'. See Details.}

\item{CONTROL_DEVICE_TYPE}{e.g. 'Other (specify)'. See Details.}

\item{CONTROL_DEVICE_TYPE_OTHER}{e.g. 'incinerator'. See Details.}

\item{C_EMISSION_FACTOR_BASIS_Y16B}{e.g. 'NA'. See Details.}

\item{C_EMISS_FACTOR_BASIS_Y16B_OTHR}{e.g. 'NA'. See Details.}

\item{VOL_FLOW_DISCHARGED_IS_SUB}{e.g. 'NA'. See Details.}

\item{VOLUMETRIC_FLOW_DISCHARGED}{e.g. 'NA'. See Details.}

\item{VOL_FLOW_DISCHARGED_NUM_SUB}{e.g. 'NA'. See Details.}

\item{VOL_FLOW_DISCHARGED_METHOD}{e.g. 'NA'. See Details.}

\item{NUM_OF_VENTING_EVNTS_IS_SUB}{e.g. 'NA'. See Details.}

\item{NUMBER_OF_VENTING_EVENTS}{e.g. 'NA'. See Details.}

\item{NUM_OF_VENTING_EVNTS_NUM_SUB}{e.g. 'NA'. See Details.}

\item{VENTING_TIME}{e.g. 'NA'. See Details.}

\item{CH4_MOLE_FRACTION}{e.g. 'NA'. See Details.}

\item{CH4_MOLE_FRAC_NUM_SUBSTITUTED}{e.g. 'NA'. See Details.}

\item{CH4_MOLE_FRACTION_METHOD}{e.g. 'NA'. See Details.}

\item{CO2_MOLE_FRAC_IS_SUBSTITUTED}{e.g. 'NA'. See Details.}

\item{CO2_MOLE_FRACTION}{e.g. 'NA'. See Details.}

\item{CO2_MOLE_FRAC_NUM_SUBSTITUTED}{e.g. 'NA'. See Details.}

\item{CO2_MOLE_FRACTION_METHOD}{e.g. 'NA'. See Details.}

\item{N2O_MOLE_FRAC_IS_SUBSTITUTED}{e.g. 'NA'. See Details.}

\item{N2O_MOLE_FRACTION}{e.g. 'NA'. See Details.}

\item{N2O_MOLE_FRAC_NUM_SUBSTITUTED}{e.g. 'NA'. See Details.}

\item{CH4_MOLE_FRACTION_METHOD_OTHER}{e.g. 'NA'. See Details.}

\item{CO2_MOLE_FRACTION_METHOD_OTHER}{e.g. 'NA'. See Details.}

\item{N2O_MOLE_FRACTION_METHOD_OTHER}{e.g. 'NA'. See Details.}

\item{VOL_FLOW_DISCHARGED_MTHD_OTHER}{e.g. 'NA'. See Details.}

\item{CO2_CALCULATION_METHOD}{e.g. 'Equation Y-16a and Y-17'. See Details.}

\item{MASS_CO2}{e.g. '766.8'. See Details.}

\item{CH4_EMISSIONS}{e.g. '0.9'. See Details.}

\item{CO2_EMISSION_FACTOR_BASIS_Y14}{e.g. 'NA'. See Details.}

\item{CO2_EMIS_FACTOR_BASIS_Y14_OTHR}{e.g. 'NA'. See Details.}

\item{CH4_EMISSION_FACTOR_BASIS_Y14}{e.g. 'NA'. See Details.}

\item{CH4_EMM_FACTOR_BASIS_Y14_OTHER}{e.g. 'NA'. See Details.}

\item{CH4_EMISSION_FACTOR_BASIS_Y17}{e.g. 'Used default emission factor'. See Details.}

\item{CH4_EMISS_FACTOR_BAS_Y17_OTHR}{e.g. 'NA'. See Details.}

\item{C_EMISSION_FACTOR_BASIS_Y16A}{e.g. 'Used default emission factor'. See Details.}

\item{C_EMISS_FACTOR_BASIS_Y16A_OTHR}{e.g. 'NA'. See Details.}

\item{CO2_EMM_FACTOR_BASIS_Y16B}{e.g. 'NA'. See Details.}

\item{CO2_EMISS_FACTOR_BAS_Y16B_OTHR}{e.g. 'NA'. See Details.}
}
\description{
Retrieve y asphaltblowing info data from ghg database
}

