% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_y_cems_details.R
\name{ghg_y_cems_details}
\alias{ghg_y_cems_details}
\title{Retrieve y cems details data from ghg database}
\usage{
ghg_y_cems_details(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  CML_NAME = NULL, CML_DESCRIPTION = NULL, CML_CONFIGURATION_TYPE = NULL,
  CEMS_BIOGENIC_CO2_EMISSIONS = NULL, CEMS_NONBIOGENIC_CO2_EMISSIONS = NULL,
  CEMS_CO2_EMISSIONS = NULL, CEMS_CH4_EMISSIONS = NULL,
  CEMS_N2O_EMISSIONS = NULL, CML_TOT_NUM_OF_OPERATING_HOURS = NULL,
  TIER_4_SUB_OPR_HRS_CO2_CONC = NULL, TIER_4_SUB_OPR_HRS_GAS_FLOW_RT = NULL,
  TIER4_SUB_OPR_HRS_MOIST_CONT = NULL, CEMS_CO2_COMBUSTION_EMISSIONS = NULL,
  CEMS_METHODOLOGY_START_DATE = NULL, CEMS_METHODOLOGY_END_DATE = NULL,
  CEMS_FUEL = NULL, CEMS_QUARTER_NAME = NULL,
  CEMS_CO2_EMISSIONS_BY_QUARTER = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000411'. See Details.}

\item{REPORTING_YEAR}{e.g. '2013'. See Details.}

\item{CML_NAME}{e.g. 'P010CEMS'. See Details.}

\item{CML_DESCRIPTION}{e.g. 'FCC CEMs between regenerator and ESP stack (CO boiler was removed 10/2010)'. See Details.}

\item{CML_CONFIGURATION_TYPE}{e.g. 'Single process/process unit exhausts to dedicated stack'. See Details.}

\item{CEMS_BIOGENIC_CO2_EMISSIONS}{e.g. '0'. See Details.}

\item{CEMS_NONBIOGENIC_CO2_EMISSIONS}{e.g. '270918.8'. See Details.}

\item{CEMS_CO2_EMISSIONS}{e.g. '270918.8'. See Details.}

\item{CEMS_CH4_EMISSIONS}{e.g. 'NA'. See Details.}

\item{CEMS_N2O_EMISSIONS}{e.g. 'NA'. See Details.}

\item{CML_TOT_NUM_OF_OPERATING_HOURS}{e.g. '8760'. See Details.}

\item{TIER_4_SUB_OPR_HRS_CO2_CONC}{e.g. '376'. See Details.}

\item{TIER_4_SUB_OPR_HRS_GAS_FLOW_RT}{e.g. '0'. See Details.}

\item{TIER4_SUB_OPR_HRS_MOIST_CONT}{e.g. '0'. See Details.}

\item{CEMS_CO2_COMBUSTION_EMISSIONS}{e.g. '0'. See Details.}

\item{CEMS_METHODOLOGY_START_DATE}{e.g. '01-JAN-13'. See Details.}

\item{CEMS_METHODOLOGY_END_DATE}{e.g. '31-DEC-13'. See Details.}

\item{CEMS_FUEL}{e.g. 'NA'. See Details.}

\item{CEMS_QUARTER_NAME}{e.g. 'First Quarter'. See Details.}

\item{CEMS_CO2_EMISSIONS_BY_QUARTER}{e.g. '66603.4'. See Details.}
}
\description{
Retrieve y cems details data from ghg database
}

