% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_y_cokecalciner_info.R
\name{ghg_y_cokecalciner_info}
\alias{ghg_y_cokecalciner_info}
\title{Retrieve y cokecalciner info data from ghg database}
\usage{
ghg_y_cokecalciner_info(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, UNIT_NAME = NULL, UNIT_TYPE = NULL,
  CO2_EMISSIONS = NULL, CO2_CALCULATION_METHOD = NULL,
  CH4_EMISSIONS = NULL, CH4_CALCULATION_METHOD = NULL,
  CH4_EMISSION_FACTOR_BASIS = NULL,
  CH4_EMISSION_FACTOR_BASIS_OTHR = NULL, N2O_EMISSIONS = NULL,
  N2O_CALCULATION_METHOD = NULL, N2O_EMISSION_FACTOR_BASIS = NULL,
  N2O_EMISSION_FACTOR_BASIS_OTHR = NULL,
  C_CONTENT_GREEN_COKE_METHOD = NULL,
  C_CONTENT_GREEN_COKE_MTHD_OTHR = NULL,
  C_CONTENT_MARKETABLE_COKE_MTHD = NULL,
  C_CONT_MARKETBL_COKE_MTHD_OTHR = NULL, COKE_DUST_RECYCLING = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1006012'. See Details.}

\item{REPORTING_YEAR}{e.g. '2010'. See Details.}

\item{FACILITY_NAME}{e.g. 'PHILLIPS 66 CO - LAKE CHARLES REFINERY'. See Details.}

\item{UNIT_NAME}{e.g. 'No. 2 Calciner'. See Details.}

\item{UNIT_TYPE}{e.g. 'Coke Calcining Unit'. See Details.}

\item{CO2_EMISSIONS}{e.g. '125072.9'. See Details.}

\item{CO2_CALCULATION_METHOD}{e.g. 'Equation Y-13'. See Details.}

\item{CH4_EMISSIONS}{e.g. '3.66'. See Details.}

\item{CH4_CALCULATION_METHOD}{e.g. 'Equation Y-9 with a default emission factor'. See Details.}

\item{CH4_EMISSION_FACTOR_BASIS}{e.g. 'NA'. See Details.}

\item{CH4_EMISSION_FACTOR_BASIS_OTHR}{e.g. 'NA'. See Details.}

\item{N2O_EMISSIONS}{e.g. '0.733'. See Details.}

\item{N2O_CALCULATION_METHOD}{e.g. 'Equation Y-10 with a default emission factor'. See Details.}

\item{N2O_EMISSION_FACTOR_BASIS}{e.g. 'NA'. See Details.}

\item{N2O_EMISSION_FACTOR_BASIS_OTHR}{e.g. 'NA'. See Details.}

\item{C_CONTENT_GREEN_COKE_METHOD}{e.g. 'ASTM D5373-08'. See Details.}

\item{C_CONTENT_GREEN_COKE_MTHD_OTHR}{e.g. 'NA'. See Details.}

\item{C_CONTENT_MARKETABLE_COKE_MTHD}{e.g. 'ASTM D5373-08'. See Details.}

\item{C_CONT_MARKETBL_COKE_MTHD_OTHR}{e.g. 'NA'. See Details.}

\item{COKE_DUST_RECYCLING}{e.g. 'None of the dust is recycled'. See Details.}
}
\description{
Retrieve y cokecalciner info data from ghg database
}
