% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_y_crackingcokingreforming_info.R
\name{ghg_y_crackingcokingreforming_info}
\alias{ghg_y_crackingcokingreforming_info}
\title{Retrieve y crackingcokingreforming info data from ghg database}
\usage{
ghg_y_crackingcokingreforming_info(FACILITY_ID = NULL,
  REPORTING_YEAR = NULL, CH4_EMISSION_FACTOR_BASIS_OTHR = NULL,
  N2O_EMISSIONS = NULL, N2O_CALCULATION_METHOD = NULL,
  N2O_EMISSION_FACTOR_BASIS = NULL, N2O_EMISSION_FACTOR_BASIS_OTHR = NULL,
  FLEXICOKER_EMISS_IN_SUBPART_C = NULL,
  VOLUMETRIC_FLOW_RT_EXHAUST_GAS = NULL,
  PERCENT_CO2_EXHAUST_GAS_METHOD = NULL,
  PERCENT_CO_EXHAUST_GAS_METHOD = NULL,
  PERCENT_O2_EXHAUST_GAS_METHOD = NULL,
  PERCENT_N2_EXHAUST_GAS_METHOD = NULL, PCT_N2_EXHAUST_GAS_MTHD_OTHR = NULL,
  PERCENT_N2_INLET_GAS_METHOD = NULL, FACILITY_NAME = NULL,
  UNIT_NAME = NULL, UNIT_TYPE = NULL, CO2_EMISSIONS = NULL,
  CO2_CALCULATION_METHOD = NULL, CH4_EMISSIONS = NULL,
  CH4_CALCULATION_METHOD = NULL, CH4_EMISSION_FACTOR_BASIS = NULL,
  PERCENT_N2_INLET_GAS_MTHD_OTHR = NULL, C_CONTENT_BASIS_Y8 = NULL,
  C_CONTENT_BASIS_Y8_OTHER = NULL, COKE_BURN_OFF_QUANTITY = NULL,
  COKE_BURN_OFF_QUANTITY_UOM = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000025'. See Details.}

\item{REPORTING_YEAR}{e.g. '2012'. See Details.}

\item{CH4_EMISSION_FACTOR_BASIS_OTHR}{e.g. 'NA'. See Details.}

\item{N2O_EMISSIONS}{e.g. '0'. See Details.}

\item{N2O_CALCULATION_METHOD}{e.g. 'Equation Y-10 with a default emission factor'. See Details.}

\item{N2O_EMISSION_FACTOR_BASIS}{e.g. 'NA'. See Details.}

\item{N2O_EMISSION_FACTOR_BASIS_OTHR}{e.g. 'NA'. See Details.}

\item{FLEXICOKER_EMISS_IN_SUBPART_C}{e.g. 'NA'. See Details.}

\item{VOLUMETRIC_FLOW_RT_EXHAUST_GAS}{e.g. 'NA'. See Details.}

\item{PERCENT_CO2_EXHAUST_GAS_METHOD}{e.g. 'NA'. See Details.}

\item{PERCENT_CO_EXHAUST_GAS_METHOD}{e.g. 'NA'. See Details.}

\item{PERCENT_O2_EXHAUST_GAS_METHOD}{e.g. 'NA'. See Details.}

\item{PERCENT_N2_EXHAUST_GAS_METHOD}{e.g. 'NA'. See Details.}

\item{PCT_N2_EXHAUST_GAS_MTHD_OTHR}{e.g. 'NA'. See Details.}

\item{PERCENT_N2_INLET_GAS_METHOD}{e.g. 'NA'. See Details.}

\item{FACILITY_NAME}{e.g. 'PLACID REFINING CO LLC - PORT ALLEN REFINERY'. See Details.}

\item{UNIT_NAME}{e.g. 'Reformer'. See Details.}

\item{UNIT_TYPE}{e.g. 'Catalytic Reforming Unit'. See Details.}

\item{CO2_EMISSIONS}{e.g. '28.6'. See Details.}

\item{CO2_CALCULATION_METHOD}{e.g. 'Equation Y-11 - 98.253(e)(3)'. See Details.}

\item{CH4_EMISSIONS}{e.g. '0'. See Details.}

\item{CH4_CALCULATION_METHOD}{e.g. 'Equation Y-9 with a default emission factor'. See Details.}

\item{CH4_EMISSION_FACTOR_BASIS}{e.g. 'NA'. See Details.}

\item{PERCENT_N2_INLET_GAS_MTHD_OTHR}{e.g. 'NA'. See Details.}

\item{C_CONTENT_BASIS_Y8}{e.g. 'NA'. See Details.}

\item{C_CONTENT_BASIS_Y8_OTHER}{e.g. 'NA'. See Details.}

\item{COKE_BURN_OFF_QUANTITY}{e.g. '8287'. See Details.}

\item{COKE_BURN_OFF_QUANTITY_UOM}{e.g. 'kg/cycle or kg/measurement period'. See Details.}
}
\description{
Retrieve y crackingcokingreforming info data from ghg database
}

