% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_y_delayedcoking_info.R
\name{ghg_y_delayedcoking_info}
\alias{ghg_y_delayedcoking_info}
\title{Retrieve y delayedcoking info data from ghg database}
\usage{
ghg_y_delayedcoking_info(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, TOTAL_NUMBER_OF_UNITS = NULL,
  CH4_EMISSIONS = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000025'. See Details.}

\item{REPORTING_YEAR}{e.g. '2012'. See Details.}

\item{FACILITY_NAME}{e.g. 'PLACID REFINING CO LLC - PORT ALLEN REFINERY'. See Details.}

\item{TOTAL_NUMBER_OF_UNITS}{e.g. 'NA'. See Details.}

\item{CH4_EMISSIONS}{e.g. '0'. See Details.}
}
\description{
Retrieve y delayedcoking info data from ghg database
}

