% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_y_emissions_by_source.R
\name{ghg_y_emissions_by_source}
\alias{ghg_y_emissions_by_source}
\title{Retrieve y emissions by source data from ghg database}
\usage{
ghg_y_emissions_by_source(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, BLOWDOWN_CH4_EMISSIONS = NULL,
  EQUIP_LEAK_CH4_EMISSIONS = NULL, LOADINGOPS_CH4_EMISSIONS = NULL,
  UNSTABALIZEDCRUDE_CH4_EMISS = NULL, STORAGETANKS_CH4_EMISSIONS = NULL,
  SOUR_GAS_CO2_EMISSIONS = NULL, DELAYED_COKING_CH4_EMISSIONS = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000025'. See Details.}

\item{REPORTING_YEAR}{e.g. '2012'. See Details.}

\item{FACILITY_NAME}{e.g. 'PLACID REFINING CO LLC - PORT ALLEN REFINERY'. See Details.}

\item{BLOWDOWN_CH4_EMISSIONS}{e.g. 'NA'. See Details.}

\item{EQUIP_LEAK_CH4_EMISSIONS}{e.g. '10.5'. See Details.}

\item{LOADINGOPS_CH4_EMISSIONS}{e.g. '0'. See Details.}

\item{UNSTABALIZEDCRUDE_CH4_EMISS}{e.g. 'NA'. See Details.}

\item{STORAGETANKS_CH4_EMISSIONS}{e.g. '0'. See Details.}

\item{SOUR_GAS_CO2_EMISSIONS}{e.g. 'NA'. See Details.}

\item{DELAYED_COKING_CH4_EMISSIONS}{e.g. '0'. See Details.}
}
\description{
Retrieve y emissions by source data from ghg database
}

