% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_y_equipmentleak_info.R
\name{ghg_y_equipmentleak_info}
\alias{ghg_y_equipmentleak_info}
\title{Retrieve y equipmentleak info data from ghg database}
\usage{
ghg_y_equipmentleak_info(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, CH4_CALCULATION_METHOD = NULL,
  CH4_EMISSIONS = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000025'. See Details.}

\item{REPORTING_YEAR}{e.g. '2012'. See Details.}

\item{FACILITY_NAME}{e.g. 'PLACID REFINING CO LLC - PORT ALLEN REFINERY'. See Details.}

\item{CH4_CALCULATION_METHOD}{e.g. '98.253(l)(2) method'. See Details.}

\item{CH4_EMISSIONS}{e.g. '10.5'. See Details.}
}
\description{
Retrieve y equipmentleak info data from ghg database
}

