% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_y_loadingoperations_info.R
\name{ghg_y_loadingoperations_info}
\alias{ghg_y_loadingoperations_info}
\title{Retrieve y loadingoperations info data from ghg database}
\usage{
ghg_y_loadingoperations_info(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, CH4_EMISSIONS = NULL, VESSEL_ID = NULL,
  VESSEL_TYPE_OTHER = NULL, VESSEL_TYPE = NULL, MATERIAL_ID = NULL,
  CONTROL_SYSTEM_TYPE = NULL, CONTROL_SYSTEM_TYPE_OTHER = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000025'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{FACILITY_NAME}{e.g. 'PLACID REFINING CO LLC - PORT ALLEN REFINERY'. See Details.}

\item{CH4_EMISSIONS}{e.g. '0'. See Details.}

\item{VESSEL_ID}{e.g. 'NA'. See Details.}

\item{VESSEL_TYPE_OTHER}{e.g. 'NA'. See Details.}

\item{VESSEL_TYPE}{e.g. 'NA'. See Details.}

\item{MATERIAL_ID}{e.g. 'NA'. See Details.}

\item{CONTROL_SYSTEM_TYPE}{e.g. 'NA'. See Details.}

\item{CONTROL_SYSTEM_TYPE_OTHER}{e.g. 'NA'. See Details.}
}
\description{
Retrieve y loadingoperations info data from ghg database
}
