% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_y_processvent_info.R
\name{ghg_y_processvent_info}
\alias{ghg_y_processvent_info}
\title{Retrieve y processvent info data from ghg database}
\usage{
ghg_y_processvent_info(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  VOLUMETRIC_FLOW_IS_SUBSTITUTED = NULL, VOLUMETRIC_FLOW_DISCHARGED = NULL,
  VOLUMETRIC_FLOW_NUMBER_SUB = NULL, CH4_EMISSIONS = NULL,
  CO2_EMISSIONS = NULL, CONTROL_DEVICE_TYPE = NULL, N2O_EMISSIONS = NULL,
  ASSOCIATED_OPERATION = NULL, VOLUMETRIC_FLOW_METHOD_OTHER = NULL,
  CONTROL_DEVICE_TYPE_OTHER = NULL, ASSOCIATED_OPERATION_OTHER = NULL,
  UNIT_NAME = NULL, UNIT_TYPE = NULL, CH4_MOLE_FRAC_IS_SUBSTITUTED = NULL,
  CH4_MOLE_FRACTION = NULL, CH4_MOLE_FRAC_NUMBER_SUB = NULL,
  CH4_MOLE_FRACTION_METHOD = NULL, CO2_MOLE_FRAC_IS_SUBSTITUTED = NULL,
  CO2_MOLE_FRACTION = NULL, CO2_MOLE_FRAC_NUM_SUBSTITUTED = NULL,
  CO2_MOLE_FRACTION_METHOD = NULL, N2O_MOLE_FRAC_IS_SUBSTITUTED = NULL,
  N2O_MOLE_FRACTION = NULL, N2O_MOLE_FRAC_NUM_SUBSTITUTED = NULL,
  N2O_MOLE_FRACTION_METHOD = NULL, CH4_MOLE_FRACTION_METHOD_OTHER = NULL,
  CO2_MOLE_FRACTION_METHOD_OTHER = NULL,
  N2O_MOLE_FRACTION_METHOD_OTHER = NULL, NUMBER_OF_VENTING_EVENTS = NULL,
  TOTAL_VENTING_TIME = NULL, FACILITY_NAME = NULL,
  VOLUMETRIC_FLOW_METHOD = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000099'. See Details.}

\item{REPORTING_YEAR}{e.g. '2011'. See Details.}

\item{VOLUMETRIC_FLOW_IS_SUBSTITUTED}{e.g. 'NA'. See Details.}

\item{VOLUMETRIC_FLOW_DISCHARGED}{e.g. '13516583'. See Details.}

\item{VOLUMETRIC_FLOW_NUMBER_SUB}{e.g. 'NA'. See Details.}

\item{CH4_EMISSIONS}{e.g. '0.03'. See Details.}

\item{CO2_EMISSIONS}{e.g. '711'. See Details.}

\item{CONTROL_DEVICE_TYPE}{e.g. 'Other (specify)'. See Details.}

\item{N2O_EMISSIONS}{e.g. '0.004'. See Details.}

\item{ASSOCIATED_OPERATION}{e.g. 'Catalytic reforming unit - continuous regeneration'. See Details.}

\item{VOLUMETRIC_FLOW_METHOD_OTHER}{e.g. 'NA'. See Details.}

\item{CONTROL_DEVICE_TYPE_OTHER}{e.g. 'Heater firebox'. See Details.}

\item{ASSOCIATED_OPERATION_OTHER}{e.g. 'Catalytic reforming unit - continuous regeneration'. See Details.}

\item{UNIT_NAME}{e.g. 'Vent1'. See Details.}

\item{UNIT_TYPE}{e.g. 'Process Vent'. See Details.}

\item{CH4_MOLE_FRAC_IS_SUBSTITUTED}{e.g. 'NA'. See Details.}

\item{CH4_MOLE_FRACTION}{e.g. '1e-04'. See Details.}

\item{CH4_MOLE_FRAC_NUMBER_SUB}{e.g. 'NA'. See Details.}

\item{CH4_MOLE_FRACTION_METHOD}{e.g. 'Engineering estimates/process knowledge'. See Details.}

\item{CO2_MOLE_FRAC_IS_SUBSTITUTED}{e.g. 'NA'. See Details.}

\item{CO2_MOLE_FRACTION}{e.g. '0.9999'. See Details.}

\item{CO2_MOLE_FRAC_NUM_SUBSTITUTED}{e.g. 'NA'. See Details.}

\item{CO2_MOLE_FRACTION_METHOD}{e.g. 'Engineering estimates/process knowledge'. See Details.}

\item{N2O_MOLE_FRAC_IS_SUBSTITUTED}{e.g. 'NA'. See Details.}

\item{N2O_MOLE_FRACTION}{e.g. '1e-05'. See Details.}

\item{N2O_MOLE_FRAC_NUM_SUBSTITUTED}{e.g. 'NA'. See Details.}

\item{N2O_MOLE_FRACTION_METHOD}{e.g. 'Engineering estimates/process knowledge'. See Details.}

\item{CH4_MOLE_FRACTION_METHOD_OTHER}{e.g. 'NA'. See Details.}

\item{CO2_MOLE_FRACTION_METHOD_OTHER}{e.g. 'NA'. See Details.}

\item{N2O_MOLE_FRACTION_METHOD_OTHER}{e.g. 'NA'. See Details.}

\item{NUMBER_OF_VENTING_EVENTS}{e.g. '365'. See Details.}

\item{TOTAL_VENTING_TIME}{e.g. 'Hours'. See Details.}

\item{FACILITY_NAME}{e.g. 'Marathon Petroleum Company LP'. See Details.}

\item{VOLUMETRIC_FLOW_METHOD}{e.g. 'Engineering calculation'. See Details.}
}
\description{
Retrieve y processvent info data from ghg database
}

