% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_y_storagetank_info.R
\name{ghg_y_storagetank_info}
\alias{ghg_y_storagetank_info}
\title{Retrieve y storagetank info data from ghg database}
\usage{
ghg_y_storagetank_info(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  CH4_VAPORCOMP_UNSTABCRUDE_MTHD = NULL, FACILITY_NAME = NULL,
  CH4_MOLE_FRAC_BAS_UNSTCRD_OTHR = NULL,
  CH4_CALC_MTHD_STRGTANKS_GENL = NULL,
  CH4_EMISSIONS_STRGTANKS_GENL = NULL,
  CH4_MOLE_FRAC_BAS_UNSTCRD = NULL,
  CH4_CALC_MTHD_UNSTABILIZEDCRD = NULL,
  CH4_EMISS_STRGTANKS_UNSTCRD = NULL,
  CH4_VPRCOMP_UNSTABILIZEDCRUDE = NULL,
  CH4_VPRCOMP_UNSTCRD_NUM_SUB = NULL, CH4_GASGENRATE_UNSTCRUDE = NULL,
  CH4_GASGENRATE_UNSTCRD_NUM_SUB = NULL,
  CH4_GASGENRATE_UNSTCRUDE_MTHD = NULL,
  CH4_GASGENRT_UNSTCRD_MTHD_OTHR = NULL,
  CH4_VPRCOMP_UNSTCRD_MTHD_OTHR = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000025'. See Details.}

\item{REPORTING_YEAR}{e.g. '2012'. See Details.}

\item{CH4_VAPORCOMP_UNSTABCRUDE_MTHD}{e.g. 'NA'. See Details.}

\item{FACILITY_NAME}{e.g. 'PLACID REFINING CO LLC - PORT ALLEN REFINERY'. See Details.}

\item{CH4_MOLE_FRAC_BAS_UNSTCRD_OTHR}{e.g. 'NA'. See Details.}

\item{CH4_CALC_MTHD_STRGTANKS_GENL}{e.g. 'Equation Y-22'. See Details.}

\item{CH4_EMISSIONS_STRGTANKS_GENL}{e.g. '0'. See Details.}

\item{CH4_MOLE_FRAC_BAS_UNSTCRD}{e.g. 'NA'. See Details.}

\item{CH4_CALC_MTHD_UNSTABILIZEDCRD}{e.g. 'NA'. See Details.}

\item{CH4_EMISS_STRGTANKS_UNSTCRD}{e.g. 'NA'. See Details.}

\item{CH4_VPRCOMP_UNSTABILIZEDCRUDE}{e.g. 'NA'. See Details.}

\item{CH4_VPRCOMP_UNSTCRD_NUM_SUB}{e.g. 'NA'. See Details.}

\item{CH4_GASGENRATE_UNSTCRUDE}{e.g. 'NA'. See Details.}

\item{CH4_GASGENRATE_UNSTCRD_NUM_SUB}{e.g. 'NA'. See Details.}

\item{CH4_GASGENRATE_UNSTCRUDE_MTHD}{e.g. 'NA'. See Details.}

\item{CH4_GASGENRT_UNSTCRD_MTHD_OTHR}{e.g. 'NA'. See Details.}

\item{CH4_VPRCOMP_UNSTCRD_MTHD_OTHR}{e.g. 'NA'. See Details.}
}
\description{
Retrieve y storagetank info data from ghg database
}
