% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_y_subpart_level_information.R
\name{ghg_y_subpart_level_information}
\alias{ghg_y_subpart_level_information}
\title{Retrieve y subpart level information data from ghg database}
\usage{
ghg_y_subpart_level_information(FACILITY_ID = NULL,
  REPORTING_YEAR = NULL, GHG_QUANTITY = NULL, FACILITY_NAME = NULL,
  GHG_NAME = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000025'. See Details.}

\item{REPORTING_YEAR}{e.g. '2012'. See Details.}

\item{GHG_QUANTITY}{e.g. '206127.1'. See Details.}

\item{FACILITY_NAME}{e.g. 'PLACID REFINING CO LLC - PORT ALLEN REFINERY'. See Details.}

\item{GHG_NAME}{e.g. 'Carbon Dioxide'. See Details.}
}
\description{
Retrieve y subpart level information data from ghg database
}
