% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_y_sulfurrecovery_info.R
\name{ghg_y_sulfurrecovery_info}
\alias{ghg_y_sulfurrecovery_info}
\title{Retrieve y sulfurrecovery info data from ghg database}
\usage{
ghg_y_sulfurrecovery_info(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  CO2_MOLE_FRAC_IS_SUBSTITUTED = NULL, CO2_MOLE_FRACTION = NULL,
  CO2_MOLE_FRAC_NUM_SUBSTITUTED = NULL, CO2_MOLE_FRACTION_METHOD = NULL,
  N2O_MOLE_FRAC_IS_SUBSTITUTED = NULL, N2O_MOLE_FRACTION = NULL,
  N2O_MOLE_FRAC_NUM_SUBSTITUTED = NULL, N2O_MOLE_FRACTION_METHOD = NULL,
  CH4_MOLE_FRACTION_METHOD_OTHER = NULL,
  CO2_MOLE_FRACTION_METHOD_OTHER = NULL, RECYCLED_TAIL_GAS_INCLUDED = NULL,
  VOL_SOUR_GAS_NUM_SUBSTITUTED = NULL, VOLUME_SOUR_GAS_METHOD = NULL,
  SOUR_GAS_C_MOLE_FRAC_METHOD = NULL, SOUR_GAS_C_MOLE_FRAC_MTHD_OTHR = NULL,
  RECYC_TAIL_GAS_FLWRT_INCLUDED = NULL,
  RECYC_TAIL_GAS_CO2_CORREC_USED = NULL, RECYCLED_TAIL_GAS_METHOD = NULL,
  RECYCLED_TAIL_GAS_METHOD_OTHER = NULL, C_MOLE_FRACTION_METHOD = NULL,
  C_MOLE_FRACTION_METHOD_OTHER = NULL, CORRECTION_BASIS = NULL,
  CORRECTION_APPROACH = NULL, CORRECTION_APPROACH_OTHER = NULL,
  FACILITY_NAME = NULL, UNIT_NAME = NULL, UNIT_TYPE = NULL,
  CO2_EMISSIONS = NULL, CO2_CALCULATION_METHOD = NULL,
  VOLUMETRIC_FLOW_DISCHARGED = NULL, VOL_FLOW_DISCHARGED_METHOD = NULL,
  VOL_FLOW_DISCHARGED_MTHD_OTHR = NULL, NUMBER_OF_VENTING_EVENTS = NULL,
  VENTING_TIME = NULL, CH4_MOLE_FRAC_IS_SUBSTITUTED = NULL,
  CH4_MOLE_FRACTION = NULL, CH4_MOLE_FRAC_NUM_SUBSTITUTED = NULL,
  CH4_MOLE_FRACTION_METHOD = NULL, TAIL_GAS_CORR_FCTR = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1000025'. See Details.}

\item{REPORTING_YEAR}{e.g. '2012'. See Details.}

\item{CO2_MOLE_FRAC_IS_SUBSTITUTED}{e.g. 'NA'. See Details.}

\item{CO2_MOLE_FRACTION}{e.g. 'NA'. See Details.}

\item{CO2_MOLE_FRAC_NUM_SUBSTITUTED}{e.g. 'NA'. See Details.}

\item{CO2_MOLE_FRACTION_METHOD}{e.g. 'NA'. See Details.}

\item{N2O_MOLE_FRAC_IS_SUBSTITUTED}{e.g. 'NA'. See Details.}

\item{N2O_MOLE_FRACTION}{e.g. 'NA'. See Details.}

\item{N2O_MOLE_FRAC_NUM_SUBSTITUTED}{e.g. 'NA'. See Details.}

\item{N2O_MOLE_FRACTION_METHOD}{e.g. 'NA'. See Details.}

\item{CH4_MOLE_FRACTION_METHOD_OTHER}{e.g. 'NA'. See Details.}

\item{CO2_MOLE_FRACTION_METHOD_OTHER}{e.g. 'NA'. See Details.}

\item{RECYCLED_TAIL_GAS_INCLUDED}{e.g. 'N'. See Details.}

\item{VOL_SOUR_GAS_NUM_SUBSTITUTED}{e.g. '0'. See Details.}

\item{VOLUME_SOUR_GAS_METHOD}{e.g. 'Both sour water gas and acid gas are fed to the SRU and are monitored using continuous flow monitors as described in 40 CFR 98.253(f)(2).'. See Details.}

\item{SOUR_GAS_C_MOLE_FRAC_METHOD}{e.g. 'Other (specify)'. See Details.}

\item{SOUR_GAS_C_MOLE_FRAC_MTHD_OTHR}{e.g. 'Default value for mole fraction of 0.20 as allowed per 40CFR98.253(f)(4)'. See Details.}

\item{RECYC_TAIL_GAS_FLWRT_INCLUDED}{e.g. 'N'. See Details.}

\item{RECYC_TAIL_GAS_CO2_CORREC_USED}{e.g. 'NA'. See Details.}

\item{RECYCLED_TAIL_GAS_METHOD}{e.g. 'NA'. See Details.}

\item{RECYCLED_TAIL_GAS_METHOD_OTHER}{e.g. 'NA'. See Details.}

\item{C_MOLE_FRACTION_METHOD}{e.g. 'NA'. See Details.}

\item{C_MOLE_FRACTION_METHOD_OTHER}{e.g. 'NA'. See Details.}

\item{CORRECTION_BASIS}{e.g. 'NA'. See Details.}

\item{CORRECTION_APPROACH}{e.g. 'NA'. See Details.}

\item{CORRECTION_APPROACH_OTHER}{e.g. 'NA'. See Details.}

\item{FACILITY_NAME}{e.g. 'PLACID REFINING CO LLC - PORT ALLEN REFINERY'. See Details.}

\item{UNIT_NAME}{e.g. 'Sulfur Recovery Plant 3'. See Details.}

\item{UNIT_TYPE}{e.g. 'Sulfur Recovery Plant'. See Details.}

\item{CO2_EMISSIONS}{e.g. '2508.2'. See Details.}

\item{CO2_CALCULATION_METHOD}{e.g. 'Equation Y-12'. See Details.}

\item{VOLUMETRIC_FLOW_DISCHARGED}{e.g. 'NA'. See Details.}

\item{VOL_FLOW_DISCHARGED_METHOD}{e.g. 'NA'. See Details.}

\item{VOL_FLOW_DISCHARGED_MTHD_OTHR}{e.g. 'NA'. See Details.}

\item{NUMBER_OF_VENTING_EVENTS}{e.g. 'NA'. See Details.}

\item{VENTING_TIME}{e.g. 'NA'. See Details.}

\item{CH4_MOLE_FRAC_IS_SUBSTITUTED}{e.g. 'NA'. See Details.}

\item{CH4_MOLE_FRACTION}{e.g. 'NA'. See Details.}

\item{CH4_MOLE_FRAC_NUM_SUBSTITUTED}{e.g. 'NA'. See Details.}

\item{CH4_MOLE_FRACTION_METHOD}{e.g. 'NA'. See Details.}

\item{TAIL_GAS_CORR_FCTR}{e.g. 'NA'. See Details.}
}
\description{
Retrieve y sulfurrecovery info data from ghg database
}

