% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_z_co2_content_method.R
\name{ghg_z_co2_content_method}
\alias{ghg_z_co2_content_method}
\title{Retrieve z co2 content method data from ghg database}
\usage{
ghg_z_co2_content_method(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  MONTH_NAME = NULL, MASS_CONSUMED_SUB = NULL, GRABSAMPLEMETHOD = NULL,
  GRABSAMPLECO2_SUB = NULL, UNIT_TYPE = NULL, UNIT_NAME = NULL,
  FACILITY_NAME = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1003087'. See Details.}

\item{REPORTING_YEAR}{e.g. '2013'. See Details.}

\item{MONTH_NAME}{e.g. 'JANUARY'. See Details.}

\item{MASS_CONSUMED_SUB}{e.g. 'N'. See Details.}

\item{GRABSAMPLEMETHOD}{e.g. 'NA'. See Details.}

\item{GRABSAMPLECO2_SUB}{e.g. 'N'. See Details.}

\item{UNIT_TYPE}{e.g. 'WET-PROCESS PHOSPHORIC ACID LINE'. See Details.}

\item{UNIT_NAME}{e.g. 'MUSTANG-BADGER PLANT'. See Details.}

\item{FACILITY_NAME}{e.g. 'SIMPLOT PHOSPHATES'. See Details.}
}
\description{
Retrieve z co2 content method data from ghg database
}

