% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_z_facility_info.R
\name{ghg_z_facility_info}
\alias{ghg_z_facility_info}
\title{Retrieve z facility info data from ghg database}
\usage{
ghg_z_facility_info(FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  WET_PROCESS_LINES_NUMBER = NULL, FACILITY_NAME = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1001154'. See Details.}

\item{REPORTING_YEAR}{e.g. '2010'. See Details.}

\item{WET_PROCESS_LINES_NUMBER}{e.g. 'NA'. See Details.}

\item{FACILITY_NAME}{e.g. 'AGRIUM US INC'. See Details.}
}
\description{
Retrieve z facility info data from ghg database
}
