% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_z_subpart_level_information.R
\name{ghg_z_subpart_level_information}
\alias{ghg_z_subpart_level_information}
\title{Retrieve z subpart level information data from ghg database}
\usage{
ghg_z_subpart_level_information(FACILITY_ID = NULL,
  REPORTING_YEAR = NULL, GHG_QUANTITY = NULL, GHG_NAME = NULL,
  FACILITY_NAME = NULL)
}
\arguments{
\item{FACILITY_ID}{e.g. '1001154'. See Details.}

\item{REPORTING_YEAR}{e.g. '2010'. See Details.}

\item{GHG_QUANTITY}{e.g. '19991'. See Details.}

\item{GHG_NAME}{e.g. 'CARBON DIOXIDE'. See Details.}

\item{FACILITY_NAME}{e.g. 'AGRIUM US INC'. See Details.}
}
\description{
Retrieve z subpart level information data from ghg database
}
