% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igms_ogd_cfda.R
\name{igms_ogd_cfda}
\alias{igms_ogd_cfda}
\title{Retrieve ogd cfda data from igms database}
\usage{
igms_ogd_cfda(CFDA_SEQ = NULL, CFDA_CODE = NULL,
  CFDA_DESCRIPTION = NULL)
}
\arguments{
\item{CFDA_SEQ}{e.g. '208520'. See Details.}

\item{CFDA_CODE}{e.g. '66.121'. See Details.}

\item{CFDA_DESCRIPTION}{e.g. 'Puget Sound Protection and Restoration:  Tribal Implementation Assistance Program'. See Details.}
}
\description{
Retrieve ogd cfda data from igms database
}
