% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igms_ogd_ef_grant_drpt_cfda.R
\name{igms_ogd_ef_grant_drpt_cfda}
\alias{igms_ogd_ef_grant_drpt_cfda}
\title{Retrieve ogd ef grant drpt cfda data from igms database}
\usage{
igms_ogd_ef_grant_drpt_cfda(FK_CFDA_SEQ = NULL, FK_GRANT_SEQ = NULL)
}
\arguments{
\item{FK_CFDA_SEQ}{e.g. '208520'. See Details.}

\item{FK_GRANT_SEQ}{e.g. '522271097'. See Details.}
}
\description{
Retrieve ogd ef grant drpt cfda data from igms database
}

