% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igms_ogd_ef_grant_drpt_sa.R
\name{igms_ogd_ef_grant_drpt_sa}
\alias{igms_ogd_ef_grant_drpt_sa}
\title{Retrieve ogd ef grant drpt sa data from igms database}
\usage{
igms_ogd_ef_grant_drpt_sa(FK_GRANT_SEQ = NULL, FK_STAT_AUTH_SEQ = NULL)
}
\arguments{
\item{FK_GRANT_SEQ}{e.g. '522264659'. See Details.}

\item{FK_STAT_AUTH_SEQ}{e.g. '2432384'. See Details.}
}
\description{
Retrieve ogd ef grant drpt sa data from igms database
}
