% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igms_ogd_ef_grant_sum_rpt.R
\name{igms_ogd_ef_grant_sum_rpt}
\alias{igms_ogd_ef_grant_sum_rpt}
\title{Retrieve ogd ef grant sum rpt data from igms database}
\usage{
igms_ogd_ef_grant_sum_rpt(GRANT_FAMILY_SEQ = NULL, GRANT_FAMILY = NULL,
  AWARD_DATE = NULL, PROJ_BEG_DATE = NULL, PROJ_END_DATE = NULL,
  CUMULATIVE_AWARD = NULL, NUMBER_OF_AMENDMENTS = NULL)
}
\arguments{
\item{GRANT_FAMILY_SEQ}{e.g. '170466001'. See Details.}

\item{GRANT_FAMILY}{e.g. '99125706'. See Details.}

\item{AWARD_DATE}{e.g. '18-FEB-14'. See Details.}

\item{PROJ_BEG_DATE}{e.g. '01-OCT-13'. See Details.}

\item{PROJ_END_DATE}{e.g. '30-SEP-16'. See Details.}

\item{CUMULATIVE_AWARD}{e.g. '43523329'. See Details.}

\item{NUMBER_OF_AMENDMENTS}{e.g. '8'. See Details.}
}
\description{
Retrieve ogd ef grant sum rpt data from igms database
}

