% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igms_ogd_stat_authority.R
\name{igms_ogd_stat_authority}
\alias{igms_ogd_stat_authority}
\title{Retrieve ogd stat authority data from igms database}
\usage{
igms_ogd_stat_authority(STAT_AUTH_SEQ = NULL, STAT_AUTH_CODE = NULL,
  STAT_AUTH_DESCRIPTION = NULL)
}
\arguments{
\item{STAT_AUTH_SEQ}{e.g. '2432114'. See Details.}

\item{STAT_AUTH_CODE}{e.g. '0E'. See Details.}

\item{STAT_AUTH_DESCRIPTION}{e.g. 'National Environmental Educ. Act: Sec. 6'. See Details.}
}
\description{
Retrieve ogd stat authority data from igms database
}
