% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt_geo_collect_mth_lk.R
\name{lrt_geo_collect_mth_lk}
\alias{lrt_geo_collect_mth_lk}
\title{Retrieve geo collect mth lk data from lrt database}
\usage{
lrt_geo_collect_mth_lk(COLLECT_MTH_CODE = NULL, COLLECT_DESC = NULL)
}
\arguments{
\item{COLLECT_MTH_CODE}{e.g. '001'. See Details.}

\item{COLLECT_DESC}{e.g. 'ADDRESS MATCHING-HOUSE NUMBER'. See Details.}
}
\description{
Retrieve geo collect mth lk data from lrt database
}
