% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt_geo_ref_point_lk.R
\name{lrt_geo_ref_point_lk}
\alias{lrt_geo_ref_point_lk}
\title{Retrieve geo ref point lk data from lrt database}
\usage{
lrt_geo_ref_point_lk(REF_POINT_CODE = NULL, REF_POINT_DESC = NULL)
}
\arguments{
\item{REF_POINT_CODE}{e.g. '001'. See Details.}

\item{REF_POINT_DESC}{e.g. 'UNKNOWN'. See Details.}
}
\description{
Retrieve geo ref point lk data from lrt database
}
