% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcs-icis_pcs_admin_penalty_order.R
\name{pcs.icis_pcs_admin_penalty_order}
\alias{pcs.icis_pcs_admin_penalty_order}
\title{Retrieve pcs admin penalty order data from pcs-icis database}
\usage{
pcs.icis_pcs_admin_penalty_order(NPDES = NULL, ENFOR_ACTION_CODE = NULL,
  ENFOR_ACTION_TYPE = NULL, ENFOR_ACTION_DATE = NULL,
  PROPOSED_ORDER_ISSUED_DATE = NULL, ADMINISTRATIVE_PENALTY_CLASS = NULL,
  HEARING_REQUEST = NULL, FINAL_ORDER_DATE = NULL,
  PENALTY_AMOUNT_ASSESSED = NULL, CLASS_II_APPEAL_FILED = NULL,
  JUDICIAL_APPEAL_FILED_DATE = NULL, PENALTY_COLLECTED_DATE = NULL)
}
\arguments{
\item{NPDES}{e.g. 'AK0000272'. See Details.}

\item{ENFOR_ACTION_CODE}{e.g. '21'. See Details.}

\item{ENFOR_ACTION_TYPE}{e.g. 'E'. See Details.}

\item{ENFOR_ACTION_DATE}{e.g. '25-NOV-91'. See Details.}

\item{PROPOSED_ORDER_ISSUED_DATE}{e.g. 'NA'. See Details.}

\item{ADMINISTRATIVE_PENALTY_CLASS}{e.g. 'NA'. See Details.}

\item{HEARING_REQUEST}{e.g. 'NA'. See Details.}

\item{FINAL_ORDER_DATE}{e.g. 'NA'. See Details.}

\item{PENALTY_AMOUNT_ASSESSED}{e.g. '0'. See Details.}

\item{CLASS_II_APPEAL_FILED}{e.g. 'NA'. See Details.}

\item{JUDICIAL_APPEAL_FILED_DATE}{e.g. 'NA'. See Details.}

\item{PENALTY_COLLECTED_DATE}{e.g. 'NA'. See Details.}
}
\description{
Retrieve pcs admin penalty order data from pcs-icis database
}

