% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcs-icis_pcs_cmpl_schd_viol.R
\name{pcs.icis_pcs_cmpl_schd_viol}
\alias{pcs.icis_pcs_cmpl_schd_viol}
\title{Retrieve pcs cmpl schd viol data from pcs-icis database}
\usage{
pcs.icis_pcs_cmpl_schd_viol(NPDES = NULL, CMPL_SCHD_NUM = NULL,
  CMPL_SCHD_DATA_SRC_CODE = NULL, CMPL_SCHD_EVT = NULL,
  CMPL_SCHD_VIOL_CODE = NULL, CMPL_SCHD_VIOL_DATE = NULL,
  CS_RNC_DETECT_CODE = NULL, CS_RNC_DETECT_DATE = NULL,
  CS_RNC_RESOL_CODE = NULL, CS_RNC_RESOL_DATE = NULL)
}
\arguments{
\item{NPDES}{e.g. 'AK0000272'. See Details.}

\item{CMPL_SCHD_NUM}{e.g. 'DB'. See Details.}

\item{CMPL_SCHD_DATA_SRC_CODE}{e.g. '0790'. See Details.}

\item{CMPL_SCHD_EVT}{e.g. '61099'. See Details.}

\item{CMPL_SCHD_VIOL_CODE}{e.g. 'C20'. See Details.}

\item{CMPL_SCHD_VIOL_DATE}{e.g. '01-JUN-92'. See Details.}

\item{CS_RNC_DETECT_CODE}{e.g. 'N'. See Details.}

\item{CS_RNC_DETECT_DATE}{e.g. '01-JUL-92'. See Details.}

\item{CS_RNC_RESOL_CODE}{e.g. '5'. See Details.}

\item{CS_RNC_RESOL_DATE}{e.g. '19-FEB-93'. See Details.}
}
\description{
Retrieve pcs cmpl schd viol data from pcs-icis database
}
