% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcs-icis_pcs_code_desc.R
\name{pcs.icis_pcs_code_desc}
\alias{pcs.icis_pcs_code_desc}
\title{Retrieve pcs code desc data from pcs-icis database}
\usage{
pcs.icis_pcs_code_desc(TABLE_ID = NULL, CODE = NULL, DESCRIPT = NULL)
}
\arguments{
\item{TABLE_ID}{e.g. '000'. See Details.}

\item{CODE}{e.g. '030'. See Details.}

\item{DESCRIPT}{e.g. 'ENFORCEMENT ACTION CODES                          002032SYENAC'. See Details.}
}
\description{
Retrieve pcs code desc data from pcs-icis database
}
