% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcs-icis_pcs_effl_lim.R
\name{pcs.icis_pcs_effl_lim}
\alias{pcs.icis_pcs_effl_lim}
\title{Retrieve pcs effl lim data from pcs-icis database}
\usage{
pcs.icis_pcs_effl_lim(NPDES = NULL, DISCHARGE_NUM = NULL,
  REPORT_DESIG = NULL, PIPE_SET_QUALIFIER = NULL, LIMIT_TYPE = NULL,
  PARAM_CODE = NULL, MONITORING_LOC = NULL, SEASON_NUM = NULL,
  MODIF_NUM = NULL, MODIF_PERIOD_START_DATE = NULL,
  MODIF_PERIOD_END_DATE = NULL, CHANGE_OF_LIMIT_STAT = NULL,
  CONTESTED_PARAM_IND = NULL, DOCKET_NUMBER = NULL)
}
\arguments{
\item{NPDES}{e.g. 'AK0000272'. See Details.}

\item{DISCHARGE_NUM}{e.g. '001'. See Details.}

\item{REPORT_DESIG}{e.g. 'A'. See Details.}

\item{PIPE_SET_QUALIFIER}{e.g. '9'. See Details.}

\item{LIMIT_TYPE}{e.g. '5'. See Details.}

\item{PARAM_CODE}{e.g. '00545'. See Details.}

\item{MONITORING_LOC}{e.g. '1'. See Details.}

\item{SEASON_NUM}{e.g. '0'. See Details.}

\item{MODIF_NUM}{e.g. '0'. See Details.}

\item{MODIF_PERIOD_START_DATE}{e.g. '01-APR-03'. See Details.}

\item{MODIF_PERIOD_END_DATE}{e.g. '31-MAR-08'. See Details.}

\item{CHANGE_OF_LIMIT_STAT}{e.g. 'NA'. See Details.}

\item{CONTESTED_PARAM_IND}{e.g. 'NA'. See Details.}

\item{DOCKET_NUMBER}{e.g. 'NA'. See Details.}
}
\description{
Retrieve pcs effl lim data from pcs-icis database
}

