% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcs-icis_pcs_effl_lim_concentr.R
\name{pcs.icis_pcs_effl_lim_concentr}
\alias{pcs.icis_pcs_effl_lim_concentr}
\title{Retrieve pcs effl lim concentr data from pcs-icis database}
\usage{
pcs.icis_pcs_effl_lim_concentr(NPDES = NULL, DISCHARGE_NUM = NULL,
  REPORT_DESIG = NULL, PIPE_SET_QUALIFIER = NULL, LIMIT_TYPE = NULL,
  PARAM_CODE = NULL, MONITORING_LOC = NULL, SEASON_NUM = NULL,
  MODIF_NUM = NULL, CONC_MAX_LIMIT = NULL, CONC_AVG_LIMIT = NULL,
  CONC_MIN_LIMIT = NULL, CONC_UNIT_CODE = NULL,
  STAT_LIM_AVG_OVERRIDE = NULL, STAT_LIM_MIN_OVERRIDE = NULL,
  CONC_MIN_LIM_STAT_BASE = NULL, CONC_AVG_LIM_STAT_BASE = NULL,
  CONC_MAX_LIM_STAT_BASE = NULL)
}
\arguments{
\item{NPDES}{e.g. 'AK0000272'. See Details.}

\item{DISCHARGE_NUM}{e.g. 'FAC'. See Details.}

\item{REPORT_DESIG}{e.g. 'S'. See Details.}

\item{PIPE_SET_QUALIFIER}{e.g. '2'. See Details.}

\item{LIMIT_TYPE}{e.g. '5'. See Details.}

\item{PARAM_CODE}{e.g. '00545'. See Details.}

\item{MONITORING_LOC}{e.g. '1'. See Details.}

\item{SEASON_NUM}{e.g. '0'. See Details.}

\item{MODIF_NUM}{e.g. '0'. See Details.}

\item{CONC_MAX_LIMIT}{e.g. 'ADDMON'. See Details.}

\item{CONC_AVG_LIMIT}{e.g. 'ADDMON'. See Details.}

\item{CONC_MIN_LIMIT}{e.g. 'DELMON'. See Details.}

\item{CONC_UNIT_CODE}{e.g. '25'. See Details.}

\item{STAT_LIM_AVG_OVERRIDE}{e.g. 'NA'. See Details.}

\item{STAT_LIM_MIN_OVERRIDE}{e.g. 'NA'. See Details.}

\item{CONC_MIN_LIM_STAT_BASE}{e.g. 'NA'. See Details.}

\item{CONC_AVG_LIM_STAT_BASE}{e.g. 'MK'. See Details.}

\item{CONC_MAX_LIM_STAT_BASE}{e.g. 'DD'. See Details.}
}
\description{
Retrieve pcs effl lim concentr data from pcs-icis database
}
