% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcs-icis_pcs_effl_lim_qty.R
\name{pcs.icis_pcs_effl_lim_qty}
\alias{pcs.icis_pcs_effl_lim_qty}
\title{Retrieve pcs effl lim qty data from pcs-icis database}
\usage{
pcs.icis_pcs_effl_lim_qty(NPDES = NULL, DISCHARGE_NUM = NULL,
  REPORT_DESIG = NULL, PIPE_SET_QUALIFIER = NULL, LIMIT_TYPE = NULL,
  PARAM_CODE = NULL, MONITORING_LOC = NULL, SEASON_NUM = NULL,
  MODIF_NUM = NULL, QTY_MAX_LIMIT = NULL, QTY_AVG_LIMIT = NULL,
  QTY_UNIT_CODE = NULL, QTY_AVG_LIM_STAT_BASE = NULL,
  QTY_MAX_LIM_STAT_BASE = NULL)
}
\arguments{
\item{NPDES}{e.g. 'AK0000272'. See Details.}

\item{DISCHARGE_NUM}{e.g. '001'. See Details.}

\item{REPORT_DESIG}{e.g. 'B'. See Details.}

\item{PIPE_SET_QUALIFIER}{e.g. '1'. See Details.}

\item{LIMIT_TYPE}{e.g. '5'. See Details.}

\item{PARAM_CODE}{e.g. '00145'. See Details.}

\item{MONITORING_LOC}{e.g. 'O'. See Details.}

\item{SEASON_NUM}{e.g. '0'. See Details.}

\item{MODIF_NUM}{e.g. '0'. See Details.}

\item{QTY_MAX_LIMIT}{e.g. 'OPTMON'. See Details.}

\item{QTY_AVG_LIMIT}{e.g. 'DELMON'. See Details.}

\item{QTY_UNIT_CODE}{e.g. '76'. See Details.}

\item{QTY_AVG_LIM_STAT_BASE}{e.g. 'NA'. See Details.}

\item{QTY_MAX_LIM_STAT_BASE}{e.g. 'MP'. See Details.}
}
\description{
Retrieve pcs effl lim qty data from pcs-icis database
}

