% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcs-icis_pcs_enfor_action.R
\name{pcs.icis_pcs_enfor_action}
\alias{pcs.icis_pcs_enfor_action}
\title{Retrieve pcs enfor action data from pcs-icis database}
\usage{
pcs.icis_pcs_enfor_action(NPDES = NULL, EA_CODE = NULL,
  EA_TYPE = NULL, EA_DATE = NULL, EA_STAT_CODE = NULL,
  EA_STAT_DATE = NULL, EA_RESP_DUE_DATE = NULL,
  EA_DOCKET_NUMBER = NULL, EA_DATA_SRC_CODE = NULL)
}
\arguments{
\item{NPDES}{e.g. 'AK0000272'. See Details.}

\item{EA_CODE}{e.g. '21'. See Details.}

\item{EA_TYPE}{e.g. 'E'. See Details.}

\item{EA_DATE}{e.g. '06-OCT-88'. See Details.}

\item{EA_STAT_CODE}{e.g. 'CL'. See Details.}

\item{EA_STAT_DATE}{e.g. '01-APR-89'. See Details.}

\item{EA_RESP_DUE_DATE}{e.g. 'NA'. See Details.}

\item{EA_DOCKET_NUMBER}{e.g. '10880605309'. See Details.}

\item{EA_DATA_SRC_CODE}{e.g. 'NA'. See Details.}
}
\description{
Retrieve pcs enfor action data from pcs-icis database
}
