% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcs-icis_pcs_evidentiary_hearing_event.R
\name{pcs.icis_pcs_evidentiary_hearing_event}
\alias{pcs.icis_pcs_evidentiary_hearing_event}
\title{Retrieve pcs evidentiary hearing event data from pcs-icis database}
\usage{
pcs.icis_pcs_evidentiary_hearing_event(NPDES = NULL, EVENT_CODE = NULL,
  EVENT_DATE = NULL)
}
\arguments{
\item{NPDES}{e.g. 'AK0000272'. See Details.}

\item{EVENT_CODE}{e.g. '07099'. See Details.}

\item{EVENT_DATE}{e.g. '05-NOV-80'. See Details.}
}
\description{
Retrieve pcs evidentiary hearing event data from pcs-icis database
}
