% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcs-icis_pcs_industrial_user_info.R
\name{pcs.icis_pcs_industrial_user_info}
\alias{pcs.icis_pcs_industrial_user_info}
\title{Retrieve pcs industrial user info data from pcs-icis database}
\usage{
pcs.icis_pcs_industrial_user_info(NPDES = NULL, INSP_DATE = NULL,
  INSP_TYPE = NULL, SIGNIFICANT_INDUSTRIAL_USERS = NULL,
  CATEGORICAL_INDUSTRIAL_USERS = NULL,
  SIUS_WITHOUT_CTL_MECHANISM = NULL,
  SIUS_IN_SNC_WITH_PRETR_STDS = NULL, SIUS_NOT_INSP_SAMPLED = NULL,
  SIUS_SNC_SELF_MONITOR = NULL, SIUS_SNC_SELF_NOT_INSP = NULL)
}
\arguments{
\item{NPDES}{e.g. 'AK0022551'. See Details.}

\item{INSP_DATE}{e.g. '24-MAY-88'. See Details.}

\item{INSP_TYPE}{e.g. 'F'. See Details.}

\item{SIGNIFICANT_INDUSTRIAL_USERS}{e.g. '11'. See Details.}

\item{CATEGORICAL_INDUSTRIAL_USERS}{e.g. 'NA'. See Details.}

\item{SIUS_WITHOUT_CTL_MECHANISM}{e.g. '0'. See Details.}

\item{SIUS_IN_SNC_WITH_PRETR_STDS}{e.g. '0'. See Details.}

\item{SIUS_NOT_INSP_SAMPLED}{e.g. '0'. See Details.}

\item{SIUS_SNC_SELF_MONITOR}{e.g. '0'. See Details.}

\item{SIUS_SNC_SELF_NOT_INSP}{e.g. '0'. See Details.}
}
\description{
Retrieve pcs industrial user info data from pcs-icis database
}
