% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcs-icis_pcs_inspection.R
\name{pcs.icis_pcs_inspection}
\alias{pcs.icis_pcs_inspection}
\title{Retrieve pcs inspection data from pcs-icis database}
\usage{
pcs.icis_pcs_inspection(NPDES = NULL, INSP_DATE = NULL, INSP_TYPE = NULL,
  INSP_CODE = NULL)
}
\arguments{
\item{NPDES}{e.g. 'AK0000272'. See Details.}

\item{INSP_DATE}{e.g. '25-MAY-89'. See Details.}

\item{INSP_TYPE}{e.g. 'C'. See Details.}

\item{INSP_CODE}{e.g. 'R'. See Details.}
}
\description{
Retrieve pcs inspection data from pcs-icis database
}

