% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcs-icis_pcs_pci_audit.R
\name{pcs.icis_pcs_pci_audit}
\alias{pcs.icis_pcs_pci_audit}
\title{Retrieve pcs pci audit data from pcs-icis database}
\usage{
pcs.icis_pcs_pci_audit(NPDES = NULL, INSP_DATE = NULL, INSP_TYPE = NULL,
  PERMIT_MOD_FOR_PRETR_DATE = NULL, ADOPT_LOCAL_LIMITS = NULL,
  TECH_EVAL_LOCAL_LIMIT = NULL, SLUDGE_DISPOSAL_METHOD = NULL)
}
\arguments{
\item{NPDES}{e.g. 'AK0022551'. See Details.}

\item{INSP_DATE}{e.g. '26-JUN-89'. See Details.}

\item{INSP_TYPE}{e.g. 'F'. See Details.}

\item{PERMIT_MOD_FOR_PRETR_DATE}{e.g. '16-SEP-85'. See Details.}

\item{ADOPT_LOCAL_LIMITS}{e.g. 'N'. See Details.}

\item{TECH_EVAL_LOCAL_LIMIT}{e.g. 'Y'. See Details.}

\item{SLUDGE_DISPOSAL_METHOD}{e.g. 'NA'. See Details.}
}
\description{
Retrieve pcs pci audit data from pcs-icis database
}

