% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcs-icis_pcs_permit_event.R
\name{pcs.icis_pcs_permit_event}
\alias{pcs.icis_pcs_permit_event}
\title{Retrieve pcs permit event data from pcs-icis database}
\usage{
pcs.icis_pcs_permit_event(NPDES = NULL, TRACKING_EVENT_CODE = NULL,
  EVENT_ACTUAL_DATE = NULL)
}
\arguments{
\item{NPDES}{e.g. 'AK0000272'. See Details.}

\item{TRACKING_EVENT_CODE}{e.g. 'P1099'. See Details.}

\item{EVENT_ACTUAL_DATE}{e.g. '27-NOV-00'. See Details.}
}
\description{
Retrieve pcs permit event data from pcs-icis database
}

