% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcs-icis_pcs_permit_facility.R
\name{pcs.icis_pcs_permit_facility}
\alias{pcs.icis_pcs_permit_facility}
\title{Retrieve pcs permit facility data from pcs-icis database}
\usage{
pcs.icis_pcs_permit_facility(NPDES = NULL, NAME_1 = NULL,
  NAME_2 = NULL, MAJOR_DISCHARGE_INDICATOR = NULL, REGION = NULL,
  CITY_CODE = NULL, CITY_NAME = NULL, COUNTY_CODE = NULL,
  COUNTY_NAME = NULL, SIC_CODE = NULL, LOCATION_NAME = NULL,
  LOCATION_STREET_1 = NULL, LOCATION_STREET_2 = NULL,
  LOCATION_CITY = NULL, LOCATION_STATE = NULL,
  LOCATION_ZIP_CODE = NULL, LOCATION_TELEPHONE_NUMBER = NULL,
  MAILING_NAME = NULL, MAILING_STREET_1 = NULL,
  MAILING_STREET_2 = NULL, MAILING_CITY = NULL, MAILING_STATE = NULL,
  MAILING_ZIP_CODE = NULL, INACTIVE_CODE = NULL,
  PERMIT_EXPIRED_DATE = NULL, PERMIT_ISSUED_DATE = NULL,
  DRAFT_PERM_PUB_NOTIF_DATE = NULL, APPL_RECEIVED_DATE = NULL,
  INACTIVE_DATE = NULL, TYPE_OF_OWNERSHIP = NULL,
  COGN_OFFICIAL = NULL, COGN_OFFICIAL_TELE_NUMBER = NULL,
  INDUSTRY_CLASS = NULL, ORIGINL_PERMIT_ISSUE_DATE = NULL,
  TYPE_OF_PERMIT_ISSUED = NULL, RIVER_BASIN = NULL,
  RECEIVING_WATERS = NULL, CURR_YEAR_MANL_QNCR_STAT = NULL,
  CURR_YEAR_AUTO_QNCR_STAT = NULL, PREV_YEAR_MANL_QNCR_STAT = NULL,
  PREV_YEAR_AUTO_QNCR_STAT = NULL, FEDERAL_GRANT_IND = NULL,
  FINAL_LIMITS_IND = NULL, FLOW_RATE = NULL,
  PRETREATMENT_CODE = NULL, NMP_FINAN_STAT_CODE = NULL,
  NMP_FINAL_SCHED = NULL, NMP_QUARTER = NULL, LATITUDE = NULL,
  LONGITUDE = NULL, CODE_OF_ACCURACY = NULL, METHOD = NULL,
  DATUM = NULL, SCALE = NULL, DESCRIPT = NULL,
  USGS_HYDRO_BASIN_CODE = NULL, STREAM_SEGMENT = NULL,
  MILEAGE_IND = NULL, RECVNG_STREAM_CLASS_CODE = NULL)
}
\arguments{
\item{NPDES}{e.g. 'AK0000272'. See Details.}

\item{NAME_1}{e.g. 'ALYESKA SEAFOODS INC'. See Details.}

\item{NAME_2}{e.g. 'S: UNALASKA (FISH MEAL/SURIMI)'. See Details.}

\item{MAJOR_DISCHARGE_INDICATOR}{e.g. 'M'. See Details.}

\item{REGION}{e.g. '10'. See Details.}

\item{CITY_CODE}{e.g. '92820'. See Details.}

\item{CITY_NAME}{e.g. 'UNALASKA'. See Details.}

\item{COUNTY_CODE}{e.g. '010'. See Details.}

\item{COUNTY_NAME}{e.g. 'ALEUTIAN ISLANDS DI'. See Details.}

\item{SIC_CODE}{e.g. '2091'. See Details.}

\item{LOCATION_NAME}{e.g. 'ALYESKA SEAFOODS INC'. See Details.}

\item{LOCATION_STREET_1}{e.g. 'S: UNALASKA (FISH MEAL/SURIMI)'. See Details.}

\item{LOCATION_STREET_2}{e.g. '551 WEST BROADWAY AVE'. See Details.}

\item{LOCATION_CITY}{e.g. 'UNALASKA'. See Details.}

\item{LOCATION_STATE}{e.g. 'AK'. See Details.}

\item{LOCATION_ZIP_CODE}{e.g. '99685'. See Details.}

\item{LOCATION_TELEPHONE_NUMBER}{e.g. '9075811211'. See Details.}

\item{MAILING_NAME}{e.g. 'ALYESKA SEAFOODS INC'. See Details.}

\item{MAILING_STREET_1}{e.g. 'P.O. BOX 31359'. See Details.}

\item{MAILING_STREET_2}{e.g. 'NA'. See Details.}

\item{MAILING_CITY}{e.g. 'SEATTLE'. See Details.}

\item{MAILING_STATE}{e.g. 'WA'. See Details.}

\item{MAILING_ZIP_CODE}{e.g. '98103'. See Details.}

\item{INACTIVE_CODE}{e.g. 'A'. See Details.}

\item{PERMIT_EXPIRED_DATE}{e.g. '31-MAR-08'. See Details.}

\item{PERMIT_ISSUED_DATE}{e.g. '05-FEB-03'. See Details.}

\item{DRAFT_PERM_PUB_NOTIF_DATE}{e.g. '21-SEP-01'. See Details.}

\item{APPL_RECEIVED_DATE}{e.g. '27-NOV-00'. See Details.}

\item{INACTIVE_DATE}{e.g. 'NA'. See Details.}

\item{TYPE_OF_OWNERSHIP}{e.g. 'PRI'. See Details.}

\item{COGN_OFFICIAL}{e.g. 'WINN BRINDLE, GENERAL MANAGER'. See Details.}

\item{COGN_OFFICIAL_TELE_NUMBER}{e.g. '2065472100'. See Details.}

\item{INDUSTRY_CLASS}{e.g. 'R'. See Details.}

\item{ORIGINL_PERMIT_ISSUE_DATE}{e.g. '14-APR-75'. See Details.}

\item{TYPE_OF_PERMIT_ISSUED}{e.g. 'E'. See Details.}

\item{RIVER_BASIN}{e.g. '160600'. See Details.}

\item{RECEIVING_WATERS}{e.g. 'UNALASKA BAY, ILIULIUK HARBOR'. See Details.}

\item{CURR_YEAR_MANL_QNCR_STAT}{e.g. 'NA'. See Details.}

\item{CURR_YEAR_AUTO_QNCR_STAT}{e.g. 'NA'. See Details.}

\item{PREV_YEAR_MANL_QNCR_STAT}{e.g. 'NA'. See Details.}

\item{PREV_YEAR_AUTO_QNCR_STAT}{e.g. 'NA'. See Details.}

\item{FEDERAL_GRANT_IND}{e.g. 'NA'. See Details.}

\item{FINAL_LIMITS_IND}{e.g. 'F'. See Details.}

\item{FLOW_RATE}{e.g. 'NA'. See Details.}

\item{PRETREATMENT_CODE}{e.g. 'NA'. See Details.}

\item{NMP_FINAN_STAT_CODE}{e.g. 'NA'. See Details.}

\item{NMP_FINAL_SCHED}{e.g. 'NA'. See Details.}

\item{NMP_QUARTER}{e.g. 'NA'. See Details.}

\item{LATITUDE}{e.g. '+5352440'. See Details.}

\item{LONGITUDE}{e.g. '-16632250'. See Details.}

\item{CODE_OF_ACCURACY}{e.g. 'NA'. See Details.}

\item{METHOD}{e.g. 'A'. See Details.}

\item{DATUM}{e.g. '1'. See Details.}

\item{SCALE}{e.g. 'U'. See Details.}

\item{DESCRIPT}{e.g. '01099'. See Details.}

\item{USGS_HYDRO_BASIN_CODE}{e.g. 'NA'. See Details.}

\item{STREAM_SEGMENT}{e.g. 'NA'. See Details.}

\item{MILEAGE_IND}{e.g. 'NA'. See Details.}

\item{RECVNG_STREAM_CLASS_CODE}{e.g. 'NA'. See Details.}
}
\description{
Retrieve pcs permit facility data from pcs-icis database
}
