% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcs-icis_pcs_pipe_sched.R
\name{pcs.icis_pcs_pipe_sched}
\alias{pcs.icis_pcs_pipe_sched}
\title{Retrieve pcs pipe sched data from pcs-icis database}
\usage{
pcs.icis_pcs_pipe_sched(NPDES = NULL, DISCHARGE_NUM = NULL,
  REPORT_DESIG = NULL, PIPE_SET_QUALIFIER = NULL,
  OUTFALL_TYPE_CODE = NULL, INIT_LIMITS_START_DATE = NULL,
  INIT_LIMITS_END_DATE = NULL, INTERIM_LIMITS_START_DATE = NULL,
  INTERIM_LIMITS_END_DATE = NULL, FINAL_LIMITS_START_DATE = NULL,
  FINAL_LIMITS_END_DATE = NULL, PIPE_DESC = NULL,
  INIT_RPT_DATE = NULL, UNITS_IN_RPT_PERIOD = NULL,
  REPORTING_UNITS = NULL, PIPE_INACTIVE_CODE = NULL,
  PIPE_INACTIVE_DATE = NULL, INIT_EPA_SUBMISSION_DATE = NULL,
  UNITS_IN_EPA_SUBM_PERIOD = NULL, EPA_SUBMISSION_UNITS = NULL,
  INIT_STATE_SUBMISSION_DATE = NULL, UNITS_IN_STATE_SUBM_PERIOD = NULL,
  STATE_SUBMISSION_UNITS = NULL)
}
\arguments{
\item{NPDES}{e.g. 'AK0000272'. See Details.}

\item{DISCHARGE_NUM}{e.g. 'SUM'. See Details.}

\item{REPORT_DESIG}{e.g. 'A'. See Details.}

\item{PIPE_SET_QUALIFIER}{e.g. '2'. See Details.}

\item{OUTFALL_TYPE_CODE}{e.g. 'NA'. See Details.}

\item{INIT_LIMITS_START_DATE}{e.g. 'NA'. See Details.}

\item{INIT_LIMITS_END_DATE}{e.g. 'NA'. See Details.}

\item{INTERIM_LIMITS_START_DATE}{e.g. 'NA'. See Details.}

\item{INTERIM_LIMITS_END_DATE}{e.g. 'NA'. See Details.}

\item{FINAL_LIMITS_START_DATE}{e.g. '31-MAY-96'. See Details.}

\item{FINAL_LIMITS_END_DATE}{e.g. '30-MAY-01'. See Details.}

\item{PIPE_DESC}{e.g. 'COMBINED BOD OF ALL STREAMS'. See Details.}

\item{INIT_RPT_DATE}{e.g. '01-JUN-96'. See Details.}

\item{UNITS_IN_RPT_PERIOD}{e.g. '1'. See Details.}

\item{REPORTING_UNITS}{e.g. 'M'. See Details.}

\item{PIPE_INACTIVE_CODE}{e.g. 'A'. See Details.}

\item{PIPE_INACTIVE_DATE}{e.g. 'NA'. See Details.}

\item{INIT_EPA_SUBMISSION_DATE}{e.g. '10-JUL-96'. See Details.}

\item{UNITS_IN_EPA_SUBM_PERIOD}{e.g. '1'. See Details.}

\item{EPA_SUBMISSION_UNITS}{e.g. 'M'. See Details.}

\item{INIT_STATE_SUBMISSION_DATE}{e.g. 'NA'. See Details.}

\item{UNITS_IN_STATE_SUBM_PERIOD}{e.g. '0'. See Details.}

\item{STATE_SUBMISSION_UNITS}{e.g. 'NA'. See Details.}
}
\description{
Retrieve pcs pipe sched data from pcs-icis database
}
