% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcs-icis_pcs_pretreatment_perf_summary.R
\name{pcs.icis_pcs_pretreatment_perf_summary}
\alias{pcs.icis_pcs_pretreatment_perf_summary}
\title{Retrieve pcs pretreatment perf summary data from pcs-icis database}
\usage{
pcs.icis_pcs_pretreatment_perf_summary(NPDES = NULL,
  REPORT_START_DATE = NULL, REPORT_END_DATE = NULL,
  SIUS_IN_SNC_COMPL_SCHED = NULL, CIVIL_OR_CRIM_JUDICIAL_ACTS = NULL,
  NOVS_AND_AOS_AGAINST_SIUS = NULL, IUS_THAT_PAID_PENALTIES = NULL,
  SIUS_WITH_PUBLISHED_VIOLS = NULL)
}
\arguments{
\item{NPDES}{e.g. 'AK0022551'. See Details.}

\item{REPORT_START_DATE}{e.g. '15-OCT-86'. See Details.}

\item{REPORT_END_DATE}{e.g. '15-OCT-87'. See Details.}

\item{SIUS_IN_SNC_COMPL_SCHED}{e.g. '0'. See Details.}

\item{CIVIL_OR_CRIM_JUDICIAL_ACTS}{e.g. '0'. See Details.}

\item{NOVS_AND_AOS_AGAINST_SIUS}{e.g. '2'. See Details.}

\item{IUS_THAT_PAID_PENALTIES}{e.g. '0'. See Details.}

\item{SIUS_WITH_PUBLISHED_VIOLS}{e.g. '0'. See Details.}
}
\description{
Retrieve pcs pretreatment perf summary data from pcs-icis database
}
