% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcs-icis_pcs_single_event_viol.R
\name{pcs.icis_pcs_single_event_viol}
\alias{pcs.icis_pcs_single_event_viol}
\title{Retrieve pcs single event viol data from pcs-icis database}
\usage{
pcs.icis_pcs_single_event_viol(NPDES = NULL, SINGLE_EVENT_VIOL_CODE = NULL,
  SINGLE_EVENT_VIOL_DATE = NULL, QNCR_RNC_DET_CODE = NULL,
  QNCR_RNC_DET_DATE = NULL, QNCR_RNC_RESOL_CODE = NULL,
  QNCR_RNC_RESOL_DATE = NULL)
}
\arguments{
\item{NPDES}{e.g. 'AL0024210'. See Details.}

\item{SINGLE_EVENT_VIOL_CODE}{e.g. '02099'. See Details.}

\item{SINGLE_EVENT_VIOL_DATE}{e.g. '13-OCT-94'. See Details.}

\item{QNCR_RNC_DET_CODE}{e.g. 'D'. See Details.}

\item{QNCR_RNC_DET_DATE}{e.g. '13-OCT-94'. See Details.}

\item{QNCR_RNC_RESOL_CODE}{e.g. 'A'. See Details.}

\item{QNCR_RNC_RESOL_DATE}{e.g. 'NA'. See Details.}
}
\description{
Retrieve pcs single event viol data from pcs-icis database
}

