% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcs-icis_pcs_sludge.R
\name{pcs.icis_pcs_sludge}
\alias{pcs.icis_pcs_sludge}
\title{Retrieve pcs sludge data from pcs-icis database}
\usage{
pcs.icis_pcs_sludge(NPDES = NULL, ANNUAL_DRY_SLUDGE_PROD = NULL,
  SLUDGE_FACILITY_IND = NULL, SLUDGE_COMMERCIAL_HANDLER = NULL,
  HANDLER_STREET_1 = NULL, HANDLER_STREET_2 = NULL, HANDLER_CITY = NULL,
  HANDLER_STATE = NULL, HANDLER_ZIP_CODE = NULL,
  SLUDGE_RELATED_PERMIT_NUM = NULL, SLUDGE_CLASS_FACILITY_IND = NULL)
}
\arguments{
\item{NPDES}{e.g. 'AK0021377'. See Details.}

\item{ANNUAL_DRY_SLUDGE_PROD}{e.g. '115'. See Details.}

\item{SLUDGE_FACILITY_IND}{e.g. 'B'. See Details.}

\item{SLUDGE_COMMERCIAL_HANDLER}{e.g. 'NA'. See Details.}

\item{HANDLER_STREET_1}{e.g. 'NA'. See Details.}

\item{HANDLER_STREET_2}{e.g. 'NA'. See Details.}

\item{HANDLER_CITY}{e.g. 'NA'. See Details.}

\item{HANDLER_STATE}{e.g. 'NA'. See Details.}

\item{HANDLER_ZIP_CODE}{e.g. 'NA'. See Details.}

\item{SLUDGE_RELATED_PERMIT_NUM}{e.g. 'AK0021377'. See Details.}

\item{SLUDGE_CLASS_FACILITY_IND}{e.g. '2'. See Details.}
}
\description{
Retrieve pcs sludge data from pcs-icis database
}

