% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radinfo_rad_facility.R
\name{radinfo_rad_facility}
\alias{radinfo_rad_facility}
\title{Retrieve rad facility data from radinfo database}
\usage{
radinfo_rad_facility(RAD_SYS_ID = NULL, FACILITY_REGISTRY_ID = NULL,
  EF_PGM_SOURCE = NULL, PRIMARY_NAME = NULL, LOCATION_ADDRESS = NULL,
  SUPPLEMENTAL_LOCATION = NULL, CITY_NAME = NULL, COUNTY_NAME = NULL,
  STATE_CODE = NULL, POSTAL_CODE = NULL, STATE_NAME = NULL,
  COUNTRY_NAME = NULL, EPA_REGION_CODE = NULL,
  CONGRESSIONAL_DIST_NUM = NULL, HUC_CODE = NULL,
  FED_FACILITY_CODE = NULL, TRIBAL_LAND_CODE = NULL,
  TRIBAL_LAND_NAME = NULL, SOURCE_DATA = NULL, URL = NULL,
  URL_LINK_DESCRIPTION = NULL, FRS_UPDATE_DATE = NULL,
  RAD_CHANGE_DATE = NULL)
}
\arguments{
\item{RAD_SYS_ID}{e.g. 'RAD200000001'. See Details.}

\item{FACILITY_REGISTRY_ID}{e.g. '110000609066'. See Details.}

\item{EF_PGM_SOURCE}{e.g. 'NA'. See Details.}

\item{PRIMARY_NAME}{e.g. 'U S A F WILLIAMS AIR FORCE BASE'. See Details.}

\item{LOCATION_ADDRESS}{e.g. '6001 S POWER BLDG 1'. See Details.}

\item{SUPPLEMENTAL_LOCATION}{e.g. 'NA'. See Details.}

\item{CITY_NAME}{e.g. 'MESA'. See Details.}

\item{COUNTY_NAME}{e.g. 'MARICOPA'. See Details.}

\item{STATE_CODE}{e.g. 'AZ'. See Details.}

\item{POSTAL_CODE}{e.g. '852060901'. See Details.}

\item{STATE_NAME}{e.g. 'ARIZONA'. See Details.}

\item{COUNTRY_NAME}{e.g. 'USA'. See Details.}

\item{EPA_REGION_CODE}{e.g. '09'. See Details.}

\item{CONGRESSIONAL_DIST_NUM}{e.g. '06'. See Details.}

\item{HUC_CODE}{e.g. '15050100'. See Details.}

\item{FED_FACILITY_CODE}{e.g. 'Y'. See Details.}

\item{TRIBAL_LAND_CODE}{e.g. 'NA'. See Details.}

\item{TRIBAL_LAND_NAME}{e.g. 'NA'. See Details.}

\item{SOURCE_DATA}{e.g. 'NA'. See Details.}

\item{URL}{e.g. 'NA'. See Details.}

\item{URL_LINK_DESCRIPTION}{e.g. 'NA'. See Details.}

\item{FRS_UPDATE_DATE}{e.g. 'NA'. See Details.}

\item{RAD_CHANGE_DATE}{e.g. 'NA'. See Details.}
}
\description{
Retrieve rad facility data from radinfo database
}

