% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radinfo_rad_facility_type.R
\name{radinfo_rad_facility_type}
\alias{radinfo_rad_facility_type}
\title{Retrieve rad facility type data from radinfo database}
\usage{
radinfo_rad_facility_type(CIT_REF_CODE = NULL, SUBPART_ID = NULL,
  SEC_CIT_REF_FLAG = NULL, FACILITY_TYPE = NULL)
}
\arguments{
\item{CIT_REF_CODE}{e.g. '40CFR191'. See Details.}

\item{SUBPART_ID}{e.g. 'NA'. See Details.}

\item{SEC_CIT_REF_FLAG}{e.g. 'Y'. See Details.}

\item{FACILITY_TYPE}{e.g. 'WIPP Facility/Respository'. See Details.}
}
\description{
Retrieve rad facility type data from radinfo database
}

