% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radinfo_rad_geo_location.R
\name{radinfo_rad_geo_location}
\alias{radinfo_rad_geo_location}
\title{Retrieve rad geo location data from radinfo database}
\usage{
radinfo_rad_geo_location(RAD_SYS_ID = NULL, LATITUDE_MEASURE = NULL,
  LONGITUDE_MEASURE = NULL, HORIZONTAL_ACCURACY_MEASURE = NULL,
  SUB_ID = NULL, VERTICAL_MEASURE = NULL,
  VERTICAL_COLLECT_METHOD_CODE = NULL, VERTICAL_REFERENCE_DATA_CODE = NULL,
  VERTICAL_ACCURACY_MEASURE = NULL, HORIZONTAL_COLLECT_METHOD_CODE = NULL,
  SOURCE_MAP_SCALE_NUMBER = NULL, COORDINATE_DATA_SOURCE_CODE = NULL,
  HORIZONTAL_REFER_DATUM_CODE = NULL, REFERENCE_POINT_CODE = NULL,
  DATA_COLLECTION_DATE = NULL, LOCATION_COMMENTS_TEXT = NULL,
  GEOMETRIC_TYPE_CODE = NULL, EF_PGM_SOURCE = NULL, RAD_OVERRIDE = NULL,
  RAD_CHANGE_DATE = NULL, SUB_TYPE_CODE = NULL)
}
\arguments{
\item{RAD_SYS_ID}{e.g. 'RAD200000059'. See Details.}

\item{LATITUDE_MEASURE}{e.g. '32.36'. See Details.}

\item{LONGITUDE_MEASURE}{e.g. '-103.79'. See Details.}

\item{HORIZONTAL_ACCURACY_MEASURE}{e.g. '100'. See Details.}

\item{SUB_ID}{e.g. 'NA'. See Details.}

\item{VERTICAL_MEASURE}{e.g. 'NA'. See Details.}

\item{VERTICAL_COLLECT_METHOD_CODE}{e.g. 'NA'. See Details.}

\item{VERTICAL_REFERENCE_DATA_CODE}{e.g. 'NA'. See Details.}

\item{VERTICAL_ACCURACY_MEASURE}{e.g. 'NA'. See Details.}

\item{HORIZONTAL_COLLECT_METHOD_CODE}{e.g. '027'. See Details.}

\item{SOURCE_MAP_SCALE_NUMBER}{e.g. 'NA'. See Details.}

\item{COORDINATE_DATA_SOURCE_CODE}{e.g. '082'. See Details.}

\item{HORIZONTAL_REFER_DATUM_CODE}{e.g. '001'. See Details.}

\item{REFERENCE_POINT_CODE}{e.g. '025'. See Details.}

\item{DATA_COLLECTION_DATE}{e.g. '14-SEP-01'. See Details.}

\item{LOCATION_COMMENTS_TEXT}{e.g. 'NA'. See Details.}

\item{GEOMETRIC_TYPE_CODE}{e.g. '001'. See Details.}

\item{EF_PGM_SOURCE}{e.g. 'NA'. See Details.}

\item{RAD_OVERRIDE}{e.g. 'NA'. See Details.}

\item{RAD_CHANGE_DATE}{e.g. 'NA'. See Details.}

\item{SUB_TYPE_CODE}{e.g. 'NA'. See Details.}
}
\description{
Retrieve rad geo location data from radinfo database
}

