% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radinfo_rad_regulation.R
\name{radinfo_rad_regulation}
\alias{radinfo_rad_regulation}
\title{Retrieve rad regulation data from radinfo database}
\usage{
radinfo_rad_regulation(CIT_REF_CODE = NULL, STATUTE = NULL,
  STAT_ACRONYM = NULL, TITLE_ID = NULL, REG_TITLE = NULL,
  PART_ID = NULL, CFR_PART = NULL, SECTION_ID = NULL,
  CRF_SECTION = NULL, SUBPART_ID = NULL, CFR_SUBPART = NULL, URL = NULL,
  RAD_CHANGE_DATE = NULL)
}
\arguments{
\item{CIT_REF_CODE}{e.g. '40CFR191'. See Details.}

\item{STATUTE}{e.g. 'Atomic Energy Act'. See Details.}

\item{STAT_ACRONYM}{e.g. 'AEA'. See Details.}

\item{TITLE_ID}{e.g. '40'. See Details.}

\item{REG_TITLE}{e.g. 'Protection of Environment'. See Details.}

\item{PART_ID}{e.g. '191'. See Details.}

\item{CFR_PART}{e.g. 'Environmental Radiation Protection Standards for Management and Disposal of Spent Nuclear Fuel, High-Level and Transuranic Radioactive Waste'. See Details.}

\item{SECTION_ID}{e.g. 'NA'. See Details.}

\item{CRF_SECTION}{e.g. 'NA'. See Details.}

\item{SUBPART_ID}{e.g. 'NA'. See Details.}

\item{CFR_SUBPART}{e.g. 'NA'. See Details.}

\item{URL}{e.g. 'NA'. See Details.}

\item{RAD_CHANGE_DATE}{e.g. 'NA'. See Details.}
}
\description{
Retrieve rad regulation data from radinfo database
}

