% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radinfo_rad_regulatory_prog.R
\name{radinfo_rad_regulatory_prog}
\alias{radinfo_rad_regulatory_prog}
\title{Retrieve rad regulatory prog data from radinfo database}
\usage{
radinfo_rad_regulatory_prog(RAD_SYS_ID = NULL, CIT_REF_CODE = NULL,
  SEC_CIT_REF_FLAG = NULL, SUBPART_ID = NULL, PROG_FAC_TYPE = NULL,
  PROG_FAC_STATUS = NULL, OVERSIGHT_AGENCY = NULL,
  ENFORCEMENT_AGENCY = NULL, OPERATING_ORGANIZATION = NULL,
  RAD_CHANGE_DATE = NULL)
}
\arguments{
\item{RAD_SYS_ID}{e.g. 'RAD200000001'. See Details.}

\item{CIT_REF_CODE}{e.g. '40CFR300'. See Details.}

\item{SEC_CIT_REF_FLAG}{e.g. 'N'. See Details.}

\item{SUBPART_ID}{e.g. 'NA'. See Details.}

\item{PROG_FAC_TYPE}{e.g. 'RAD NPL Facility'. See Details.}

\item{PROG_FAC_STATUS}{e.g. 'NA'. See Details.}

\item{OVERSIGHT_AGENCY}{e.g. 'NA'. See Details.}

\item{ENFORCEMENT_AGENCY}{e.g. 'NA'. See Details.}

\item{OPERATING_ORGANIZATION}{e.g. 'NA'. See Details.}

\item{RAD_CHANGE_DATE}{e.g. 'NA'. See Details.}
}
\description{
Retrieve rad regulatory prog data from radinfo database
}

