% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radnet_bgm_onsite_treatment.R
\name{radnet_bgm_onsite_treatment}
\alias{radnet_bgm_onsite_treatment}
\title{Retrieve bgm onsite treatment data from radnet database}
\usage{
radnet_bgm_onsite_treatment(HANDLER_ID = NULL, ACTIVITY_LOCATION = NULL,
  SOURCE_TYPE = NULL, SEQ_NUMBER = NULL, HZ_PG = NULL,
  SYS_PG_NUM_SEQ = NULL, MANAGEMENT_METHOD_OWNER = NULL,
  MANAGEMENT_METHOD = NULL, SYS_TDR_QTY = NULL, MANAGED_TONS_CALC = NULL,
  USERID = NULL, LAST_CHANGE = NULL)
}
\arguments{
\item{HANDLER_ID}{e.g. 'AK1570028646'. See Details.}

\item{ACTIVITY_LOCATION}{e.g. 'AK'. See Details.}

\item{SOURCE_TYPE}{e.g. 'R'. See Details.}

\item{SEQ_NUMBER}{e.g. '1'. See Details.}

\item{HZ_PG}{e.g. '53'. See Details.}

\item{SYS_PG_NUM_SEQ}{e.g. '1'. See Details.}

\item{MANAGEMENT_METHOD_OWNER}{e.g. 'HQ'. See Details.}

\item{MANAGEMENT_METHOD}{e.g. 'H129'. See Details.}

\item{SYS_TDR_QTY}{e.g. '6'. See Details.}

\item{MANAGED_TONS_CALC}{e.g. '0.003'. See Details.}

\item{USERID}{e.g. 'IYWCVT'. See Details.}

\item{LAST_CHANGE}{e.g. '09-NOV-02'. See Details.}
}
\description{
Retrieve bgm onsite treatment data from radnet database
}

