% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radnet_erm_ana_proc.R
\name{radnet_erm_ana_proc}
\alias{radnet_erm_ana_proc}
\title{Retrieve erm ana proc data from radnet database}
\usage{
radnet_erm_ana_proc(ANA_PROC_NUM = NULL, ANA_PROC_ID = NULL,
  PROC_TYPE_ID = NULL, ANA_PROC_NAME = NULL)
}
\arguments{
\item{ANA_PROC_NUM}{e.g. '1'. See Details.}

\item{ANA_PROC_ID}{e.g. 'AIRBETA'. See Details.}

\item{PROC_TYPE_ID}{e.g. 'AIRBETA'. See Details.}

\item{ANA_PROC_NAME}{e.g. 'Gross beta in air'. See Details.}
}
\description{
Retrieve erm ana proc data from radnet database
}

