% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radnet_erm_analysis.R
\name{radnet_erm_analysis}
\alias{radnet_erm_analysis}
\title{Retrieve erm analysis data from radnet database}
\usage{
radnet_erm_analysis(ANA_NUM = NULL, SAMP_NUM = NULL,
  ANA_PROC_NUM = NULL, ANA_SIZE = NULL, ANA_UNIT = NULL,
  ANA_SIZE_2 = NULL, ANA_UNIT_2 = NULL, ANA_COMMENT = NULL)
}
\arguments{
\item{ANA_NUM}{e.g. '642'. See Details.}

\item{SAMP_NUM}{e.g. '230'. See Details.}

\item{ANA_PROC_NUM}{e.g. '9'. See Details.}

\item{ANA_SIZE}{e.g. '3.5'. See Details.}

\item{ANA_UNIT}{e.g. 'L'. See Details.}

\item{ANA_SIZE_2}{e.g. 'NA'. See Details.}

\item{ANA_UNIT_2}{e.g. 'NA'. See Details.}

\item{ANA_COMMENT}{e.g. 'NA'. See Details.}
}
\description{
Retrieve erm analysis data from radnet database
}
