% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radnet_erm_analyte.R
\name{radnet_erm_analyte}
\alias{radnet_erm_analyte}
\title{Retrieve erm analyte data from radnet database}
\usage{
radnet_erm_analyte(ANALYTE_ID = NULL, ANA_TYPE = NULL,
  HALF_LIFE = NULL, HALF_LIFE_TIME_UNIT = NULL, ANALYTE_NAME = NULL,
  CRS_ID = NULL)
}
\arguments{
\item{ANALYTE_ID}{e.g. 'AC228'. See Details.}

\item{ANA_TYPE}{e.g. 'R'. See Details.}

\item{HALF_LIFE}{e.g. '6.13'. See Details.}

\item{HALF_LIFE_TIME_UNIT}{e.g. 'H'. See Details.}

\item{ANALYTE_NAME}{e.g. 'Actinium-228'. See Details.}

\item{CRS_ID}{e.g. 'NA'. See Details.}
}
\description{
Retrieve erm analyte data from radnet database
}
