% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radnet_erm_count.R
\name{radnet_erm_count}
\alias{radnet_erm_count}
\title{Retrieve erm count data from radnet database}
\usage{
radnet_erm_count(ANA_NUM = NULL, RUN_NUM = NULL, DET_NUM = NULL,
  RUN_START = NULL, DURATION = NULL)
}
\arguments{
\item{ANA_NUM}{e.g. '642'. See Details.}

\item{RUN_NUM}{e.g. '1'. See Details.}

\item{DET_NUM}{e.g. '11'. See Details.}

\item{RUN_START}{e.g. '18-OCT-78'. See Details.}

\item{DURATION}{e.g. '50'. See Details.}
}
\description{
Retrieve erm count data from radnet database
}

