% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radnet_erm_det_type.R
\name{radnet_erm_det_type}
\alias{radnet_erm_det_type}
\title{Retrieve erm det type data from radnet database}
\usage{
radnet_erm_det_type(DET_NUM = NULL, DET_ID = NULL,
  DET_TYPE_ID = NULL)
}
\arguments{
\item{DET_NUM}{e.g. '1'. See Details.}

\item{DET_ID}{e.g. 'UNK'. See Details.}

\item{DET_TYPE_ID}{e.g. 'UNK'. See Details.}
}
\description{
Retrieve erm det type data from radnet database
}
