% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radnet_erm_location.R
\name{radnet_erm_location}
\alias{radnet_erm_location}
\title{Retrieve erm location data from radnet database}
\usage{
radnet_erm_location(LOC_NUM = NULL, STATE_ABBR = NULL,
  CITY_NAME = NULL, SURFACE_WATER_SOURCE = NULL, STATION = NULL)
}
\arguments{
\item{LOC_NUM}{e.g. '1'. See Details.}

\item{STATE_ABBR}{e.g. 'AL'. See Details.}

\item{CITY_NAME}{e.g. 'MONTGOMERY'. See Details.}

\item{SURFACE_WATER_SOURCE}{e.g. 'NA'. See Details.}

\item{STATION}{e.g. 'MONTGOMERY'. See Details.}
}
\description{
Retrieve erm location data from radnet database
}
