% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radnet_erm_matrix.R
\name{radnet_erm_matrix}
\alias{radnet_erm_matrix}
\title{Retrieve erm matrix data from radnet database}
\usage{
radnet_erm_matrix(MAT_NUM = NULL, MAT_ID = NULL, MAT_DESC = NULL)
}
\arguments{
\item{MAT_NUM}{e.g. '7'. See Details.}

\item{MAT_ID}{e.g. 'PASTEURIZED MILK'. See Details.}

\item{MAT_DESC}{e.g. 'Milk'. See Details.}
}
\description{
Retrieve erm matrix data from radnet database
}
