% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radnet_erm_project.R
\name{radnet_erm_project}
\alias{radnet_erm_project}
\title{Retrieve erm project data from radnet database}
\usage{
radnet_erm_project(PROJ_NUM = NULL, STUDY_NUM = NULL, PROJ_ID = NULL,
  PROJ_NAME = NULL)
}
\arguments{
\item{PROJ_NUM}{e.g. '40'. See Details.}

\item{STUDY_NUM}{e.g. '1'. See Details.}

\item{PROJ_ID}{e.g. 'PMN 1978-4'. See Details.}

\item{PROJ_NAME}{e.g. 'PMN 1978-4'. See Details.}
}
\description{
Retrieve erm project data from radnet database
}

