% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radnet_erm_result.R
\name{radnet_erm_result}
\alias{radnet_erm_result}
\title{Retrieve erm result data from radnet database}
\usage{
radnet_erm_result(ANA_NUM = NULL, RESULT_ID = NULL,
  ANALYTE_ID = NULL, RESULT_AMOUNT = NULL, CSU = NULL, MDC = NULL,
  RESULT_UNIT = NULL, RESULT_DATE = NULL, RESULT_IN_SI = NULL,
  CSU_IN_SI = NULL, MDC_IN_SI = NULL, SI_UNIT = NULL)
}
\arguments{
\item{ANA_NUM}{e.g. '642'. See Details.}

\item{RESULT_ID}{e.g. '1'. See Details.}

\item{ANALYTE_ID}{e.g. 'BA140'. See Details.}

\item{RESULT_AMOUNT}{e.g. '-6'. See Details.}

\item{CSU}{e.g. '4'. See Details.}

\item{MDC}{e.g. 'NA'. See Details.}

\item{RESULT_UNIT}{e.g. 'PCI/L'. See Details.}

\item{RESULT_DATE}{e.g. '13-OCT-78'. See Details.}

\item{RESULT_IN_SI}{e.g. '-0.22'. See Details.}

\item{CSU_IN_SI}{e.g. '0.15'. See Details.}

\item{MDC_IN_SI}{e.g. 'NA'. See Details.}

\item{SI_UNIT}{e.g. 'BQ/L'. See Details.}
}
\description{
Retrieve erm result data from radnet database
}
