% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radnet_erm_sample.R
\name{radnet_erm_sample}
\alias{radnet_erm_sample}
\title{Retrieve erm sample data from radnet database}
\usage{
radnet_erm_sample(SAMP_NUM = NULL, PROJ_NUM = NULL, SAMP_ID = NULL,
  CLIENT_ID = NULL, LOC_NUM = NULL, COLLECT_START = NULL,
  COLLECT_END = NULL, SAMP_SIZE = NULL, SAMP_UNIT = NULL,
  MAT_ID = NULL, SAMP_DESC = NULL, SAMP_COMMENT = NULL)
}
\arguments{
\item{SAMP_NUM}{e.g. '230'. See Details.}

\item{PROJ_NUM}{e.g. '40'. See Details.}

\item{SAMP_ID}{e.g. '7800230'. See Details.}

\item{CLIENT_ID}{e.g. 'NA'. See Details.}

\item{LOC_NUM}{e.g. '23'. See Details.}

\item{COLLECT_START}{e.g. '13-OCT-78'. See Details.}

\item{COLLECT_END}{e.g. '13-OCT-78'. See Details.}

\item{SAMP_SIZE}{e.g. 'NA'. See Details.}

\item{SAMP_UNIT}{e.g. 'NA'. See Details.}

\item{MAT_ID}{e.g. 'PASTEURIZED MILK'. See Details.}

\item{SAMP_DESC}{e.g. 'NA'. See Details.}

\item{SAMP_COMMENT}{e.g. 'SHEET-#1067'. See Details.}
}
\description{
Retrieve erm sample data from radnet database
}
