% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radnet_erm_study.R
\name{radnet_erm_study}
\alias{radnet_erm_study}
\title{Retrieve erm study data from radnet database}
\usage{
radnet_erm_study(STUDY_NUM = NULL, STUDY_ID = NULL, STUDY_NAME = NULL)
}
\arguments{
\item{STUDY_NUM}{e.g. '1'. See Details.}

\item{STUDY_ID}{e.g. 'PMN'. See Details.}

\item{STUDY_NAME}{e.g. 'RadNet Pasteurized Milk Network'. See Details.}
}
\description{
Retrieve erm study data from radnet database
}

