% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radnet_lcr_sample.R
\name{radnet_lcr_sample}
\alias{radnet_lcr_sample}
\title{Retrieve lcr sample data from radnet database}
\usage{
radnet_lcr_sample(PWSID = NULL, SAMPLE_ID = NULL,
  SAMPLING_END_DATE = NULL, SAMPLING_START_DATE = NULL,
  RECONCILIATION_ID = NULL, PRIMACY_AGENCY_CODE = NULL, EPA_REGION = NULL)
}
\arguments{
\item{PWSID}{e.g. '010106001'. See Details.}

\item{SAMPLE_ID}{e.g. '0101060011712/31/104'. See Details.}

\item{SAMPLING_END_DATE}{e.g. '31-DEC-10'. See Details.}

\item{SAMPLING_START_DATE}{e.g. '01-JAN-10'. See Details.}

\item{RECONCILIATION_ID}{e.g. 'NA'. See Details.}

\item{PRIMACY_AGENCY_CODE}{e.g. '01'. See Details.}

\item{EPA_REGION}{e.g. '01'. See Details.}
}
\description{
Retrieve lcr sample data from radnet database
}

