% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radnet_lcr_sample_result.R
\name{radnet_lcr_sample_result}
\alias{radnet_lcr_sample_result}
\title{Retrieve lcr sample result data from radnet database}
\usage{
radnet_lcr_sample_result(PWSID = NULL, SAMPLE_ID = NULL,
  PRIMACY_AGENCY_CODE = NULL, EPA_REGION = NULL, SAR_ID = NULL,
  CONTAMINANT_CODE = NULL, RESULT_SIGN_CODE = NULL,
  SAMPLE_MEASURE = NULL, UNIT_OF_MEASURE = NULL)
}
\arguments{
\item{PWSID}{e.g. '010106001'. See Details.}

\item{SAMPLE_ID}{e.g. '0101060011712/31/104'. See Details.}

\item{PRIMACY_AGENCY_CODE}{e.g. '01'. See Details.}

\item{EPA_REGION}{e.g. '01'. See Details.}

\item{SAR_ID}{e.g. '16494764'. See Details.}

\item{CONTAMINANT_CODE}{e.g. 'PB90'. See Details.}

\item{RESULT_SIGN_CODE}{e.g. 'E'. See Details.}

\item{SAMPLE_MEASURE}{e.g. '0.001'. See Details.}

\item{UNIT_OF_MEASURE}{e.g. 'mg/L'. See Details.}
}
\description{
Retrieve lcr sample result data from radnet database
}
