% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcrainfo_rcr_area_event.R
\name{rcrainfo_rcr_area_event}
\alias{rcrainfo_rcr_area_event}
\title{Retrieve rcr area event data from rcrainfo database}
\usage{
rcrainfo_rcr_area_event(AREA_HANDLER_ID = NULL, AREA_SEQ = NULL,
  EVENT_HANDLER_ID = NULL, EVENT_OWN = NULL, CA_EVENT_CODE = NULL,
  EVENT_SEQ = NULL, RESPONSIBLE_AGENCY = NULL, ACTIVITY_LOCATION = NULL)
}
\arguments{
\item{AREA_HANDLER_ID}{e.g. 'AK0000374959'. See Details.}

\item{AREA_SEQ}{e.g. '1'. See Details.}

\item{EVENT_HANDLER_ID}{e.g. 'AK0000374959'. See Details.}

\item{EVENT_OWN}{e.g. 'HQ'. See Details.}

\item{CA_EVENT_CODE}{e.g. 'CA750IN'. See Details.}

\item{EVENT_SEQ}{e.g. '1'. See Details.}

\item{RESPONSIBLE_AGENCY}{e.g. 'E'. See Details.}

\item{ACTIVITY_LOCATION}{e.g. 'AK'. See Details.}
}
\description{
Retrieve rcr area event data from rcrainfo database
}

