% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcrainfo_rcr_authcodes.R
\name{rcrainfo_rcr_authcodes}
\alias{rcrainfo_rcr_authcodes}
\title{Retrieve rcr authcodes data from rcrainfo database}
\usage{
rcrainfo_rcr_authcodes(OWNER = NULL, AUTHORITY_TYPE = NULL,
  ACTIVE_STATUS = NULL, AUTHORITY_DESC = NULL)
}
\arguments{
\item{OWNER}{e.g. 'HQ'. See Details.}

\item{AUTHORITY_TYPE}{e.g. 'A'. See Details.}

\item{ACTIVE_STATUS}{e.g. 'Y'. See Details.}

\item{AUTHORITY_DESC}{e.g. 'Consent Order'. See Details.}
}
\description{
Retrieve rcr authcodes data from rcrainfo database
}
