% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcrainfo_rcr_ca_area.R
\name{rcrainfo_rcr_ca_area}
\alias{rcrainfo_rcr_ca_area}
\title{Retrieve rcr ca area data from rcrainfo database}
\usage{
rcrainfo_rcr_ca_area(HANDLER_ID = NULL, AREA_SEQ = NULL,
  AREA_FACILITYWIDE_IND = NULL, REGULATED_UNIT_IND = NULL,
  AREA_NAME = NULL, AIR_RELEASE_IND = NULL,
  GROUNDWATER_RELEASE_IND = NULL, SOIL_RELEASE_IND = NULL,
  SURFACE_WATER_RELEASE_IND = NULL, EPA_OWNER = NULL,
  EPA_PERSON_ID = NULL, STATE_OWNER = NULL, STATE_PERSON_ID = NULL)
}
\arguments{
\item{HANDLER_ID}{e.g. 'AK0000374959'. See Details.}

\item{AREA_SEQ}{e.g. '1'. See Details.}

\item{AREA_FACILITYWIDE_IND}{e.g. 'Y'. See Details.}

\item{REGULATED_UNIT_IND}{e.g. 'NA'. See Details.}

\item{AREA_NAME}{e.g. 'ENTIRE FACILITY'. See Details.}

\item{AIR_RELEASE_IND}{e.g. 'NA'. See Details.}

\item{GROUNDWATER_RELEASE_IND}{e.g. 'NA'. See Details.}

\item{SOIL_RELEASE_IND}{e.g. 'NA'. See Details.}

\item{SURFACE_WATER_RELEASE_IND}{e.g. 'NA'. See Details.}

\item{EPA_OWNER}{e.g. 'NA'. See Details.}

\item{EPA_PERSON_ID}{e.g. 'NA'. See Details.}

\item{STATE_OWNER}{e.g. 'NA'. See Details.}

\item{STATE_PERSON_ID}{e.g. 'NA'. See Details.}
}
\description{
Retrieve rcr ca area data from rcrainfo database
}

