% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcrainfo_rcr_ca_authority.R
\name{rcrainfo_rcr_ca_authority}
\alias{rcrainfo_rcr_ca_authority}
\title{Retrieve rcr ca authority data from rcrainfo database}
\usage{
rcrainfo_rcr_ca_authority(HANDLER_ID = NULL, EFFECTIVE_DATE = NULL,
  RESPONSIBLE_AGENCY = NULL, ISSUE_DATE = NULL, REVOKE_DATE = NULL,
  REPOSITORY = NULL, ACTIVITY_LOCATION = NULL, OWNER = NULL,
  AUTHORITY_TYPE = NULL, PERSON_OWNER = NULL, PERSON_ID = NULL,
  LEAD_PROGRAM = NULL, SUB_ORGANIZATION_OWNER = NULL,
  SUB_ORGANIZATION = NULL)
}
\arguments{
\item{HANDLER_ID}{e.g. 'AK0000374959'. See Details.}

\item{EFFECTIVE_DATE}{e.g. '19-OCT-00'. See Details.}

\item{RESPONSIBLE_AGENCY}{e.g. 'E'. See Details.}

\item{ISSUE_DATE}{e.g. 'NA'. See Details.}

\item{REVOKE_DATE}{e.g. 'NA'. See Details.}

\item{REPOSITORY}{e.g. 'NA'. See Details.}

\item{ACTIVITY_LOCATION}{e.g. 'AK'. See Details.}

\item{OWNER}{e.g. 'HQ'. See Details.}

\item{AUTHORITY_TYPE}{e.g. 'X'. See Details.}

\item{PERSON_OWNER}{e.g. 'NA'. See Details.}

\item{PERSON_ID}{e.g. 'NA'. See Details.}

\item{LEAD_PROGRAM}{e.g. 'NA'. See Details.}

\item{SUB_ORGANIZATION_OWNER}{e.g. 'NA'. See Details.}

\item{SUB_ORGANIZATION}{e.g. 'NA'. See Details.}
}
\description{
Retrieve rcr ca authority data from rcrainfo database
}
