% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcrainfo_rcr_ca_event.R
\name{rcrainfo_rcr_ca_event}
\alias{rcrainfo_rcr_ca_event}
\title{Retrieve rcr ca event data from rcrainfo database}
\usage{
rcrainfo_rcr_ca_event(HANDLER_ID = NULL, EVENT_SEQ = NULL,
  RESPONSIBLE_AGENCY = NULL, ACTIVITY_LOCATION = NULL, ACTUAL_DATE = NULL,
  OWNER = NULL, CA_EVENT_CODE = NULL, SCHEDULE_DATE_ORIG = NULL,
  SCHEDULE_DATE_NEW = NULL, BEST_DATE = NULL, PERSON_OWNER = NULL,
  PERSON_ID = NULL, SUB_ORGANIZATION_OWNER = NULL,
  SUB_ORGANIZATION = NULL)
}
\arguments{
\item{HANDLER_ID}{e.g. 'AK0000374959'. See Details.}

\item{EVENT_SEQ}{e.g. '1'. See Details.}

\item{RESPONSIBLE_AGENCY}{e.g. 'E'. See Details.}

\item{ACTIVITY_LOCATION}{e.g. 'AK'. See Details.}

\item{ACTUAL_DATE}{e.g. '20-AUG-97'. See Details.}

\item{OWNER}{e.g. 'HQ'. See Details.}

\item{CA_EVENT_CODE}{e.g. 'CA750IN'. See Details.}

\item{SCHEDULE_DATE_ORIG}{e.g. 'NA'. See Details.}

\item{SCHEDULE_DATE_NEW}{e.g. 'NA'. See Details.}

\item{BEST_DATE}{e.g. '20-AUG-97'. See Details.}

\item{PERSON_OWNER}{e.g. '10'. See Details.}

\item{PERSON_ID}{e.g. 'R0DSD'. See Details.}

\item{SUB_ORGANIZATION_OWNER}{e.g. '10'. See Details.}

\item{SUB_ORGANIZATION}{e.g. 'PMT'. See Details.}
}
\description{
Retrieve rcr ca event data from rcrainfo database
}

