% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcrainfo_rcr_event_authority.R
\name{rcrainfo_rcr_event_authority}
\alias{rcrainfo_rcr_event_authority}
\title{Retrieve rcr event authority data from rcrainfo database}
\usage{
rcrainfo_rcr_event_authority(EVENT_ID = NULL, EVENT_OWN = NULL,
  EVENT_CODE = NULL, EVENT_SEQ = NULL, EVENT_AGENCY = NULL,
  EVENT_LOCATION = NULL, AUTHORITY_ID = NULL, AUTHORITY_OWN = NULL,
  AUTHORITY_TYPE = NULL, EFFECTIVE_DATE = NULL,
  AUTHORITY_AGENCY = NULL, AUTHORITY_LOCATION = NULL)
}
\arguments{
\item{EVENT_ID}{e.g. 'AK0000374959'. See Details.}

\item{EVENT_OWN}{e.g. 'HQ'. See Details.}

\item{EVENT_CODE}{e.g. 'CA750IN'. See Details.}

\item{EVENT_SEQ}{e.g. '1'. See Details.}

\item{EVENT_AGENCY}{e.g. 'E'. See Details.}

\item{EVENT_LOCATION}{e.g. 'AK'. See Details.}

\item{AUTHORITY_ID}{e.g. 'AK0000374959'. See Details.}

\item{AUTHORITY_OWN}{e.g. 'HQ'. See Details.}

\item{AUTHORITY_TYPE}{e.g. 'X'. See Details.}

\item{EFFECTIVE_DATE}{e.g. '19-OCT-00'. See Details.}

\item{AUTHORITY_AGENCY}{e.g. 'E'. See Details.}

\item{AUTHORITY_LOCATION}{e.g. 'AK'. See Details.}
}
\description{
Retrieve rcr event authority data from rcrainfo database
}
