% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcrainfo_rcr_event_unit.R
\name{rcrainfo_rcr_event_unit}
\alias{rcrainfo_rcr_event_unit}
\title{Retrieve rcr event unit data from rcrainfo database}
\usage{
rcrainfo_rcr_event_unit(HANDLER_ID = NULL, SERIES_SEQ = NULL,
  OWNER = NULL, PERMIT_EVENT_CODE = NULL, EVENT_SEQ = NULL,
  RESPONSIBLE_AGENCY = NULL, ACTIVITY_LOCATION = NULL,
  UD_HANDLER_ID = NULL, UD_UNIT_SEQ = NULL, UNIT_DETAIL_SEQ = NULL)
}
\arguments{
\item{HANDLER_ID}{e.g. 'AK1210022157'. See Details.}

\item{SERIES_SEQ}{e.g. '500'. See Details.}

\item{OWNER}{e.g. 'HQ'. See Details.}

\item{PERMIT_EVENT_CODE}{e.g. 'CL310'. See Details.}

\item{EVENT_SEQ}{e.g. '1'. See Details.}

\item{RESPONSIBLE_AGENCY}{e.g. 'E'. See Details.}

\item{ACTIVITY_LOCATION}{e.g. 'AK'. See Details.}

\item{UD_HANDLER_ID}{e.g. 'AK1210022157'. See Details.}

\item{UD_UNIT_SEQ}{e.g. '2'. See Details.}

\item{UNIT_DETAIL_SEQ}{e.g. '1'. See Details.}
}
\description{
Retrieve rcr event unit data from rcrainfo database
}
