% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcrainfo_rcr_hbasic.R
\name{rcrainfo_rcr_hbasic}
\alias{rcrainfo_rcr_hbasic}
\title{Retrieve rcr hbasic data from rcrainfo database}
\usage{
rcrainfo_rcr_hbasic(HANDLER_ID = NULL, FACILITY_IDENTIFIER = NULL,
  HANDLER_NAME = NULL, REGION = NULL, STATE = NULL)
}
\arguments{
\item{HANDLER_ID}{e.g. 'AK0000000323'. See Details.}

\item{FACILITY_IDENTIFIER}{e.g. 'NA'. See Details.}

\item{HANDLER_NAME}{e.g. 'USDA FS CHUGACH NF CORDOVA FOREST WORK C'. See Details.}

\item{REGION}{e.g. '10'. See Details.}

\item{STATE}{e.g. 'AK'. See Details.}
}
\description{
Retrieve rcr hbasic data from rcrainfo database
}

