% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcrainfo_rcr_hcertification.R
\name{rcrainfo_rcr_hcertification}
\alias{rcrainfo_rcr_hcertification}
\title{Retrieve rcr hcertification data from rcrainfo database}
\usage{
rcrainfo_rcr_hcertification(HANDLER_ID = NULL, ACTIVITY_LOCATION = NULL,
  SOURCE_TYPE = NULL, SEQ_NUMBER = NULL, CERT_SEQ = NULL,
  CERT_SIGNED_DATE = NULL, CERT_TITLE = NULL, CERT_FIRST_NAME = NULL,
  CERT_MIDDLE_INITIAL = NULL, CERT_LAST_NAME = NULL)
}
\arguments{
\item{HANDLER_ID}{e.g. 'AK0000006569'. See Details.}

\item{ACTIVITY_LOCATION}{e.g. 'AK'. See Details.}

\item{SOURCE_TYPE}{e.g. 'N'. See Details.}

\item{SEQ_NUMBER}{e.g. '2'. See Details.}

\item{CERT_SEQ}{e.g. '1'. See Details.}

\item{CERT_SIGNED_DATE}{e.g. '10-DEC-99'. See Details.}

\item{CERT_TITLE}{e.g. 'SH%26E SPECIALIST'. See Details.}

\item{CERT_FIRST_NAME}{e.g. 'JAMES'. See Details.}

\item{CERT_MIDDLE_INITIAL}{e.g. 'R'. See Details.}

\item{CERT_LAST_NAME}{e.g. 'KANTOLA'. See Details.}
}
\description{
Retrieve rcr hcertification data from rcrainfo database
}

