% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcrainfo_rcr_hnaics.R
\name{rcrainfo_rcr_hnaics}
\alias{rcrainfo_rcr_hnaics}
\title{Retrieve rcr hnaics data from rcrainfo database}
\usage{
rcrainfo_rcr_hnaics(NAICS_SEQ = NULL, SEQ_NUMBER = NULL,
  ACTIVITY_LOCATION = NULL, HANDLER_ID = NULL, SOURCE_TYPE = NULL,
  NAICS_OWNER = NULL, NAICS_CODE = NULL)
}
\arguments{
\item{NAICS_SEQ}{e.g. '1'. See Details.}

\item{SEQ_NUMBER}{e.g. '3'. See Details.}

\item{ACTIVITY_LOCATION}{e.g. 'IA'. See Details.}

\item{HANDLER_ID}{e.g. 'IAR000002923'. See Details.}

\item{SOURCE_TYPE}{e.g. 'I'. See Details.}

\item{NAICS_OWNER}{e.g. 'HQ'. See Details.}

\item{NAICS_CODE}{e.g. '44111'. See Details.}
}
\description{
Retrieve rcr hnaics data from rcrainfo database
}

