% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcrainfo_rcr_hother_permit.R
\name{rcrainfo_rcr_hother_permit}
\alias{rcrainfo_rcr_hother_permit}
\title{Retrieve rcr hother permit data from rcrainfo database}
\usage{
rcrainfo_rcr_hother_permit(HANDLER_ID = NULL, ACTIVITY_LOCATION = NULL,
  SOURCE_TYPE = NULL, SEQ_NUMBER = NULL, OTHER_PERMIT_OWNER = NULL,
  OTHER_PERMIT_TYPE = NULL, OTHER_PERMIT_NUMBER = NULL,
  OTHER_PERMIT_DESC = NULL)
}
\arguments{
\item{HANDLER_ID}{e.g. 'AK1210022157'. See Details.}

\item{ACTIVITY_LOCATION}{e.g. 'AK'. See Details.}

\item{SOURCE_TYPE}{e.g. 'A'. See Details.}

\item{SEQ_NUMBER}{e.g. '1'. See Details.}

\item{OTHER_PERMIT_OWNER}{e.g. 'HQ'. See Details.}

\item{OTHER_PERMIT_TYPE}{e.g. 'E'. See Details.}

\item{OTHER_PERMIT_NUMBER}{e.g. '8421BA005'. See Details.}

\item{OTHER_PERMIT_DESC}{e.g. '8421BA005'. See Details.}
}
\description{
Retrieve rcr hother permit data from rcrainfo database
}

