% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcrainfo_rcr_howner_operator.R
\name{rcrainfo_rcr_howner_operator}
\alias{rcrainfo_rcr_howner_operator}
\title{Retrieve rcr howner operator data from rcrainfo database}
\usage{
rcrainfo_rcr_howner_operator(HANDLER_ID = NULL, ACTIVITY_LOCATION = NULL,
  SOURCE_TYPE = NULL, SEQ_NUMBER = NULL, OWNER_OPERATOR_SEQ = NULL,
  OWNER_OPERATOR_INDICATOR = NULL, OWNER_OPERATOR_NAME = NULL,
  OWNER_OPERATOR_TYPE = NULL, DATE_BECAME_CURRENT = NULL,
  DATE_ENDED_CURRENT = NULL, STREET_NO = NULL, STREET1 = NULL,
  STREET2 = NULL, CITY = NULL, STATE = NULL, COUNTRY = NULL,
  ZIP = NULL, PHONE = NULL)
}
\arguments{
\item{HANDLER_ID}{e.g. 'AK0000000323'. See Details.}

\item{ACTIVITY_LOCATION}{e.g. 'AK'. See Details.}

\item{SOURCE_TYPE}{e.g. 'N'. See Details.}

\item{SEQ_NUMBER}{e.g. '1'. See Details.}

\item{OWNER_OPERATOR_SEQ}{e.g. '1'. See Details.}

\item{OWNER_OPERATOR_INDICATOR}{e.g. 'CO'. See Details.}

\item{OWNER_OPERATOR_NAME}{e.g. 'US FOREST SERVICE CHUGACH NF'. See Details.}

\item{OWNER_OPERATOR_TYPE}{e.g. 'F'. See Details.}

\item{DATE_BECAME_CURRENT}{e.g. 'NA'. See Details.}

\item{DATE_ENDED_CURRENT}{e.g. 'NA'. See Details.}

\item{STREET_NO}{e.g. 'NA'. See Details.}

\item{STREET1}{e.g. '3301 C ST STE 300'. See Details.}

\item{STREET2}{e.g. 'NA'. See Details.}

\item{CITY}{e.g. 'ANCHORAGE'. See Details.}

\item{STATE}{e.g. 'AK'. See Details.}

\item{COUNTRY}{e.g. 'NA'. See Details.}

\item{ZIP}{e.g. '995033998'. See Details.}

\item{PHONE}{e.g. '9072712500'. See Details.}
}
\description{
Retrieve rcr howner operator data from rcrainfo database
}

