% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcrainfo_rcr_hreport_univ.R
\name{rcrainfo_rcr_hreport_univ}
\alias{rcrainfo_rcr_hreport_univ}
\title{Retrieve rcr hreport univ data from rcrainfo database}
\usage{
rcrainfo_rcr_hreport_univ(HANDLER_ID = NULL, ACTIVITY_LOCATION = NULL,
  SOURCE_TYPE = NULL, SEQ_NUMBER = NULL, HANDLER_NAME = NULL,
  NON_NOTIFIER = NULL, RECEIVE_DATE = NULL, REPORT_CYCLE = NULL,
  ACCESSIBILITY = NULL, REGION = NULL, STATE = NULL,
  EXTRACT_FLAG = NULL, ACTIVE_SITE = NULL, LOCATION_STREET_NO = NULL,
  LOCATION_STREET1 = NULL, LOCATION_STREET2 = NULL,
  LOCATION_CITY = NULL, LOCATION_STATE = NULL, LOCATION_ZIP = NULL,
  LOCATION_COUNTY_CODE = NULL, LOCATION_COUNTY_NAME = NULL,
  LOCATION_COUNTRY = NULL, STATE_DISTRICT_OWNER = NULL,
  STATE_DISTRICT = NULL, LAND_TYPE = NULL, MAIL_STREET_NO = NULL,
  MAIL_STREET1 = NULL, MAIL_STREET2 = NULL, MAIL_CITY = NULL,
  MAIL_STATE = NULL, MAIL_ZIP = NULL, MAIL_COUNTRY = NULL,
  CONTACT_NAME = NULL, CONTACT_STREET_NO = NULL,
  CONTACT_STREET1 = NULL, CONTACT_STREET2 = NULL,
  CONTACT_CITY = NULL, CONTACT_STATE = NULL, CONTACT_ZIP = NULL,
  CONTACT_COUNTRY = NULL, CONTACT_PHONE_AND_EXT = NULL,
  CONTACT_FAX = NULL, CONTACT_EMAIL_ADDRESS = NULL,
  CONTACT_TITLE = NULL, OWNER_NAME = NULL, OWNER_TYPE = NULL,
  OWNER_SEQ = NULL, OPERATOR_NAME = NULL, OPERATOR_TYPE = NULL,
  OPERATOR_SEQ = NULL, NAIC1 = NULL, NAIC2 = NULL, NAIC3 = NULL,
  NAIC4 = NULL, IN_HANDLER_UNIVERSES = NULL, IN_A_UNIVERSE = NULL,
  FED_WASTE_GENERATOR_OWNER = NULL, FED_WASTE_GENERATOR = NULL,
  STATE_WASTE_GENERATOR_OWNER = NULL, STATE_WASTE_GENERATOR = NULL,
  GENSTATUS = NULL, SHORT_TERM_GENERATOR = NULL, IMPORTER = NULL,
  MIXED_WASTE_GENERATOR = NULL, TRANSPORTER = NULL,
  TRANSFER_FACILITY = NULL, RECYCLER = NULL,
  ONSITE_BURNER_EXEMPT = NULL, FURNACE_EXEMPTION = NULL,
  UNDERGROUND_INJECTION = NULL, OFF_SITE_RECEIPT = NULL,
  UNIVWASTE = NULL, UNIVERSAL_WASTE_DEST_FACILITY = NULL,
  USED_OIL = NULL, FEDERAL_UNIVERSAL_WASTE = NULL,
  AS_FEDERALLY_REGULATED_TSDF = NULL, AS_CONVERTER_TSDF = NULL,
  AS_STATE_REGULATED_TSDF = NULL, FEDERAL_INDICATOR = NULL,
  HSM = NULL, SUBPART_K = NULL, COMMERCIAL_TSD = NULL,
  TSD_TYPE = NULL, GPRA_PERMIT = NULL, GPRA_RENEWAL = NULL,
  PERMIT_RENEWAL_WRKLD = NULL, PERMWRKLD = NULL, PERMPROG = NULL,
  PCWRKLD = NULL, CLOSWRKLD = NULL, GPRA_CA = NULL, CAWRKLD = NULL,
  SUBJCA = NULL, SUBJCA_NON_TSD = NULL, SUBJCA_TSD_3004 = NULL,
  SUBJCA_TSD_DISCRETION = NULL, NCAPS = NULL, EC_INDICATOR = NULL,
  IC_INDICATOR = NULL, CA725_INDICATOR = NULL,
  CA750_INDICATOR = NULL, OPERATING_TSDF = NULL,
  FULL_ENFORCEMENT = NULL, SNC = NULL, UNADDRESSED_SNC = NULL,
  ADDRESSED_SNC = NULL, SNC_WITH_COMP_SCHED = NULL,
  FA_REQUIRED = NULL, HHANDLER_LAST_CHANGE = NULL)
}
\arguments{
\item{HANDLER_ID}{e.g. 'AK0000000323'. See Details.}

\item{ACTIVITY_LOCATION}{e.g. 'AK'. See Details.}

\item{SOURCE_TYPE}{e.g. 'N'. See Details.}

\item{SEQ_NUMBER}{e.g. '1'. See Details.}

\item{HANDLER_NAME}{e.g. 'USDA FS CHUGACH NF CORDOVA FOREST WORK C'. See Details.}

\item{NON_NOTIFIER}{e.g. 'NA'. See Details.}

\item{RECEIVE_DATE}{e.g. '17-SEP-93'. See Details.}

\item{REPORT_CYCLE}{e.g. 'NA'. See Details.}

\item{ACCESSIBILITY}{e.g. 'NA'. See Details.}

\item{REGION}{e.g. '10'. See Details.}

\item{STATE}{e.g. 'AK'. See Details.}

\item{EXTRACT_FLAG}{e.g. 'X'. See Details.}

\item{ACTIVE_SITE}{e.g. 'H----'. See Details.}

\item{LOCATION_STREET_NO}{e.g. 'NA'. See Details.}

\item{LOCATION_STREET1}{e.g. 'COR LEFEVRE ST & COPPER RIV HW'. See Details.}

\item{LOCATION_STREET2}{e.g. 'Y'. See Details.}

\item{LOCATION_CITY}{e.g. 'CORDOVA'. See Details.}

\item{LOCATION_STATE}{e.g. 'AK'. See Details.}

\item{LOCATION_ZIP}{e.g. '99574'. See Details.}

\item{LOCATION_COUNTY_CODE}{e.g. 'AK261'. See Details.}

\item{LOCATION_COUNTY_NAME}{e.g. 'VALDEZ-CORDOVA'. See Details.}

\item{LOCATION_COUNTRY}{e.g. 'NA'. See Details.}

\item{STATE_DISTRICT_OWNER}{e.g. 'NA'. See Details.}

\item{STATE_DISTRICT}{e.g. 'NA'. See Details.}

\item{LAND_TYPE}{e.g. 'F'. See Details.}

\item{MAIL_STREET_NO}{e.g. 'NA'. See Details.}

\item{MAIL_STREET1}{e.g. 'PO BOX 280'. See Details.}

\item{MAIL_STREET2}{e.g. 'NA'. See Details.}

\item{MAIL_CITY}{e.g. 'CORDOVA'. See Details.}

\item{MAIL_STATE}{e.g. 'AK'. See Details.}

\item{MAIL_ZIP}{e.g. '99574'. See Details.}

\item{MAIL_COUNTRY}{e.g. 'NA'. See Details.}

\item{CONTACT_NAME}{e.g. 'SARAH  BAKER'. See Details.}

\item{CONTACT_STREET_NO}{e.g. 'NA'. See Details.}

\item{CONTACT_STREET1}{e.g. '3301 C ST STE 300'. See Details.}

\item{CONTACT_STREET2}{e.g. 'NA'. See Details.}

\item{CONTACT_CITY}{e.g. 'ANCHORAGE'. See Details.}

\item{CONTACT_STATE}{e.g. 'AK'. See Details.}

\item{CONTACT_ZIP}{e.g. '995033998'. See Details.}

\item{CONTACT_COUNTRY}{e.g. 'US'. See Details.}

\item{CONTACT_PHONE_AND_EXT}{e.g. '9072712556'. See Details.}

\item{CONTACT_FAX}{e.g. 'NA'. See Details.}

\item{CONTACT_EMAIL_ADDRESS}{e.g. 'NA'. See Details.}

\item{CONTACT_TITLE}{e.g. 'NA'. See Details.}

\item{OWNER_NAME}{e.g. 'US FOREST SERVICE CHUGACH NF'. See Details.}

\item{OWNER_TYPE}{e.g. 'F'. See Details.}

\item{OWNER_SEQ}{e.g. '1'. See Details.}

\item{OPERATOR_NAME}{e.g. 'NA'. See Details.}

\item{OPERATOR_TYPE}{e.g. 'NA'. See Details.}

\item{OPERATOR_SEQ}{e.g. 'NA'. See Details.}

\item{NAIC1}{e.g. 'NA'. See Details.}

\item{NAIC2}{e.g. 'NA'. See Details.}

\item{NAIC3}{e.g. 'NA'. See Details.}

\item{NAIC4}{e.g. 'NA'. See Details.}

\item{IN_HANDLER_UNIVERSES}{e.g. 'Y'. See Details.}

\item{IN_A_UNIVERSE}{e.g. 'Y'. See Details.}

\item{FED_WASTE_GENERATOR_OWNER}{e.g. 'HQ'. See Details.}

\item{FED_WASTE_GENERATOR}{e.g. '3'. See Details.}

\item{STATE_WASTE_GENERATOR_OWNER}{e.g. 'NA'. See Details.}

\item{STATE_WASTE_GENERATOR}{e.g. 'NA'. See Details.}

\item{GENSTATUS}{e.g. 'CEG'. See Details.}

\item{SHORT_TERM_GENERATOR}{e.g. 'N'. See Details.}

\item{IMPORTER}{e.g. 'N'. See Details.}

\item{MIXED_WASTE_GENERATOR}{e.g. 'N'. See Details.}

\item{TRANSPORTER}{e.g. 'N'. See Details.}

\item{TRANSFER_FACILITY}{e.g. 'N'. See Details.}

\item{RECYCLER}{e.g. 'N'. See Details.}

\item{ONSITE_BURNER_EXEMPT}{e.g. 'N'. See Details.}

\item{FURNACE_EXEMPTION}{e.g. 'N'. See Details.}

\item{UNDERGROUND_INJECTION}{e.g. 'N'. See Details.}

\item{OFF_SITE_RECEIPT}{e.g. 'N'. See Details.}

\item{UNIVWASTE}{e.g. 'N'. See Details.}

\item{UNIVERSAL_WASTE_DEST_FACILITY}{e.g. 'N'. See Details.}

\item{USED_OIL}{e.g. 'NNNNNNN'. See Details.}

\item{FEDERAL_UNIVERSAL_WASTE}{e.g. 'N'. See Details.}

\item{AS_FEDERALLY_REGULATED_TSDF}{e.g. '------'. See Details.}

\item{AS_CONVERTER_TSDF}{e.g. '------'. See Details.}

\item{AS_STATE_REGULATED_TSDF}{e.g. '---------'. See Details.}

\item{FEDERAL_INDICATOR}{e.g. 'LO-'. See Details.}

\item{HSM}{e.g. 'N'. See Details.}

\item{SUBPART_K}{e.g. '----'. See Details.}

\item{COMMERCIAL_TSD}{e.g. 'N'. See Details.}

\item{TSD_TYPE}{e.g. '-----'. See Details.}

\item{GPRA_PERMIT}{e.g. 'N'. See Details.}

\item{GPRA_RENEWAL}{e.g. 'N'. See Details.}

\item{PERMIT_RENEWAL_WRKLD}{e.g. '------'. See Details.}

\item{PERMWRKLD}{e.g. '------'. See Details.}

\item{PERMPROG}{e.g. '------'. See Details.}

\item{PCWRKLD}{e.g. '------'. See Details.}

\item{CLOSWRKLD}{e.g. '------'. See Details.}

\item{GPRA_CA}{e.g. 'N'. See Details.}

\item{CAWRKLD}{e.g. 'N'. See Details.}

\item{SUBJCA}{e.g. 'N'. See Details.}

\item{SUBJCA_NON_TSD}{e.g. 'N'. See Details.}

\item{SUBJCA_TSD_3004}{e.g. 'N'. See Details.}

\item{SUBJCA_TSD_DISCRETION}{e.g. 'N'. See Details.}

\item{NCAPS}{e.g. 'N'. See Details.}

\item{EC_INDICATOR}{e.g. 'N'. See Details.}

\item{IC_INDICATOR}{e.g. 'N'. See Details.}

\item{CA725_INDICATOR}{e.g. 'N'. See Details.}

\item{CA750_INDICATOR}{e.g. 'N'. See Details.}

\item{OPERATING_TSDF}{e.g. '------'. See Details.}

\item{FULL_ENFORCEMENT}{e.g. '------'. See Details.}

\item{SNC}{e.g. 'N'. See Details.}

\item{UNADDRESSED_SNC}{e.g. 'N'. See Details.}

\item{ADDRESSED_SNC}{e.g. 'N'. See Details.}

\item{SNC_WITH_COMP_SCHED}{e.g. 'N'. See Details.}

\item{FA_REQUIRED}{e.g. '-----'. See Details.}

\item{HHANDLER_LAST_CHANGE}{e.g. '15-SEP-00'. See Details.}
}
\description{
Retrieve rcr hreport univ data from rcrainfo database
}
