% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcrainfo_rcr_hstate_activity.R
\name{rcrainfo_rcr_hstate_activity}
\alias{rcrainfo_rcr_hstate_activity}
\title{Retrieve rcr hstate activity data from rcrainfo database}
\usage{
rcrainfo_rcr_hstate_activity(HANDLER_ID = NULL, ACTIVITY_LOCATION = NULL,
  SOURCE_TYPE = NULL, SEQ_NUMBER = NULL, STATE_ACTIVITY_OWNER = NULL,
  STATE_ACTIVITY_TYPE = NULL)
}
\arguments{
\item{HANDLER_ID}{e.g. 'AL0000005819'. See Details.}

\item{ACTIVITY_LOCATION}{e.g. 'AL'. See Details.}

\item{SOURCE_TYPE}{e.g. 'N'. See Details.}

\item{SEQ_NUMBER}{e.g. '8'. See Details.}

\item{STATE_ACTIVITY_OWNER}{e.g. 'AL'. See Details.}

\item{STATE_ACTIVITY_TYPE}{e.g. 'UOGOG'. See Details.}
}
\description{
Retrieve rcr hstate activity data from rcrainfo database
}

