% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcrainfo_rcr_huniversal_waste.R
\name{rcrainfo_rcr_huniversal_waste}
\alias{rcrainfo_rcr_huniversal_waste}
\title{Retrieve rcr huniversal waste data from rcrainfo database}
\usage{
rcrainfo_rcr_huniversal_waste(HANDLER_ID = NULL,
  ACTIVITY_LOCATION = NULL, SOURCE_TYPE = NULL, SEQ_NUMBER = NULL,
  UNIVERSAL_WASTE_OWNER = NULL, UNIVERSAL_WASTE_TYPE = NULL,
  ACCUMULATED = NULL, GENERATED = NULL)
}
\arguments{
\item{HANDLER_ID}{e.g. 'AK1210022157'. See Details.}

\item{ACTIVITY_LOCATION}{e.g. 'AK'. See Details.}

\item{SOURCE_TYPE}{e.g. 'R'. See Details.}

\item{SEQ_NUMBER}{e.g. '9'. See Details.}

\item{UNIVERSAL_WASTE_OWNER}{e.g. 'HQ'. See Details.}

\item{UNIVERSAL_WASTE_TYPE}{e.g. 'P'. See Details.}

\item{ACCUMULATED}{e.g. 'Y'. See Details.}

\item{GENERATED}{e.g. 'Y'. See Details.}
}
\description{
Retrieve rcr huniversal waste data from rcrainfo database
}
