% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcrainfo_rcr_hwaste_code.R
\name{rcrainfo_rcr_hwaste_code}
\alias{rcrainfo_rcr_hwaste_code}
\title{Retrieve rcr hwaste code data from rcrainfo database}
\usage{
rcrainfo_rcr_hwaste_code(HANDLER_ID = NULL, ACTIVITY_LOCATION = NULL,
  SOURCE_TYPE = NULL, SEQ_NUMBER = NULL, WASTE_CODE_OWNER = NULL,
  WASTE_CODE = NULL)
}
\arguments{
\item{HANDLER_ID}{e.g. 'AK0000000323'. See Details.}

\item{ACTIVITY_LOCATION}{e.g. 'AK'. See Details.}

\item{SOURCE_TYPE}{e.g. 'N'. See Details.}

\item{SEQ_NUMBER}{e.g. '1'. See Details.}

\item{WASTE_CODE_OWNER}{e.g. 'HQ'. See Details.}

\item{WASTE_CODE}{e.g. 'D002'. See Details.}
}
\description{
Retrieve rcr hwaste code data from rcrainfo database
}
