% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcrainfo_rcr_lu_county.R
\name{rcrainfo_rcr_lu_county}
\alias{rcrainfo_rcr_lu_county}
\title{Retrieve rcr lu county data from rcrainfo database}
\usage{
rcrainfo_rcr_lu_county(COUNTY_CODE = NULL, COUNTY_NAME = NULL,
  ACTIVE_STATUS = NULL)
}
\arguments{
\item{COUNTY_CODE}{e.g. 'AK000'. See Details.}

\item{COUNTY_NAME}{e.g. 'OUT-OF-STATE'. See Details.}

\item{ACTIVE_STATUS}{e.g. 'Y'. See Details.}
}
\description{
Retrieve rcr lu county data from rcrainfo database
}
