% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcrainfo_rcr_lu_generator_status.R
\name{rcrainfo_rcr_lu_generator_status}
\alias{rcrainfo_rcr_lu_generator_status}
\title{Retrieve rcr lu generator status data from rcrainfo database}
\usage{
rcrainfo_rcr_lu_generator_status(OWNER = NULL,
  WASTE_GENERATOR_STATUS_CODE = NULL, ACTIVE_STATUS = NULL,
  WASTE_GENERATOR_STATUS_DESC = NULL)
}
\arguments{
\item{OWNER}{e.g. '02'. See Details.}

\item{WASTE_GENERATOR_STATUS_CODE}{e.g. 'N'. See Details.}

\item{ACTIVE_STATUS}{e.g. 'N'. See Details.}

\item{WASTE_GENERATOR_STATUS_DESC}{e.g. 'NA'. See Details.}
}
\description{
Retrieve rcr lu generator status data from rcrainfo database
}
