% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcrainfo_rcr_lu_legal_operating_status.R
\name{rcrainfo_rcr_lu_legal_operating_status}
\alias{rcrainfo_rcr_lu_legal_operating_status}
\title{Retrieve rcr lu legal operating status data from rcrainfo database}
\usage{
rcrainfo_rcr_lu_legal_operating_status(OWNER = NULL,
  LEG_OP_STATUS_CODE = NULL, ACTIVE_STATUS = NULL,
  LEG_OP_STATUS_DESC = NULL, STRANGE_BUT_TRUE_FLAG = NULL,
  SUBJECT_TO_INSPECTION = NULL, PERMIT_PROGRESS = NULL,
  PERMITTING_WRKLD = NULL, CLOSURE_WRKLD = NULL,
  POST_CLOSURE_WRKLD = NULL, SUBJECT_TO_CORRECTIVE_ACT = NULL,
  CORRECTIVE_ACTION_WRKLD = NULL, FULL_ENFORCEMENT = NULL,
  OPERATING_TSDF = NULL, SUBJCA_TSD_3004 = NULL,
  SUBJCA_TSD_DISCRETION = NULL, SUBJCA_NON_TSD = NULL,
  GPRA_PERMIT_2006_ACCOMPLISHED = NULL,
  AS_FEDERALLY_REGULATED_TSDF = NULL, AS_CONVERTER_TSDF = NULL,
  AS_STATE_REGULATED_TSDF = NULL, PERMIT_RENEWAL_WRKLD = NULL)
}
\arguments{
\item{OWNER}{e.g. 'HQ'. See Details.}

\item{LEG_OP_STATUS_CODE}{e.g. 'DLCP'. See Details.}

\item{ACTIVE_STATUS}{e.g. 'N'. See Details.}

\item{LEG_OP_STATUS_DESC}{e.g. 'DELISTED - CLOSED WITH WASTE IN PLACE'. See Details.}

\item{STRANGE_BUT_TRUE_FLAG}{e.g. 'N'. See Details.}

\item{SUBJECT_TO_INSPECTION}{e.g. 'N'. See Details.}

\item{PERMIT_PROGRESS}{e.g. '4'. See Details.}

\item{PERMITTING_WRKLD}{e.g. 'N'. See Details.}

\item{CLOSURE_WRKLD}{e.g. 'N'. See Details.}

\item{POST_CLOSURE_WRKLD}{e.g. 'N'. See Details.}

\item{SUBJECT_TO_CORRECTIVE_ACT}{e.g. 'N'. See Details.}

\item{CORRECTIVE_ACTION_WRKLD}{e.g. 'N'. See Details.}

\item{FULL_ENFORCEMENT}{e.g. 'N'. See Details.}

\item{OPERATING_TSDF}{e.g. 'N'. See Details.}

\item{SUBJCA_TSD_3004}{e.g. 'N'. See Details.}

\item{SUBJCA_TSD_DISCRETION}{e.g. 'N'. See Details.}

\item{SUBJCA_NON_TSD}{e.g. '4'. See Details.}

\item{GPRA_PERMIT_2006_ACCOMPLISHED}{e.g. 'N'. See Details.}

\item{AS_FEDERALLY_REGULATED_TSDF}{e.g. 'N'. See Details.}

\item{AS_CONVERTER_TSDF}{e.g. 'N'. See Details.}

\item{AS_STATE_REGULATED_TSDF}{e.g. 'N'. See Details.}

\item{PERMIT_RENEWAL_WRKLD}{e.g. 'N'. See Details.}
}
\description{
Retrieve rcr lu legal operating status data from rcrainfo database
}
