% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcrainfo_rcr_lu_naics.R
\name{rcrainfo_rcr_lu_naics}
\alias{rcrainfo_rcr_lu_naics}
\title{Retrieve rcr lu naics data from rcrainfo database}
\usage{
rcrainfo_rcr_lu_naics(OWNER = NULL, NAICS_CODE = NULL,
  ACTIVE_STATUS = NULL, NAICS_DESC = NULL, NAICS_CYCLE = NULL)
}
\arguments{
\item{OWNER}{e.g. 'HQ'. See Details.}

\item{NAICS_CODE}{e.g. '5419'. See Details.}

\item{ACTIVE_STATUS}{e.g. 'N'. See Details.}

\item{NAICS_DESC}{e.g. 'OTHER PROFESSIONAL, SCIENTIFIC, AND TECHNICAL SERVICES'. See Details.}

\item{NAICS_CYCLE}{e.g. '2002'. See Details.}
}
\description{
Retrieve rcr lu naics data from rcrainfo database
}
