% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcrainfo_rcr_lu_state_activity.R
\name{rcrainfo_rcr_lu_state_activity}
\alias{rcrainfo_rcr_lu_state_activity}
\title{Retrieve rcr lu state activity data from rcrainfo database}
\usage{
rcrainfo_rcr_lu_state_activity(OWNER = NULL, ACTIVITY_TYPE = NULL,
  ACTIVE_STATUS = NULL, ACTIVITY_DESC = NULL)
}
\arguments{
\item{OWNER}{e.g. '07'. See Details.}

\item{ACTIVITY_TYPE}{e.g. 'ADMIN'. See Details.}

\item{ACTIVE_STATUS}{e.g. 'Y'. See Details.}

\item{ACTIVITY_DESC}{e.g. 'ADMINISTRATIVE NUMBER ASSIGNED FOR EPA INTERNAL USE ONLY'. See Details.}
}
\description{
Retrieve rcr lu state activity data from rcrainfo database
}

