% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcrainfo_rcr_lu_universal_waste.R
\name{rcrainfo_rcr_lu_universal_waste}
\alias{rcrainfo_rcr_lu_universal_waste}
\title{Retrieve rcr lu universal waste data from rcrainfo database}
\usage{
rcrainfo_rcr_lu_universal_waste(OWNER = NULL,
  UNIVERSAL_WASTE_TYPE = NULL, ACTIVE_STATUS = NULL,
  UNIVERSAL_WASTE_DESC = NULL)
}
\arguments{
\item{OWNER}{e.g. '02'. See Details.}

\item{UNIVERSAL_WASTE_TYPE}{e.g. 'C'. See Details.}

\item{ACTIVE_STATUS}{e.g. 'N'. See Details.}

\item{UNIVERSAL_WASTE_DESC}{e.g. 'Consumer Electronics'. See Details.}
}
\description{
Retrieve rcr lu universal waste data from rcrainfo database
}
