% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcrainfo_rcr_lu_waste_code.R
\name{rcrainfo_rcr_lu_waste_code}
\alias{rcrainfo_rcr_lu_waste_code}
\title{Retrieve rcr lu waste code data from rcrainfo database}
\usage{
rcrainfo_rcr_lu_waste_code(OWNER = NULL, WASTE_CODE = NULL,
  CODE_TYPE = NULL, WASTE_CODE_DESC = NULL, ACTIVE_STATUS = NULL,
  BR_LOAD_ACTIVE_STATUS = NULL)
}
\arguments{
\item{OWNER}{e.g. '01'. See Details.}

\item{WASTE_CODE}{e.g. 'P134'. See Details.}

\item{CODE_TYPE}{e.g. 'X'. See Details.}

\item{WASTE_CODE_DESC}{e.g. 'NA'. See Details.}

\item{ACTIVE_STATUS}{e.g. 'Y'. See Details.}

\item{BR_LOAD_ACTIVE_STATUS}{e.g. 'Y'. See Details.}
}
\description{
Retrieve rcr lu waste code data from rcrainfo database
}

