% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcrainfo_rcr_permit_event_codes.R
\name{rcrainfo_rcr_permit_event_codes}
\alias{rcrainfo_rcr_permit_event_codes}
\title{Retrieve rcr permit event codes data from rcrainfo database}
\usage{
rcrainfo_rcr_permit_event_codes(OWNER = NULL, PERMIT_EVENT_CODE = NULL,
  ACTIVE_STATUS = NULL, EVENT_DESC = NULL)
}
\arguments{
\item{OWNER}{e.g. '02'. See Details.}

\item{PERMIT_EVENT_CODE}{e.g. 'CL381'. See Details.}

\item{ACTIVE_STATUS}{e.g. 'Y'. See Details.}

\item{EVENT_DESC}{e.g. 'DATE INSPECTED TO COMFIRM CLOSURE'. See Details.}
}
\description{
Retrieve rcr permit event codes data from rcrainfo database
}

