% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcrainfo_rcr_pevent.R
\name{rcrainfo_rcr_pevent}
\alias{rcrainfo_rcr_pevent}
\title{Retrieve rcr pevent data from rcrainfo database}
\usage{
rcrainfo_rcr_pevent(HANDLER_ID = NULL, SERIES_SEQ = NULL,
  EVENT_SEQ = NULL, RESPONSIBLE_AGENCY = NULL,
  ACTIVITY_LOCATION = NULL, OWNER = NULL, PERMIT_EVENT_CODE = NULL,
  ACTUAL_DATE = NULL, SCHEDULE_DATE_ORIG = NULL,
  SCHEDULE_DATE_NEW = NULL, BEST_DATE = NULL,
  SUB_ORGANIZATION_OWNER = NULL, SUB_ORGANIZATION = NULL,
  PERSON_OWNER = NULL, PERSON_ID = NULL)
}
\arguments{
\item{HANDLER_ID}{e.g. 'AK1210022157'. See Details.}

\item{SERIES_SEQ}{e.g. '500'. See Details.}

\item{EVENT_SEQ}{e.g. '1'. See Details.}

\item{RESPONSIBLE_AGENCY}{e.g. 'E'. See Details.}

\item{ACTIVITY_LOCATION}{e.g. 'AK'. See Details.}

\item{OWNER}{e.g. 'HQ'. See Details.}

\item{PERMIT_EVENT_CODE}{e.g. 'OP100'. See Details.}

\item{ACTUAL_DATE}{e.g. '26-MAY-89'. See Details.}

\item{SCHEDULE_DATE_ORIG}{e.g. 'NA'. See Details.}

\item{SCHEDULE_DATE_NEW}{e.g. 'NA'. See Details.}

\item{BEST_DATE}{e.g. '26-MAY-89'. See Details.}

\item{SUB_ORGANIZATION_OWNER}{e.g. 'NA'. See Details.}

\item{SUB_ORGANIZATION}{e.g. 'NA'. See Details.}

\item{PERSON_OWNER}{e.g. '10'. See Details.}

\item{PERSON_ID}{e.g. 'JML'. See Details.}
}
\description{
Retrieve rcr pevent data from rcrainfo database
}
