% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcrainfo_rcr_process_codes.R
\name{rcrainfo_rcr_process_codes}
\alias{rcrainfo_rcr_process_codes}
\title{Retrieve rcr process codes data from rcrainfo database}
\usage{
rcrainfo_rcr_process_codes(OWNER = NULL, PROCESS_CODE = NULL,
  UOM_OWN = NULL, UOM_TYPE = NULL, ACTIVE_STATUS = NULL,
  PROCESS_TYPE = NULL, PROCESS_DESC = NULL)
}
\arguments{
\item{OWNER}{e.g. 'HQ'. See Details.}

\item{PROCESS_CODE}{e.g. 'T85'. See Details.}

\item{UOM_OWN}{e.g. 'HQ'. See Details.}

\item{UOM_TYPE}{e.g. 'V'. See Details.}

\item{ACTIVE_STATUS}{e.g. 'Y'. See Details.}

\item{PROCESS_TYPE}{e.g. 'BOILER - INDUSTRIAL FURNACE'. See Details.}

\item{PROCESS_DESC}{e.g. 'COKE OVEN'. See Details.}
}
\description{
Retrieve rcr process codes data from rcrainfo database
}

