% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcrainfo_rcr_pseries.R
\name{rcrainfo_rcr_pseries}
\alias{rcrainfo_rcr_pseries}
\title{Retrieve rcr pseries data from rcrainfo database}
\usage{
rcrainfo_rcr_pseries(HANDLER_ID = NULL, SERIES_SEQ = NULL,
  SERIES_NAME = NULL, RESPONSIBLE_OWNER = NULL,
  RESPONSIBLE_PERSON = NULL)
}
\arguments{
\item{HANDLER_ID}{e.g. 'AK1210022157'. See Details.}

\item{SERIES_SEQ}{e.g. '500'. See Details.}

\item{SERIES_NAME}{e.g. 'PERMIT1'. See Details.}

\item{RESPONSIBLE_OWNER}{e.g. 'NA'. See Details.}

\item{RESPONSIBLE_PERSON}{e.g. 'NA'. See Details.}
}
\description{
Retrieve rcr pseries data from rcrainfo database
}
