% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcrainfo_rcr_punit.R
\name{rcrainfo_rcr_punit}
\alias{rcrainfo_rcr_punit}
\title{Retrieve rcr punit data from rcrainfo database}
\usage{
rcrainfo_rcr_punit(HANDLER_ID = NULL, UNIT_SEQ = NULL,
  UNIT_NAME = NULL)
}
\arguments{
\item{HANDLER_ID}{e.g. 'AK1210022157'. See Details.}

\item{UNIT_SEQ}{e.g. '9'. See Details.}

\item{UNIT_NAME}{e.g. 'BLDG 955'. See Details.}
}
\description{
Retrieve rcr punit data from rcrainfo database
}
