% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcrainfo_rcr_punit_detail.R
\name{rcrainfo_rcr_punit_detail}
\alias{rcrainfo_rcr_punit_detail}
\title{Retrieve rcr punit detail data from rcrainfo database}
\usage{
rcrainfo_rcr_punit_detail(HANDLER_ID = NULL, UNIT_SEQ = NULL,
  UNIT_DETAIL_SEQ = NULL, EFFECTIVE_DATE = NULL, CAPACITY = NULL,
  NUMBER_OF_UNITS = NULL, CAPACITY_TYPE = NULL,
  COMMERCIAL_STATUS_CODE = NULL, LE_OWN = NULL, LEG_OP_STATUS_CODE = NULL,
  UOM_OWN = NULL, UOM_TYPE = NULL, PROCESS_OWN = NULL,
  PROCESS_CODE = NULL, STANDARDIZED_PERMIT_IND = NULL)
}
\arguments{
\item{HANDLER_ID}{e.g. 'AK1210022157'. See Details.}

\item{UNIT_SEQ}{e.g. '13'. See Details.}

\item{UNIT_DETAIL_SEQ}{e.g. '1'. See Details.}

\item{EFFECTIVE_DATE}{e.g. '26-JAN-06'. See Details.}

\item{CAPACITY}{e.g. '10500'. See Details.}

\item{NUMBER_OF_UNITS}{e.g. '1'. See Details.}

\item{CAPACITY_TYPE}{e.g. 'O'. See Details.}

\item{COMMERCIAL_STATUS_CODE}{e.g. '0'. See Details.}

\item{LE_OWN}{e.g. 'HQ'. See Details.}

\item{LEG_OP_STATUS_CODE}{e.g. 'ISCC'. See Details.}

\item{UOM_OWN}{e.g. 'HQ'. See Details.}

\item{UOM_TYPE}{e.g. 'G'. See Details.}

\item{PROCESS_OWN}{e.g. 'HQ'. See Details.}

\item{PROCESS_CODE}{e.g. 'S01'. See Details.}

\item{STANDARDIZED_PERMIT_IND}{e.g. 'N'. See Details.}
}
\description{
Retrieve rcr punit detail data from rcrainfo database
}

