% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdwis_violation.R
\name{sdwis_contaminant}
\alias{sdwis_contaminant}
\title{Get codes or names for SDWIS contaminants.}
\usage{
sdwis_contaminant(name = NULL, code = NULL, full = FALSE)
}
\arguments{
\item{name}{Name of contaminant to match}

\item{code}{Code of conaminant}

\item{full}{if TRUE, return a full data.frame with more information}
}
\description{
name -> code is 1 -> many (using fuzzy matching), while code -> name is 1 -> 1.
}
