% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdwis_enforcement_action.R
\name{sdwis_enforcement_action}
\alias{sdwis_enforcement_action}
\title{Retrieve enforcement action data from sdwis database}
\usage{
sdwis_enforcement_action(PWSID = NULL, ENFORCEMENT_ID = NULL,
  ORIGINATOR_CODE = NULL, ENFORCEMENT_DATE = NULL,
  ENFORCEMENT_ACTION_TYPE_CODE = NULL, ENFORCEMENT_COMMENT_TEXT = NULL)
}
\arguments{
\item{PWSID}{e.g. '010106001'. See Details.}

\item{ENFORCEMENT_ID}{e.g. '0106001IF9/27/2010'. See Details.}

\item{ORIGINATOR_CODE}{e.g. 'R'. See Details.}

\item{ENFORCEMENT_DATE}{e.g. '27-SEP-10'. See Details.}

\item{ENFORCEMENT_ACTION_TYPE_CODE}{e.g. 'EIF'. See Details.}

\item{ENFORCEMENT_COMMENT_TEXT}{e.g. 'Lead and Copper Results Notification - Action auto-generated by DIME based upon Public Notice records.'. See Details.}
}
\description{
Retrieve enforcement action data from sdwis database
}
