% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdwis_geographic_area.R
\name{sdwis_geographic_area}
\alias{sdwis_geographic_area}
\title{Retrieve geographic area data from sdwis database}
\usage{
sdwis_geographic_area(PWSID = NULL, GEO_ID = NULL,
  PRIMACY_AGENCY_CODE = NULL, EPA_REGION = NULL,
  PWS_ACTIVITY_CODE = NULL, PWS_TYPE_CODE = NULL, TRIBAL_CODE = NULL,
  STATE_SERVED = NULL, ANSI_ENTITY_CODE = NULL,
  ZIP_CODE_SERVED = NULL, CITY_SERVED = NULL, AREA_TYPE_CODE = NULL,
  COUNTY_SERVED = NULL)
}
\arguments{
\item{PWSID}{e.g. '010106001'. See Details.}

\item{GEO_ID}{e.g. '9431858'. See Details.}

\item{PRIMACY_AGENCY_CODE}{e.g. '01'. See Details.}

\item{EPA_REGION}{e.g. '01'. See Details.}

\item{PWS_ACTIVITY_CODE}{e.g. 'A'. See Details.}

\item{PWS_TYPE_CODE}{e.g. 'CWS'. See Details.}

\item{TRIBAL_CODE}{e.g. '020'. See Details.}

\item{STATE_SERVED}{e.g. 'NA'. See Details.}

\item{ANSI_ENTITY_CODE}{e.g. 'NA'. See Details.}

\item{ZIP_CODE_SERVED}{e.g. 'NA'. See Details.}

\item{CITY_SERVED}{e.g. 'NA'. See Details.}

\item{AREA_TYPE_CODE}{e.g. 'TR'. See Details.}

\item{COUNTY_SERVED}{e.g. 'NA'. See Details.}
}
\description{
Retrieve geographic area data from sdwis database
}
