% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdwis_service_area.R
\name{sdwis_service_area}
\alias{sdwis_service_area}
\title{Retrieve service area data from sdwis database}
\usage{
sdwis_service_area(PWSID = NULL, PRIMACY_AGENCY_CODE = NULL,
  EPA_REGION = NULL, PWS_ACTIVITY_CODE = NULL, PWS_TYPE_CODE = NULL,
  SERVICE_AREA_TYPE_CODE = NULL, IS_PRIMARY_SERVICE_AREA_CODE = NULL)
}
\arguments{
\item{PWSID}{e.g. '010106001'. See Details.}

\item{PRIMACY_AGENCY_CODE}{e.g. '01'. See Details.}

\item{EPA_REGION}{e.g. '01'. See Details.}

\item{PWS_ACTIVITY_CODE}{e.g. 'A'. See Details.}

\item{PWS_TYPE_CODE}{e.g. 'CWS'. See Details.}

\item{SERVICE_AREA_TYPE_CODE}{e.g. 'RA'. See Details.}

\item{IS_PRIMARY_SERVICE_AREA_CODE}{e.g. 'Y'. See Details.}
}
\description{
Retrieve service area data from sdwis database
}
