% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdwis_violation.R
\name{sdwis_violation}
\alias{sdwis_violation}
\title{Retrieve violation data from sdwis database}
\usage{
sdwis_violation(PWSID = NULL, VIOLATION_ID = NULL, FACILITY_ID = NULL,
  POPULATION_SERVED_COUNT = NULL, NPM_CANDIDATE = NULL,
  PWS_ACTIVITY_CODE = NULL, PWS_DEACTIVATION_DATE = NULL,
  PRIMARY_SOURCE_CODE = NULL, POP_CAT_5_CODE = NULL,
  PRIMACY_AGENCY_CODE = NULL, EPA_REGION = NULL, PWS_TYPE_CODE = NULL,
  VIOLATION_CODE = NULL, VIOLATION_CATEGORY_CODE = NULL,
  IS_HEALTH_BASED_IND = NULL, CONTAMINANT_CODE = NULL,
  COMPLIANCE_STATUS_CODE = NULL, VIOL_MEASURE = NULL,
  UNIT_OF_MEASURE = NULL, STATE_MCL = NULL, IS_MAJOR_VIOL_IND = NULL,
  SEVERITY_IND_CNT = NULL, COMPL_PER_BEGIN_DATE = NULL,
  COMPL_PER_END_DATE = NULL, LATEST_ENFORCEMENT_ID = NULL,
  RTC_ENFORCEMENT_ID = NULL, RTC_DATE = NULL,
  PUBLIC_NOTIFICATION_TIER = NULL, ORIGINATOR_CODE = NULL,
  SAMPLE_RESULT_ID = NULL, CORRECTIVE_ACTION_ID = NULL, RULE_CODE = NULL,
  RULE_GROUP_CODE = NULL, RULE_FAMILY_CODE = NULL)
}
\arguments{
\item{PWSID}{e.g. '010109005'. See Details.}

\item{VIOLATION_ID}{e.g. '010900526310042011'. See Details.}

\item{FACILITY_ID}{e.g. '3'. See Details.}

\item{POPULATION_SERVED_COUNT}{e.g. '37860'. See Details.}

\item{NPM_CANDIDATE}{e.g. 'N'. See Details.}

\item{PWS_ACTIVITY_CODE}{e.g. 'A'. See Details.}

\item{PWS_DEACTIVATION_DATE}{e.g. 'NA'. See Details.}

\item{PRIMARY_SOURCE_CODE}{e.g. 'SWP'. See Details.}

\item{POP_CAT_5_CODE}{e.g. '4'. See Details.}

\item{PRIMACY_AGENCY_CODE}{e.g. '01'. See Details.}

\item{EPA_REGION}{e.g. '01'. See Details.}

\item{PWS_TYPE_CODE}{e.g. 'CWS'. See Details.}

\item{VIOLATION_CODE}{e.g. '26'. See Details.}

\item{VIOLATION_CATEGORY_CODE}{e.g. 'MR'. See Details.}

\item{IS_HEALTH_BASED_IND}{e.g. 'N'. See Details.}

\item{CONTAMINANT_CODE}{e.g. '3100'. See Details.}

\item{COMPLIANCE_STATUS_CODE}{e.g. 'R'. See Details.}

\item{VIOL_MEASURE}{e.g. 'NA'. See Details.}

\item{UNIT_OF_MEASURE}{e.g. 'NA'. See Details.}

\item{STATE_MCL}{e.g. 'NA'. See Details.}

\item{IS_MAJOR_VIOL_IND}{e.g. 'NA'. See Details.}

\item{SEVERITY_IND_CNT}{e.g. 'NA'. See Details.}

\item{COMPL_PER_BEGIN_DATE}{e.g. '01-APR-11'. See Details.}

\item{COMPL_PER_END_DATE}{e.g. '30-APR-11'. See Details.}

\item{LATEST_ENFORCEMENT_ID}{e.g. '0109005IF2/16/2012'. See Details.}

\item{RTC_ENFORCEMENT_ID}{e.g. '0109005OX2/16/2012'. See Details.}

\item{RTC_DATE}{e.g. '16-FEB-12'. See Details.}

\item{PUBLIC_NOTIFICATION_TIER}{e.g. '3'. See Details.}

\item{ORIGINATOR_CODE}{e.g. 'R'. See Details.}

\item{SAMPLE_RESULT_ID}{e.g. 'NA'. See Details.}

\item{CORRECTIVE_ACTION_ID}{e.g. 'NA'. See Details.}

\item{RULE_CODE}{e.g. '110'. See Details.}

\item{RULE_GROUP_CODE}{e.g. '100'. See Details.}

\item{RULE_FAMILY_CODE}{e.g. '110'. See Details.}
}
\description{
Retrieve violation data from sdwis database
}

