% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdwis_violation_enf_assoc.R
\name{sdwis_violation_enf_assoc}
\alias{sdwis_violation_enf_assoc}
\title{Retrieve violation enf assoc data from sdwis database}
\usage{
sdwis_violation_enf_assoc(PWSID = NULL, ENFORCEMENT_ID = NULL,
  VIOLATION_ID = NULL)
}
\arguments{
\item{PWSID}{e.g. '010109005'. See Details.}

\item{ENFORCEMENT_ID}{e.g. '0109005IE7/28/2011'. See Details.}

\item{VIOLATION_ID}{e.g. '010900526310042011'. See Details.}
}
\description{
Retrieve violation enf assoc data from sdwis database
}
