% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdwis_water_system.R
\name{sdwis_water_system}
\alias{sdwis_water_system}
\title{Retrieve water system data from sdwis database}
\usage{
sdwis_water_system(PWSID = NULL, PWS_NAME = NULL,
  NPM_CANDIDATE = NULL, PRIMACY_AGENCY_CODE = NULL,
  EPA_REGION = NULL, SEASON_BEGIN_DATE = NULL,
  SEASON_END_DATE = NULL, PWS_ACTIVITY_CODE = NULL,
  PWS_DEACTIVATION_DATE = NULL, PWS_TYPE_CODE = NULL,
  DBPR_SCHEDULE_CAT_CODE = NULL, CDS_ID = NULL, GW_SW_CODE = NULL,
  LT2_SCHEDULE_CAT_CODE = NULL, OWNER_TYPE_CODE = NULL,
  POPULATION_SERVED_COUNT = NULL, POP_CAT_2_CODE = NULL,
  POP_CAT_3_CODE = NULL, POP_CAT_4_CODE = NULL,
  POP_CAT_5_CODE = NULL, POP_CAT_11_CODE = NULL, PRIMACY_TYPE = NULL,
  PRIMARY_SOURCE_CODE = NULL, IS_GRANT_ELIGIBLE_IND = NULL,
  IS_WHOLESALER_IND = NULL, IS_SCHOOL_OR_DAYCARE_IND = NULL,
  SERVICE_CONNECTIONS_COUNT = NULL, SUBMISSION_STATUS_CODE = NULL,
  ORG_NAME = NULL, ADMIN_NAME = NULL, EMAIL_ADDR = NULL,
  PHONE_NUMBER = NULL, PHONE_EXT_NUMBER = NULL, FAX_NUMBER = NULL,
  ALT_PHONE_NUMBER = NULL, ADDRESS_LINE1 = NULL,
  ADDRESS_LINE2 = NULL, CITY_NAME = NULL, ZIP_CODE = NULL,
  COUNTRY_CODE = NULL, STATE_CODE = NULL,
  SOURCE_WATER_PROTECTION_CODE = NULL,
  SOURCE_PROTECTION_BEGIN_DATE = NULL, OUTSTANDING_PERFORMER = NULL,
  OUTSTANDING_PERFORM_BEGIN_DATE = NULL, CITIES_SERVED = NULL,
  COUNTIES_SERVED = NULL)
}
\arguments{
\item{PWSID}{e.g. '010106001'. See Details.}

\item{PWS_NAME}{e.g. 'Mashantucket Pequot Water System'. See Details.}

\item{NPM_CANDIDATE}{e.g. 'Y'. See Details.}

\item{PRIMACY_AGENCY_CODE}{e.g. '01'. See Details.}

\item{EPA_REGION}{e.g. '01'. See Details.}

\item{SEASON_BEGIN_DATE}{e.g. 'NA'. See Details.}

\item{SEASON_END_DATE}{e.g. 'NA'. See Details.}

\item{PWS_ACTIVITY_CODE}{e.g. 'A'. See Details.}

\item{PWS_DEACTIVATION_DATE}{e.g. 'NA'. See Details.}

\item{PWS_TYPE_CODE}{e.g. 'CWS'. See Details.}

\item{DBPR_SCHEDULE_CAT_CODE}{e.g. 'NA'. See Details.}

\item{CDS_ID}{e.g. 'NA'. See Details.}

\item{GW_SW_CODE}{e.g. 'SW'. See Details.}

\item{LT2_SCHEDULE_CAT_CODE}{e.g. 'NA'. See Details.}

\item{OWNER_TYPE_CODE}{e.g. 'N'. See Details.}

\item{POPULATION_SERVED_COUNT}{e.g. '39552'. See Details.}

\item{POP_CAT_2_CODE}{e.g. '2'. See Details.}

\item{POP_CAT_3_CODE}{e.g. '2'. See Details.}

\item{POP_CAT_4_CODE}{e.g. '2'. See Details.}

\item{POP_CAT_5_CODE}{e.g. '4'. See Details.}

\item{POP_CAT_11_CODE}{e.g. '6'. See Details.}

\item{PRIMACY_TYPE}{e.g. 'Tribal'. See Details.}

\item{PRIMARY_SOURCE_CODE}{e.g. 'GU'. See Details.}

\item{IS_GRANT_ELIGIBLE_IND}{e.g. 'Y'. See Details.}

\item{IS_WHOLESALER_IND}{e.g. 'Y'. See Details.}

\item{IS_SCHOOL_OR_DAYCARE_IND}{e.g. 'N'. See Details.}

\item{SERVICE_CONNECTIONS_COUNT}{e.g. '143'. See Details.}

\item{SUBMISSION_STATUS_CODE}{e.g. 'Y'. See Details.}

\item{ORG_NAME}{e.g. 'Mashantucket Pequot Tribal Utilities'. See Details.}

\item{ADMIN_NAME}{e.g. 'Rodney Butler'. See Details.}

\item{EMAIL_ADDR}{e.g. 'NA'. See Details.}

\item{PHONE_NUMBER}{e.g. '8603966740'. See Details.}

\item{PHONE_EXT_NUMBER}{e.g. 'NA'. See Details.}

\item{FAX_NUMBER}{e.g. 'NA'. See Details.}

\item{ALT_PHONE_NUMBER}{e.g. 'NA'. See Details.}

\item{ADDRESS_LINE1}{e.g. 'Tribal Office'. See Details.}

\item{ADDRESS_LINE2}{e.g. 'Indiantown Rd, - PO BOX 3060'. See Details.}

\item{CITY_NAME}{e.g. 'MASHANTUCKET'. See Details.}

\item{ZIP_CODE}{e.g. '06339-3060'. See Details.}

\item{COUNTRY_CODE}{e.g. 'US'. See Details.}

\item{STATE_CODE}{e.g. 'CT'. See Details.}

\item{SOURCE_WATER_PROTECTION_CODE}{e.g. 'NA'. See Details.}

\item{SOURCE_PROTECTION_BEGIN_DATE}{e.g. 'NA'. See Details.}

\item{OUTSTANDING_PERFORMER}{e.g. 'NA'. See Details.}

\item{OUTSTANDING_PERFORM_BEGIN_DATE}{e.g. 'NA'. See Details.}

\item{CITIES_SERVED}{e.g. 'Not Reported'. See Details.}

\item{COUNTIES_SERVED}{e.g. 'Not Reported'. See Details.}
}
\description{
Retrieve water system data from sdwis database
}
