% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdwis_water_system_facility.R
\name{sdwis_water_system_facility}
\alias{sdwis_water_system_facility}
\title{Retrieve water system facility data from sdwis database}
\usage{
sdwis_water_system_facility(PWSID = NULL, PRIMACY_AGENCY_CODE = NULL,
  EPA_REGION = NULL, FACILITY_ID = NULL, FACILITY_NAME = NULL,
  STATE_FACILITY_ID = NULL, FACILITY_ACTIVITY_CODE = NULL,
  FACILITY_DEACTIVATION_DATE = NULL, FACILITY_TYPE_CODE = NULL,
  SUBMISSION_STATUS_CODE = NULL, IS_SOURCE_IND = NULL,
  WATER_TYPE_CODE = NULL, AVAILABILITY_CODE = NULL,
  SELLER_TREATMENT_CODE = NULL, SELLER_PWSID = NULL,
  SELLER_PWS_NAME = NULL, FILTRATION_STATUS_CODE = NULL,
  PWS_ACTIVITY_CODE = NULL, PWS_DEACTIVATION_DATE = NULL,
  PWS_TYPE_CODE = NULL, IS_SOURCE_TREATED_IND = NULL)
}
\arguments{
\item{PWSID}{e.g. '010106001'. See Details.}

\item{PRIMACY_AGENCY_CODE}{e.g. '01'. See Details.}

\item{EPA_REGION}{e.g. '01'. See Details.}

\item{FACILITY_ID}{e.g. '0106001GW007'. See Details.}

\item{FACILITY_NAME}{e.g. 'Well #2a (Southern Wellfield)'. See Details.}

\item{STATE_FACILITY_ID}{e.g. '010106001GW007'. See Details.}

\item{FACILITY_ACTIVITY_CODE}{e.g. 'A'. See Details.}

\item{FACILITY_DEACTIVATION_DATE}{e.g. 'NA'. See Details.}

\item{FACILITY_TYPE_CODE}{e.g. 'WL'. See Details.}

\item{SUBMISSION_STATUS_CODE}{e.g. 'Y'. See Details.}

\item{IS_SOURCE_IND}{e.g. 'Y'. See Details.}

\item{WATER_TYPE_CODE}{e.g. 'GW'. See Details.}

\item{AVAILABILITY_CODE}{e.g. 'P'. See Details.}

\item{SELLER_TREATMENT_CODE}{e.g. 'NA'. See Details.}

\item{SELLER_PWSID}{e.g. 'NA'. See Details.}

\item{SELLER_PWS_NAME}{e.g. 'NA'. See Details.}

\item{FILTRATION_STATUS_CODE}{e.g. 'NA'. See Details.}

\item{PWS_ACTIVITY_CODE}{e.g. 'A'. See Details.}

\item{PWS_DEACTIVATION_DATE}{e.g. 'NA'. See Details.}

\item{PWS_TYPE_CODE}{e.g. 'CWS'. See Details.}

\item{IS_SOURCE_TREATED_IND}{e.g. 'NA'. See Details.}
}
\description{
Retrieve water system facility data from sdwis database
}

