% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envirofacts.R
\docType{data}
\name{services}
\alias{services}
\title{Available Envirofacts services}
\format{An object of class \code{character} of length 13.}
\usage{
services
}
\description{
Available Envirofacts services
}
\keyword{data}
