% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_tri_chem_activity.R
\name{tri_tri_chem_activity}
\alias{tri_tri_chem_activity}
\title{Retrieve tri chem activity data from tri database}
\usage{
tri_tri_chem_activity(DOC_CTRL_NUM = NULL, PRODUCE = NULL,
  IMPORTED = NULL, USED_PROCESSED = NULL, SALE_DISTRIBUTION = NULL,
  BYPRODUCT = NULL, MANUFACTURE_IMPURITY = NULL, REACTANT = NULL,
  FORMULATION_COMPONENT = NULL, ARTICLE_COMPONENT = NULL,
  REPACKAGING = NULL, PROCESS_IMPURITY = NULL,
  CHEM_PROCESSING_AID = NULL, MANUFACTURE_AID = NULL,
  ANCILLARY = NULL)
}
\arguments{
\item{DOC_CTRL_NUM}{e.g. '1300140000011'. See Details.}

\item{PRODUCE}{e.g. '1'. See Details.}

\item{IMPORTED}{e.g. '0'. See Details.}

\item{USED_PROCESSED}{e.g. '0'. See Details.}

\item{SALE_DISTRIBUTION}{e.g. '0'. See Details.}

\item{BYPRODUCT}{e.g. '1'. See Details.}

\item{MANUFACTURE_IMPURITY}{e.g. '0'. See Details.}

\item{REACTANT}{e.g. '0'. See Details.}

\item{FORMULATION_COMPONENT}{e.g. '0'. See Details.}

\item{ARTICLE_COMPONENT}{e.g. '1'. See Details.}

\item{REPACKAGING}{e.g. '0'. See Details.}

\item{PROCESS_IMPURITY}{e.g. '0'. See Details.}

\item{CHEM_PROCESSING_AID}{e.g. '0'. See Details.}

\item{MANUFACTURE_AID}{e.g. '1'. See Details.}

\item{ANCILLARY}{e.g. '0'. See Details.}
}
\description{
Retrieve tri chem activity data from tri database
}
