% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_tri_chem_info.R
\name{tri_tri_chem_info}
\alias{tri_tri_chem_info}
\title{Retrieve tri chem info data from tri database}
\usage{
tri_tri_chem_info(TRI_CHEM_ID = NULL, CHEM_NAME = NULL,
  ACTIVE_DATE = NULL, INACTIVE_DATE = NULL, CAAC_IND = NULL,
  CARC_IND = NULL, R3350_IND = NULL, METAL_IND = NULL, FEDS_IND = NULL,
  CLASSIFICATION = NULL, PBT_START_YEAR = NULL, PBT_END_YEAR = NULL,
  NO_DECIMALS = NULL, UNIT_OF_MEASURE = NULL, CAS_REGISTRY_NUMBER = NULL,
  SRS_ID = NULL, DEFAULT_PERCENTAGE_TO_81C = NULL,
  DEFAULT_PERCENTAGE_TO_81D = NULL, DEFAULT_PERCENTAGE_TO_87 = NULL)
}
\arguments{
\item{TRI_CHEM_ID}{e.g. '000050000'. See Details.}

\item{CHEM_NAME}{e.g. 'FORMALDEHYDE'. See Details.}

\item{ACTIVE_DATE}{e.g. '1987'. See Details.}

\item{INACTIVE_DATE}{e.g. '9999'. See Details.}

\item{CAAC_IND}{e.g. '1'. See Details.}

\item{CARC_IND}{e.g. '1'. See Details.}

\item{R3350_IND}{e.g. '0'. See Details.}

\item{METAL_IND}{e.g. '0'. See Details.}

\item{FEDS_IND}{e.g. '0'. See Details.}

\item{CLASSIFICATION}{e.g. '0'. See Details.}

\item{PBT_START_YEAR}{e.g. 'NA'. See Details.}

\item{PBT_END_YEAR}{e.g. 'NA'. See Details.}

\item{NO_DECIMALS}{e.g. 'NA'. See Details.}

\item{UNIT_OF_MEASURE}{e.g. 'Pounds'. See Details.}

\item{CAS_REGISTRY_NUMBER}{e.g. '     50-00-0'. See Details.}

\item{SRS_ID}{e.g. '1008'. See Details.}

\item{DEFAULT_PERCENTAGE_TO_81C}{e.g. '0'. See Details.}

\item{DEFAULT_PERCENTAGE_TO_81D}{e.g. '8'. See Details.}

\item{DEFAULT_PERCENTAGE_TO_87}{e.g. '92'. See Details.}
}
\description{
Retrieve tri chem info data from tri database
}

