% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_tri_code_desc.R
\name{tri_tri_code_desc}
\alias{tri_tri_code_desc}
\title{Retrieve tri code desc data from tri database}
\usage{
tri_tri_code_desc(TABLE_ID = NULL, CODE = NULL, DESCRIPT = NULL)
}
\arguments{
\item{TABLE_ID}{e.g. '1'. See Details.}

\item{CODE}{e.g. '-1'. See Details.}

\item{DESCRIPT}{e.g. '-1'. See Details.}
}
\description{
Retrieve tri code desc data from tri database
}
