% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_tri_county.R
\name{tri_tri_county}
\alias{tri_tri_county}
\title{Retrieve tri county data from tri database}
\usage{
tri_tri_county(ZIP_CODE = NULL, COUNTY_NAME = NULL)
}
\arguments{
\item{ZIP_CODE}{e.g. '00401'. See Details.}

\item{COUNTY_NAME}{e.g. 'WESTCHESTER'. See Details.}
}
\description{
Retrieve tri county data from tri database
}

