% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_tri_energy_recovery.R
\name{tri_tri_energy_recovery}
\alias{tri_tri_energy_recovery}
\title{Retrieve tri energy recovery data from tri database}
\usage{
tri_tri_energy_recovery(DOC_CTRL_NUM = NULL,
  ONSITE_ENERGY_PROC_CODE = NULL)
}
\arguments{
\item{DOC_CTRL_NUM}{e.g. '1300140000011'. See Details.}

\item{ONSITE_ENERGY_PROC_CODE}{e.g. 'U09'. See Details.}
}
\description{
Retrieve tri energy recovery data from tri database
}
