% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_tri_facility_db.R
\name{tri_tri_facility_db}
\alias{tri_tri_facility_db}
\title{Retrieve tri facility db data from tri database}
\usage{
tri_tri_facility_db(TRI_FACILITY_ID = NULL, ASGN_DB_IND = NULL,
  DB_NUM = NULL)
}
\arguments{
\item{TRI_FACILITY_ID}{e.g. '00602BXTRF111CO'. See Details.}

\item{ASGN_DB_IND}{e.g. '0'. See Details.}

\item{DB_NUM}{e.g. '180532558'. See Details.}
}
\description{
Retrieve tri facility db data from tri database
}

