% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_tri_facility_db_history.R
\name{tri_tri_facility_db_history}
\alias{tri_tri_facility_db_history}
\title{Retrieve tri facility db history data from tri database}
\usage{
tri_tri_facility_db_history(TRI_FACILITY_ID = NULL,
  REPORTING_YEAR = NULL, ASGN_DB_IND = NULL, DB_NUM = NULL)
}
\arguments{
\item{TRI_FACILITY_ID}{e.g. '00602BXTRF111CO'. See Details.}

\item{REPORTING_YEAR}{e.g. '1988'. See Details.}

\item{ASGN_DB_IND}{e.g. '0'. See Details.}

\item{DB_NUM}{e.g. '080532558'. See Details.}
}
\description{
Retrieve tri facility db history data from tri database
}
