% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_tri_facility_history.R
\name{tri_tri_facility_history}
\alias{tri_tri_facility_history}
\title{Retrieve tri facility history data from tri database}
\usage{
tri_tri_facility_history(TRI_FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  FACILITY_NAME = NULL, STREET_ADDRESS = NULL, CITY_NAME = NULL,
  COUNTY_NAME = NULL, STATE_COUNTY_FIPS_CODE = NULL, STATE_ABBR = NULL,
  ZIP_CODE = NULL, REGION = NULL, MAIL_NAME = NULL,
  MAIL_STREET_ADDRESS = NULL, MAIL_CITY = NULL, MAIL_STATE_ABBR = NULL,
  MAIL_PROVINCE = NULL, MAIL_COUNTRY = NULL, MAIL_ZIP_CODE = NULL,
  ASGN_FEDERAL_IND = NULL, ASGN_AGENCY = NULL, PARENT_CO_DB_NUM = NULL,
  PARENT_CO_NAME = NULL, FAC_LATITUDE = NULL, FAC_LONGITUDE = NULL,
  PREF_LATITUDE = NULL, PREF_LONGITUDE = NULL, PREF_ACCURACY = NULL,
  PREF_COLLECT_METH = NULL, PREF_DESC_CATEGORY = NULL,
  PREF_HORIZONTAL_DATUM = NULL, PREF_SOURCE_SCALE = NULL,
  PREF_QA_CODE = NULL, ASGN_PARTIAL_IND = NULL,
  ASGN_PUBLIC_CONTACT = NULL, ASGN_PUBLIC_PHONE = NULL,
  ASGN_PUBLIC_CONTACT_EMAIL = NULL, BIA_CODE = NULL,
  STANDARDIZED_PARENT_COMPANY = NULL, ASGN_PUBLIC_PHONE_EXT = NULL)
}
\arguments{
\item{TRI_FACILITY_ID}{e.g. '00602BXTRF111CO'. See Details.}

\item{REPORTING_YEAR}{e.g. '1988'. See Details.}

\item{FACILITY_NAME}{e.g. 'BAXTER FENWAL DIV.,'. See Details.}

\item{STREET_ADDRESS}{e.g. '111 COLON ST.,'. See Details.}

\item{CITY_NAME}{e.g. 'AGUADA'. See Details.}

\item{COUNTY_NAME}{e.g. 'AGUADA'. See Details.}

\item{STATE_COUNTY_FIPS_CODE}{e.g. 'NA'. See Details.}

\item{STATE_ABBR}{e.g. 'PR'. See Details.}

\item{ZIP_CODE}{e.g. '00602'. See Details.}

\item{REGION}{e.g. '2'. See Details.}

\item{MAIL_NAME}{e.g. 'NA'. See Details.}

\item{MAIL_STREET_ADDRESS}{e.g. 'NA'. See Details.}

\item{MAIL_CITY}{e.g. 'NA'. See Details.}

\item{MAIL_STATE_ABBR}{e.g. 'NA'. See Details.}

\item{MAIL_PROVINCE}{e.g. 'NA'. See Details.}

\item{MAIL_COUNTRY}{e.g. 'NA'. See Details.}

\item{MAIL_ZIP_CODE}{e.g. 'NA'. See Details.}

\item{ASGN_FEDERAL_IND}{e.g. 'C'. See Details.}

\item{ASGN_AGENCY}{e.g. 'NA'. See Details.}

\item{PARENT_CO_DB_NUM}{e.g. 'NA'. See Details.}

\item{PARENT_CO_NAME}{e.g. 'BAXTER SALES CORP.'. See Details.}

\item{FAC_LATITUDE}{e.g. 'NA'. See Details.}

\item{FAC_LONGITUDE}{e.g. 'NA'. See Details.}

\item{PREF_LATITUDE}{e.g. 'NA'. See Details.}

\item{PREF_LONGITUDE}{e.g. 'NA'. See Details.}

\item{PREF_ACCURACY}{e.g. 'NA'. See Details.}

\item{PREF_COLLECT_METH}{e.g. 'NA'. See Details.}

\item{PREF_DESC_CATEGORY}{e.g. 'NA'. See Details.}

\item{PREF_HORIZONTAL_DATUM}{e.g. 'NA'. See Details.}

\item{PREF_SOURCE_SCALE}{e.g. 'NA'. See Details.}

\item{PREF_QA_CODE}{e.g. 'NA'. See Details.}

\item{ASGN_PARTIAL_IND}{e.g. 'NA'. See Details.}

\item{ASGN_PUBLIC_CONTACT}{e.g. 'NA'. See Details.}

\item{ASGN_PUBLIC_PHONE}{e.g. 'NA'. See Details.}

\item{ASGN_PUBLIC_CONTACT_EMAIL}{e.g. 'NA'. See Details.}

\item{BIA_CODE}{e.g. 'NA'. See Details.}

\item{STANDARDIZED_PARENT_COMPANY}{e.g. 'NA'. See Details.}

\item{ASGN_PUBLIC_PHONE_EXT}{e.g. 'NA'. See Details.}
}
\description{
Retrieve tri facility history data from tri database
}

