% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_tri_facility_npdes.R
\name{tri_tri_facility_npdes}
\alias{tri_tri_facility_npdes}
\title{Retrieve tri facility npdes data from tri database}
\usage{
tri_tri_facility_npdes(TRI_FACILITY_ID = NULL, ASGN_NPDES_IND = NULL,
  NPDES_NUM = NULL)
}
\arguments{
\item{TRI_FACILITY_ID}{e.g. '00602SMRTMRD115'. See Details.}

\item{ASGN_NPDES_IND}{e.g. '0'. See Details.}

\item{NPDES_NUM}{e.g. 'NA'. See Details.}
}
\description{
Retrieve tri facility npdes data from tri database
}

