% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_tri_facility_npdes_history.R
\name{tri_tri_facility_npdes_history}
\alias{tri_tri_facility_npdes_history}
\title{Retrieve tri facility npdes history data from tri database}
\usage{
tri_tri_facility_npdes_history(TRI_FACILITY_ID = NULL,
  REPORTING_YEAR = NULL, ASGN_NPDES_IND = NULL, NPDES_NUM = NULL)
}
\arguments{
\item{TRI_FACILITY_ID}{e.g. '00602SMRTMRD115'. See Details.}

\item{REPORTING_YEAR}{e.g. '2004'. See Details.}

\item{ASGN_NPDES_IND}{e.g. '0'. See Details.}

\item{NPDES_NUM}{e.g. 'NA'. See Details.}
}
\description{
Retrieve tri facility npdes history data from tri database
}

