% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_tri_facility_rcra.R
\name{tri_tri_facility_rcra}
\alias{tri_tri_facility_rcra}
\title{Retrieve tri facility rcra data from tri database}
\usage{
tri_tri_facility_rcra(TRI_FACILITY_ID = NULL, ASGN_RCRA_IND = NULL,
  RCRA_NUM = NULL)
}
\arguments{
\item{TRI_FACILITY_ID}{e.g. '00602BXTRF111CO'. See Details.}

\item{ASGN_RCRA_IND}{e.g. '0'. See Details.}

\item{RCRA_NUM}{e.g. 'PRD000706473'. See Details.}
}
\description{
Retrieve tri facility rcra data from tri database
}

