% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_tri_facility_uic.R
\name{tri_tri_facility_uic}
\alias{tri_tri_facility_uic}
\title{Retrieve tri facility uic data from tri database}
\usage{
tri_tri_facility_uic(TRI_FACILITY_ID = NULL, ASGN_UIC_IND = NULL,
  UIC_NUM = NULL)
}
\arguments{
\item{TRI_FACILITY_ID}{e.g. '00602SMRTMRD115'. See Details.}

\item{ASGN_UIC_IND}{e.g. '0'. See Details.}

\item{UIC_NUM}{e.g. 'NA'. See Details.}
}
\description{
Retrieve tri facility uic data from tri database
}
