% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_tri_facility_uic_history.R
\name{tri_tri_facility_uic_history}
\alias{tri_tri_facility_uic_history}
\title{Retrieve tri facility uic history data from tri database}
\usage{
tri_tri_facility_uic_history(TRI_FACILITY_ID = NULL, REPORTING_YEAR = NULL,
  ASGN_UIC_IND = NULL, UIC_NUM = NULL)
}
\arguments{
\item{TRI_FACILITY_ID}{e.g. '00602SMRTMRD115'. See Details.}

\item{REPORTING_YEAR}{e.g. '2009'. See Details.}

\item{ASGN_UIC_IND}{e.g. '0'. See Details.}

\item{UIC_NUM}{e.g. 'NA'. See Details.}
}
\description{
Retrieve tri facility uic history data from tri database
}

