% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_tri_form_r_schedule_one.R
\name{tri_tri_form_r_schedule_one}
\alias{tri_tri_form_r_schedule_one}
\title{Retrieve tri form r schedule one data from tri database}
\usage{
tri_tri_form_r_schedule_one(DOC_CTRL_NUM = NULL,
  ENVIRONMENTAL_MEDIUM = NULL, SEQUENCE_NUM = NULL,
  OFF_SITE_AMOUNT_SEQUENCE = NULL, TYPE_OF_WASTE_MANAGEMENT = NULL,
  TOTAL_RELEASE = NULL, RELEASE_NA = NULL, CALCULATED_TEQ = NULL,
  DIOXIN_CONGENER1 = NULL, DIOXIN_CONGENER2 = NULL,
  DIOXIN_CONGENER3 = NULL, DIOXIN_CONGENER4 = NULL,
  DIOXIN_CONGENER5 = NULL, DIOXIN_CONGENER6 = NULL,
  DIOXIN_CONGENER7 = NULL, DIOXIN_CONGENER8 = NULL,
  DIOXIN_CONGENER9 = NULL, DIOXIN_CONGENER10 = NULL,
  DIOXIN_CONGENER11 = NULL, DIOXIN_CONGENER12 = NULL,
  DIOXIN_CONGENER13 = NULL, DIOXIN_CONGENER14 = NULL,
  DIOXIN_CONGENER15 = NULL, DIOXIN_CONGENER16 = NULL,
  DIOXIN_CONGENER17 = NULL)
}
\arguments{
\item{DOC_CTRL_NUM}{e.g. '1307209856564'. See Details.}

\item{ENVIRONMENTAL_MEDIUM}{e.g. 'POTW'. See Details.}

\item{SEQUENCE_NUM}{e.g. '1'. See Details.}

\item{OFF_SITE_AMOUNT_SEQUENCE}{e.g. '1'. See Details.}

\item{TYPE_OF_WASTE_MANAGEMENT}{e.g. 'NA'. See Details.}

\item{TOTAL_RELEASE}{e.g. 'NA'. See Details.}

\item{RELEASE_NA}{e.g. '1'. See Details.}

\item{CALCULATED_TEQ}{e.g. '0'. See Details.}

\item{DIOXIN_CONGENER1}{e.g. 'NA'. See Details.}

\item{DIOXIN_CONGENER2}{e.g. 'NA'. See Details.}

\item{DIOXIN_CONGENER3}{e.g. 'NA'. See Details.}

\item{DIOXIN_CONGENER4}{e.g. 'NA'. See Details.}

\item{DIOXIN_CONGENER5}{e.g. 'NA'. See Details.}

\item{DIOXIN_CONGENER6}{e.g. 'NA'. See Details.}

\item{DIOXIN_CONGENER7}{e.g. 'NA'. See Details.}

\item{DIOXIN_CONGENER8}{e.g. 'NA'. See Details.}

\item{DIOXIN_CONGENER9}{e.g. 'NA'. See Details.}

\item{DIOXIN_CONGENER10}{e.g. 'NA'. See Details.}

\item{DIOXIN_CONGENER11}{e.g. 'NA'. See Details.}

\item{DIOXIN_CONGENER12}{e.g. 'NA'. See Details.}

\item{DIOXIN_CONGENER13}{e.g. 'NA'. See Details.}

\item{DIOXIN_CONGENER14}{e.g. 'NA'. See Details.}

\item{DIOXIN_CONGENER15}{e.g. 'NA'. See Details.}

\item{DIOXIN_CONGENER16}{e.g. 'NA'. See Details.}

\item{DIOXIN_CONGENER17}{e.g. 'NA'. See Details.}
}
\description{
Retrieve tri form r schedule one data from tri database
}
