% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_tri_off_site_transfer_location.R
\name{tri_tri_off_site_transfer_location}
\alias{tri_tri_off_site_transfer_location}
\title{Retrieve tri off site transfer location data from tri database}
\usage{
tri_tri_off_site_transfer_location(DOC_CTRL_NUM = NULL,
  TRANSFER_LOC_NUM = NULL, OFF_SITE_NAME = NULL,
  OFF_SITE_STREET_ADDRESS = NULL, CITY_NAME = NULL,
  COUNTY_NAME = NULL, STATE_ABBR = NULL, PROVINCE = NULL,
  ZIP_CODE = NULL, COUNTRY_CODE = NULL, RCRA_NUM = NULL,
  CONTROLLED_LOC = NULL)
}
\arguments{
\item{DOC_CTRL_NUM}{e.g. '1300140000011'. See Details.}

\item{TRANSFER_LOC_NUM}{e.g. '1'. See Details.}

\item{OFF_SITE_NAME}{e.g. 'METALCHEM, INC US ZINC'. See Details.}

\item{OFF_SITE_STREET_ADDRESS}{e.g. '1725 WASHINGTON ROAD'. See Details.}

\item{CITY_NAME}{e.g. 'PITTSBURGH'. See Details.}

\item{COUNTY_NAME}{e.g. 'ALLEGHENY'. See Details.}

\item{STATE_ABBR}{e.g. 'PA'. See Details.}

\item{PROVINCE}{e.g. 'NA'. See Details.}

\item{ZIP_CODE}{e.g. '15241'. See Details.}

\item{COUNTRY_CODE}{e.g. 'NA'. See Details.}

\item{RCRA_NUM}{e.g. 'NA'. See Details.}

\item{CONTROLLED_LOC}{e.g. '0'. See Details.}
}
\description{
Retrieve tri off site transfer location data from tri database
}
