% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_tri_onsite_waste_treatment_met.R
\name{tri_tri_onsite_waste_treatment_met}
\alias{tri_tri_onsite_waste_treatment_met}
\title{Retrieve tri onsite waste treatment met data from tri database}
\usage{
tri_tri_onsite_waste_treatment_met(DOC_CTRL_NUM = NULL,
  WASTESTREAM_SEQ_NUM = NULL, TREATMENT_SEQUENCE = NULL,
  TREATMENT_METHOD_CODE = NULL)
}
\arguments{
\item{DOC_CTRL_NUM}{e.g. '1300140000011'. See Details.}

\item{WASTESTREAM_SEQ_NUM}{e.g. '1'. See Details.}

\item{TREATMENT_SEQUENCE}{e.g. '1'. See Details.}

\item{TREATMENT_METHOD_CODE}{e.g. 'A07'. See Details.}
}
\description{
Retrieve tri onsite waste treatment met data from tri database
}

