% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_tri_onsite_wastestream.R
\name{tri_tri_onsite_wastestream}
\alias{tri_tri_onsite_wastestream}
\title{Retrieve tri onsite wastestream data from tri database}
\usage{
tri_tri_onsite_wastestream(DOC_CTRL_NUM = NULL,
  WASTESTREAM_SEQ_NUM = NULL, WASTESTREAM_CODE = NULL,
  INFLUENT_CONC_RANGE = NULL, SEQUENTIAL_TREAT_87_90 = NULL,
  TREATMENT_EFFICIENCY_EST = NULL, TREATMENT_EFFICIENCY_EST_NA = NULL,
  OPERATING_DATA_IND = NULL, EFFICIENCY_RANGE_CODE = NULL)
}
\arguments{
\item{DOC_CTRL_NUM}{e.g. '1300140000011'. See Details.}

\item{WASTESTREAM_SEQ_NUM}{e.g. '1'. See Details.}

\item{WASTESTREAM_CODE}{e.g. 'A'. See Details.}

\item{INFLUENT_CONC_RANGE}{e.g. '2'. See Details.}

\item{SEQUENTIAL_TREAT_87_90}{e.g. 'NA'. See Details.}

\item{TREATMENT_EFFICIENCY_EST}{e.g. '98'. See Details.}

\item{TREATMENT_EFFICIENCY_EST_NA}{e.g. '0'. See Details.}

\item{OPERATING_DATA_IND}{e.g. '0'. See Details.}

\item{EFFICIENCY_RANGE_CODE}{e.g. 'NA'. See Details.}
}
\description{
Retrieve tri onsite wastestream data from tri database
}
