% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_tri_potw_location.R
\name{tri_tri_potw_location}
\alias{tri_tri_potw_location}
\title{Retrieve tri potw location data from tri database}
\usage{
tri_tri_potw_location(DOC_CTRL_NUM = NULL, POTW_LOC_NUM = NULL,
  POTW_NAME = NULL, POTW_STREET_ADDRESS = NULL, CITY_NAME = NULL,
  COUNTY_NAME = NULL, STATE_ABBR = NULL, ZIP_CODE = NULL)
}
\arguments{
\item{DOC_CTRL_NUM}{e.g. '1300140000011'. See Details.}

\item{POTW_LOC_NUM}{e.g. '1'. See Details.}

\item{POTW_NAME}{e.g. 'GREATER JOHNSTOWN REDEVELOPMENT AUTHORITY'. See Details.}

\item{POTW_STREET_ADDRESS}{e.g. '401 WASHINGTON STREET'. See Details.}

\item{CITY_NAME}{e.g. 'JOHNSTOWN'. See Details.}

\item{COUNTY_NAME}{e.g. 'CAMBRIA'. See Details.}

\item{STATE_ABBR}{e.g. 'PA'. See Details.}

\item{ZIP_CODE}{e.g. '15901'. See Details.}
}
\description{
Retrieve tri potw location data from tri database
}

