% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_tri_recycling_process.R
\name{tri_tri_recycling_process}
\alias{tri_tri_recycling_process}
\title{Retrieve tri recycling process data from tri database}
\usage{
tri_tri_recycling_process(DOC_CTRL_NUM = NULL,
  ONSITE_RECYCLING_PROC_CODE = NULL)
}
\arguments{
\item{DOC_CTRL_NUM}{e.g. '1300140000011'. See Details.}

\item{ONSITE_RECYCLING_PROC_CODE}{e.g. 'R30'. See Details.}
}
\description{
Retrieve tri recycling process data from tri database
}
