% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_tri_reporting_form.R
\name{tri_tri_reporting_form}
\alias{tri_tri_reporting_form}
\title{Retrieve tri reporting form data from tri database}
\usage{
tri_tri_reporting_form(DOC_CTRL_NUM = NULL, ACTIVE_STATUS = NULL,
  TRI_FACILITY_ID = NULL, TRI_CHEM_ID = NULL, FORM_TYPE_IND = NULL,
  REPORTING_YEAR = NULL, CERTIF_NAME = NULL,
  CERTIF_OFFICIAL_TITLE = NULL, PARTIAL_FAC = NULL,
  ENTIRE_FAC = NULL, MAX_AMOUNT_OF_CHEM = NULL,
  CERTIF_SIGNATURE = NULL, CERTIF_DATE_SIGNED = NULL,
  POSTMARK_DATE = NULL, PUBLIC_CONTACT_PERSON = NULL,
  PUBLIC_CONTACT_PHONE = NULL, ADDITIONAL_DATA_IND = NULL,
  ONE_TIME_RELEASE_QTY = NULL, ONE_TIME_RELEASE_QTY_NA = NULL,
  PRODUCTION_RATIO = NULL, PRODUCTION_RATIO_NA = NULL,
  MIXTURE_NAME = NULL, RECEIVED_DATE = NULL, REVISION_NA = NULL,
  ORIG_POSTMARK = NULL, ORIG_RECEIVED = NULL, FEDERAL_FAC_IND = NULL,
  GOCO_FLAG = NULL, GENERIC_CHEM_NAME = NULL, CAS_CHEM_NAME = NULL,
  TRADE_SECRET_IND = NULL, SANITIZED_IND = NULL,
  DIOXIN_DISTRIBUTION_NA = NULL, DIOXIN_DISTRIBUTION_1 = NULL,
  DIOXIN_DISTRIBUTION_2 = NULL, DIOXIN_DISTRIBUTION_3 = NULL,
  DIOXIN_DISTRIBUTION_4 = NULL, DIOXIN_DISTRIBUTION_5 = NULL,
  DIOXIN_DISTRIBUTION_6 = NULL, DIOXIN_DISTRIBUTION_7 = NULL,
  DIOXIN_DISTRIBUTION_8 = NULL, DIOXIN_DISTRIBUTION_9 = NULL,
  DIOXIN_DISTRIBUTION_10 = NULL, DIOXIN_DISTRIBUTION_11 = NULL,
  DIOXIN_DISTRIBUTION_12 = NULL, DIOXIN_DISTRIBUTION_13 = NULL,
  DIOXIN_DISTRIBUTION_14 = NULL, DIOXIN_DISTRIBUTION_15 = NULL,
  DIOXIN_DISTRIBUTION_16 = NULL, DIOXIN_DISTRIBUTION_17 = NULL,
  PUBLIC_CONTACT_EMAIL = NULL, REVISION_CODE_1 = NULL,
  REVISION_CODE_2 = NULL, MEDIA_TYPE = NULL, PUBLIC_PHONE_EXT = NULL,
  PROD_RATIO_OR_ACTIVITY = NULL)
}
\arguments{
\item{DOC_CTRL_NUM}{e.g. '1300140000011'. See Details.}

\item{ACTIVE_STATUS}{e.g. '1'. See Details.}

\item{TRI_FACILITY_ID}{e.g. '15902CCKRN75BRI'. See Details.}

\item{TRI_CHEM_ID}{e.g. 'N982'. See Details.}

\item{FORM_TYPE_IND}{e.g. 'L'. See Details.}

\item{REPORTING_YEAR}{e.g. '2000'. See Details.}

\item{CERTIF_NAME}{e.g. 'LARRY E. CAPROUS'. See Details.}

\item{CERTIF_OFFICIAL_TITLE}{e.g. 'PRESIDENT'. See Details.}

\item{PARTIAL_FAC}{e.g. '0'. See Details.}

\item{ENTIRE_FAC}{e.g. '1'. See Details.}

\item{MAX_AMOUNT_OF_CHEM}{e.g. '04'. See Details.}

\item{CERTIF_SIGNATURE}{e.g. 'O'. See Details.}

\item{CERTIF_DATE_SIGNED}{e.g. '23-FEB-01'. See Details.}

\item{POSTMARK_DATE}{e.g. '26-FEB-01'. See Details.}

\item{PUBLIC_CONTACT_PERSON}{e.g. 'BARRY R. HEIDER'. See Details.}

\item{PUBLIC_CONTACT_PHONE}{e.g. '814 535 3293'. See Details.}

\item{ADDITIONAL_DATA_IND}{e.g. '0'. See Details.}

\item{ONE_TIME_RELEASE_QTY}{e.g. '0'. See Details.}

\item{ONE_TIME_RELEASE_QTY_NA}{e.g. '0'. See Details.}

\item{PRODUCTION_RATIO}{e.g. '1.25'. See Details.}

\item{PRODUCTION_RATIO_NA}{e.g. '0'. See Details.}

\item{MIXTURE_NAME}{e.g. 'NA'. See Details.}

\item{RECEIVED_DATE}{e.g. '28-FEB-01'. See Details.}

\item{REVISION_NA}{e.g. '0'. See Details.}

\item{ORIG_POSTMARK}{e.g. '26-FEB-01'. See Details.}

\item{ORIG_RECEIVED}{e.g. '28-FEB-01'. See Details.}

\item{FEDERAL_FAC_IND}{e.g. '0'. See Details.}

\item{GOCO_FLAG}{e.g. '0'. See Details.}

\item{GENERIC_CHEM_NAME}{e.g. 'NA'. See Details.}

\item{CAS_CHEM_NAME}{e.g. 'ZINC COMPOUNDS'. See Details.}

\item{TRADE_SECRET_IND}{e.g. '0'. See Details.}

\item{SANITIZED_IND}{e.g. '0'. See Details.}

\item{DIOXIN_DISTRIBUTION_NA}{e.g. '0'. See Details.}

\item{DIOXIN_DISTRIBUTION_1}{e.g. 'NA'. See Details.}

\item{DIOXIN_DISTRIBUTION_2}{e.g. 'NA'. See Details.}

\item{DIOXIN_DISTRIBUTION_3}{e.g. 'NA'. See Details.}

\item{DIOXIN_DISTRIBUTION_4}{e.g. 'NA'. See Details.}

\item{DIOXIN_DISTRIBUTION_5}{e.g. 'NA'. See Details.}

\item{DIOXIN_DISTRIBUTION_6}{e.g. 'NA'. See Details.}

\item{DIOXIN_DISTRIBUTION_7}{e.g. 'NA'. See Details.}

\item{DIOXIN_DISTRIBUTION_8}{e.g. 'NA'. See Details.}

\item{DIOXIN_DISTRIBUTION_9}{e.g. 'NA'. See Details.}

\item{DIOXIN_DISTRIBUTION_10}{e.g. 'NA'. See Details.}

\item{DIOXIN_DISTRIBUTION_11}{e.g. 'NA'. See Details.}

\item{DIOXIN_DISTRIBUTION_12}{e.g. 'NA'. See Details.}

\item{DIOXIN_DISTRIBUTION_13}{e.g. 'NA'. See Details.}

\item{DIOXIN_DISTRIBUTION_14}{e.g. 'NA'. See Details.}

\item{DIOXIN_DISTRIBUTION_15}{e.g. 'NA'. See Details.}

\item{DIOXIN_DISTRIBUTION_16}{e.g. 'NA'. See Details.}

\item{DIOXIN_DISTRIBUTION_17}{e.g. 'NA'. See Details.}

\item{PUBLIC_CONTACT_EMAIL}{e.g. 'NA'. See Details.}

\item{REVISION_CODE_1}{e.g. 'NA'. See Details.}

\item{REVISION_CODE_2}{e.g. 'NA'. See Details.}

\item{MEDIA_TYPE}{e.g. 'M'. See Details.}

\item{PUBLIC_PHONE_EXT}{e.g. 'NA'. See Details.}

\item{PROD_RATIO_OR_ACTIVITY}{e.g. 'NA'. See Details.}
}
\description{
Retrieve tri reporting form data from tri database
}
