% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_tri_source_reduct_qty.R
\name{tri_tri_source_reduct_qty}
\alias{tri_tri_source_reduct_qty}
\title{Retrieve tri source reduct qty data from tri database}
\usage{
tri_tri_source_reduct_qty(DOC_CTRL_NUM = NULL,
  RECYC_ONSITE_CURR_YR_QTY = NULL, RECYC_ONSITE_CURR_YR_NA = NULL,
  RECYC_ONSITE_PREV_YR_QTY = NULL, RECYC_ONSITE_PREV_YR_NA = NULL,
  RECYC_ONSITE_FOLL_YR_QTY = NULL, RECYC_ONSITE_FOLL_YR_NA = NULL,
  RECYC_ONSITE_SECD_YR_QTY = NULL, RECYC_ONSITE_SECD_YR_NA = NULL,
  RECYC_OFFSITE_CURR_YR_QTY = NULL, RECYC_OFFSITE_CURR_YR_NA = NULL,
  RECYC_OFFSITE_PREV_YR_QTY = NULL, RECYC_OFFSITE_PREV_YR_NA = NULL,
  RECYC_OFFSITE_FOLL_YR_QTY = NULL, RECYC_OFFSITE_FOLL_YR_NA = NULL,
  RECYC_OFFSITE_SECD_YR_QTY = NULL, RECYC_OFFSITE_SECD_YR_NA = NULL,
  ENERGY_ONSITE_CURR_YR_QTY = NULL, ENERGY_ONSITE_CURR_YR_NA = NULL,
  ENERGY_ONSITE_PREV_YR_QTY = NULL, ENERGY_ONSITE_PREV_YR_NA = NULL,
  ENERGY_ONSITE_FOLL_YR_QTY = NULL, ENERGY_ONSITE_FOLL_YR_NA = NULL,
  ENERGY_ONSITE_SECD_YR_QTY = NULL, ENERGY_ONSITE_SECD_YR_NA = NULL,
  ENERGY_OFFSITE_CURR_YR_QTY = NULL, ENERGY_OFFSITE_CURR_YR_NA = NULL,
  ENERGY_OFFSITE_PREV_YR_QTY = NULL, ENERGY_OFFSITE_PREV_YR_NA = NULL,
  ENERGY_OFFSITE_FOLL_YR_QTY = NULL, ENERGY_OFFSITE_FOLL_YR_NA = NULL,
  ENERGY_OFFSITE_SECD_YR_QTY = NULL, ENERGY_OFFSITE_SECD_YR_NA = NULL,
  TREATED_ONSITE_CURR_YR_QTY = NULL, TREATED_ONSITE_CURR_YR_NA = NULL,
  TREATED_ONSITE_PREV_YR_QTY = NULL, TREATED_ONSITE_PREV_YR_NA = NULL,
  TREATED_ONSITE_FOLL_YR_QTY = NULL, TREATED_ONSITE_FOLL_YR_NA = NULL,
  TREATED_ONSITE_SECD_YR_QTY = NULL, TREATED_ONSITE_SECD_YR_NA = NULL,
  TREATED_OFFSITE_CURR_YR_QTY = NULL,
  TREATED_OFFSITE_CURR_YR_NA = NULL,
  TREATED_OFFSITE_PREV_YR_QTY = NULL,
  TREATED_OFFSITE_PREV_YR_NA = NULL,
  TREATED_OFFSITE_FOLL_YR_QTY = NULL,
  TREATED_OFFSITE_FOLL_YR_NA = NULL,
  TREATED_OFFSITE_SECD_YR_QTY = NULL,
  TREATED_OFFSITE_SECD_YR_NA = NULL, REL_CURR_YR_QTY = NULL,
  REL_CURR_YR_NA = NULL, REL_PREV_YR_QTY = NULL,
  REL_PREV_YR_NA = NULL, REL_FOLL_YR_QTY = NULL,
  REL_FOLL_YR_NA = NULL, REL_SECD_YR_QTY = NULL,
  REL_SECD_YR_NA = NULL, REL_81A_CURR_YR_QTY = NULL,
  REL_81A_CURR_YR_NA = NULL, REL_81A_PREV_YR_QTY = NULL,
  REL_81A_PREV_YR_NA = NULL, REL_81A_FOLL_YR_QTY = NULL,
  REL_81A_FOLL_YR_NA = NULL, REL_81A_SECD_YR_QTY = NULL,
  REL_81A_SECD_YR_NA = NULL, REL_81B_CURR_YR_QTY = NULL,
  REL_81B_CURR_YR_NA = NULL, REL_81B_PREV_YR_QTY = NULL,
  REL_81B_PREV_YR_NA = NULL, REL_81B_FOLL_YR_QTY = NULL,
  REL_81B_FOLL_YR_NA = NULL, REL_81B_SECD_YR_QTY = NULL,
  REL_81B_SECD_YR_NA = NULL, REL_81C_CURR_YR_QTY = NULL,
  REL_81C_CURR_YR_NA = NULL, REL_81C_PREV_YR_QTY = NULL,
  REL_81C_PREV_YR_NA = NULL, REL_81C_FOLL_YR_QTY = NULL,
  REL_81C_FOLL_YR_NA = NULL, REL_81C_SECD_YR_QTY = NULL,
  REL_81C_SECD_YR_NA = NULL, REL_81D_CURR_YR_QTY = NULL,
  REL_81D_CURR_YR_NA = NULL, REL_81D_PREV_YR_QTY = NULL,
  REL_81D_PREV_YR_NA = NULL, REL_81D_FOLL_YR_QTY = NULL,
  REL_81D_FOLL_YR_NA = NULL, REL_81D_SECD_YR_QTY = NULL,
  REL_81D_SECD_YR_NA = NULL)
}
\arguments{
\item{DOC_CTRL_NUM}{e.g. '1300140000011'. See Details.}

\item{RECYC_ONSITE_CURR_YR_QTY}{e.g. '70000'. See Details.}

\item{RECYC_ONSITE_CURR_YR_NA}{e.g. '0'. See Details.}

\item{RECYC_ONSITE_PREV_YR_QTY}{e.g. '65000'. See Details.}

\item{RECYC_ONSITE_PREV_YR_NA}{e.g. '0'. See Details.}

\item{RECYC_ONSITE_FOLL_YR_QTY}{e.g. '70000'. See Details.}

\item{RECYC_ONSITE_FOLL_YR_NA}{e.g. '0'. See Details.}

\item{RECYC_ONSITE_SECD_YR_QTY}{e.g. '70000'. See Details.}

\item{RECYC_ONSITE_SECD_YR_NA}{e.g. '0'. See Details.}

\item{RECYC_OFFSITE_CURR_YR_QTY}{e.g. '174020'. See Details.}

\item{RECYC_OFFSITE_CURR_YR_NA}{e.g. '0'. See Details.}

\item{RECYC_OFFSITE_PREV_YR_QTY}{e.g. '241502'. See Details.}

\item{RECYC_OFFSITE_PREV_YR_NA}{e.g. '0'. See Details.}

\item{RECYC_OFFSITE_FOLL_YR_QTY}{e.g. '170000'. See Details.}

\item{RECYC_OFFSITE_FOLL_YR_NA}{e.g. '0'. See Details.}

\item{RECYC_OFFSITE_SECD_YR_QTY}{e.g. '160000'. See Details.}

\item{RECYC_OFFSITE_SECD_YR_NA}{e.g. '0'. See Details.}

\item{ENERGY_ONSITE_CURR_YR_QTY}{e.g. '0'. See Details.}

\item{ENERGY_ONSITE_CURR_YR_NA}{e.g. '0'. See Details.}

\item{ENERGY_ONSITE_PREV_YR_QTY}{e.g. '0'. See Details.}

\item{ENERGY_ONSITE_PREV_YR_NA}{e.g. '0'. See Details.}

\item{ENERGY_ONSITE_FOLL_YR_QTY}{e.g. '0'. See Details.}

\item{ENERGY_ONSITE_FOLL_YR_NA}{e.g. '0'. See Details.}

\item{ENERGY_ONSITE_SECD_YR_QTY}{e.g. '0'. See Details.}

\item{ENERGY_ONSITE_SECD_YR_NA}{e.g. '0'. See Details.}

\item{ENERGY_OFFSITE_CURR_YR_QTY}{e.g. '0'. See Details.}

\item{ENERGY_OFFSITE_CURR_YR_NA}{e.g. '0'. See Details.}

\item{ENERGY_OFFSITE_PREV_YR_QTY}{e.g. '0'. See Details.}

\item{ENERGY_OFFSITE_PREV_YR_NA}{e.g. '0'. See Details.}

\item{ENERGY_OFFSITE_FOLL_YR_QTY}{e.g. '0'. See Details.}

\item{ENERGY_OFFSITE_FOLL_YR_NA}{e.g. '0'. See Details.}

\item{ENERGY_OFFSITE_SECD_YR_QTY}{e.g. '0'. See Details.}

\item{ENERGY_OFFSITE_SECD_YR_NA}{e.g. '0'. See Details.}

\item{TREATED_ONSITE_CURR_YR_QTY}{e.g. '200'. See Details.}

\item{TREATED_ONSITE_CURR_YR_NA}{e.g. '0'. See Details.}

\item{TREATED_ONSITE_PREV_YR_QTY}{e.g. '0'. See Details.}

\item{TREATED_ONSITE_PREV_YR_NA}{e.g. '0'. See Details.}

\item{TREATED_ONSITE_FOLL_YR_QTY}{e.g. '250'. See Details.}

\item{TREATED_ONSITE_FOLL_YR_NA}{e.g. '0'. See Details.}

\item{TREATED_ONSITE_SECD_YR_QTY}{e.g. '250'. See Details.}

\item{TREATED_ONSITE_SECD_YR_NA}{e.g. '0'. See Details.}

\item{TREATED_OFFSITE_CURR_YR_QTY}{e.g. '0'. See Details.}

\item{TREATED_OFFSITE_CURR_YR_NA}{e.g. '0'. See Details.}

\item{TREATED_OFFSITE_PREV_YR_QTY}{e.g. '0'. See Details.}

\item{TREATED_OFFSITE_PREV_YR_NA}{e.g. '0'. See Details.}

\item{TREATED_OFFSITE_FOLL_YR_QTY}{e.g. '0'. See Details.}

\item{TREATED_OFFSITE_FOLL_YR_NA}{e.g. '0'. See Details.}

\item{TREATED_OFFSITE_SECD_YR_QTY}{e.g. '0'. See Details.}

\item{TREATED_OFFSITE_SECD_YR_NA}{e.g. '0'. See Details.}

\item{REL_CURR_YR_QTY}{e.g. '190'. See Details.}

\item{REL_CURR_YR_NA}{e.g. '0'. See Details.}

\item{REL_PREV_YR_QTY}{e.g. '220'. See Details.}

\item{REL_PREV_YR_NA}{e.g. '0'. See Details.}

\item{REL_FOLL_YR_QTY}{e.g. '180'. See Details.}

\item{REL_FOLL_YR_NA}{e.g. '0'. See Details.}

\item{REL_SECD_YR_QTY}{e.g. '170'. See Details.}

\item{REL_SECD_YR_NA}{e.g. '0'. See Details.}

\item{REL_81A_CURR_YR_QTY}{e.g. 'NA'. See Details.}

\item{REL_81A_CURR_YR_NA}{e.g. 'NA'. See Details.}

\item{REL_81A_PREV_YR_QTY}{e.g. 'NA'. See Details.}

\item{REL_81A_PREV_YR_NA}{e.g. 'NA'. See Details.}

\item{REL_81A_FOLL_YR_QTY}{e.g. 'NA'. See Details.}

\item{REL_81A_FOLL_YR_NA}{e.g. 'NA'. See Details.}

\item{REL_81A_SECD_YR_QTY}{e.g. 'NA'. See Details.}

\item{REL_81A_SECD_YR_NA}{e.g. 'NA'. See Details.}

\item{REL_81B_CURR_YR_QTY}{e.g. 'NA'. See Details.}

\item{REL_81B_CURR_YR_NA}{e.g. 'NA'. See Details.}

\item{REL_81B_PREV_YR_QTY}{e.g. 'NA'. See Details.}

\item{REL_81B_PREV_YR_NA}{e.g. 'NA'. See Details.}

\item{REL_81B_FOLL_YR_QTY}{e.g. 'NA'. See Details.}

\item{REL_81B_FOLL_YR_NA}{e.g. 'NA'. See Details.}

\item{REL_81B_SECD_YR_QTY}{e.g. 'NA'. See Details.}

\item{REL_81B_SECD_YR_NA}{e.g. 'NA'. See Details.}

\item{REL_81C_CURR_YR_QTY}{e.g. 'NA'. See Details.}

\item{REL_81C_CURR_YR_NA}{e.g. 'NA'. See Details.}

\item{REL_81C_PREV_YR_QTY}{e.g. 'NA'. See Details.}

\item{REL_81C_PREV_YR_NA}{e.g. 'NA'. See Details.}

\item{REL_81C_FOLL_YR_QTY}{e.g. 'NA'. See Details.}

\item{REL_81C_FOLL_YR_NA}{e.g. 'NA'. See Details.}

\item{REL_81C_SECD_YR_QTY}{e.g. 'NA'. See Details.}

\item{REL_81C_SECD_YR_NA}{e.g. 'NA'. See Details.}

\item{REL_81D_CURR_YR_QTY}{e.g. 'NA'. See Details.}

\item{REL_81D_CURR_YR_NA}{e.g. 'NA'. See Details.}

\item{REL_81D_PREV_YR_QTY}{e.g. 'NA'. See Details.}

\item{REL_81D_PREV_YR_NA}{e.g. 'NA'. See Details.}

\item{REL_81D_FOLL_YR_QTY}{e.g. 'NA'. See Details.}

\item{REL_81D_FOLL_YR_NA}{e.g. 'NA'. See Details.}

\item{REL_81D_SECD_YR_QTY}{e.g. 'NA'. See Details.}

\item{REL_81D_SECD_YR_NA}{e.g. 'NA'. See Details.}
}
\description{
Retrieve tri source reduct qty data from tri database
}
