% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_tri_submission_naics.R
\name{tri_tri_submission_naics}
\alias{tri_tri_submission_naics}
\title{Retrieve tri submission naics data from tri database}
\usage{
tri_tri_submission_naics(TRI_FACILITY_ID = NULL, DOC_CTRL_NUM = NULL,
  NAICS_CODE = NULL, NAICS_SEQUENCE_NUM = NULL, PRIMARY_IND = NULL)
}
\arguments{
\item{TRI_FACILITY_ID}{e.g. '00602BXTRF111CO'. See Details.}

\item{DOC_CTRL_NUM}{e.g. '1389035655796'. See Details.}

\item{NAICS_CODE}{e.g. '325412'. See Details.}

\item{NAICS_SEQUENCE_NUM}{e.g. '1'. See Details.}

\item{PRIMARY_IND}{e.g. '1'. See Details.}
}
\description{
Retrieve tri submission naics data from tri database
}

