% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_tri_submission_sic.R
\name{tri_tri_submission_sic}
\alias{tri_tri_submission_sic}
\title{Retrieve tri submission sic data from tri database}
\usage{
tri_tri_submission_sic(TRI_FACILITY_ID = NULL, DOC_CTRL_NUM = NULL,
  SIC_CODE = NULL, SIC_SEQUENCE_NUM = NULL, PRIMARY_IND = NULL)
}
\arguments{
\item{TRI_FACILITY_ID}{e.g. '00602BXTRF111CO'. See Details.}

\item{DOC_CTRL_NUM}{e.g. '1388025052414'. See Details.}

\item{SIC_CODE}{e.g. 'NA'. See Details.}

\item{SIC_SEQUENCE_NUM}{e.g. '2'. See Details.}

\item{PRIMARY_IND}{e.g. '0'. See Details.}
}
\description{
Retrieve tri submission sic data from tri database
}
