% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_tri_transfer_qty.R
\name{tri_tri_transfer_qty}
\alias{tri_tri_transfer_qty}
\title{Retrieve tri transfer qty data from tri database}
\usage{
tri_tri_transfer_qty(DOC_CTRL_NUM = NULL, TRANSFER_LOC_NUM = NULL,
  OFF_SITE_AMOUNT_SEQUENCE = NULL, TYPE_OF_WASTE_MANAGEMENT = NULL,
  TRANSFER_RANGE_CODE = NULL, TOTAL_TRANSFER = NULL,
  TRANSFER_EST_NA = NULL, TRANSFER_BASIS_EST_CODE = NULL,
  POTW_PERCENTAGE_TO_81C = NULL, POTW_PERCENTAGE_TO_81D = NULL,
  POTW_PERCENTAGE_TO_87 = NULL, POTW_PERCENTAGE_PROVIDED = NULL)
}
\arguments{
\item{DOC_CTRL_NUM}{e.g. '1300140000011'. See Details.}

\item{TRANSFER_LOC_NUM}{e.g. '0'. See Details.}

\item{OFF_SITE_AMOUNT_SEQUENCE}{e.g. '0'. See Details.}

\item{TYPE_OF_WASTE_MANAGEMENT}{e.g. 'P91'. See Details.}

\item{TRANSFER_RANGE_CODE}{e.g. '3'. See Details.}

\item{TOTAL_TRANSFER}{e.g. 'NA'. See Details.}

\item{TRANSFER_EST_NA}{e.g. '0'. See Details.}

\item{TRANSFER_BASIS_EST_CODE}{e.g. 'M '. See Details.}

\item{POTW_PERCENTAGE_TO_81C}{e.g. 'NA'. See Details.}

\item{POTW_PERCENTAGE_TO_81D}{e.g. 'NA'. See Details.}

\item{POTW_PERCENTAGE_TO_87}{e.g. 'NA'. See Details.}

\item{POTW_PERCENTAGE_PROVIDED}{e.g. 'NA'. See Details.}
}
\description{
Retrieve tri transfer qty data from tri database
}

