% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_tri_tribe_desc.R
\name{tri_tri_tribe_desc}
\alias{tri_tri_tribe_desc}
\title{Retrieve tri tribe desc data from tri database}
\usage{
tri_tri_tribe_desc(BIA_CODE = NULL, TRIBE = NULL, REGION = NULL,
  EPA_TRIBE_ID = NULL, STATE_ABBR = NULL)
}
\arguments{
\item{BIA_CODE}{e.g. '001'. See Details.}

\item{TRIBE}{e.g. 'Eastern Band of Cherokee Indians'. See Details.}

\item{REGION}{e.g. '4'. See Details.}

\item{EPA_TRIBE_ID}{e.g. '000000076'. See Details.}

\item{STATE_ABBR}{e.g. 'NC'. See Details.}
}
\description{
Retrieve tri tribe desc data from tri database
}
