% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_tri_water_stream.R
\name{tri_tri_water_stream}
\alias{tri_tri_water_stream}
\title{Retrieve tri water stream data from tri database}
\usage{
tri_tri_water_stream(DOC_CTRL_NUM = NULL, WATER_SEQUENCE_NUM = NULL,
  STREAM_NAME = NULL, STORM_WATER_NA = NULL,
  STORM_WATER_PERCENT = NULL, REACH_CODE = NULL)
}
\arguments{
\item{DOC_CTRL_NUM}{e.g. '1300140000011'. See Details.}

\item{WATER_SEQUENCE_NUM}{e.g. '1'. See Details.}

\item{STREAM_NAME}{e.g. 'NA'. See Details.}

\item{STORM_WATER_NA}{e.g. '0'. See Details.}

\item{STORM_WATER_PERCENT}{e.g. 'NA'. See Details.}

\item{REACH_CODE}{e.g. 'NA'. See Details.}
}
\description{
Retrieve tri water stream data from tri database
}
