% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_tri_zip_code.R
\name{tri_tri_zip_code}
\alias{tri_tri_zip_code}
\title{Retrieve tri zip code data from tri database}
\usage{
tri_tri_zip_code(ZIP_CODE = NULL, CITY_NAME = NULL, STATE_ABBR = NULL,
  REGION = NULL, COUNTRY_NAME = NULL, TRI_CENTROID_LAT = NULL,
  TRI_CENTROID_LONG = NULL)
}
\arguments{
\item{ZIP_CODE}{e.g. '00401'. See Details.}

\item{CITY_NAME}{e.g. 'PLEASANTVILLE'. See Details.}

\item{STATE_ABBR}{e.g. 'NY'. See Details.}

\item{REGION}{e.g. '2'. See Details.}

\item{COUNTRY_NAME}{e.g. 'NA'. See Details.}

\item{TRI_CENTROID_LAT}{e.g. '410806'. See Details.}

\item{TRI_CENTROID_LONG}{e.g. '734742'. See Details.}
}
\description{
Retrieve tri zip code data from tri database
}

